ManifoldsBase.jl – an interface for manifolds

The interface for a manifold is provided in the lightweight package ManifoldsBase.jl separate from the collection of manifolds in here. You can easily implement your algorithms and even your own manifolds just using the interface.

ManifoldsBase.jl – an interface for manifolds

Types and functions

DefaultManifold

Allocation

Additionally the AbstractDecoratorManifold is provided as well as the ArrayManifold as a specific example of such a decorator.

Types and functions

The following functions are currently available from the interface. If a manifold that you implement for your own package fits this interface, we happily look forward to a Pull Request to add it here.

ManifoldsBase.AbstractInverseRetractionMethod — Type

AbstractInverseRetractionMethod

Abstract type for methods for inverting a retraction (see inverse_retract).

ManifoldsBase.AbstractRetractionMethod — Type

AbstractRetractionMethod

Abstract type for methods for retracting a tangent vector to a manifold.

ManifoldsBase.CoTVector — Type

CoTVector

Type for a cotangent vector of a manifold. While a Manifold does not necessarily require this type, for example when it is implemented for Vectors or Matrix type elements, this type can be used for more complicated representations, semantic verification, or even dispatch for different representations of cotangent vectors and their types on a manifold.

ManifoldsBase.ExponentialRetraction — Type

ExponentialRetraction

Retraction using the exponential map.

ManifoldsBase.LogarithmicInverseRetraction — Type

LogarithmicInverseRetraction

Inverse retraction using the logarithmic map.

ManifoldsBase.MPoint — Type

MPoint

Type for a point on a manifold. While a Manifold does not necessarily require this type, for example when it is implemented for Vectors or Matrix type elements, this type can be used for more complicated representations, semantic verification, or even dispatch for different representations of points on a manifold.

ManifoldsBase.Manifold — Type

Manifold

A manifold type. The Manifold is used to dispatch to different functions on a manifold, usually as the first argument of the function. Examples are the exponential and logarithmic maps as well as more general functions that are built on them like the geodesic.

ManifoldsBase.ParallelTransport — Type

ParallelTransport <: AbstractVectorTransportMethod

Specify to use parallel transport as vector transport method within vector_transport_to, vector_transport_direction, or vector_transport_along.

ManifoldsBase.PolarInverseRetraction — Type

PolarInverseRetraction <: AbstractInverseRetractionMethod

Inverse retractions that are based on a singular value decomposition of the matrix / matrices for point and tangent vector on a Manifold

ManifoldsBase.PolarRetraction — Type

PolarRetraction <: AbstractRetractionMethod

Retractions that are based on singular value decompositions of the matrix / matrices for point and tangent vector on a Manifold

ManifoldsBase.ProjectionInverseRetraction — Type

ProjectionInverseRetraction <: AbstractInverseRetractionMethod

Inverse retractions that are based on a projection (or its inversion).

ManifoldsBase.ProjectionRetraction — Type

ProjectionRetraction <: AbstractRetractionMethod

Retractions that are based on projection and usually addition in the embedding.

ManifoldsBase.ProjectionTransport — Type

ProjectionTransport <: AbstractVectorTransportMethod

Specify to use projection onto tangent space as vector transport method within vector_transport_to, vector_transport_direction, or vector_transport_along. See project_tangent for details.

ManifoldsBase.QRInverseRetraction — Type

QRInverseRetraction <: AbstractInverseRetractionMethod

Inverse retractions that are based on a QR decomposition of the matrix / matrices for point and tangent vector on a Manifold

ManifoldsBase.QRRetraction — Type

QRRetraction <: AbstractRetractionMethod

Retractions that are based on a QR decomposition of the matrix / matrices for point and tangent vector on a Manifold

ManifoldsBase.TVector — Type

TVector

Type for a tangent vector of a manifold. While a Manifold does not necessarily require this type, for example when it is implemented for Vectors or Matrix type elements, this type can be used for more complicated representations, semantic verification, or even dispatch for different representations of tangent vectors and their types on a manifold.

Base.exp — Method

exp(M::Manifold, p, X)
exp(M::Manifold, p, X, t::Real = 1)

Compute the exponential map of tangent vector X, optionally scaled by t, at point p from manifold the Manifold M.

Base.isapprox — Method

isapprox(M::Manifold, p, X, Y; kwargs...)

Check if vectors X and Y tangent at p from Manifold M are approximately equal.

Keyword arguments can be used to specify tolerances.

Base.isapprox — Method

isapprox(M::Manifold, p, q; kwargs...)

Check if points p and q from Manifold M are approximately equal.

Keyword arguments can be used to specify tolerances.

Base.log — Method

log(M::Manifold, p, q)

Compute the logarithmic map of point q at base point p on the Manifold M.

LinearAlgebra.norm — Method

norm(M::Manifold, p, X)

Compute the norm of tangent vector X at point p from a Manifold M. By default this is computed using inner.

ManifoldsBase.allocate — Method

allocate(a)
allocate(a, dims::Int...)
allocate(a, dims::Tuple)
allocate(a, T::Type)
allocate(a, T::Type, dims::Int...)
allocate(a, T::Type, dims::Tuple)

Allocate an object similar to a. It is similar to function similar, although instead of working only on the outermost layer of a nested structure, it maps recursively through outer layers and calls similar on the innermost array-like object only. Type T is the new number element type number_eltype, if it is not given the element type of a is retained. The dims argument can be given for non-nested allocation and is forwarded to the function similar.

ManifoldsBase.base_manifold — Function

base_manifold(M::Manifold, depth = Val(-1))

Return the internally stored Manifold for decorated manifold M and the base manifold for vector bundles or power manifolds. The optional parameter depth can be used to remove only the first depth many decorators and return the Manifold from that level, whether its decorated or not. Any negative value deactivates this depth limit.

ManifoldsBase.check_manifold_point — Method

check_manifold_point(M::Manifold, p; kwargs...) -> Union{Nothing,String}

Return nothing when p is a point on the Manifold M. Otherwise, return an error with description why the point does not belong to manifold M.

By default, check_manifold_point returns nothing, i.e. if no checks are implemented, the assumption is to be optimistic for a point not deriving from the MPoint type.

ManifoldsBase.check_tangent_vector — Method

check_tangent_vector(M::Manifold, p, X; kwargs...) -> Union{Nothing,String}

Check whether X is a valid tangent vector in the tangent space of p on the Manifold M. An implementation should first call check_manifold_point(M, p; kwargs...) and then validate X. If it is not a tangent vector, an error string should be returned.

By default, check_tangent_vector returns nothing, i.e. if no checks are implemented, the assumption is to be optimistic for tangent vectors not deriving from the TVector type.

ManifoldsBase.distance — Method

distance(M::Manifold, p, q)

Shortest distance between the points p and q on the Manifold M.

ManifoldsBase.exp! — Method

exp!(M::Manifold, q, p, X)
exp!(M::Manifold, q, p, X, t::Real = 1)

Compute the exponential map of tangent vector X, optionally scaled by t, at point p from manifold the Manifold M. The result is saved to q.

ManifoldsBase.geodesic — Method

geodesic(M::Manifold, p, X) -> Function

Get the geodesic with initial point p and velocity X on the Manifold M. The geodesic is the curve of constant velocity that is locally distance-minimizing. This function returns a function of (time) t.

geodesic(M::Manifold, x, v, t::Real)
geodesic(M::Manifold, x, v, T::AbstractVector) -> AbstractVector

Return the point at time t or points at times t in T along the geodesic.

ManifoldsBase.hat — Method

hat(M::Manifold, p, Xⁱ)

Given a basis $e_i$ on the tangent space at a point p and tangent component vector $X^i$, compute the equivalent vector representation $X=X^i e_i$, where Einstein summation notation is used:

\[∧ : X^i ↦ X^i e_i\]

For array manifolds, this converts a vector representation of the tangent vector to an array representation. The vee map is the hat map's inverse.

ManifoldsBase.injectivity_radius — Method

injectivity_radius(M::Manifold, p)

Return the distance $d$ such that exp(M, p, X) is injective for all tangent vectors shorter than $d$ (i.e. has an inverse).

injectivity_radius(M::Manifold)

Infimum of the injectivity radius of all manifold points.

injectivity_radius(M::Manifold[, x], method::AbstractRetractionMethod)
injectivity_radius(M::Manifold, x, method::AbstractRetractionMethod)

Distance $d$ such that retract(M, p, X, method) is injective for all tangent vectors shorter than $d$ (i.e. has an inverse) for point p if provided or all manifold points otherwise.

ManifoldsBase.inner — Method

inner(M::Manifold, p, X, Y)

Compute the inner product of tangent vectors X and Y at point p from the Manifold M.

ManifoldsBase.inverse_retract! — Method

inverse_retract!(M::Manifold, X, p, q[, method::AbstractInverseRetractionMethod])

Compute the inverse retraction, a cheaper, approximate version of the logarithmic map), of points p and q on the Manifold M. Result is saved to X.

Inverse retraction method can be specified by the last argument, defaulting to LogarithmicInverseRetraction. See the documentation of respective manifolds for available methods.

ManifoldsBase.inverse_retract — Method

inverse_retract(M::Manifold, x, y)
inverse_retract(M::Manifold, x, y, method::AbstractInverseRetractionMethod

Compute the inverse retraction, a cheaper, approximate version of the logarithmic map), of points p and q on the Manifold M.

Inverse retraction method can be specified by the last argument, defaulting to LogarithmicInverseRetraction. See the documentation of respective manifolds for available methods.

ManifoldsBase.is_manifold_point — Function

is_manifold_point(M::Manifold, p, throw_error = false; kwargs...)

Return whether p is a valid point on the Manifold M.

If throw_error is false, the function returns either true or false. If throw_error is true, the function either returns true or throws an error. By default the function calls check_manifold_point(M, p; kwargs...) and checks whether the returned value is nothing or an error.

ManifoldsBase.is_tangent_vector — Function

is_tangent_vector(M::Manifold, p, X, throw_error = false; kwargs...)

Return whether X is a valid tangent vector at point p on the Manifold M. Returns either true or false.

If throw_error is false, the function returns either true or false. If throw_error is true, the function either returns true or throws an error. By default the function calls check_tangent_vector(M, p, X; kwargs...) and checks whether the returned value is nothing or an error.

ManifoldsBase.log! — Method

log!(M::Manifold, X, p, q)

Compute the logarithmic map of point q at base point p on the Manifold M. THe result is saved to X.

ManifoldsBase.manifold_dimension — Method

manifold_dimension(M::Manifold)

The dimension $n=\dim_{\mathcal M}$ of real space $\mathbb R^n$ to which the neighborhood of each point of the Manifold M is homeomorphic.

ManifoldsBase.number_eltype — Method

number_eltype(x)

Numeric element type of the a nested representation of a point or a vector. To be used in conjuntion with allocate or allocate_result.

ManifoldsBase.project_point! — Method

project_point!(M::Manifold, q, p)

Project point p from the ambient space onto the Manifold M. The point q is overwritten by the projection. The function works only for selected embedded manifolds and is not required to return the closest point.

ManifoldsBase.project_point — Method

project_point(M::Manifold, p)

Project point pfrom the ambient space onto the Manifold M. The function works only for selected embedded manifolds and is not required to return the closest point.

ManifoldsBase.project_tangent! — Method

project_tangent!(M::Manifold, Y, p, X)

Project ambient space representation of a vector X to a tangent vector at point p on the Manifold M. The result is saved in vector Y.

The function works only for selected embedded manifolds and is not required to return the closest vector.

ManifoldsBase.project_tangent — Method

project_tangent(M::Manifold, p, X)

Project ambient space representation of a vector X to a tangent vector at point p on the Manifold M.

The function works only for selected embedded manifolds and is not required to return the closest vector.

ManifoldsBase.representation_size — Method

representation_size(M::Manifold)

The size of an array representing a point on Manifold M.

ManifoldsBase.retract! — Method

retract!(M::Manifold, q, p, X)
retract!(M::Manifold, q, p, X, t::Real=1)
retract!(M::Manifold, q, p, X, method::AbstractRetractionMethod)
retract!(M::Manifold, q, p, X, t::Real=1, method::AbstractRetractionMethod)

Compute a retraction, a cheaper, approximate version of the exponential map, from p into direction X, scaled by t, on the Manifold manifold M. Result is saved to q.

Retraction method can be specified by the last argument, defaulting to ExponentialRetraction. See the documentation of respective manifolds for available methods.

ManifoldsBase.retract — Method

retract(M::Manifold, p, X)
retract(M::Manifold, p, X, t::Real=1)
retract(M::Manifold, p, X, method::AbstractRetractionMethod)
retract(M::Manifold, p, X, t::Real=1, method::AbstractRetractionMethod)

Compute a retraction, a cheaper, approximate version of the exponential map, from p into direction X, scaled by t, on the Manifold M.

Retraction method can be specified by the last argument, defaulting to ExponentialRetraction. See the documentation of respective manifolds for available methods.

ManifoldsBase.shortest_geodesic — Method

shortest_geodesic(M::Manifold, p, q) -> Function

Get a geodesic $\gamma_{p,q}(t)$ whose length is the shortest path between the points pand q, where $\gamma_{p,q}(0)=p$ and $\gamma_{p,q}(1)=q$. When there are multiple shortest geodesics, there is no guarantee which will be returned.

This function returns a function of time, which may be a Real or an AbstractVector.

shortest_geodesic(M::Manifold, p, q, t::Real)
shortest_geodesic(M::Manifold, p, q, T::AbstractVector) -> AbstractVector

Return the point at time t or points at times t in T along the shortest geodesic.

ManifoldsBase.vector_transport_along! — Method

vector_transport_along!(M::Manifold, Y, p, X, c)
vector_transport_along!(M::Manifold, Y, p, X, c, method::AbstractVectorTransportMethod)

Transport a vector X from a point p along the curve c such that c(0) is equal to p to the point c(1) using the method, which defaults to ParallelTransport. The result is saved to Y.

ManifoldsBase.vector_transport_along — Method

vector_transport_along(M::Manifold, p, X, c)
vector_transport_along(M::Manifold, p, X, c, method::AbstractVectorTransportMethod)

Transport a vector X from a point p along the curve c such that c(0) is equal to p to the point c(1) using the method, which defaults to ParallelTransport.

ManifoldsBase.vector_transport_direction! — Method

vector_transport_direction!(M::Manifold, Y, p, X, d)
vector_transport_direction!(M::Manifold, Y, p, X, d, method::AbstractVectorTransportMethod)

Transport a vector X from a point p in the direction indicated by the tangent vector d at point p. The result is saved to Y. By default, exp and vector_transport_to! are used with the method, which defaults to ParallelTransport.

ManifoldsBase.vector_transport_direction — Method

vector_transport_direction(M::Manifold, p, X, d)
vector_transport_direction(M::Manifold, p, X, d, method::AbstractVectorTransportMethod)

Transport a vector X from a point p in the direction indicated by the tangent vector d at point p. By default, exp and vector_transport_to! are used with the method, which defaults to ParallelTransport.

ManifoldsBase.vector_transport_to! — Method

vector_transport_to!(M::Manifold, Y, p, X, q, method::ProjectionTransport)

Transport a vector X from the tangent space at p on a Manifold M by interpreting it as an element of the embedding and then projecting it onto the tangent space at q. This method requires project_tangent.

ManifoldsBase.vector_transport_to! — Method

vector_transport_to!(M::Manifold, Y, p, X, q)
vector_transport_to!(M::Manifold, Y, p, X, q, method::AbstractVectorTransportMethod)

Compute the vector transport of vector X at point p to point q. The result is saved to Y. By default, the AbstractVectorTransportMethod method is ParallelTransport.

ManifoldsBase.vector_transport_to — Method

vector_transport_to(M::Manifold, p, X, q)
vector_transport_to(M::Manifold, p, X, q, method::AbstractVectorTransportMethod)

Compute the vector transport of vector X at point p to point q. By default, the AbstractVectorTransportMethod method is ParallelTransport.

ManifoldsBase.vee — Method

vee(M::Manifold, p, X)

Given a basis $e_i$ on the tangent space at a point p and tangent vector X, compute the vector components $X^i$, such that $X = X^i e_i$, where Einstein summation notation is used:

\[\vee : X^i e_i ↦ X^i\]

For array manifolds, this converts an array representation of the tangent vector to a vector representation. The hat map is the vee map's inverse.

ManifoldsBase.zero_tangent_vector! — Method

zero_tangent_vector!(M::Manifold, X, p)

Save to X a vector such that retracting X to the Manifold M at p produces p.

ManifoldsBase.zero_tangent_vector — Method

zero_tangent_vector(M::Manifold, p)

Return the tangent vector from the tangent space at p on the Manifold M, that represents the zero vector, i.e. such that a retraction at p produces p.

ManifoldsBase.AbstractEstimationMethod — Type

AbstractEstimationMethod

Abstract type for defining statistical estimation methods.

ManifoldsBase.AbstractVectorTransportMethod — Type

AbstractVectorTransportMethod

Abstract type for methods for transporting vectors.

ManifoldsBase.OutOfInjectivityRadiusError — Type

OutOfInjectivityRadiusError

An error thrown when a function (for example logarithmic map or inverse_retract) is given arguments outside of its injectivity_radius.

Base.angle — Method

angle(M::Manifold, p, X, Y)

Compute the angle between tangent vectors X and Y at point p from the Manifold M with respect to the inner product from inner.

ManifoldsBase.allocate_result — Method

allocate_result(M::Manifold, f, x...)

Allocate an array for the result of function f on Manifold M and arguments x... for implementing the non-modifying operation using the modifying operation.

Usefulness of passing a function is demonstrated by methods that allocate results of musical isomorphisms.

ManifoldsBase.allocate_result_type — Method

allocate_result_type(M::Manifold, f, args::NTuple{N,Any}) where N

Return type of element of the array that will represent the result of function f and the Manifold M on given arguments args (passed as a tuple).

DefaultManifold

DefaultManifold is a simplified version of Euclidean and demonstrates a basic interface implementation.

ManifoldsBase.DefaultManifold — Type

DefaultManifold <: Manifold

This default manifold illustrates the main features of the interface and provides a skeleton to build one's own manifold. It is a simplified/shortened variant of Euclidean from Manifolds.jl.

This manifold further illustrates how to type your manifold points and tangent vectors. Note that the interface does not require this, but it might be handy in debugging and educative situations to verify correctness of involved variabes.

Allocation

Non-mutating functions in Manifolds.jl are typically implemented using mutating variants. Allocation of new points is performed using a custom mechanism that relies on the following functions:

allocate that allocates a new point or vector similar to the given one. This function behaves like similar for simple representations of points and vectors (for example Array{Float64}). For more complex types, such as nested representations of PowerManifold (see NestedPowerRepresentation), FVector types, checked types like ArrayMPoint and more it operates differently. While similar only concerns itself with the higher level of nested structures, allocate maps itself through all levels of nesting until a simple array of numbers is reached and then calls similar. The difference can be most easily seen in the following example:

julia> x = similar([[1.0], [2.0]])
2-element Array{Array{Float64,1},1}:
 #undef
 #undef

julia> y = Manifolds.allocate([[1.0], [2.0]])
2-element Array{Array{Float64,1},1}:
 [6.90031725726027e-310]
 [6.9003678131654e-310]

julia> x[1]
ERROR: UndefRefError: access to undefined reference
Stacktrace:
 [1] getindex(::Array{Array{Float64,1},1}, ::Int64) at ./array.jl:744
 [2] top-level scope at REPL[12]:1

julia> y[1]
1-element Array{Float64,1}:
 6.90031725726027e-310

allocate_result allocates a result of a particular function (for example [exp], [flat], etc.) on a particular manifold with particular arguments. It takes into account the possibility that different arguments may have different numeric number_eltype types thorough the ManifoldsBase.allocate_result_type function.