Differentiation

Documentation for Manifolds.jl's methods and types for finite differences and automatic differentiation.

Differentiation backends

Manifolds.AbstractDiffBackend — Type

AbstractDiffBackend

An abstract type for diff backends. See FiniteDifferencesBackend for an example.

Manifolds.FiniteDifferencesBackend — Type

FiniteDifferencesBackend(method::FiniteDifferenceMethod = central_fdm(5, 1))

Differentiation backend based on the FiniteDifferences package.

Manifolds.diff_backend! — Method

diff_backend!(backend::AbstractDiffBackend)

Set current backend for differentiation to backend.

Manifolds.diff_backend — Method

diff_backend() -> AbstractDiffBackend

Get the current differentiation backend.

Manifolds.diff_backends — Method

diff_backends() -> Vector{AbstractDiffBackend}

Get vector of currently valid differentiation backends.

Manifolds.CurrentDiffBackend — Type

CurrentDiffBackend(backend::AbstractDiffBackend)

A mutable struct for storing the current differentiation backend in a global constant _current_diff_backend.

See also

AbstractDiffBackend, diff_backend, diff_backend!

Manifolds._derivative — Method

_derivative(f, t[, backend::AbstractDiffBackend])

Compute the derivative of a callable f at time t computed using the given backend, an object of type AbstractDiffBackend. If the backend is not explicitly specified, it is obtained using the function diff_backend.

Note

Not specifying the backend explicitly will usually result in a type instability and decreased performance.

Manifolds._gradient — Method

_gradient(f, p[, backend::AbstractDiffBackend])

Compute the gradient of a callable f at point p computed using the given backend, an object of type AbstractDiffBackend. If the backend is not explicitly specified, it is obtained using the function diff_backend.

Note

Not specifying the backend explicitly will usually result in a type instability and decreased performance.

Manifolds._hessian — Method

_hessian(f, p[, backend::AbstractDiffBackend])

Compute the Hessian of a callable f at point p computed using the given backend, an object of type AbstractDiffBackend. If the backend is not explicitly specified, it is obtained using the function diff_backend.

Note

Not specifying the backend explicitly will usually result in a type instability and decreased performance.

Manifolds._hessian_vector_product — Method

_hessian_vector_product(f, p, X[, backend::AbstractDiffBackend])

Compute product of the Hessian of a callable f at point p computed using the given backend, an object of type AbstractDiffBackend, and the vector X. If the backend is not explicitly specified, it is obtained using the function diff_backend.

Note

Not specifying the backend explicitly will usually result in a type instability and decreased performance.

Manifolds._jacobian — Method

_jacobian(f, p[, backend::AbstractDiffBackend])

Compute the Jacobian of a callable f at point p computed using the given backend, an object of type AbstractDiffBackend. If the backend is not explicitly specified, it is obtained using the function diff_backend.

Note

Not specifying the backend explicitly will usually result in a type instability and decreased performance.

Manifolds._current_diff_backend — Constant

_current_diff_backend

The instance of [CurrentDiffBackend] that stores the globally default differentiation backend.

Manifolds._diff_backends — Constant

_diff_backends

A vector of valid AbstractDiffBackend.

Riemannian differentiation backends

Manifolds.AbstractRiemannianDiffBackend — Type

AbstractRiemannianDiffBackend

An abstract type for diff backends. See RiemannianONBDiffBackend for an example.

Manifolds.RiemannianONBDiffBackend — Type

RiemannianONBDiffBackend(
    diff_backend::AbstractDiffBackend
    retraction::AbstractRetractionMethod
    inverse_retraction::AbstractInverseRetractionMethod
    basis::Union{AbstractOrthonormalBasis,CachedBasis{<:AbstractOrthonormalBasis}},
) <: AbstractRiemannianDiffBackend

Riemannian differentiation based on differentiation in an AbstractOrthonormalBasis basis using specified retraction, inverse_retraction and using backend diff_backend.

Manifolds.CurrentRiemannianDiffBackend — Type

CurrentRiemannianDiffBackend(backend::AbstractRiemannianDiffBackend)

A mutable struct for storing the current Riemannian differentiation backend in a global constant _current_rdiff_backend.

See also

AbstractRiemannianDiffBackend, rdiff_backend, [rdiff_backend!]

Manifolds.RiemannianProjectionGradientBackend — Type

RiemannianProjectionGradientBackend(
    diff_backend::AbstractDiffBackend
) <: AbstractRiemannianDiffBackend

Riemannian differentiation based on differentiation in the ambient space and projection to the given manifold. Differentiation in the ambient space is performed using the backend diff_backend.

Only valid for manifolds that are embedded in a special way in the Euclidean space. See [Absil2008], Section 3.6.1 for details.

Manifolds.differential — Method

differential(f::Any, c::Curve, t::Real, backend::AbstractDiffBackend = rdiff_backend())

Compute the Riemannian differential of a curve of type c represented by function f at time t using the given backend.

Manifolds.gradient — Method

gradient(f::Any, rf::RealField, p, backend::AbstractRiemannianDiffBackend = rdiff_backend())

Compute the Riemannian gradient of a real field of type rf represented by function f at point p using the given backend.

Manifolds.hessian — Method

hessian(f::Any, rf::RealField, p, backend::AbstractRiemannianDiffBackend = rdiff_backend())

Compute the Riemannian Hessian of a real field of type rf represented by function f at point p using the given backend.

Manifolds.hessian — Method

hessian(f, ft::RealField, p, backend::RiemannianONBDiffBackend)

Compute the Riemannian Hessian using the Euclidean Hessian according to Proposition 5.5.4 from [Absil2008] (generalized to arbitrary retractions).

Manifolds.hessian_vector_product — Method

hessian_vector_product(f, ft::RealField, p, X, backend::AbstractRiemannianDiffBackend)

Compute the product of Riemannian Hessian of a real field of type rf represented by function f at point p and the tangent vector X at point p using the given backend.

Manifolds.jacobian — Method

jacobian(f::Any, mt::AbstractMap, p, backend::AbstractRiemannianDiffBackend = rdiff_backend())

Compute the Riemannian Jacobian of a map of type mt represented by function f at point p using the given backend.

Manifolds.rdifferential_backend! — Method

rdifferential_backend!(backend::AbstractRiemannianDiffBackend)

Set current Riemannian differential backend for differentiation to backend.

Manifolds.rdifferential_backend — Method

rdifferential_backend() -> AbstractRiemannianDiffBackend

Get the current differentiation backend for Riemannian differentials.

Manifolds.rgradient_backend! — Method

rgradient_backend!(backend::AbstractRiemannianDiffBackend)

Set current Riemannian gradient backend for differentiation to backend.

Manifolds.rgradient_backend — Method

rgradient_backend() -> AbstractRiemannianDiffBackend

Get the current differentiation backend for Riemannian gradients.

Manifolds.rhessian_backend! — Method

rhessian_backend!(backend::AbstractRiemannianDiffBackend)

Set current Riemannian Hessian backend for differentiation to backend.

Manifolds.rhessian_backend — Method

rhessian_backend() -> AbstractRiemannianDiffBackend

Get the current differentiation backend for Riemannian Hessians.

Manifolds.rjacobian_backend! — Method

rjacobian_backend!(backend::AbstractRiemannianDiffBackend)

Set current Riemannian Jacobian backend for differentiation to backend.

Manifolds.rjacobian_backend — Method

rjacobian_backend() -> AbstractRiemannianDiffBackend

Get the current differentiation backend for Riemannian Jacobians.

Manifolds._current_rgradient_backend — Constant

_current_rgradient_backend

The instance of CurrentRiemannianGradientBackend that stores the globally default differentiation backend.

Absil2008

Absil, P. A., et al. Optimization Algorithms on Matrix Manifolds. 2008.

Absil2008

Absil, P. A., et al. Optimization Algorithms on Matrix Manifolds. 2008.