Internals

Internal Documentation

Documentation for Manifolds.jl's internal methods and types.

Manifolds.SizedAbstractArray — Type.

SizedAbstractArray{Tuple{dims...}}(array)

Wraps an AbstractArray with a static size, so to take advantage of the (faster) methods defined by the static array package. The size is checked once upon construction to determine if the number of elements (length) match, but the array may be reshaped.

Manifolds.ziptuples — Function.

ziptuples(a, b)

Zips tuples a and b in a fast, type-stable way. If they have different lengths, the result is trimmed to the length of the shorter tuple.

ziptuples(a, b, c)

Zips tuples a, b and c in a fast, type-stable way. If they have different lengths, the result is trimmed to the length of the shorter tuple.

ziptuples(a, b, c, d)

Zips tuples a, b, c and d in a fast, type-stable way. If they have different lengths, the result is trimmed to the length of the shorter tuple.

ziptuples(a, b, c, d, e)

Zips tuples a, b, c, d and e in a fast, type-stable way. If they have different lengths, the result is trimmed to the length of the shorter tuple.

Manifolds.size_to_tuple — Function.

size_to_tuple(::Type{S}) where S<:Tuple

Converts a size given by Tuple{N, M, ...} into a tuple (N, M, ...).

Manifolds.select_from_tuple — Function.

select_from_tuple(t::NTuple{N, Any}, positions::Val{P})

Selects elements of tuple t at positions specified by the second argument. For example select_from_tuple(("a", "b", "c"), Val((3, 1, 1))) returns ("c", "a", "a").

Manifolds.find_pv — Function.

find_pv(x...)

A = find_pv(x...) returns the first ProductArray among the arguments.