Testing

Documentation for testing utilities for Manifolds.jl. The function test_manifold can be used to verify that your manifold correctly implements the Manifolds.jl interface.

Manifolds.ManifoldTests.test_manifold — Method

test_manifold(
    M::Manifold,
    pts::AbstractVector;
    args,
)

Test general properties of manifold M, given at least three different points that lie on it (contained in pts).

Arguments

basis_has_specialized_diagonalizing_get = false: if true, assumes that DiagonalizingOrthonormalBasis given in basis_types has get_coordinates and get_vector that work without caching.

basis_types_to_from = (): basis types that will be tested based on get_coordinates and get_vector.

basis_types_vecs = () : basis types that will be tested based on get_vectors.

default_inverse_retraction_method = ManifoldsBase.LogarithmicInverseRetraction(): default method for inverse retractions (log).

default_retraction_method = ManifoldsBase.ExponentialRetraction(): default method for retractions (exp).

exp_log_atol_multiplier = 0: change absolute tolerance of exp/log tests (0 use default, i.e. deactivate atol and use rtol).

exp_log_rtol_multiplier = 1: change the relative tolerance of exp/log tests (1 use default). This is deactivated if the exp_log_atol_multiplier is nonzero.

expected_dimension_type = Integer: expected type of value returned by manifold_dimension.

inverse_retraction_methods = []: inverse retraction methods that will be tested.

is_mutating = true: whether mutating variants of functions should be tested.

is_point_atol_multiplier = 0: determines atol of is_manifold_point checks.

is_tangent_atol_multiplier = 0: determines atol of is_tangent_vector checks.

mid_point12 = shortest_geodesic(M, pts[1], pts[2], 0.5): if not nothing, then check that mid_point(M, pts[1], pts[2]) is approximately equal to mid_point12.

point_distributions = [] : point distributions to test.

rand_tvector_atol_multiplier = 0 : chage absolute tolerance in testing random vectors (0 use default, i.e. deactivate atol and use rtol) random tangent vectors are tangent vectors.

retraction_atol_multiplier = 0: change absolute tolerance of (inverse) retraction tests (0 use default, i.e. deactivate atol and use rtol).

retraction_rtol_multiplier = 1: change the relative tolerance of (inverse) retraction tests (1 use default). This is deactivated if the exp_log_atol_multiplier is nonzero.

retraction_methods = []: retraction methods that will be tested.

test_exp_log = true: if true, checkthat exp is the inverse of log.

test_forward_diff = true: if true, automatic differentiation using ForwardDiff is tested.

test_injectivity_radius = true: whether implementation of injectivity_radius should be tested.

test_is_tangent: if true check that the default_inverse_retraction_method actually returns valid tangent vectors.

test_musical_isomorphisms = false : test musical isomorphisms.

test_mutating_rand = false : test the mutating random function for points on manifolds.

test_project_point = false: test projections onto the manifold.

test_project_tangent = false : test projections on tangent spaces.

test_representation_size = true : test repersentation size of points/tvectprs.

test_reverse_diff = true: if true, automatic differentiation using ReverseDiff is tested.

test_tangent_vector_broadcasting = true : test boradcasting operators on TangentSpace.

test_vector_spaces = true : test Vector bundle of this manifold.

test_vector_transport = false : test vector transport.

test_vee_hat = false: test vee and hat functions.

tvector_distributions = [] : tangent vector distributions to test.

vector_transport_methods = []: vector transport methods that should be tested.

Manifolds.ManifoldTests.find_eps — Method

find_eps(x...)

Find an appropriate tolerance for given points or tangent vectors, or their types.