Array manifold

Array manifold

Manifolds.ArrayManifold — Type.

ArrayManifold{M <: Manifold} <: Manifold

A manifold to encapsulate manifolds working on array representations of MPoints and TVectors in a transparent way, such that for these manifolds its not necessary to introduce explicit types for the points and tangent vectors, but they are encapusalted/stripped automatically when needed.