Circle

Manifolds.Circle — Type

Circle{𝔽} <: AbstractManifold{𝔽}

The circle $𝕊^1$ is a manifold here represented by real-valued points in $[-π,π)$ or complex-valued points $z ∈ ℂ$ of absolute value $\lvert z\rvert = 1$.

Constructor

Circle(𝔽=ℝ)

Generate the ℝ-valued Circle represented by angles, which alternatively can be set to use the AbstractNumbers 𝔽=ℂ to obtain the circle represented by ℂ-valued circle of unit numbers.

Base.exp — Method

exp(M::Circle, p, X)

Compute the exponential map on the Circle.

\[\exp_p X = (p+X)_{2π},\]

where $(⋅)_{2π}$ is the (symmetric) remainder with respect to division by $2π$, i.e. in $[-π,π)$.

For the complex-valued case, the same formula as for the Sphere $𝕊^1$ is applied to values in the complex plane.

Base.log — Method

log(M::Circle, p, q)

Compute the logarithmic map on the Circle M.

\[\log_p q = (q-p)_{2π},\]

where $(⋅)_{2π}$ is the (symmetric) remainder with respect to division by $2π$, i.e. in $[-π,π)$.

For the complex-valued case, the same formula as for the Sphere $𝕊^1$ is applied to values in the complex plane.

Base.rand — Method

Random.rand(M::Circle{ℝ}; vector_at = nothing, σ::Real=1.0)

If vector_at is nothing, return a random point on the Circle $\mathbb S^1$ by picking a random element from $[-\pi,\pi)$ uniformly.

If vector_at is not nothing, return a random tangent vector from the tangent space of the point vector_at on the Circle by using a normal distribution with mean 0 and standard deviation σ.

Manifolds.complex_dot — Method

complex_dot(a, b)

Compute the inner product of two (complex) numbers with in the complex plane.

Manifolds.manifold_volume — Method

manifold_volume(M::Circle)

Return the volume of the Circle M, i.e. $2π$.

Manifolds.sym_rem — Method

sym_rem(x,[T=π])

Compute symmetric remainder of x with respect to the interall 2*T, i.e. (x+T)%2T, where the default for T is $π$

Manifolds.volume_density — Method

volume_density(::Circle, p, X)

Return volume density of Circle, i.e. 1.

ManifoldsBase.check_point — Method

check_point(M::Circle, p)

Check whether p is a point on the Circle M. For the real-valued case, p is an angle and hence it checks that $p ∈ [-π,π)$. for the complex-valued case, it is a unit number, $p ∈ ℂ$ with $\lvert p \rvert = 1$.

ManifoldsBase.check_vector — Method

check_vector(M::Circle, p, X; kwargs...)

Check whether X is a tangent vector in the tangent space of p on the Circle M. For the real-valued case represented by angles, all X are valid, since the tangent space is the whole real line. For the complex-valued case X has to lie on the line parallel to the tangent line at p in the complex plane, i.e. their inner product has to be zero.

ManifoldsBase.distance — Method

distance(M::Circle, p, q)

Compute the distance on the Circle M between p and q. For the real-valued case this is the absolute value of the symmetric remainder of p and q, for the complex-valued case it is the angle between both complex numbers in the Gaussian plane.

ManifoldsBase.embed — Method

embed(M::Circle{ℂ}, p, X)

Embed a tangent vector X at p on the complex Circle M in the ambient space. It returns X.

ManifoldsBase.embed — Method

embed(M::Circle{ℂ}, p)

Embed a point p on the complex Circle M in the ambient space. It returns p.

ManifoldsBase.get_coordinates — Method

get_coordinates(M::Circle{ℂ}, p, X, B::DefaultOrthonormalBasis)

Return tangent vector coordinates in the Lie algebra of the Circle.

ManifoldsBase.get_embedding — Method

get_embedding(M::Circle{ℂ})

Get the ambient space of the complex Circle M, which is ℂ.

ManifoldsBase.get_vector_orthonormal — Method

get_vector(M::Circle{ℂ}, p, X, B::DefaultOrthonormalBasis)

Return tangent vector from the coordinates in the Lie algebra of the Circle.

ManifoldsBase.injectivity_radius — Method

injectivity_radius(M::Circle[, p])

Return the injectivity radius on the Circle M, i.e. $π$.

ManifoldsBase.inner — Method

inner(M::Circle, p, X, Y)

Compute the inner product of the two tangent vectors X,Y from the tangent plane at p on the Circle M using the restriction of the metric from the embedding, i.e.

\[g_p(X,Y) = X*Y\]

for the real case and

\[g_p(X,Y) = Y^\mathrm{T}X\]

for the complex case interpreting complex numbers in the Gaussian plane.

ManifoldsBase.is_flat — Method

is_flat(::Circle)

Return true. Circle is a flat manifold.

ManifoldsBase.manifold_dimension — Method

manifold_dimension(M::Circle)

Return the dimension of the Circle M, i.e. $\dim(𝕊^1) = 1$.

ManifoldsBase.parallel_transport_to — Method

 parallel_transport_to(M::Circle, p, X, q)

Compute the parallel transport of X from the tangent space at p to the tangent space at q on the Circle M. For the real-valued case this results in the identity. For the complex-valued case, the formula is the same as for the Sphere(1) in the complex plane.

\[\mathcal P_{q←p} X = X - \frac{⟨\log_p q,X⟩_p}{d^2_{ℂ}(p,q)} \bigl(\log_p q + \log_q p \bigr),\]

where log denotes the logarithmic map on M.

ManifoldsBase.project — Method

project(M::Circle, p, X)

Project a value X onto the tangent space of the point p on the Circle M.

For the complex valued case X is projected onto the line in the complex plane that is parallel to the tangent to p on the unit circle and contains 0.

ManifoldsBase.project — Method

project(M::Circle{ℂ}, p)

Project a point p onto the complex Circle M, i.e. the unit circle in the complex plane.

Statistics.mean — Method

mean(M::Circle{ℂ}, x::AbstractVector[, w::AbstractWeights])

Compute the Riemannian mean of x of points on the Circle $𝕊^1$, represented by complex numbers, i.e. embedded in the complex plane. Computing the sum

\[s = \sum_{i=1}^n x_i\]

the mean is the angle of the complex number $s$, so represented in the complex plane as $\frac{s}{\lvert s \rvert}$ whenever $s \neq 0$.

If the sum $s=0$, the mean is not unique. For example for opposite points or equally spaced angles.

Statistics.mean — Method

mean(M::Circle{ℝ}, x::AbstractVector[, w::AbstractWeights])

Compute the Riemannian mean of x of points on the Circle $𝕊^1$, represented by real numbers, i.e. the angular mean

\[\operatorname{atan}\Bigl( \sum_{i=1}^n w_i\sin(x_i), \sum_{i=1}^n w_i\sin(x_i) \Bigr).\]

Circle{𝔽} <: AbstractManifold{𝔽}
The circle is a manifold here represented by real-valued points in or complex-valued points of absolute value .

sym_rem(x,[T=π])
Compute symmetric remainder of x with respect to the interall 2*T, i.e. (x+T)%2T, where the default for T is

get_coordinates(M::Circle{ℂ}, p, X, B::DefaultOrthonormalBasis)
Return tangent vector coordinates in the Lie algebra of the Circle.

get_coordinates(M::Rotations, p, X) get_coordinates(M::OrthogonalMatrices, p, X) get_coordinates(M::UnitaryMatrices, p, X)
Extract the unique tangent vector components at point p on Rotations from the matrix representation X of the tangent vector.

get_coordinates(M::UnitaryMatrices, p, X, B::DefaultOrthonormalBasis)
Extract the unique tangent vector coordinates at point p on UnitaryMatrices from the skew-Hermitian tangent vector X.

get_vector(M::OrthogonalMatrices, p, Xⁱ, B::DefaultOrthogonalBasis) get_vector(M::Rotations, p, Xⁱ, B::DefaultOrthogonalBasis)
Convert the unique tangent vector components Xⁱ at point p on Rotations or OrthogonalMatrices to the matrix representation of the tangent vector.

get_vector(M::UnitaryMatrices, p, c, B::DefaultOrthonormalBasis)
Convert tangent coordinates c at point p on UnitaryMatrices to the skew-Hermitian tangent vector X.

get_coordinates(M::Grasmmann{ℝ}, p, X, B::DefaultOrthonormalBasis)
Given a point p on the Grassmann manifold M in Stiefel representation, i.e. compute the coordinates representing the tangent vector X with respect to the DefaultOrthonormalBasis.

get_vector(M::Grasmmann{ℝ}, p, c, B::DefaultOrthonormalBasis)
Given a point p on the Grassmann manifold M in Stiefel representation, i.e. reconstruct a tangent vector with respect to the DefaultOrthonormalBasis given coefficients .

get_coordinates(M::HeisenbergMatrices, p, X, ::DefaultOrthonormalBasis{ℝ,TangentSpaceType})
Get coordinates of tangent vector X at point p from the HeisenbergMatrices M.

get_vector(M::HeisenbergMatrices, p, Xⁱ, ::DefaultOrthonormalBasis{ℝ,TangentSpaceType})
Get tangent vector with coordinates Xⁱ at point p from the HeisenbergMatrices M.

get_coordinates(M::Hyperbolic, p, X, ::DefaultOrthonormalBasis)
Compute the coordinates of the vector X with respect to an orthonormal basis of the tangent space at p.

get_vector(M::Hyperbolic, p, c, ::DefaultOrthonormalBasis)
Compute the tangent vector from the coordinates c with respect to an orthonormal basis of the tangent space at p.

get_coordinates(::PositiveNumbers, p, X, ::DefaultOrthonormalBasis{ℝ})
Compute the coordinate of vector X which is tangent to p on the PositiveNumbers manifold.

get_vector(::PositiveNumbers, p, c, ::DefaultOrthonormalBasis{ℝ})
Compute the vector with coordinate c which is tangent to p on the PositiveNumbers manifold.

get_coordinates(M::AbstractProjectiveSpace, p, X, B::DefaultOrthonormalBasis{ℝ})
Represent the tangent vector at point from the AbstractProjectiveSpace in an orthonormal basis.

get_vector(M::AbstractProjectiveSpace, p, X, B::DefaultOrthonormalBasis{ℝ})
Convert a one-dimensional vector of coefficients in the basis B of the tangent space at on the AbstractProjectiveSpace to a tangent vector at .

get_coordinates(M::Segre{𝔽, V}, p, X, ::DefaultOrthonormalBasis; kwargs...)
Get coordinates of X in the tangent space using a DefaultOrthonormalBasis on each factor.

get_coordinates(M::Segre{𝔽, V}, p, v, ::DefaultOrthonormalBasis; kwargs...)
Get coordinates of X in the tangent space using a DefaultOrthonormalBasis on each factor.

get_vector( M::Segre{𝔽, V}, p, c, DefaultOrthonormalBasis; kwargs...)
Get tangent vector X from coordinates in the tangent space using a DefaultOrthonormalBasis on each factor.

get_vector(M::Segre{𝔽, V}, p, c, ::DefaultOrthonormalBasis; kwargs...)
Get tangent vector X from coordinates in the tangent space using a DefaultOrthonormalBasis on each factor.

get_coordinates(M::AbstractSphere{ℝ}, p, X, B::DefaultOrthonormalBasis)
Represent the tangent vector X at point p from the AbstractSphere M in an orthonormal basis by rotating the hyperplane containing X to a hyperplane whose normal is the -axis.

get_vector(M::AbstractSphere{ℝ}, p, c, B::DefaultOrthonormalBasis)
Convert a one-dimensional vector of coefficients c with respect to an orthonormal basis B of the tangent space at p on the AbstractSphere M to a tangent vector Y at p.

rand(::SymplecticMatrices; vector_at=nothing, σ::Real=1.0)
Generate a random point on or a random tangent vector if vector_at is set to a point .