Embedded manifold

Some manifolds can easily be defined by using a certain embedding. For example the Sphere(n) is embedded in Euclidean(n+1). Similar to the metric and MetricManifold, an embedding is often implicitly assumed. We introduce the embedded manifolds hence as an AbstractDecoratorManifold.

This decorator enables to use such an embedding in an transparent way. Different types of embeddings can be distinguished using the AbstractEmbeddingType.

The embedding also covers representation of tangent vectors. For both points and tangent vectors the function embed returns their representation in the embedding. For any point or vector in the embedding the functions project_point and project_tangent can be used to obtain the closest point on the manifold and tangent vector in the tangent space, respectively. A specific example where embed might be useful, is for example a Lie group, where tangent vectors are often represented in the Lie algebra. Then their representation is different from the representation in the embedding.

Isometric Embeddings

For isometric embeddings the type AbstractIsometricEmbeddingType can be used to avoid reimplementing the metric. See Sphere or Hyperbolic for example. Here, the exponential map, the logarithmic map, the retraction and its inverse are set to :intransparent, i.e. they have to be implemented.

Furthermore, the TransparentIsometricEmbedding type even states that the exponential and logarithmic maps as well as retractions and vector transports of the embedding can be used for the embedded manifold as well. See SymmetricMatrices for an example.

In both cases of course check_manifold_point and check_tangent_vector have to be implemented.

Technical Details

Semantically we use the idea of the embedding to efficiently implement a manifold by not having to implement those functions that are already given by its embedding. Hence we decorate in some sense the manifold we implement. Still, technically base_manifold returns the embedding as long as EmbeddedManifold is used. For the abstract case, AbstractEmbeddedManifold the base manfiold might differ. Note that internally base_manifold uses decorated_manifold for one step of removing multiple decorators.

Clearly get_embedding always returns the embedding.

Types

Manifolds.AbstractEmbeddedManifold — Type

AbstractEmbeddedManifold{T<:AbstractEmbeddingType} <: AbstractDecoratorManifold

An abstract type for embedded manifolds, which acts as an AbstractDecoratorManifold. The functions of the manifold that is embedded can hence be just passed on to the embedding. The embedding is further specified by an AbstractEmbeddingType.

This means, that technically an embedded manifold is a decorator for the embedding, i.e. functions of this type get, in the semi-transparent way of the AbstractDecoratorManifold, passed on to the embedding.

Manifolds.AbstractEmbeddingType — Type

AbstractEmbeddingType

A type used to specify properties of an AbstractEmbeddedManifold.

Manifolds.AbstractIsometricEmbeddingType — Type

AbstractIsometricEmbeddingType <: AbstractEmbeddingType

Characterizes an embedding as isometric. For this case the inner product is passed from the embedded manifold to the embedding.

Manifolds.EmbeddedManifold — Type

EmbeddedManifold{MT <: Manifold, NT <: Manifold, ET} <: AbstractEmbeddedManifold{ET}

A type to represent that a Manifold M of type MT is indeed an emebedded manifold and embedded into the manifold N of type NT. Based on the AbstractEmbeddingType ET, this introduces methods for M by passing them through to embedding N.

Fields

manifold the manifold that is an embedded manifold

embedding a second manifold, the first one is embedded into

Constructor

EmbeddedManifold(M, N, e=TransparentIsometricEmbedding())

Generate the EmbeddedManifold of the Manifold M into the Manifold N with AbstractEmbeddingType e that by default is the most transparent TransparentIsometricEmbedding

Manifolds.TransparentIsometricEmbedding — Type

TransparentIsometricEmbedding <: AbstractIsometricEmbeddingType

Specify that an embedding is the default isometric embedding. This even inherits logarithmic and exponential map as well as retraction and inverse retractions from the embedding.

For an example, see SymmetricMatrices which are isometrically embedded in the Euclidean space of matrices but also inherit exponential and logarithmic maps.

Functions

Manifolds.embed — Method

embed(M::AbstractEmbeddedManifold, p)

return the embedded representation of a point p on the AbstractEmbeddedManifold M.

embed(M::AbstractEmbeddedManifold, p, X)

return the embedded representation of a tangent vector X at point p on the AbstractEmbeddedManifold M.

Manifolds.get_embedding — Method

get_embedding(M::AbstractEmbeddedManifold)

Return the Manifold N an AbstractEmbeddedManifold is embedded into.