Heisenberg matrices

Manifolds.HeisenbergMatrices — Type

HeisenbergMatrices{T} <: AbstractDecoratorManifold{𝔽}

Heisenberg matrices HeisenbergMatrices(n) is the manifold of $(n+2)×(n+2)$ matrices [BP08]

\[\begin{bmatrix} 1 & \mathbf{a} & c \\ \mathbf{0}_n & I_n & \mathbf{b} \\ 0 & \mathbf{0}_n^\mathrm{T} & 1 \end{bmatrix}\]

where $I_n$ is the $n×n$ unit matrix, $\mathbf{a}$ is a row vector of length $n$, $\mathbf{b}$ is a column vector of length $n$, $\mathbf{0}_n$ is the column zero vector of length $n$, and $c$ is a real number.

It is a submanifold of Euclidean(n+2, n+2) and the manifold of the HeisenbergGroup.

Constructor

HeisenbergMatrices(n::Int; parameter::Symbol=:type)

Generate the manifold of $(n+2)×(n+2)$ Heisenberg matrices.

ManifoldsBase.Weingarten — Method

Y = Weingarten(M::HeisenbergMatrices, p, X, V)
Weingarten!(M::HeisenbergMatrices, Y, p, X, V)

Compute the Weingarten map $\mathcal W_p$ at p on the HeisenbergMatrices M with respect to the tangent vector $X \in T_p\mathcal M$ and the normal vector $V \in N_p\mathcal M$.

Since this a flat space by itself, the result is always the zero tangent vector.

ManifoldsBase.is_flat — Method

is_flat(::HeisenbergMatrices)

Return true. HeisenbergMatrices is a flat manifold.

ManifoldsBase.manifold_dimension — Method

manifold_dimension(M::HeisenbergMatrices)

Return the dimension of HeisenbergMatrices(n), which is equal to $2n+1$.