Invertible matrices

Manifolds.InvertibleMatrices — Type

InvertibleMatrices{𝔽,T} <: AbstractDecoratorManifold{𝔽}

The AbstractManifold consisting of the real- or complex-valued invertible matrices, that is the set

\[\bigl\{p ∈ 𝔽^{n×n}\ \big|\ \det(p) \neq 0 \bigr\},\]

where the field $𝔽 ∈ \{ ℝ, ℂ\}$.

Constructor

InvertibleMatrices(n::Int, field::AbstractNumbers=ℝ)

Generate the manifold of $n×n$ invertible matrices.

Base.rand — Method

Random.rand(M::InvertibleMatrices; vector_at=nothing, kwargs...)

If vector_at is nothing, return a random point on the InvertibleMatrices manifold M by using rand in the embedding.

If vector_at is not nothing, return a random tangent vector from the tangent space of the point vector_at on the InvertibleMatrices by using by using rand in the embedding.

ManifoldsBase.Weingarten — Method

Y = Weingarten(M::InvertibleMatrices, p, X, V)
Weingarten!(M::InvertibleMatrices, Y, p, X, V)

Compute the Weingarten map $\mathcal W_p$ at p on the InvertibleMatrices M with respect to the tangent vector $X \in T_p\mathcal M$ and the normal vector $V \in N_p\mathcal M$.

Since this a flat space by itself, the result is always the zero tangent vector.

ManifoldsBase.check_point — Method

check_point(M::InvertibleMatrices{n,𝔽}, p; kwargs...)

Check whether p is a valid manifold point on the InvertibleMatrices M, i.e. whether p is an invertible matrix of size (n,n) with values from the corresponding AbstractNumbers 𝔽.

ManifoldsBase.check_vector — Method

check_vector(M::InvertibleMatrices{n,𝔽}, p, X; kwargs... )

Check whether X is a tangent vector to manifold point p on the InvertibleMatrices M, which are all matrices of size $n×n$ its values have to be from the correct AbstractNumbers.

ManifoldsBase.is_flat — Method

is_flat(::InvertibleMatrices)

Return true. InvertibleMatrices is a flat manifold.

ManifoldsBase.manifold_dimension — Method

manifold_dimension(M::InvertibleMatrices{n,𝔽})

Return the dimension of the InvertibleMatrices matrix M over the number system 𝔽, which is the same dimension as its embedding, the Euclidean(n, n; field=𝔽).

InvertibleMatrices{𝔽,T} <: AbstractDecoratorManifold{𝔽}
The AbstractManifold consisting of the real- or complex-valued invertible matrices, that is the set

manifold_dimension(M::InvertibleMatrices{n,𝔽})
Return the dimension of the InvertibleMatrices matrix M over the number system 𝔽, which is the same dimension as its embedding, the Euclidean(n, n; field=𝔽).

manifold_dimension(M::CenteredMatrices)
Return the manifold dimension of the CenteredMatrices m-by-n matrix M over the number system 𝔽, i.e.

manifold_dimension(M::CholeskySpace)
Return the manifold dimension for the CholeskySpace M, i.e.

manifold_dimension(M::Circle)
Return the dimension of the Circle M, i.e. .

manifold_dimension(M::DeterminantOneMatrices{n,𝔽})
Return the dimension of the DeterminantOneMatrices matrix M over the number system 𝔽, which is one dimension less than its embedding, the Euclidean(n, n; field=𝔽).

manifold_dimension(M::Elliptope)
returns the dimension of Elliptope M , i.e.

manifold_dimension(M::EssentialManifold{is_signed, ℝ})
Return the manifold dimension of the EssentialManifold, which is 5[TD17].

manifold_dimension(M::Euclidean)
Return the manifold dimension of the Euclidean M, i.e. the product of all array dimensions and the real_dimension of the underlying number system.

manifold_dimension(M::FixedRankMatrices)
Return the manifold dimension for the 𝔽-valued FixedRankMatrices M of dimension mxn of rank k, namely

manifold_dimension(M::Flag)
Return dimension of flag manifold .

manifold_dimension(M::GeneralizedGrassmann)
Return the dimension of the GeneralizedGrassmann(n,k,𝔽) manifold M, i.e.

manifold_dimension(M::GeneralizedStiefel)
Return the dimension of the GeneralizedStiefel manifold M=.

manifold_dimension(M::Rotations) manifold_dimension(M::OrthogonalMatrices)
Return the dimension of the manifold orthogonal matrices and of the manifold of rotations

manifold_dimension(M::SpecialUnitaryMatrices)
Return the dimension of the manifold of special unitary matrices.

manifold_dimension(M::UnitaryMatrices{n,ℂ}) where {n}
Return the dimension of the manifold unitary matrices.

manifold_dimension(M::UnitaryMatrices{<:Any,ℍ})
Return the dimension of the manifold unitary matrices.

manifold_dimension(N::GraphManifold{G,𝔽,M,EdgeManifold})
returns the manifold dimension of the GraphManifold N on the edges of a graph , i.e.

manifold_dimension(N::GraphManifold{G,𝔽,M,VertexManifold})
returns the manifold dimension of the GraphManifold N on the vertices of a graph , i.e.

manifold_dimension(M::Grassmann)
Return the dimension of the Grassmann(n,k,𝔽) manifold M, i.e.

manifold_dimension(M::HeisenbergMatrices)
Return the dimension of HeisenbergMatrices(n), which is equal to .

manifold_dimension(M::Hyperbolic)
Return the dimension of the hyperbolic space manifold , i.e. .

manifold_dimension(M::Hyperrectangle)
Return the manifold dimension of the Hyperrectangle M, i.e. the product of all array dimensions.

manifold_dimension(M::MultinomialDoubleStochastic)
returns the dimension of the MultinomialDoubleStochastic manifold namely

manifold_dimension(M::MultinomialSymmetric)
returns the dimension of the MultinomialSymmetric manifold namely

manifold_dimension(M::ParametricSurface)
Return the dimension of the parameter space of M.

manifold_dimension(M::PositiveNumbers)
Return the dimension of the PositiveNumbers M, i.e. of the 1-dimensional hyperbolic space,

manifold_dimension(M::ProbabilitySimplex)
Returns the manifold dimension of the probability simplex in , i.e.

manifold_dimension(M::AbstractProjectiveSpace{𝔽}) where {𝔽}
Return the real dimension of the AbstractProjectiveSpace M, respectively i.e. the real dimension of the embedding minus the real dimension of the field 𝔽.

manifold_dimension(M::KendallsPreShapeSpace)
Return the dimension of the KendallsPreShapeSpace manifold M.

manifold_dimension(M::KendallsShapeSpace)
Return the dimension of the KendallsShapeSpace manifold M.

manifold_dimension(M::SkewHermitianMatrices)
Return the dimension of the SkewHermitianMatrices matrix M over the number system 𝔽, i.e.

manifold_dimension(M::Spectrahedron)
returns the dimension of Spectrahedron M, i.e.

manifold_dimension(M::AbstractSphere)
Return the dimension of the AbstractSphere M, respectively i.e. the dimension of the embedding -1.

manifold_dimension(M::SphereSymmetricMatrices{<:Any,𝔽})
Return the manifold dimension of the SphereSymmetricMatrices n-by-n symmetric matrix M of unit Frobenius norm over the number system 𝔽, i.e.

manifold_dimension(M::Stiefel)
Return the dimension of the Stiefel manifold M=.

manifold_dimension(M::SymmetricMatrices{n,𝔽})
Return the dimension of the SymmetricMatrices matrix M over the number system 𝔽, i.e.

manifold_dimension(M::SymmetricPositiveDefinite)
returns the dimension of SymmetricPositiveDefinite M , i.e.

manifold_dimension(M::SymmetricPositiveSemidefiniteFixedRank)
Return the dimension of the SymmetricPositiveSemidefiniteFixedRank matrix M over the number system 𝔽, i.e.

rand(::SymplecticMatrices; vector_at=nothing, σ::Real=1.0)
Generate a random point on or a random tangent vector if vector_at is set to a point .

manifold_dimension(::SymplecticMatrices)
Returns the dimension of the symplectic manifold embedded in , i.e.

manifold_dimension(::SymplecticGrassmann)
Return the dimension of the SymplecticGrassmann(2n,2k), which is

manifold_dimension(::SymplecticStiefel)
Returns the dimension of the symplectic Stiefel manifold embedded in , i.e. [BZ21]

manifold_dimension(M::EmbeddedTorus)
Return the dimension of the EmbeddedTorus M that is 2.

manifold_dimension(::Tucker)
The dimension of the manifold of tensors of multilinear rank , i.e.