Multinomial symmetric matrices

Manifolds.MultinomialSymmetric — Type

MultinomialSymmetric{T} <: AbstractMultinomialDoublyStochastic

The multinomial symmetric matrices manifold consists of all symmetric $n×n$ matrices with positive entries such that each column sums to one, i.e.

\[\begin{aligned} \mathcal{SP}(n) \coloneqq \bigl\{p ∈ ℝ^{n×n}\ \big|\ &p_{i,j} > 0 \text{ for all } i=1,…,n, j=1,…,m,\\ & p^\mathrm{T} = p,\\ & p\mathbf{1}_n = \mathbf{1}_n \bigr\}, \end{aligned}\]

where $\mathbf{1}_n$ is the vector of length $n$ containing ones.

It is modeled as IsometricallyEmbeddedManifoldType. via the AbstractMultinomialDoublyStochastic type, since it shares a few functions also with AbstractMultinomialDoublyStochastic, most and foremost projection of a point from the embedding onto the manifold.

The tangent space can be written as

\[T_p\mathcal{SP}(n) \coloneqq \bigl\{ X ∈ ℝ^{n×n}\ \big|\ X = X^{\mathrm{T}} \text{ and } X\mathbf{1}_n = \mathbf{0}_n \bigr\},\]

where $\mathbf{0}_n$ is the vector of length $n$ containing zeros.

More details can be found in Section IV [DH19].

Constructor

MultinomialSymmetric(n)

Generate the manifold of matrices $ℝ^{n×n}$ that are doubly stochastic and symmetric.

Base.rand — Method

rand(::MultinomialSymmetric; vector_at=nothing, σ::Real=1.0, kwargs...)

Generate random points on the MultinomialSymmetric manifold or tangent vectors at the point vector_at if that is not nothing.

Let $n×n$ denote the matrix dimension of the MultinomialSymmetric.

When vector_at is nothing, this is done by generating a random matrix rand(n, n) with positive entries and projecting it onto the manifold. The kwargs... are passed to this projection.

When vector_at is not nothing, a random matrix in the ambient space is generated and projected onto the tangent space

ManifoldDiff.riemannian_Hessian — Method

Y = riemannian_Hessian(M::MultinomialSymmetric, p, G, H, X)
riemannian_Hessian!(M::MultinomialSymmetric, Y, p, G, H, X)

Compute the Riemannian Hessian $\operatorname{Hess} f(p)[X]$ given the Euclidean gradient $∇ f(\tilde p)$ in G and the Euclidean Hessian $∇^2 f(\tilde p)[\tilde X]$ in H, where $\tilde p, \tilde X$ are the representations of $p,X$ in the embedding,.

The Riemannian Hessian can be computed as stated in Corollary 3 [DH19].

ManifoldsBase.check_point — Method

check_point(M::MultinomialSymmetric, p)

Checks whether p is a valid point on the MultinomialSymmetric(m,n) M, i.e. is a symmetric matrix with positive entries whose rows sum to one.

ManifoldsBase.check_vector — Method

check_vector(M::MultinomialSymmetric p, X; kwargs...)

Checks whether X is a valid tangent vector to p on the MultinomialSymmetric M. This means, that p is valid, that X is of correct dimension, symmetric, and sums to zero along any row.

ManifoldsBase.is_flat — Method

is_flat(::MultinomialSymmetric)

Return false. MultinomialSymmetric is not a flat manifold.

ManifoldsBase.manifold_dimension — Method

manifold_dimension(M::MultinomialSymmetric)

returns the dimension of the MultinomialSymmetric manifold namely

\[\operatorname{dim}_{\mathcal{SP}(n)} = \frac{n(n-1)}{2}.\]

ManifoldsBase.project — Method

project(M::MultinomialSymmetric, p, Y)

Project Y onto the tangent space at p on the MultinomialSymmetric M, return the result in X.

The formula from [DH19], Sec. VI reads

\[ \operatorname{proj}_p(Y) = Y - (α\mathbf{1}_n^{\mathrm{T}} + \mathbf{1}_n α^{\mathrm{T}}) ⊙ p,\]

where $⊙$ denotes the Hadamard or elementwise product and $\mathbb{1}_n$ is the vector of length $n$ containing ones. The two vector $α ∈ ℝ^{n×n}$ is given by solving

\[ (I_n+p)α = Y\mathbf{1},\]

where $I_n$ is the $n×n$ unit matrix and $\mathbf{1}_n$ is the vector of length $n$ containing ones.

ManifoldsBase.retract — Method

retract(M::MultinomialSymmetric, p, X, ::ProjectionRetraction)

compute a projection based retraction by projecting $p⊙\exp(X⨸p)$ back onto the manifold, where $⊙,⨸$ are elementwise multiplication and division, respectively. Similarly, $\exp$ refers to the elementwise exponentiation.

Literature

[DH19]

A. Douik and B. Hassibi. Manifold Optimization Over the Set of Doubly Stochastic Matrices: A Second-Order Geometry. IEEE Transactions on Signal Processing 67, 5761–5774 (2019), arXiv:1802.02628.

MultinomialSymmetric{T} <: AbstractMultinomialDoublyStochastic
The multinomial symmetric matrices manifold consists of all symmetric matrices with positive entries such that each column sums to one, i.e.

check_vector(M::CenteredMatrices, p, X; kwargs... )
Check whether X is a tangent vector to manifold point p on the CenteredMatrices M, i.e. that X is a matrix of size (m, n) whose columns sum to zero and its values are from the correct AbstractNumbers.

check_vector(M::Circle, p, X; kwargs...)
Check whether X is a tangent vector in the tangent space of p on the Circle M.

check_vector(M::Elliptope, q, Y; kwargs... )
Check whether is a tangent vector to on the Elliptope M, i.e. Y has to be of same dimension as q and a has to be a symmetric matrix with zero diagonal.

check_vector(M::Flag, p::AbstractMatrix, X::AbstractMatrix; kwargs... )
Check whether X is a tangent vector to point p on the Flag manifold M in the Stiefel representation, i.e. that X is a matrix of the form

check_vector(M::Flag, p::OrthogonalPoint, X::OrthogonalTangentVector; kwargs... )
Check whether X is a tangent vector to point p on the Flag manifold M in the orthogonal matrix representation, i.e. that X is block-skew-symmetric with zero diagonal:

check_vector(M::UnitaryMatrices, p, X; kwargs... ) check_vector(M::OrthogonalMatrices, p, X; kwargs... ) check_vector(M::Rotations, p, X; kwargs... ) check_vector(M::GeneralUnitaryMatrices, p, X; kwargs... )
Check whether X is a tangent vector to p on the UnitaryMatrices space M, i.e. after check_point(M,p), X has to be skew symmetric (Hermitian) and orthogonal to p.

check_vector(M::GraphManifold, p, X; kwargs...)
Check whether X is a valid tangent vector at p on the GraphManifold M.

check_vector(M::ProbabilitySimplex, p, X; kwargs... )
Check whether X is a tangent vector to p on the ProbabilitySimplex M, i.e. after check_point(M,p), X has to be of same dimension as p and its elements have to sum to one.

check_vector(M::AbstractProjectiveSpace, p, X; kwargs... )
Check whether X is a tangent vector in the tangent space of p on the AbstractProjectiveSpace M.

check_vector(M::SPDFixedDeterminant, p, X; kwargs... )
Check whether X is a tangent vector to manifold point p on the SPDFixedDeterminant M, i.e. X has to be a tangent vector on SymmetricPositiveDefinite, so a symmetric matrix, and additionally fulfill .

check_vector(M::Spectrahedron, q, Y; kwargs...)
Check whether is a tangent vector to on the Spectrahedron M, i.e. after check_point of q, Y has to be of same dimension as q and a has to be a symmetric matrix with trace.

check_vector(M::AbstractSphere, p, X; kwargs... )
Check whether X is a tangent vector to p on the AbstractSphere M, i.e. after check_point(M,p), X has to be of same dimension as p and orthogonal to p.

check_vector(M::SphereSymmetricMatrices, p, X; kwargs... )
Check whether X is a tangent vector to manifold point p on the SphereSymmetricMatrices M, i.e. X has to be a symmetric matrix of size (n,n) of unit Frobenius norm.

rand(::SymplecticMatrices; vector_at=nothing, σ::Real=1.0)
Generate a random point on or a random tangent vector if vector_at is set to a point .

check_vector(M::SymplecticMatrices, p, X; kwargs...)
Checks whether X is a valid tangent vector at p on the SymplecticMatrices M=, which requires that

check_vector(M::SymplecticMatrices, p, X; kwargs...)
Checks whether X is a valid tangent vector at p on the SymplecticStiefel, manifold.

check_vector(M::Torus{n}, p, X; kwargs...)
Checks whether X is a valid tangent vector to p on the Torus M.

check_vector(M::Tucker{<:Any,D}, p::TuckerPoint{T,D}, X::TuckerTangentVector) where {T,D}
Check whether a TuckerTangentVector X is is in the tangent space to the Dth order Tucker manifold M at the Dth order TuckerPoint p.