Multinomial symmetric positive definite matrices

Manifolds.MultinomialSymmetricPositiveDefinite — Type

MultinomialSymmetricPositiveDefinite <: AbstractMultinomialDoublyStochastic

The symmetric positive definite multinomial matrices manifold consists of all symmetric $n×n$ matrices with positive eigenvalues, and positive entries such that each column sums to one, i.e.

\[\begin{aligned} \mathcal{SP}^+(n) \coloneqq \bigl\{ p ∈ ℝ^{n×n}\ \big|\ &p_{i,j} > 0 \text{ for all } i=1,…,n, j=1,…,m,\\ & p^\mathrm{T} = p,\\ & p\mathbf{1}_n = \mathbf{1}_n\\ a^\mathrm{T}pa > 0 \text{ for all } a ∈ ℝ^{n}\backslash\{\mathbf{0}_n\} \bigr\}, \end{aligned}\]

where $\mathbf{1}_n$ and $\mathbr{0}_n$ are the vectors of length $n$ containing ones and zeros, respectively. More details about this manifold can be found in [DH19].

Constructor

MultinomialSymmetricPositiveDefinite(n)

Generate the manifold of matrices $\mathbb R^{n×n}$ that are symmetric, positive definite, and doubly stochastic.

Random.rand! — Method

Random.rand!(
    rng::AbstractRNG,
    M::MultinomialSymmetricPositiveDefinite,
    p::AbstractMatrix,
)

Generate a random point on MultinomialSymmetricPositiveDefinite manifold. The steps are as follows:

Generate a random totally positive matrix a. Construct a vector L of n random positive increasing real numbers. b. Construct the Vandermonde matrix V based on the sequence L. c. Perform LU factorization of V in such way that both L and U components have positive elements. d. Convert the LU factorization into LDU factorization by taking the diagonal of U and dividing U by it, V=LDU. e. Construct a new matrix R = UDL which is totally positive.

Project the totally positive matrix R onto the manifold of MultinomialDoubleStochastic matrices.

Symmetrize the projected matrix and return the result.

This method roughly follows the procedure described in https://math.stackexchange.com/questions/2773460/how-to-generate-a-totally-positive-matrix-randomly-using-software-like-maple

Literature

[DH19]

A. Douik and B. Hassibi. Manifold Optimization Over the Set of Doubly Stochastic Matrices: A Second-Order Geometry. IEEE Transactions on Signal Processing 67, 5761–5774 (2019), arXiv:1802.02628.