The Segre manifold

Manifolds.Segre — Type

Segre{𝔽,V} <: AbstractManifold{𝔽}

The Segre manifold

\[ \mathcal{S} = \operatorname{Seg}(𝔽^{n_1} \times \dots \times 𝔽^{n_d})\]

is the set of rank-one tensors in $𝔽^{n_1} \otimes \dots \otimes 𝔽^{n_d}$.

When $𝔽 = ℝ$, the Segre manifold is a normal Riemannian covering of

\[ \mathcal{P} = ℝ^{+} \times \mathbb{S}^{n_1 - 1} \times \dots \times \mathbb{S}^{n_d - 1}\]

equipped with a warped product metric. The tuple $(n_1, \dots, n_d)$ is called the valence of the manifold.

The geometry of the Segre manifold is summarized in [JSVV24]. It is named after Corrado Segre(1863–1924).

Constructor

Segre(n::Int...; field::AbstractNumbers=ℝ)

Generate a valence (n, ...) Segre manifold. Segre(n) is the same as $\mathbb{R}^{n} \setminus \{ 0 \}$.

Base.exp — Method

exp(M::Segre{ℝ, V}, p, X)

Exponential map on the Segre manifold.

Let $p ≐ (λ, x_1, …, x_d) ∈ \mathcal{S}$ and $X = (ν, u_1, …, u_d) ∈ T_p \mathcal{S}$. The exponential map is given by

\[ \operatorname{exp}_p(X) ≐ \left( \sqrt{(λ + ν)^2 + (λ m)^2},\\ x_1 \cos\mathopen{\Big(} \frac{f \lVert u_1 \rVert_{x_1}}{m} \mathclose{\Big)} + \frac{u_1}{\lVert u_1 \rVert_{x_1}} \sin\mathopen{\Big(} \frac{f \lVert u_1 \rVert_{x_1}}{m} \mathclose{\Big)},\\ …,\\ x_d \cos\mathopen{\Big(} \frac{f \lVert u_d \rVert_{x_d}}{m} \mathclose{\Big)} + \frac{u_d}{\lVert u_d \rVert_{x_d}} \sin\mathopen{\Big(} \frac{f \lVert u_d \rVert_{x_d}}{m} \mathclose{\Big)} \right),\]

where

\[ \begin{aligned} f &= \frac{\pi}{2} - \tan^{-1}\mathopen{\Big(} \frac{λ + ν}{λ m} \mathclose{\Big)},\\ m &= \sqrt{\lVert u_1 \rVert_{x_1}^2 + … + \lVert u_d \rVert_{x_d}^2}. \end{aligned}\]

If $m = 0$ and $-λ < ν$, then $\operatorname{exp}_p(v) = p + X$.

The formula is derived in proposition 3.1 in [JSVV24].

Base.log — Method

log(M::Segre{ℝ, V}, p, q)

Logarithmic map on the Segre manifold.

Let $p ≐ (λ, x_1, …, x_d)$, $q ≐ (μ, y_1, …, y_d) ∈ \mathcal{S}$. Assume $p$ and $q$ are connected by a geodesic.

Let

\[ m = \sqrt{\sphericalangle(x_1, y_1)^2 + … + \sphericalangle(x_d, y_d)^2}\]

and assume $(μ, y_1, …, y_d)$ is the representative of $q$ that minimizes $m$. Then

\[ \operatorname{log}_p(q) = \left( \mu \cos{m} - \lambda, (y_1 - ⟨x_1, y_1⟩ x_1) \frac{\mu \sphericalangle(x_1, y_1) \sin{m}}{\lambda m \sin{\sphericalangle(x_1, y_1)}}, \dots, (y_d - ⟨x_d, y_d⟩ x_d) \frac{\mu \sphericalangle(x_d, y_d) \sin{m}}{\lambda m \sin{\sphericalangle(x_d, y_d)}} \right).\]

The formula is derived in theorem 4.4 in [JSVV24].

Base.rand — Method

rand(M::Segre{ℝ, V}; vector_at=nothing)

If vector_at is nothing, return a random point on

\[ ℝ^{+} × \mathbb{S}^{n_1 - 1} ×⋯× \mathbb{S}^{n_d - 1}\]

from a log-normal distribution on $ℝ^{+}$ and a uniform distribution on $\mathbb{S}^{n_1 - 1} ×⋯× \mathbb{S}^{n_d - 1}$.

If vector_at is not nothing, return a random tangent vector from a normal distribution on the tangent space.

Manifolds.closest_representative! — Method

closest_representative!(M::Segre{ℝ, V}, p, q)

$\mathcal{S}$ is a $2^d$-sheeted Riemannian covering of

\[ \mathcal{P} = ℝ^{+} \times \mathbb{S}^{n_1 - 1} \times \dots \times \mathbb{S}^{n_d - 1}\]

with a warped product metric. Every equivalence class $q \in \mathcal{S}$ has $2^d$ representatives in $\mathcal{P}$. closest_representative!(M, q, p) changes representative of q to the one that is closest to p in $\mathcal{P}$.

Manifolds.connected_by_geodesic — Method

connected_by_geodesic(M::Segre{ℝ, V}, p, q)

$\mathcal{S}$ is not a complete manifold, i.e. not every pair p and q of points are connected by a geodesic in $\mathcal{S}$. connected_by_geodesic(M, p, q) returns true if two points, p and q, are connected by a geodesic, and otherwise returns false.

Manifolds.spherical_angle_sum — Method

spherical_angle_sum(M::Segre{ℝ, V}, p, q)

Let $p ≐ (λ, x_1, …, x_d)$, $q ≐ (μ, y_1, …, y_d) ∈ \mathcal{S}$. Then this is

\[ \sqrt{\sphericalangle(x_1, y_1)^2 + … + \sphericalangle(x_d, y_d)^2},\]

where $\sphericalangle(x_i, y_i)$ is the distance between $x_i$ and $y_i$ on the sphere $\mathbb{S}^{n_i - 1}$.

ManifoldsBase.distance — Method

distance(M::Segre{ℝ, V}, p, q)

Riemannian distance between two points p and q on the Segre manifold.

Assume $p ≐ (λ, x_1, …, x_d)$, $q ≐ (μ, y_1, …, y_d) ∈ \mathcal{S}$ are connected by a geodesic. Let

\[ m = \sqrt{\sphericalangle(x_1, y_1)^2 + … + \sphericalangle(x_d, y_d)^2}\]

and assume $(μ, y_1, …, y_d)$ is the representation of $q$ that minimizes $m$. Then

\[ \operatorname{dist}_{\mathcal{S}}(p, q) = \sqrt{λ^2 - 2 λμ\cos(m) + μ^2}.\]

ManifoldsBase.embed! — Method

embed!(M::Segre{𝔽, V}, p, X)

Embed tangent vector $X = (ν, u_1, …, u_d)$ at $p ≐ (λ, x_1, …, x_d)$ in $𝔽^{n_1 ×⋯× n_d}$ using the Kronecker product

\[ (ν, u_1, …, u_d) ↦ ν x_1 ⊗⋯⊗ x_d + λ u_1 ⊗ x_2 ⊗⋯⊗ x_d + … + λ x_1 ⊗⋯⊗ x_{d - 1} ⊗ u_d.\]

ManifoldsBase.embed — Method

embed(M::Segre{𝔽, V}, p)
embed!(M::Segre{𝔽, V}, q, p)

Embed $p ≐ (λ, x_1, …, x_d)$ in $𝔽^{n_1 ×⋯× n_d}$ using the Kronecker product

\[ (λ, x_1, …, x_d) ↦ λ x_1 ⊗⋯⊗ x_d.\]

ManifoldsBase.get_coordinates — Method

get_coordinates(M::Segre{𝔽, V}, p, X, ::DefaultOrthonormalBasis; kwargs...)

Get coordinates of X in the tangent space $T_{(λ, x_1, …, x_d)} \mathcal{S} = \mathbb{R} × T_{x_1} \mathbb{S}^{n_1 - 1} ×…× T_{x_d} \mathbb{S}^{n_d - 1}$ using a DefaultOrthonormalBasis on each factor.

ManifoldsBase.get_embedding — Method

get_embedding(M::Segre{𝔽,V})

Return the embedding of the Segre manifold $\mathcal{S}$, which is $𝔽^{n_1 ×⋯× n_d}$.

ManifoldsBase.get_vector — Method

get_vector( M::Segre{𝔽, V}, p, c, DefaultOrthonormalBasis; kwargs...)

Get tangent vector X from coordinates in the tangent space $T_{(λ, x_1, …, x_d)} \mathcal{S} = \mathbb{R} × T_{x_1} \mathbb{S}^{n_1 - 1} ×⋯× T_{x_d} \mathbb{S}^{n_d - 1}$ using a DefaultOrthonormalBasis on each factor.

ManifoldsBase.inner — Method

inner(M::Segre{ℝ, V}, p, X, Y,)

Inner product between two tangent vectors $X = (ν, u_1, …, u_d)$ and $Y = (ξ, v_1, …, v_d)$ at $p ≐ (λ, x_1, \dots, x_d)$. This inner product is obtained by embedding the Segre manifold in the space of tensors equipped with the Euclidean metric:

\[ \langle X, Y \rangle_{p} = \nu \xi + \lambda^2 (\langle u_1, v_1 \rangle_{x_1} + \dots + \langle u_d, v_d \rangle_{x_d}),\]

where $ν, ξ ∈ T_{λ} ℝ^{+} = ℝ$ and $u_i$, $v_i ∈ T_{x_i} \mathbb{S}^{n_i - 1} ⊂ ℝ^{n_i}$.

ManifoldsBase.is_point — Method

is_point(M::Segre{ℝ, V}, p; kwargs...)

Check whether p is a valid point on M, i.e. p[1] is a singleton containing a positive number and p[i + 1] is a point on Sphere(V[i]). The tolerance can be set using the kwargs....

ManifoldsBase.is_vector — Method

is_vector(M::Segre{ℝ, V}, p, X, kwargs...)

Check whether X is a tangent vector to p on M, i.e. X has to be of same dimension as p and orthogonal to p. The tolerance can be set using the kwargs....

ManifoldsBase.riemann_tensor — Method

riemann_tensor(M::Segre{ℝ, V}, p, X, Y, Z)

Riemann tensor of the Segre manifold at $p$.

$\mathcal{S}$ is locally a warped product of $ℝ^{+}$ and $\mathbb{S}^{n_1 - 1} ×⋯× \mathbb{S}^{n_d - 1}$. If $p ≐ (λ, x_1, …, x_d) ∈ \mathcal{S}$ and $X$, $Y$, $Z ∈ T_p (\mathbb{S}^{n_1 - 1} ×⋯× \mathbb{S}^{n_d - 1}) ⊂ T_p \mathcal{S}$, then

\[ R_{\mathcal{S}}(X, Y) Z = R_{\mathbb{S}^{n_1 - 1} ×⋯× \mathbb{S}^{n_d - 1}}(X, Y) Z + λ^{-2}(⟨X, Z⟩_p Y - ⟨Y, Z⟩_p X).\]

$R_{\mathcal{S}}$ is zero in the remaining (orthogonal) directions.

ManifoldsBase.sectional_curvature — Method

sectional_curvature(M::Segre{ℝ, V}, p, u, v)

Sectional curvature of the Segre manifold at $p$.

$\mathcal{S}$ is locally a warped product of $ℝ^{+}$ and $\mathbb{S}^{n_1 - 1} ×⋯× \mathbb{S}^{n_d - 1}$ If $p ≐ (λ, x_1, …, x_d) ∈ \mathcal{S}$, $u_i ∈ T_{x_i} \mathbb{S}^{n_i - 1}$, and $v_j ∈ T_{x_j} \mathbb{S}^{n_j - 1}$, then

\[ K_{\mathcal{S}}(u_i, v_j) = \frac{\delta_{i j} - 1}{\lambda^2}.\]

$K_{\mathcal{S}}$ is zero in the remaining (orthogonal) directions.

A warped metric

Manifolds.WarpedMetric — Type

WarpedMetric{A} <: AbstractMetric

The $A$-warped metric on the Segre manifold $\mathcal{S}$ is a generalization of the Euclidean metric on $\mathcal{S}$. We denote this manifold by $\mathcal{S}_A$.

Similarly to $\mathcal{S}$, when $𝔽 = ℝ$, $\mathcal{S}_A$ is a normal Riemannian covering of the product manifold

\[ ℝ^{+} × \mathbb{S}^{n_1 - 1} ×⋯× \mathbb{S}^{n_d - 1}\]

with a warped product metric, but the warping function now depends on the warping factor $A$. $A = 1$ corresponds to the usual Segre manifold. The Segre manifold is a cone in the sense that if $p \in \mathcal{S}$, then $r p \in \mathcal{S}$ for all $r \neq 0$. The tangent subspace at $p$ defined $\mathrm{d} (r p) / \mathrm{d} r$ is called the radial direction. $A < 1$ puts less weight on the directions orthogonal to the radial direction compared to $\mathcal{S}$, while $A > 1$ puts more weight on those directions.

The geometry is summarized in [JSVV24].

Constructor

WarpedMetric(A::Real)

Generate a warped product metric with warping factor A.

Base.exp — Method

exp(M::MetricManifold{ℝ, Segre{ℝ, V}, WarpedMetric{A}}, p, X)

Exponential map on the warped Segre manifold.

Let $p ≐ (λ, x_1,…, x_d) ∈ \mathcal{S}_A$ and $X = (ν, u_1,…, u_d) ∈ T_p \mathcal{S}_A$. Then the exponential map is given by

\[ \operatorname{exp}_p(X) ≐ \left( \sqrt{(λ + ν)^2 + (λ A m)^2},\\ x_1 \cos\mathopen{\Big(} \frac{f \lVert u_1 \rVert_{x_1}}{A m} \mathclose{\Big)} + \frac{u_1}{\lVert u_1 \rVert_{x_1}} \sin\mathopen{\Big(} \frac{f \lVert u_1 \rVert_{x_1}}{A m} \mathclose{\Big)},\\ …,\\ x_d \cos\mathopen{\Big(} \frac{f \lVert u_d \rVert_{x_d}}{A m} \mathclose{\Big)} + \frac{u_d}{\lVert u_d \rVert_{x_d}} \sin\mathopen{\Big(} \frac{f \lVert u_d \rVert_{x_d}}{A m} \mathclose{\Big)} \right),\]

where

\[ \begin{aligned} f &= \frac{\pi}{2} - \tan^{-1}\mathopen{\Big(} \frac{λ + ν}{λ A m} \mathclose{\Big)},\\ m &= \sqrt{\lVert u_1 \rVert_{x_1}^2 + … + \lVert u_d \rVert_{x_d}^2}. \end{aligned}\]

If $m = 0$ and $-λ < ν$, then $\operatorname{exp}_p(v) = p + X$.

The formula is derived in proposition 3.1 in [JSVV24].

Base.log — Method

log(M::MetricManifold{ℝ, Segre{ℝ,V}, WarpedMetric{A}}, p, q)

Logarithmic map on the warped Segre manifold.

Let $p ≐ (λ, x_1,…, x_d)$, $q ≐ (μ, y_1,…, y_d) ∈ \mathcal{S}_A$. Assume $p$ and $q$ are connected by a geodesic. Let

\[ m = \sqrt{\sphericalangle(x_1, y_1)^2 +… + \sphericalangle(x_d, y_d)^2}\]

and assume $(μ, y_1,…, y_d)$ is the representative of $q$ that minimizes $m$. Then

\[ \operatorname{log}_p(q) = \left( \mu \cos{m} - \lambda, (y_1 - ⟨x_1, y_1⟩ x_1) \frac{\mu \sphericalangle(x_1, y_1) \sin(A m)}{\lambda A m \sin{\sphericalangle(x_1, y_1)}}, \dots, (y_d - ⟨x_d, y_d⟩ x_d) \frac{\mu \sphericalangle(x_d, y_d) \sin(A m)}{\lambda A m \sin{\sphericalangle(x_d, y_d)}} \right).\]

The formula is derived in theorem 4.4 in [JSVV24].

ManifoldsBase.distance — Method

distance(M::MetricManifold{ℝ, Segre{ℝ,V}, WarpedMetric{A}}, p, q)

Riemannian distance between two points p and q on the warped Segre manifold.

Assume $p ≐ (λ, x_1,…, x_d)$, $q ≐ (μ, y_1,…, y_d) ∈ \mathcal{S}_A$ are connected by a geodesic. Let

\[ m = \sqrt{\sphericalangle(x_1, y_1)^2 +… + \sphericalangle(x_d, y_d)^2}\]

and assume $(μ, y_1,…, y_d)$ is the representation of $q$ that minimizes $m$. Then

\[ \operatorname{dist}_{\mathcal{S}_A}(p, q) = \sqrt{λ^2 - 2 λ μ \cos(A m) + μ^2}.\]

ManifoldsBase.get_coordinates — Method

get_coordinates(M::Segre{𝔽, V}, p, v, ::DefaultOrthonormalBasis; kwargs...)

Get coordinates of X in the tangent space $T_{(λ, x_1,…, x_d)} \mathcal{S}_A = \mathbb{R} × T_{x_1} \mathbb{S}^{n_1 - 1} ×⋯× T_{x_d} \mathbb{S}^{n_d - 1}$ using a DefaultOrthonormalBasis on each factor.

ManifoldsBase.get_vector — Method

get_vector(M::Segre{𝔽, V}, p, c, ::DefaultOrthonormalBasis; kwargs...)

Get tangent vector X from coordinates in the tangent space $T_{(λ, x_1,…, x_d)} \mathcal{S}_A = \mathbb{R} × T_{x_1} \mathbb{S}^{n_1 - 1} ×⋯× T_{x_d} \mathbb{S}^{n_d - 1}$ using a DefaultOrthonormalBasis on each factor.

ManifoldsBase.inner — Method

inner(M::MetricManifold{ℝ, Segre{ℝ, V}, WarpedMetric{A}}, p, X, Y)

Inner product between two tangent vectors $X = (ν, u_1,…, u_d)$ and $Y = (ξ, v_1,…, v_d)$ at $p \doteq (λ, x_1,…, x_d)$:

\[ ⟨X, Y⟩_{p} = ν ξ + (A λ)^2 (⟨ u_1, v_1 ⟩_{x_1} +… + ⟨u_d, v_d⟩_{x_d}),\]

where $ν$, $ξ ∈ T_{λ} ℝ^{+} = ℝ$ and $u_i$, $v_i ∈ T_{x_i} \mathbb{S}^{n_i - 1} \subset ℝ^{n_i}$.

ManifoldsBase.riemann_tensor — Method

riemann_tensor(M::MetricManifold{ℝ, Segre{ℝ,V}, WarpedMetric{A}}, p, X, Y)

Riemann tensor of the warped Segre manifold at $p$.

$\mathcal{S}_A$ is locally a warped product of $ℝ^{+}$ and $\mathbb{S}^{n_1 - 1} ×⋯× \mathbb{S}^{n_d - 1}$. If $p ≐ (λ, x_1,…, x_d) ∈ \mathcal{S}_A$ and $X$, $Y$, $Z ∈ T_{(x_1,…, x_d)} (\mathbb{S}^{n_1 - 1} ×⋯× \mathbb{S}^{n_d - 1}) \subset T_p \mathcal{S}_A$ then

\[ R_{\mathcal{S}_A}(X, Y) Z = R_{\mathbb{S}^{n_1 - 1} ×⋯× \mathbb{S}^{n_d - 1}}(X, Y) Z + λ^{-2}(⟨ X, Z ⟩_{p} Y - ⟨ Y, Z ⟩_{p} X).\]

$R_{\mathcal{S}_A}$ is zero in the remaining (orthogonal) directions.

ManifoldsBase.sectional_curvature — Method

sectional_curvature(M::MetricManifold{ℝ, Segre{ℝ,V}, WarpedMetric{A}}, p, X, Y)

Sectional curvature of the warped Segre manifold at $p$.

$\mathcal{S}_A$ is locally a warped product of $ℝ^{+}$ and $\mathbb{S}^{n_1 - 1} ×⋯× \mathbb{S}^{n_d - 1}$ If $p = (λ, x_1,…, x_d) ∈ \mathcal{S}$, $u_i ∈ T_{x_i} \mathbb{S}^{n_i - 1}$, and $v_j ∈ T_{x_j} \mathbb{S}^{n_j - 1}$, then

\[ K_{\mathcal{S}_A}(u_i, v_j) = \frac{A^{-2} \delta_{i j} - 1}{λ^{2}}.\]

$K_{\mathcal{S}_A}$ is zero in the remaining (orthogonal) directions.

Literature

[JSVV24]

S. Jacobsson, L. Swijsen, J. V. Veken and N. Vannieuwenhoven. Warped geometries of Segre-Veronese manifolds (2024), arXiv:2410.00664 [math.NA].