Maps

The following functions and types provide support for maps with manifold-valued domains and codomains. The base type is AbstractMap, which has a domain and a codomain.

Maps in Manifolds are not required to be total. That is, given a domain $M$ and codomain $N$, a function $f \colon M \to N$ is not required to be defined on all of $M$ or to cover all of $N$. Consequently, it is left to the user to ensure that sensible inputs are provided to the maps and that true, left-, and right-inverses are defined and used appropriately.

Manifolds.AbstractMap — Type

AbstractMap{D,CoD}

An abstract type that represents maps between elements of sets. Every map has a domain of type D, which is the set to which inputs belong and a codomain of type CoD, which is the set to which outputs belong. Note that maps are not required to be total. That is, the true (co)domain may be a subset of the provided (co)domain.

Every new map type must implement domain and codomain and be callable.

Manifolds.codomain — Method

codomain(m::AbstractMap)

Get the manifold to which outputs to the map m belong. By default, this is assumed to be stored in the field m.codomain.

Manifolds.domain — Method

domain(m::AbstractMap)

Get the manifold to which inputs to the map m belong. By default, this is assumed to be stored in the field m.domain.

Manifolds.Curve — Type

Curve(M::Manifold)

A type for a curve, a map from 1-D real space to a manifold, i.e. $\phi: ℝ \to M$

Manifolds.Map — Type

Map(domain, codomain)

A generic AbstractMap between domain and codomain.

Manifolds.RealField — Type

RealField(M::Manifold)

A type for a generic field, a map from a point on a manifold M to a real number.