Orthonormal bases

The following functions and types provide support for orthonormal bases of the tangent space of different manifolds. An orthonormal basis of the tangent space $T_x M$ of (real) dimension $N$ has a real-coefficient basis $e_1, e_2, \dots, e_N$ if $\mathrm{Re}(g_x(e_i, e_j)) = \delta_{ij}$ for each $i,j \in \{1, 2, \dots, N\}$ where $g_x$ is the Riemannian metric at point $x$. A vector $v$ from the tangent space $T_x M$ can be expressed as a sum $v = v^i e_i$ where coefficients $v^i$ are calculated as $v^i = \mathrm{Re}(g_x(v, e_i))$.

The main types are:

ArbitraryOrthonormalBasis, which is designed to work when no special properties of the tangent space basis are required. It is designed to make get_coordinates and get_vector fast.

DiagonalizingOrthonormalBasis, which diagonalizes the curvature tensor and makes the curvature in the selected direction equal to 0.

ProjectedOrthonormalBasis, which projects a basis of the ambient space and orthonormalizes projections to obtain a basis in a generic way.

Subtypes of AbstractPrecomputedOrthonormalBasis, which work by storing (either explicitly or implicitly) a precomputed basis at a certain point. ArbitraryOrthonormalBasis doesn't require precomputing but DiagonalizingOrthonormalBasis and DiagonalizingOrthonormalBasis usually do.

The main functions are:

get_basis precomputes a basis at a certain point.

get_coordinates returns coordinates of a tangent vector.

get_vector returns a vector for the specified coordinates.

get_vectors returns a vector of basis vectors (calling it should be avoided for high-dimensional manifolds).

Manifolds.AbstractBasis — Type

AbstractBasis{F}

Abstract type that represents a basis on a manifold or a subset of it.

The type parameter F denotes the AbstractNumbers that will be used as scalars.

Manifolds.AbstractOrthonormalBasis — Type

AbstractOrthonormalBasis{F}

Abstract type that represents an orthonormal basis on a manifold or a subset of it.

The type parameter F denotes the AbstractNumbers that will be used as scalars.

Manifolds.AbstractPrecomputedOrthonormalBasis — Type

AbstractPrecomputedOrthonormalBasis{F}

Abstract type that represents an orthonormal basis of the tangent space at a point on a manifold. Tangent vectors can be obtained using function get_vectors.

The vectors are not always fully precomputed because a partially precomputed basis may be enough for implementing get_vector and get_coordinates.

The type parameter F denotes the AbstractNumbers that will be used as scalars.

Manifolds.ArbitraryOrthonormalBasis — Type

ArbitraryOrthonormalBasis(F::AbstractNumbers = ℝ)

An arbitrary orthonormal basis on a manifold. This will usually be the fastest orthonormal basis available for a manifold.

The type parameter F denotes the AbstractNumbers that will be used as scalars.

Manifolds.DiagonalizingOrthonormalBasis — Type

DiagonalizingOrthonormalBasis(v, F::AbstractNumbers = ℝ)

An orthonormal basis Ξ as a vector of tangent vectors (of length determined by manifold_dimension) in the tangent space that diagonalizes the curvature tensor $R(u,v)w$ and where the direction v has curvature 0.

The type parameter F denotes the AbstractNumbers that will be used as scalars.

Manifolds.PrecomputedDiagonalizingOrthonormalBasis — Type

DiagonalizingOrthonormalBasis(vectors, kappas, F::AbstractNumbers = ℝ)

A precomputed orthonormal basis Ξ as a vector of tangent vectors (of length determined by manifold_dimension) in the tangent space that diagonalizes the curvature tensor $R(u,v)w$ with eigenvalues kappas and where the direction v has curvature 0.

The type parameter F denotes the AbstractNumbers that will be used as scalars.

Manifolds.PrecomputedOrthonormalBasis — Type

PrecomputedOrthonormalBasis(vectors::AbstractVector, F::AbstractNumbers = ℝ)

A precomputed orthonormal basis at a point on a manifold.

The type parameter F denotes the AbstractNumbers that will be used as scalars.

Manifolds.ProjectedOrthonormalBasis — Type

ProjectedOrthonormalBasis(method::Symbol, F::AbstractNumbers = ℝ)

An orthonormal basis that comes from orthonormalization of basis vectors of the ambient space projected onto the subspace representing the tangent space at a given point.

The type parameter F denotes the AbstractNumbers that will be used as scalars.

Available methods:

:gram_schmidt uses a modified Gram-Schmidt orthonormalization.

:svd uses SVD decomposition to orthogonalize projected vectors. The SVD-based method should be more numerically stable at the cost of an additional assumption (local metric tensor at a point where the basis is calculated has to be diagonal).

Manifolds.get_basis — Method

get_basis(M::Manifold, x, B::AbstractBasis) -> AbstractBasis

Compute the basis vectors of the tangent space at a point on manifold M represented by x.

Returned object derives from AbstractBasis and may have a field .vectors that stores tangent vectors or it may store them implicitly, in which case the function get_vectors needs to be used to retrieve the basis vectors.

See also: get_coordinates, get_vector

Manifolds.get_basis — Method

get_basis(M::Manifold, x, B::ArbitraryOrthonormalBasis)

Compute the basis vectors of an ArbitraryOrthonormalBasis.

Manifolds.get_coordinates — Method

get_coordinates(M::Manifold, x, v, B::AbstractBasis)

Compute a one-dimensional vector of coefficients of the tangent vector v at point denoted by x on manifold M in basis B.

Depending on the basis, x may not directly represent a point on the manifold. For example if a basis transported along a curve is used, x may be the coordinate along the curve.

See also: get_vector, get_basis

Manifolds.get_vector — Method

get_vector(M::Manifold, x, v, B::AbstractBasis)

Convert a one-dimensional vector of coefficients in a basis B of the tangent space at x on manifold M to a tangent vector v at x.

Depending on the basis, x may not directly represent a point on the manifold. For example if a basis transported along a curve is used, x may be the coordinate along the curve.

See also: get_coordinates, get_basis

Manifolds.get_vectors — Method

get_vectors(M::Manifold, x, B::AbstractBasis)

Get the basis vectors of basis B of the tangent space at point x.

Manifolds.number_system — Method

number_system(::AbstractBasis)

The number system used as scalars in the given basis.