Statistics

Statistics.mean! — Method

mean!(
    M::Manifold,
    y,
    x::AbstractVector,
    w::AbstractWeights,
    method::GeodesicInterpolationMethod;
    shuffle_rng=nothing,
    kwargs...,
)

Estimate the Riemannian center of mass of x in an online fashion using repeated weighted geodesic interpolation. See GeodesicInterpolationMethod for details.

If shuffle_rng is provided, it is used to shuffle the order in which the points are considered for computing the mean.

Statistics.mean! — Method

mean!(M::Manifold, y, x::AbstractVector, [w::AbstractWeights]; kwargs...)
mean!(M::Manifold, y, x::AbstractVector, [w::AbstractWeights], method::AbstractMethod; kwargs...)

Compute the mean in-place in y.

Statistics.mean — Method

mean(M::Manifold, x::AbstractVector, [w::AbstractWeights]; kwargs...)

Compute the (optionally weighted) Riemannian center of mass also known as Karcher mean of the vector x of points on the Manifold M, defined as the point that satisfies the minimizer

\[\argmin_{y\in\mathcal M} \frac{1}{2 \sum_{i=1}^n w_i} \sum_{i=1}^n w_i\mathrm{d}_{\mathcal M}^2(y,x_i),\]

where $\mathrm{d}_{\mathcal M}$ denotes the Riemannian distance.

In the general case, the GradientMethod is used to compute the mean. However, this default may be overloaded for specific manifolds.

mean(M::Manifold, x::AbstractVector, [w::AbstractWeights], method::AbstractMethod; kwargs...)

Compute the mean using the specified method.

mean(
    M::Manifold,
    x::AbstractVector,
    w::AbstractWeights,
    method::GradientMethod;
    x0=x[1],
    stop_iter=100,
    kwargs...
)

Compute the mean using the gradient descent scheme GradientMethod.

Optionally, provide x0, the starting point (by default set to the first data point). Set x0 to nothing to use the intitial value of y as the starting point. stop_iter denotes the maximal number of iterations to perform and the kwargs... are passed to isapprox to stop, when the minimal change between two iterates is small. For more stopping criteria check the Manopt.jl package and use a solver therefrom.

The algorithm is further described in [Afsari et al, 2013].

Statistics.median! — Method

median!(M::Manifold, y, x::AbstractVector, [w::AbstractWeights]; kwargs...)
median!(M::Manifold, y, x::AbstractVector, [w::AbstractWeights], method::AbstractMethod; kwargs...)

computes the median in-place in y.

Statistics.median — Method

median(M::Manifold, x::AbstractVector, [w::AbstractWeights]; kwargs...)

Compute the (optionally weighted) Riemannian median of the vector x of points on the Manifold M, defined as the point that satisfies the minimizer

\[\argmin_{y\in\mathcal M} \frac{1}{\sum_{i=1}^n w_i} \sum_{i=1}^n w_i\mathrm{d}_{\mathcal M}(y,x_i),\]

where $\mathrm{d}_{\mathcal M}$ denotes the Riemannian distance. This function is nonsmooth (i.e nondifferentiable).

In the general case, the CyclicProximalPointMethod is used to compute the median. However, this default may be overloaded for specific manifolds.

median(M::Manifold, x::AbstractVector, [w::AbstractWeights], method::AbstractMethod; kwargs...)

Compute the median using the specified method.

median(
    M::Manifold,
    x::AbstractVector,
    w::AbstractWeights,
    method::CyclicProximalPointMethod;
    x0=x[1],
    stop_iter=1000000,
    kwargs...
)

Compute the median using CyclicProximalPointMethod.

Optionally, provide x0, the starting point (by default set to the first data point). Set x0 to nothing to use the intitial value of y as the starting point. stop_iter denotes the maximal number of iterations to perform and the kwargs... are passed to isapprox to stop, when the minimal change between two iterates is small. For more stopping criteria check the Manopt.jl package and use a solver therefrom.

The algorithm is further described in Algorithm 4.3 and 4.4 in [Bačák, 2014]

Statistics.std — Function

std(M, x, m=nothing; corrected=true, kwargs...)

compute the variance of a Vector x of n data points on the Manifold M, i.e.

\[\sqrt{\frac{1}{c} \sum_{i=1}^n d_{\mathcal M}^2 (x_i,m)},\]

where c is a correction term, see Statistics.var. and m is the provideed mean of x.

Statistics.std — Function

std(M, x, w::AbstractWeights, m=nothing; corrected=false, kwargs...)

compute the (weighted) standard deviation of a Vector x of n data points on the Manifold M, i.e.

\[\sqrt{\frac{1}{c} \sum_{i=1}^n w_i d_{\mathcal M}^2 (x_i,m)},\]

where c is a correction term, see Statistics.var. The (weighted) mean of x can be specified as m, and the corrected variance can be activated by setting corrected=true. All further kwargs... are passed to the computation of the mean (if that is not provided).

Statistics.var — Function

var(M, x, m=nothing; corrected=true, kwargs...)

compute the variance of a Vector x of n data points on the Manifold M, i.e.

\[\frac{1}{c} \sum_{i=1}^n d_{\mathcal M}^2 (x_i,m),\]

where c is a correction term, see Statistics.var. and m is the provideed mean of x. The uncorrected variance can be activated by setting corrected=false.

Statistics.var — Function

var(M, x, w::AbstractWeights, m=nothing; corrected=false, kwargs...)

compute the (weighted) variance of a Vector x of n data points on the Manifold M, i.e.

\[\frac{1}{c} \sum_{i=1}^n w_i d_{\mathcal M}^2 (x_i,m),\]

where c is a correction term, see Statistics.var. The (weighted) mean of x can be specified as m, and the corrected variance can be activated by setting corrected=true. All further kwargs... are passed to the computation of the mean (if that is not provided).

StatsBase.mean_and_std — Method

mean_and_std(M, x, w; corrected = false, kwargs...) -> (mean, std)
mean_and_std(M, x; corrected = true, kwargs...) -> (mean, std)

compute the mean m and the standard deviation std s simultaneously.

StatsBase.mean_and_var — Method

mean_and_var(M, x, w; corrected = false, kwargs...) -> (mean, var)
mean_and_var(M, x; corrected = true, kwargs...) -> (mean, var)

compute the mean m and the variance v simultaneously.

Manifolds.GeodesicInterpolationMethod — Type

GeodesicInterpolationMethod <: AbstractMethod

Repeated weighted geodesic interpolation method for estimating the Riemannian center of mass.

The algorithm proceeds with the following simple online update:

\[\begin{aligned} \mu_1 &= x_1\\ t_k &= \frac{w_k}{\sum_{i=1}^k w_i}\\ \mu_{k} &= \gamma_{\mu_{k-1}}(x_k; t_k), \end{aligned}\]

where $x_k$ are points, $w_k$ are weights, $\mu_k$ is the $k$th estimate of the mean, and $\gamma_x(y; t)$ is the point at time $t$ along the shortest_geodesic between points $x,y \in \mathcal M$. The algorithm terminates when all $x_k$ have been considered. In the Euclidean case, this exactly computes the weighted mean.

The algorithm has been shown to converge asymptotically with the sample size for the following manifolds equipped with their default metrics when all sampled points are in an open geodesic ball about the mean with corresponding radius:

Euclidean: $\infty$

SymmetricPositiveDefinite [Ho et al, 2013]: $\infty$

Sphere [Salehian et al, 2015]: $\frac{\pi}{2}$

Grassmannian [Chakraborty and Vemuri, 2015]: $\frac{\pi}{4}$

Stiefel/Rotations [Chakraborty and Vemuri, 2019]: $\frac{\pi}{2 \sqrt 2}$

Literature

[Afsari et al, 2013] Afsari, B; Tron, R; Vidal, R: "On the Convergence of Gradient Descent for Finding the Riemannian Center of Mass," SIAM Journal on Control and Optimization (2013), 51(3), pp. 2230–2260. doi: 10.1137/12086282X, arxiv: 1201.0925.

[Bačák, 2014] Bačák, M: "Computing Medians and Means in Hadamard Spaces," SIAM Journal on Optimization (2014), 24(3), pp. 1542–1566, doi: 10.1137/140953393, arxiv: 1210.2145.

[Ho et al, 2013] Ho J; Cheng G; Salehian H; Vemuri B C: "Recursive Karcher expectation estimators and geometric law of large numbers," Proceedings of the 16th International Conference on Artificial Intelligence and Statistics (2013), pp. 325–332, pdf.

[Salehian et al, 2015] Salehian H; Chakraborty R; Ofori E; Vaillancourt D: "An efficient recursive estimator of the Fréchet mean on a hypersphere with applications to Medical Image Analysis," Mathematical Foundations of Computational Anatomy (2015), pdf.

[Chakraborty and Vemuri, 2015] Chakraborty R; Vemuri B C: "Recursive Fréchet Mean Computation on the Grassmannian and Its Applications to Computer Vision," Proceedings of the IEEE International Conference on Computer Vision (ICCV) (2015), pp. 4229-4237. doi: 10.1109/ICCV.2015.481, link.

[Chakraborty and Vemuri, 2019] Chakraborty R; Vemuri B C: "Statistics on the Stiefel manifold: Theory and Applications," The Annals of Statistics (2019), 47(1), pp. 415-438. doi: 10.1214/18-AOS1692, arxiv: 1708.00045.