Home

Manifolds

Package Manifolds aims to provide both a unified interface to define and use manifolds as well as a library of manifolds to use for your projects.

Interface

Manifolds.IsDecoratorManifold — Type.

IsDecoratorManifold

A Trait to mark a manifold as a decorator type. For any function that is only implemented for a decorator (i.e. a Manifold with @traitimpl IsDecoratorManifold{M}), a specific function should be implemented as a @traitfn, that transparently passes down through decorators, i.e.

@traitfn my_feature(M::MT, k...) where {MT; IsDecoratorManifold{MT}} = my_feature(M.manifold, k...)

or the shorter version

@traitfn my_feature(M::::IsDecoratorManifold, k...) = my_feature(M.manifold, k...)

such that decorators act just as pass throughs for other decorator functions and

my_feature(M::MyManifold, k...) = #... my explicit implementation

then implements the feature itself.

Manifolds.Manifold — Type.

Manifold

A manifold type. The Manifold is used to dispatch to different exponential and logarithmic maps as well as other function on manifold.

Base.exp — Method.

exp(M::Manifold, x, v, T::AbstractVector)

Exponential map of tangent vector t*v at point x from manifold M for each t in T.

Base.exp — Method.

exp(M::Manifold, x, v, t=1)

Exponential map of tangent vector t*v at point x from manifold M.

Base.isapprox — Method.

isapprox(M::Manifold, x, v, w; kwargs...)

Check if vectors v and w tangent at x from manifold M are approximately equal.

Keyword arguments can be used to specify tolerances.

Base.isapprox — Method.

isapprox(M::Manifold, x, y; kwargs...)

Check if points x and y from manifold M are approximately equal.

Keyword arguments can be used to specify tolerances.

LinearAlgebra.norm — Method.

norm(M::Manifold, x, v)

Norm of tangent vector v at point x from manifold M.

Manifolds.base_manifold — Function.

base_manifold(M::Manifold)

Strip all decorators on M, returning the underlying topological manifold. Also used for vector bundles.

Manifolds.distance — Method.

distance(M::Manifold, x, y)

Shortest distance between the points x and y on manifold M.

Manifolds.exp! — Method.

exp!(M::Manifold, y, x, v, t=1)

Exponential map of tangent vector t*v at point x from manifold M. Result is saved to y.

Manifolds.geodesic — Method.

geodesic(M::Manifold, x, v, T::AbstractVector)

Get the points for each t in T traveling from x along the geodesic with initial point x and velocity v.

Manifolds.geodesic — Method.

geodesic(M::Manifold, x, v, t)

Get the point at time t traveling from x along the geodesic with initial point x and velocity v.

Manifolds.geodesic — Method.

geodesic(M::Manifold, x, v)

Get the geodesic with initial point x and velocity v. The geodesic is the curve of constant velocity that is locally distance-minimizing. This function returns a function of time, which may be a Real or an AbstractVector.

Manifolds.injectivity_radius — Method.

injectivity_radius(M::Manifold, x, R::AbstractRetractionMethod)

Distance $d$ such that retract(M, x, v, R) is injective for all tangent vectors shorter than $d$ (has a left inverse).

Manifolds.injectivity_radius — Method.

injectivity_radius(M::Manifold, x)

Distance $d$ such that exp(M, x, v) is injective for all tangent vectors shorter than $d$ (has a left inverse).

Manifolds.injectivity_radius — Method.

injectivity_radius(M::Manifold, x)

Infimum of the injectivity radii of all manifold points.

Manifolds.inner — Method.

inner(M::Manifold, x, v, w)

Inner product of tangent vectors v and w at point x from manifold M.

Manifolds.inverse_retract! — Method.

inverse_retract!(M::Manifold, v, x, y, [method::AbstractInverseRetractionMethod=LogarithmicInverseRetraction()])

Inverse retraction (cheaper, approximate version of logarithmic map) of points x and y. Result is saved to y.

Inverse retraction method can be specified by the last argument. Please look at the documentation of respective manifolds for available methods.

Manifolds.inverse_retract — Method.

inverse_retract(M::Manifold, x, y, [method::AbstractInverseRetractionMethod])

Inverse retraction (cheaper, approximate version of logarithmic map) of points x and y from manifold M.

Inverse retraction method can be specified by the last argument. Please look at the documentation of respective manifolds for available methods.

Manifolds.is_manifold_point — Method.

is_manifold_point(M,x)

check, whether x is a valid point on the Manifold M. If it is not, an error is thrown. The default is to return true, i.e. if no checks are implmented, the assumption is to be optimistic.

Manifolds.is_tangent_vector — Method.

is_tangent_vector(M,x,v)

check, whether v is a valid tangnt vector in the tangent plane of x on the Manifold M. An implementation should first check is_manifold_point(M,x) and then validate v. If it is not a tangent vector an error should be thrown. The default is to return true, i.e. if no checks are implmented, the assumption is to be optimistic.

Manifolds.manifold_dimension — Function.

manifold_dimension(M::Manifold)

The dimension $n$ of real space $\mathbb R^n$ to which the neighborhood of each point of the manifold is homeomorphic.

Manifolds.project_point! — Method.

project_point!(M::Manifold, x)

Project point from the ambient space onto the manifold M. The point x is modified. The function works only for selected embedded manifolds and is not required to return the closest point.

Manifolds.project_point — Method.

project_point(M::Manifold, x)

Project point from the ambient space onto the manifold M. The point x is not modified. The function works only for selected embedded manifolds and is not required to return the closest point.

Manifolds.project_tangent! — Method.

project_tangent!(M::Manifold, w, x, v)

Project ambient space representation of a vector v to a tangent vector at point x from the manifold M. The result is saved in vector w.

The function works only for selected embedded manifolds and is not required to return the closest vector.

Manifolds.project_tangent — Method.

project_tangent(M::Manifold, x, v)

Project ambient space representation of a vector v to a tangent vector at point x from the manifold M.

The function works only for selected embedded manifolds and is not required to return the closest vector.

Manifolds.retract! — Method.

retract!(M::Manifold, y, x, v, [t=1], [method::AbstractRetractionMethod=ExponentialRetraction()])

Retraction (cheaper, approximate version of exponential map) of tangent vector t*v at point x from manifold M. Result is saved to y.

Retraction method can be specified by the last argument. Please look at the documentation of respective manifolds for available methods.

Manifolds.retract — Method.

retract(M::Manifold, x, v, [t=1], [method::AbstractRetractionMethod])

Retraction (cheaper, approximate version of exponential map) of tangent vector t*v at point x from manifold M.

Manifolds.shortest_geodesic — Method.

shortest_geodesic(M::Manifold, x, y, T::AbstractVector)

Get the points for each t in T traveling from x along a shortest geodesic connecting x and y, where y is reached at t=1.

Manifolds.shortest_geodesic — Method.

shortest_geodesic(M::Manifold, x, y, t)

Get the point at time t traveling from x along a shortest geodesic connecting x and y, where y is reached at t=1.

Manifolds.shortest_geodesic — Method.

shortest_geodesic(M::Manifold, x, y)

Get a geodesic with initial point x and point y at t=1 whose length is the shortest path between the two points. When there are multiple shortest geodesics, there is no guarantee which will be returned. This function returns a function of time, which may be a Real or an AbstractVector.

Manifolds.vector_transport_direction! — Method.

vector_transport_direction!(M::Manifold, vto, x, v, vdir)

Vector transport of vector v at point x in the direction indicated by the tangent vector vdir at point x. The result is saved to vto. By default, exp and vector_transport_to! are used.

Manifolds.vector_transport_direction — Method.

vector_transport_direction(M::Manifold, x, v, vdir)

Vector transport of vector v at point x in the direction indicated by the tangent vector vdir at point x.

Manifolds.vector_transport_to! — Method.

vector_transport_to!(M::Manifold, vto, x, v, y)

Vector transport of vector v at point x to point y. The result is saved to vto. By default, project_tangent! is used but this may change in the future.

Manifolds.vector_transport_to — Method.

vector_transport_to(M::Manifold, x, v, y)

Vector transport of vector v at point x to point y.

Manifolds.CoTVector — Type.

CoTVector

Type for a cotangent vector of a manifold. While a Manifold not necessarily requires this type, for example when it is implemented for Vectors or Matrix type elements, this type can be used for more complicated representations, semantic verification or even dispatch for different representations of cotangent vectors and their types on a manifold.

Manifolds.ExponentialRetraction — Type.

ExponentialRetraction

Retraction using the exponential map.

Manifolds.LogarithmicInverseRetraction — Type.

LogarithmicInverseRetraction

Inverse retraction using the logarithmic map.

Manifolds.MPoint — Type.

MPoint

Type for a point on a manifold. While a Manifold not necessarily requires this type, for example when it is implemented for Vectors or Matrix type elements, this type can be used for more complicated representations, semantic verification or even dispatch for different representations of points on a manifold.

Manifolds.OutOfInjectivityRadiusError — Type.

OutOfInjectivityRadiusError

An error thrown when a function (for example logarithmic map or inverse retraction) is given arguments outside of its injectivity radius.

Manifolds.TVector — Type.

TVector

Type for a tangent vector of a manifold. While a Manifold not necessarily requires this type, for example when it is implemented for Vectors or Matrix type elements, this type can be used for more complicated representations, semantic verification or even dispatch for different representations of tangent vectors and their types on a manifold.

Base.angle — Method.

angle(M::Manifold, x, v, w)

Angle between tangent vectors v and w at point x from manifold M.

Manifolds.hat — Method.

hat(M::Manifold, x, vⁱ)

Given a basis $e_i$ on the tangent space at a point $x$ and tangent component vector $v^i$, compute the equivalent vector representation $v=v^i e_i$, where Einstein summation notation is used:

\[\wedge: v^i \mapsto v^i e_i\]

For matrix manifolds, this converts a vector representation of the tangent vector to a matrix representation. The vee map is the hat map's inverse.

Manifolds.representation_size — Function.

representation_size(M::Manifold, [VS::VectorSpaceType])

The size of array representing a point on manifold M, Representation sizes of tangent vectors can be obtained by calling the method with the second argument.

Manifolds.similar_result — Method.

similar_result(M::Manifold, f, x...)

Allocates an array for the result of function f on manifold M and arguments x... for implementing the non-modifying operation using the modifying operation.

Usefulness of passing a function is demonstrated by methods that allocate results of musical isomorphisms.

Manifolds.similar_result_type — Method.

similar_result_type(M::Manifold, f, args::NTuple{N,Any}) where N

Returns type of element of the array that will represent the result of function f for manifold M on given arguments (passed at a tuple).

Manifolds.vector_transport_along! — Method.

vector_transport_along!(M::Manifold, vto, x, v, c)

Vector transport of vector v at point x along the curve c such that c(0) is equal to x to point c(1). The result is saved to vto.

Manifolds.vector_transport_along — Method.

vector_transport_along(M::Manifold, x, v, c)

Vector transport of vector v at point x along the curve c such that c(0) is equal to x to point c(1).

Manifolds.vee — Method.

vee(M::Manifold, x, v)

Given a basis $e_i$ on the tangent space at a point $x$ and tangent vector $v$, compute the vector components $v^i$, such that $v = v^i e_i$, where Einstein summation notation is used:

\[\vee: v^i e_i \mapsto v^i\]

For matrix manifolds, this converts a matrix representation of the tangent vector to a vector representation. The hat map is the vee map's inverse.

Decorators

Array manifold

Metric manifold

Lie Group