Manifolds

Package Manifolds aims to provide both a unified interface to define and use manifolds as well as a library of manifolds to use for your projects.

Manifolds.AbstractEstimationMethod — Type

AbstractEstimationMethod

Abstract type for defining statistical estimation methods.

Manifolds.PolarInverseRetraction — Type

PolarInverseRetraction <: AbstractInverseRetractionMethod

Inverse retractions that are based on a singular value decomposition of the matrix / matrices for point and tangent vector on a Manifold

Manifolds.PolarRetraction — Type

PolarRetraction <: AbstractRetractionMethod

Retractions that are based on singular value decompositions of the matrix / matrices for point and tangent vector on a Manifold

Manifolds.QRInverseRetraction — Type

QRInverseRetraction <: AbstractInverseRetractionMethod

Inverse retractions that are based on a QR decomposition of the matrix / matrices for point and tangent vector on a Manifold

Manifolds.QRRetraction — Type

QRRetraction <: AbstractRetractionMethod

Retractions that are based on a QR decomposition of the matrix / matrices for point and tangent vector on a Manifold

Manifolds.hat — Method

hat(M::Manifold, x, vⁱ)

Given a basis $e_i$ on the tangent space at a point $x$ and tangent component vector $v^i$, compute the equivalent vector representation $v=v^i e_i$, where Einstein summation notation is used:

\[\wedge: v^i \mapsto v^i e_i\]

For matrix manifolds, this converts a vector representation of the tangent vector to a matrix representation. The vee map is the hat map's inverse.

Manifolds.vee — Method

vee(M::Manifold, x, v)

Given a basis $e_i$ on the tangent space at a point $x$ and tangent vector $v$, compute the vector components $v^i$, such that $v = v^i e_i$, where Einstein summation notation is used:

\[\vee: v^i e_i \mapsto v^i\]

For matrix manifolds, this converts a matrix representation of the tangent vector to a vector representation. The hat map is the vee map's inverse.

Decorators

Array manifold

Metric manifold