Distributions

Following functions and types provide support for manifold-valued and tangent space-valued distributions.

Manifolds.FVectorDistribution — Type

FVectorDistribution{TSpace<:VectorBundleFibers, T}

An abstract distribution for tangent vectors at point of type T from vector space of type TSpace.

Manifolds.FVectorSupport — Type

FVectorSupport(space::Manifold, VectorBundleFibers)

Value support for tangent vector-valued distributions (values from tangent space at point x from given manifold).

Manifolds.FVectorvariate — Type

FVectorvariate

Structure that subtypes VariateForm, indicating that a single sample is a vector from a fiber of a vector bundle.

Manifolds.MPointDistribution — Type

MPointDistribution{TM<:Manifold}

An abstract distribution for points on manifold of type TM.

Manifolds.MPointSupport — Type

MPointSupport(manifold::Manifold)

Value support for manifold-valued distributions (values from given manifold M).

Manifolds.MPointvariate — Type

MPointvariate

Structure that subtypes VariateForm, indicating that a single sample is a point on a manifold.

Distributions.support — Method

support(d::FVectorDistribution)

Get the object of type FVectorSupport for the distribution d.

Manifolds.ProjectedPointDistribution — Type

ProjectedPointDistribution(m::Manifold, d, proj!, x)

Generates a random point in ambient space of m and projects it to m using function proj!. Generated arrays are of type TResult, which can be specified by providing the x argument.

Manifolds.ProjectedFVectorDistribution — Type

ProjectedFVectorDistribution(type::VectorBundleFibers, x, d, project_vector!)

Generates a random vector from ambient space of manifold type.manifold at point x and projects it to vector space of type type using function project_vector!. Generated arrays are of type TResult.