Atlases and charts

Atlases on an $n$-dimensional manifold $\mathcal M$ are collections of charts $\{(U_i, \varphi_i) \colon i \in I\}$ such that $U_i \subseteq \mathcal M$ and each chart $\varphi_i$ is a function from $U_i$ to $\mathbb{R}^n$. They provide a basic connection between manifolds and Euclidean spaces.

Most operations on manifolds in Manifolds.jl avoid operating in a chart through appropriate embeddings and formulas derived for particular manifolds, though atlases provide the most general way of working with manifolds. They are extensively used in metric-related functions on MetricManifolds.

Atlases are represented by objects of subtypes of AbstractAtlas. There are no type restrictions for indices of charts in atlases.

Operations using atlases and charts are available through the following functions:

get_chart_index can be used to select an appropriate chart for the neighborhood of a given point.

get_point_coordinates converts a point to its coordinates in a chart.

get_point converts coordinates in a chart to the point that corresponds to them.

induced_basis returns a basis of a given vector space at a point induced by a chart.

transition_map converts coordinates of a point between two charts.

Manifolds.AbstractAtlas — Type

AbstractAtlas{𝔽}

An abstract class for atlases whith charts that have values in the vector space 𝔽ⁿ for some value of n. 𝔽 is a number system determined by an AbstractNumbers object.

Manifolds.InducedBasis — Type

InducedBasis(vs::VectorSpaceType, A::AbstractAtlas, i)

The basis induced by chart with index i from an AbstractAtlas A of vector space of type vs.

See also

VectorSpaceType, AbstractBasis

Manifolds.RetractionAtlas — Type

RetractionAtlas{
    𝔽,
    TRetr<:AbstractRetractionMethod,
    TInvRetr<:AbstractInverseRetractionMethod,
    TBasis<:AbstractBasis,
} <: AbstractAtlas{𝔽}

An atlas indexed by points on a manifold, such that coordinate transformations are performed using retractions, inverse retractions, and coordinate calculation for a given basis.

See also

AbstractAtlas, AbstractInverseRetractionMethod, AbstractRetractionMethod, AbstractBasis

Manifolds.get_chart_index — Method

get_chart_index(M::AbstractManifold, A::AbstractAtlas, p)

Select a chart from an AbstractAtlas A for manifold M that is suitable for representing the neighborhood of point p. This selection should be deterministic, although different charts may be selected for arbitrarily close but distinct points.

See also

get_default_atlas

Manifolds.get_default_atlas — Method

get_default_atlas(::AbstractManifold)

Determine the default real-valued atlas for the given manifold.

Manifolds.get_point — Method

get_point(M::AbstractManifold, A::AbstractAtlas, i, a)

Calculate point at coordinates a on manifold M in chart from an AbstractAtlas A at index i.

See also

get_point_coordinates, get_chart_index

Manifolds.get_point_coordinates — Method

get_point_coordinates(M::AbstractManifold, A::AbstractAtlas, i, p)

Calculate coordinates of point p on manifold M in chart from an AbstractAtlas A at index i. Coordinates are in the number system determined by A.

See also

get_point, get_chart_index

Manifolds.induced_basis — Method

induced_basis(M::AbstractManifold, A::AbstractAtlas, i, p, VST::VectorSpaceType)

Basis of vector space of type VST at point p from manifold M induced by chart (A, i).

See also

VectorSpaceType, AbstractAtlas

Manifolds.induced_basis — Method

induced_basis(::AbstractManifold, A::AbstractAtlas, i, VST::VectorSpaceType)

Get the basis induced by chart with index i from an AbstractAtlas A of vector space of type vs. Returns an object of type InducedBasis.

Manifolds.local_metric — Method

local_metric(M::AbstractManifold, p, B::InducedBasis)

Compute the local metric tensor for vectors expressed in terms of coordinates in basis B on manifold M. The point p is not checked.

Manifolds.transition_map — Method

transition_map(M::AbstractManifold, A_from::AbstractAtlas, i_from, A_to::AbstractAtlas, i_to, a)
transition_map(M::AbstractManifold, A::AbstractAtlas, i_from, i_to, a)

Given coordinates a in chart (A_from, i_from) of a point on manifold M, returns coordinates of that point in chart (A_to, i_to). If A_from and A_to are equal, A_to can be omitted.

See also

AbstractAtlas

Cotangent space and musical isomorphisms

Related to atlases, there is also support for the cotangent space and coefficients of cotangent vectors in bases of the cotangent space.

Functions sharp and flat implement musical isomorphisms for arbitrary vector bundles.

Manifolds.RieszRepresenterCotangentVector — Type

RieszRepresenterCotangentVector(M::AbstractManifold, p, X)

Cotangent vector in Riesz representer form on manifold M at point p with Riesz representer X.

Manifolds.flat — Method

flat(M::AbstractManifold, p, X)

Compute the flat isomorphism (one of the musical isomorphisms) of tangent vector X from the vector space of type M at point p from the underlying AbstractManifold.

The function can be used for example to transform vectors from the tangent bundle to vectors from the cotangent bundle $♭ : T\mathcal M → T^{*}\mathcal M$

Manifolds.sharp — Method

sharp(M::AbstractManifold, p, ξ)

Compute the sharp isomorphism (one of the musical isomorphisms) of vector ξ from the vector space M at point p from the underlying AbstractManifold.

The function can be used for example to transform vectors from the cotangent bundle to vectors from the tangent bundle $♯ : T^{*}\mathcal M → T\mathcal M$