ManifoldsBase.jl – An Interface for Manifolds

The interface for a manifold is provided in the lightweight package `ManifoldsBase.jl separate from the collection of manifolds in here. You can easily implement your algorithms and even first own manifolds just using the interface.

The following functions are currently available from the interface. If a manifold that you implement for your own package fits this interface, we happily look forward to a Pull Request to add it here.

ManifoldsBase.CoTVector — Type

CoTVector

Type for a cotangent vector of a manifold. While a Manifold does not necessarily require this type, for example when it is implemented for Vectors or Matrix type elements, this type can be used for more complicated representations, semantic verification, or even dispatch for different representations of cotangent vectors and their types on a manifold.

ManifoldsBase.MPoint — Type

MPoint

Type for a point on a manifold. While a Manifold does not necessarily require this type, for example when it is implemented for Vectors or Matrix type elements, this type can be used for more complicated representations, semantic verification, or even dispatch for different representations of points on a manifold.

ManifoldsBase.Manifold — Type

Manifold

A manifold type. The Manifold is used to dispatch to different functions on a manifold, usually as the first argument of the function. Examples are the exponential and logarithmic maps as well as more general functions that are built on them like the geodesic.

ManifoldsBase.ParallelTransport — Type

ParallelTransport <: AbstractVectorTransportMethod

Specify to use parallel transport as vector transport method within vector_transport_to, vector_transport_direction, or vector_transport_along.

ManifoldsBase.ProjectionTransport — Type

ProjectionTransport <: AbstractVectorTransportMethod

Specify to use projection onto tangent space as vector transport method within vector_transport_to, vector_transport_direction, or vector_transport_along. See project_tangent for details.

ManifoldsBase.TVector — Type

TVector

Type for a tangent vector of a manifold. While a Manifold does not necessarily require this type, for example when it is implemented for Vectors or Matrix type elements, this type can be used for more complicated representations, semantic verification, or even dispatch for different representations of tangent vectors and their types on a manifold.

Base.exp — Method

exp(M::Manifold, x, v, t::Real = 1)
exp(M::Manifold, x, v, T::AbstractVector) -> AbstractVector

Exponential map of tangent vector t*v at point x from manifold M. t may be a scalar or elements of vector T.

Base.isapprox — Method

isapprox(M::Manifold, x, v, w; kwargs...)

Check if vectors v and w tangent at x from manifold M are approximately equal.

Keyword arguments can be used to specify tolerances.

Base.isapprox — Method

isapprox(M::Manifold, x, y; kwargs...)

Check if points x and y from manifold M are approximately equal.

Keyword arguments can be used to specify tolerances.

Base.log — Method

log(M::Manifold, x, y)

Logarithmic map of point y at base point x on Manifold M.

LinearAlgebra.norm — Method

norm(M::Manifold, x, v)

Norm of tangent vector v at point x from manifold M.

ManifoldsBase.base_manifold — Method

base_manifold(M::Manifold)

Return the internally stored manifold for decorated manifolds and the base manifold for vector bundles or power manifolds.

ManifoldsBase.check_manifold_point — Method

check_manifold_point(M::Manifold, x; kwargs...) -> Union{Nothing,String}

Return nothing when x is a point on manifold M. Otherwise, return a string with a description why the point does not belong to manifold M.

By default, check_manifold_point returns nothing, i.e. if no checks are implemented, the assumption is to be optimistic for a point not deriving from the MPoint type.

ManifoldsBase.check_tangent_vector — Method

check_tangent_vector(M::Manifold, x, v; kwargs...) -> Union{Nothing,String}

Check whether v is a valid tangent vector in the tangent plane of x on the Manifold M. An implementation should first call check_manifold_point(M, x; kwargs...) and then validate v. If it is not a tangent vector, an error string should be returned.

By default, check_tangent_vector returns nothing, i.e. if no checks are implemented, the assumption is to be optimistic for tangent vectors not deriving from the TVector type.

ManifoldsBase.distance — Method

distance(M::Manifold, x, y)

Shortest distance between the points x and y on manifold M.

ManifoldsBase.exp! — Method

exp!(M::Manifold, y, x, v, t::Real = 1)

Exponential map of tangent vector t*v at point x from manifold M. Result is saved to y.

ManifoldsBase.geodesic — Method

geodesic(M::Manifold, x, v) -> Function

Get the geodesic with initial point x and velocity v. The geodesic is the curve of constant velocity that is locally distance-minimizing. This function returns a function of time, which may be a Real or an AbstractVector.

geodesic(M::Manifold, x, v, t::Real)
geodesic(M::Manifold, x, v, T::AbstractVector) -> AbstractVector

Return the point at time t or points at times t in T along the geodesic.

ManifoldsBase.injectivity_radius — Method

injectivity_radius(M::Manifold, x)

Distance $d$ such that exp(M, x, v) is injective for all tangent vectors shorter than $d$ (i.e. has a left inverse).

injectivity_radius(M::Manifold)

Infimum of the injectivity radius of all manifold points.

injectivity_radius(M::Manifold[, x], method::AbstractRetractionMethod)

Distance $d$ such that retract(M, x, v, method) is injective for all tangent vectors shorter than $d$ (i.e. has a left inverse) for point $x$ if provided or all manifold points otherwise.

ManifoldsBase.inner — Method

inner(M::Manifold, x, v, w)

Inner product of tangent vectors v and w at point x from manifold M.

ManifoldsBase.inverse_retract! — Method

inverse_retract!(M::Manifold, v, x, y[, method::AbstractInverseRetractionMethod])

Inverse retraction (cheaper, approximate version of logarithmic map) of points x and y. Result is saved to v.

Inverse retraction method can be specified by the last argument, defaulting to LogarithmicInverseRetraction. See the documentation of respective manifolds for available methods.

ManifoldsBase.inverse_retract — Method

inverse_retract(M::Manifold, x, y)
inverse_retract(M::Manifold, x, y, method::AbstractInverseRetractionMethod

Inverse retraction (cheaper, approximate version of logarithmic map) of points x and y from manifold M.

Inverse retraction method can be specified by the last argument, defaulting to LogarithmicInverseRetraction. See the documentation of respective manifolds for available methods.

ManifoldsBase.is_decorator_manifold — Method

is_decorator_manifold(M::Manifold)

Indicate whether a manifold is a decorator manifold, i.e. whether it encapsulates a Manifold with additional features and stores internally the original manifold instance. An example is the ArrayManifold.

Certain functions are just calling themselves on the internal manifold and hence do not need to be reimplemented for decorators again, for example manifold_dimension and especially base_manifold.

It is assumed that the undecorated (base) manifold is stored in M.manifold. Alternatively, overload base_manifold.

ManifoldsBase.is_manifold_point — Function

is_manifold_point(M::Manifold, x, throw_error = false; kwargs...)

Return whether x is a valid point on the Manifold M.

If throw_error is false, the function returns either true or false. If throw_error is true, the function either returns true or throws an error. By default the function calls check_manifold_point(M, x; kwargs...) and checks whether the returned value is nothing or an error.

ManifoldsBase.is_tangent_vector — Function

is_tangent_vector(M::Manifold, x, v, throw_error = false; kwargs...)

Return whether v is a valid tangent vector at point x on the Manifold M. Returns either true or false.

The default is to return true, i.e. if no checks are implemented, the assumption is to be optimistic.

ManifoldsBase.log! — Method

log!(M::Manifold, v, x, y)

Logarithmic map of point y at base point x on Manifold M. Result is saved to v.

ManifoldsBase.manifold_dimension — Method

manifold_dimension(M::Manifold)

The dimension $n$ of real space $\mathbb R^n$ to which the neighborhood of each point of the manifold is homeomorphic.

ManifoldsBase.project_point! — Method

project_point!(M::Manifold, y, x)

Project point x from the ambient space onto the manifold M. The point y is overwritten by the projection. The function works only for selected embedded manifolds and is not required to return the closest point.

ManifoldsBase.project_point — Method

project_point(M::Manifold, x)

Project point from the ambient space onto the manifold M. The point x is not modified. The function works only for selected embedded manifolds and is not required to return the closest point.

ManifoldsBase.project_tangent! — Method

project_tangent!(M::Manifold, w, x, v)

Project ambient space representation of a vector v to a tangent vector at point x from the manifold M. The result is saved in vector w.

The function works only for selected embedded manifolds and is not required to return the closest vector.

ManifoldsBase.project_tangent — Method

project_tangent(M::Manifold, x, v)

Project ambient space representation of a vector v to a tangent vector at point x from the manifold M.

The function works only for selected embedded manifolds and is not required to return the closest vector.

ManifoldsBase.representation_size — Method

representation_size(M::Manifold)

The size of an array representing a point on manifold M.

ManifoldsBase.retract! — Method

retract!(M::Manifold, y, x, v[, t::Real=1])
retract!(M::Manifold, y, x, v[, t::Real=1], method::AbstractRetractionMethod)

Retraction (cheaper, approximate version of exponential map) of tangent vector t*v at point x from manifold M. Result is saved to y.

Retraction method can be specified by the last argument, defaulting to ExponentialRetraction. See the documentation of respective manifolds for available methods.

ManifoldsBase.retract — Method

retract(M::Manifold, x, v[, t::Real=1])
retract(M::Manifold, x, v[, t::Real=1], method::AbstractRetractionMethod)

Retraction (cheaper, approximate version of exponential map) of tangent vector t*v at point x from manifold M.

Retraction method can be specified by the last argument, defaulting to ExponentialRetraction. See the documentation of respective manifolds for available methods.

ManifoldsBase.shortest_geodesic — Method

shortest_geodesic(M::Manifold, x, y) -> Function

Get a geodesic $\gamma_x(t)$ whose length is the shortest path between the points $x$ and $y$, where $\gamma_x(0)=x$ and $\gamma_x(1)=y$. When there are multiple shortest geodesics, there is no guarantee which will be returned.

This function returns a function of time, which may be a Real or an AbstractVector.

shortest_geodesic(M::Manifold, x, y, t::Real)
shortest_geodesic(M::Manifold, x, y, T::AbstractVector) -> AbstractVector

Return the point at time $t$ or points at times $t$ in $T$ along the shortest geodesic.

ManifoldsBase.vector_transport_along! — Method

vector_transport_along!(M::Manifold, vto, x, v, c[, method::AbstractVectorTransportMethod])

Transport a vector v at point x along the curve c such that c(0) is equal to x to point c(1) using the method, which defaults to ParallelTransport. The result is saved to vto.

ManifoldsBase.vector_transport_along — Method

vector_transport_along(M::Manifold, x, v, c[, method::AbstractVectorTransportMethod])

Transport a vector v at point x along the curve c such that c(0) is equal to x to point c(1). The default method used is ParallelTransport.

ManifoldsBase.vector_transport_direction! — Method

vector_transport_direction!(M::Manifold, vto, x, v, vdir[, method::AbstractVectorTransportMethod])

Transport a vector v at point x in the direction indicated by the tangent vector vdir at point x. The result is saved to vto. By default, exp and vector_transport_to! are used with the method, which defaults to ParallelTransport.

ManifoldsBase.vector_transport_direction — Method

vector_transport_direction(M::Manifold, x, v, vdir[, method::AbstractVectorTransportMethod])

Transport a vector v at point x in the direction indicated by the tangent vector vdir at point x using the method, which defaults to ParallelTransport.

ManifoldsBase.vector_transport_to! — Method

vector_transport_to!(M::Manifold, vto, x, v, y, method::ProjectionTransport)

Transport a vector v in the tangent space at x on a Manifold M by interpreting it as an element of the embedding and then projecting it onto the tangent space at y.

ManifoldsBase.vector_transport_to! — Method

vector_transport_to!(M::Manifold, vto, x, v, y[, method::AbstractVectorTransportMethod])

Vector transport of vector v at point x to point y. The result is saved to vto. By default, the method is ParallelTransport.

ManifoldsBase.vector_transport_to — Method

vector_transport_to(M::Manifold, x, v, y[, method::AbstractVectorTransportMethod])

Transport a vector v at point x to point y using the method, which defaults to ParallelTransport.

ManifoldsBase.zero_tangent_vector! — Method

zero_tangent_vector!(M::Manifold, v, x)

Save to v a vector such that retracting v to manifold M at x produces x.

ManifoldsBase.zero_tangent_vector — Method

zero_tangent_vector(M::Manifold, x)

Vector v such that retracting v to manifold M at x produces x.

ManifoldsBase.AbstractEstimationMethod — Type

AbstractEstimationMethod

Abstract type for defining statistical estimation methods.

ManifoldsBase.AbstractInverseRetractionMethod — Type

AbstractInverseRetractionMethod

Abstract type for methods for inverting a retraction (see inverse_retract).

ManifoldsBase.AbstractRetractionMethod — Type

AbstractRetractionMethod

Abstract type for methods for retracting a tangent vector to a manifold.

ManifoldsBase.AbstractVectorTransportMethod — Type

AbstractVectorTransportMethod

Abstract type for methods for transporting vectors.

ManifoldsBase.ExponentialRetraction — Type

ExponentialRetraction

Retraction using the exponential map.

ManifoldsBase.LogarithmicInverseRetraction — Type

LogarithmicInverseRetraction

Inverse retraction using the logarithmic map.

ManifoldsBase.OutOfInjectivityRadiusError — Type

OutOfInjectivityRadiusError

An error thrown when a function (for example logarithmic map or inverse_retract) is given arguments outside of its injectivity_radius.

Base.angle — Method

angle(M::Manifold, x, v, w)

Angle between tangent vectors v and w at point x from manifold M.

ManifoldsBase.similar_result — Method

similar_result(M::Manifold, f, x...)

Allocate an array for the result of function f on manifold M and arguments x... for implementing the non-modifying operation using the modifying operation.

Usefulness of passing a function is demonstrated by methods that allocate results of musical isomorphisms.

ManifoldsBase.similar_result_type — Method

similar_result_type(M::Manifold, f, args::NTuple{N,Any}) where N

Return type of element of the array that will represent the result of function f for manifold M on given arguments args (passed as a tuple).