Cholesky Space

The Cholesky Space is a Riemannian manifold on the lower triangular matrices.

Manifolds.CholeskySpace — Type

CholeskySpace{N} <: Manifold

the manifold of lower triangular matrices with positive diagonal and a metric based on the cholesky decomposition. The formulae for this manifold are for example summarized in Table 1 of [Lin2019].

ManifoldsBase.distance — Method

distance(M,x,y)

computes the Riemannian distance on the CholeskySpace M between two matrices x, y that are lower triangular with positive diagonal. The formula reads

\[d_{\mathcal M}(x,y) = \sqrt{ \sum_{i>j} (x_{ij}-y_{ij})^2 + \sum_{j=1}^m (\log x_{jj} - \log y_{jj})^2 }\]

ManifoldsBase.exp! — Method

exp!(M,y,x,v)

compute the exponential map on the CholeskySpace M eminating from the lower triangular matrix with positive diagonal x towards the lower triangular matrix v and return the result in y. The formula reads

\[\exp_x v = \lfloor x \rfloor + \lfloor v \rfloor +\operatorname{diag}(x)\operatorname{diag}(x)\exp\bigl( \operatorname{diag}(v)\operatorname{diag}(x)^{-1}\bigr)\]

where $\lfloor x\rfloor$ denotes the strictly lower triangular matrix of $x$ and $\operatorname{diag}(x)$ the diagonal matrix of $x$

ManifoldsBase.inner — Method

inner(M,x,v,w)

computes the inner product on the CholeskySpace M at the lower triangular matric with positive diagonal x and the two tangent vectors v,w, i.e they are both lower triangular matrices with arbitrary diagonal. The formula reads

\[ g_{x}(v,w) = \sum_{i>j} v_{ij}w_{ij} + \sum_{j=1}^m v_{ii}w_{ii}x_{ii}^{-2}\]

ManifoldsBase.log! — Method

log!(M,v,x,y)

compute the exponential map on the CholeskySpace M eminating from the lower triangular matrix with positive diagonal x towards the lower triangular matrx v and return the result in y. The formula reads

\[\log_x v = \lfloor x \rfloor - \lfloor y \rfloor +\operatorname{diag}(x)\log\bigl(\operatorname{diag}(y)\operatorname{diag}(x)^{-1}\bigr)\]

where $\lfloor x\rfloor$ denotes the strictly lower triangular matrix of $x$ and $\operatorname{diag}(x)$ the diagonal matrix of $x$

ManifoldsBase.vector_transport_to! — Method

vector_transport!(M,vto,x,v,y)

parallely transport the tangent vector v at x along the geodesic to y on respect to the CholeskySpace manifold M. The formula reads

\[ \mathcal P_{x\to y}(v) = \lfloor v \rfloor + \operatorname{diag}(y)\operatorname{diag}(x)^{-1}\operatorname{diag}(v),\]

where $\lfloor\cdot\rfloor$ denotes the strictly lower triangular matrix, and $\operatorname{diag}$ extracts the diagonal matrix.

ManifoldsBase.zero_tangent_vector! — Method

zero_tangent_vector!(M,v,x)

returns the zero tangent vector on the CholeskySpace M at x in the variable v.

ManifoldsBase.check_tangent_vector — Method

check_tangent_vector(M,x,v; kwargs... )

checks whether v is a tangent vector to x on the CholeskySpace M, i.e. atfer check_manifold_point(M,x), v has to be of same dimension as x and a symmetric matrix. The tolerance for the tests can be set using the kwargs....

Literature

Lin2019

Lin, Zenhua: "Riemannian Geometry of Symmetric Positive Definite Matrices via Cholesky Decomposition", arXiv: 1908.09326.