Circle

Manifolds.Circle — Type

Circle{T} <: Manifold

The circle $\mathbb S^1$ as a manifold ere manifold represented by real-valued data in $[-\pi,\pi)$ or complex-valued data $z\in \mathbb C$ of absolute value $\lvert z\rvert = 1$.

Constructor

Circle(f=ℝ)

generates the Circle represented by angles, which alternatively can be set to use the field f=ℂ to obtain the Circle as a manifold embedded in ℂ.

Base.exp — Method

exp(M,x,v)

compute the exponential map on the Circle.

\[\exp_xv = (x+v)_{2\pi},\]

where $(\cdot)$ is the (symmetric) remainder with respect to division by $2\pi$, i.e. in $[-\pi,\pi)$.

For the complex-valued case the formula is the same as for the Sphere applied to valuedin the complex plane.

Base.log — Method

log(M,x,y)

computes the logarithmic map on the Circle M.

\[\exp_xv = (y,x)_{2\pi},\]

where $(\cdot)$ is the (symmetric) remainder with respect to division by $2\pi$, i.e. in $[-\pi,\pi)$.

For the complex-valued case the formula is the same as for the Sphere applied to valuedin the complex plane.

Manifolds.complex_dot — Method

complex_dot(a,b)

compute the inner product of two (complex) numbers with in the complex plane.

Manifolds.sym_rem — Method

sym_rem(x,[T=π])

symmetric remainder of x with respect to the interall 2*T, i.e. (x+T)%2T, where the default for T is $\pi$

ManifoldsBase.distance — Method

distance(::Circle,x,y)

compute the distance on the Circle $\mathbb S^1$, which is the absolute value of the symmetric remainder of x and y for the real-valued case and the angle between both complex numbers in the Gaussian plane for the complex-valued case.

ManifoldsBase.injectivity_radius — Method

injectivity_radius(M)
injectivity_radius(M,x)

returns the injectivity radius on the Circle M, i.e. $\pi$.

ManifoldsBase.inner — Method

inner(M::Circle, x, w, v)

compute the inner product of the two tangent vectors w,v from the tangent plane at x on the Circle $\mathbb S^1$ using the restriction of the metric from the embedding, i.e.

\[\langle v,w\rangle_x = w*v\]

for the real case and

\[\langle v,w\rangle_x = v^\mathrm{T}w\]

for the complex case interpreting complex numbers in the Gaussian plane.

ManifoldsBase.manifold_dimension — Method

manifold_dimension(M::Circle)

Return the dimension of the Circle $\mathbb S^1$, i.e. 1.

Statistics.mean — Method

mean(M::Circle, x::AbstractVector[, w::AbstractWeights])

Compute the Riemannian mean of x on the Circle $\mathbb S^1$ by the wrapped mean, i.e. the remainder of the mean modulo 2π.