Circle

Manifolds.Circle — Type

Circle{F} <: Manifold

The circle $\mathbb S^1$ as a manifold ere manifold represented by real-valued data in $[-\pi,\pi)$ or complex-valued data $z\in \mathbb C$ of absolute value $\lvert z\rvert = 1$.

Constructor

Circle(f=ℝ)

Generate the ℝ-valued Circle represented by angles, which alternatively can be set to use the AbstractNumbers f=ℂ to obtain the Circle represented by ℂ-valued Circle of unit numbers.

Base.exp — Method

exp(M::Circle, x, v)

Compute the exponential map on the Circle.

\[\exp_xv = (x+v)_{2\pi},\]

where $(\cdot)$ is the (symmetric) remainder with respect to division by $2\pi$, i.e. in $[-\pi,\pi)$.

For the complex-valued case the formula is the same as for the Sphere applied to valuedin the complex plane.

Base.log — Method

log(M::Circle, x, y)

Compute the logarithmic map on the Circle M.

\[\exp_xv = (y,x)_{2\pi},\]

where $(\cdot)$ is the (symmetric) remainder with respect to division by $2\pi$, i.e. in $[-\pi,\pi)$.

For the complex-valued case the formula is the same as for the Sphere applied to valuedin the complex plane.

Manifolds.complex_dot — Method

complex_dot(a, b)

Compute the inner product of two (complex) numbers with in the complex plane.

Manifolds.get_coordinates — Method

get_coordinates(M::Circle{ℂ}, x, v, B::ArbitraryOrthonormalBasis)

Return tangent vector coordinates in the Lie algebra of the circle.

Manifolds.get_vector — Method

get_vector(M::Circle{ℂ}, x, v, B::ArbitraryOrthonormalBasis)

Return tangent vector from the coordinates in the Lie algebra of the circle.

Manifolds.sym_rem — Method

sym_rem(x,[T=π])

Compute symmetric remainder of x with respect to the interall 2*T, i.e. (x+T)%2T, where the default for T is $\pi$

ManifoldsBase.distance — Method

distance(M::Circle, x, y)

Compute the distance on the Circle M, which is the absolute value of the symmetric remainder of x and y for the real-valued case and the angle between both complex numbers in the Gaussian plane for the complex-valued case.

ManifoldsBase.injectivity_radius — Method

injectivity_radius(M::Circle[, x])

Return the injectivity radius on the Circle M, i.e. $\pi$.

ManifoldsBase.inner — Method

inner(M::Circle, x, w, v)

Compute the inner product of the two tangent vectors w,v from the tangent plane at x on the Circle M using the restriction of the metric from the embedding, i.e.

\[g_x(v,w) = w*v\]

for the real case and

\[g_x(v,w) = v^\mathrm{T}w\]

for the complex case interpreting complex numbers in the Gaussian plane.

ManifoldsBase.manifold_dimension — Method

manifold_dimension(M::Circle)

Return the dimension of the Circle M, i.e. $\operatorname{dim}(\mathbb S^1) = 1$.

ManifoldsBase.project_point — Method

project_point(M::Circle, x)

Project a point x onto the Circle M. For the real-valued case this is the remainder with respect to modulus $2\pi$. For the complex-valued case the result is the projection of x onto the unit circle in the complex plane.

ManifoldsBase.project_tangent — Method

project_tangent(M::Circle, x, v)

Project a value v onto the tangent space of the point x on the Circle M.

For the real-valued case this is just the identity. For the complex valued case v is projected onto the line in the complex plane that is parallel to the tangent to x on the unit circle and contains 0.

ManifoldsBase.vector_transport_to — Method

vector_transport_to(M::Circle, x, v, y, ::ParallelTransport)

Compute the parallel transport of v from the tangent space at x to the tangent space at y on the Circle M. For the real-valued case this results in the identity. For the complex-valud case, the formula is the same as for the Sphere(1) in the complex plane.

\[P_{y\gets x}(v) = v - \frac{\langle \log_xy,v\rangle_x}{d^2_{\mathbb C}(x,y)} \bigl(\log_xy + \log_yx \bigr),\]

where log denotes the logarithmic map on M.

Statistics.mean — Method

mean(M::Circle, x::AbstractVector[, w::AbstractWeights])

Compute the Riemannian mean of x on the Circle $\mathbb S^1$ by the wrapped mean, i.e. the remainder of the mean modulo 2π.

ManifoldsBase.check_manifold_point — Method

check_manifold_point(M::Circle, x)

Check whether x is a point on the Circle M. For the real-valued case, x is an angle and hence it checks that $x \in [-\pi,\pi)$. for the complex-valued case its a unit number, $x \in \mathbb C$ with $\lvert x \rvert = 1$.

ManifoldsBase.check_tangent_vector — Method

check_tangent_vector(M::Circle, x, v)

Check whether v is a tangent vector in the tangent space of x on the Circle M. For the real-valued case represented by angles all v are valid, since the tangent space is the whole real line. For the complex-valued case v has to lie on the line parallel to the tangent line at x in the complex plane, i.e. the inner product is zero.