Elliptope

Manifolds.Elliptope — Type

Elliptope{T} <: AbstractDecoratorManifold{ℝ}

The Elliptope manifold, also known as the set of correlation matrices, consists of all symmetric positive semidefinite matrices of rank $k$ with unit diagonal, i.e.,

\[\begin{aligned} \mathcal E(n,k) = \bigl\{p ∈ ℝ^{n×n}\ \big|\ &a^\mathrm{T}pa \geq 0 \text{ for all } a ∈ ℝ^{n},\\ &p_{ii} = 1 \text{ for all } i=1,\ldots,n,\\ &\text{and } p = qq^{\mathrm{T}} \text{ for } q \in ℝ^{n×k} \text{ with } \operatorname{rank}(p) = \operatorname{rank}(q) = k \bigr\}. \end{aligned}\]

And this manifold is working solely on the matrices $q$. Note that this $q$ is not unique, indeed for any orthogonal matrix $A$ we have $(qA)(qA)^{\mathrm{T}} = qq^{\mathrm{T}} = p$, so the manifold implemented here is the quotient manifold. The unit diagonal translates to unit norm columns of $q$.

The tangent space at $p$, denoted $T_p\mathcal E(n,k)$, is also represented by matrices $Y\in ℝ^{n×k}$ and reads as

\[T_p\mathcal E(n,k) = \bigl\{ X ∈ ℝ^{n×n}\,|\,X = qY^{\mathrm{T}} + Yq^{\mathrm{T}} \text{ with } X_{ii} = 0 \text{ for } i=1,\ldots,n \bigr\}\]

endowed with the Euclidean metric from the embedding, i.e. from the $ℝ^{n×k}$

This manifold was for example investigated in[JBAS10].

Constructor

Elliptope(n::Int, k::Int; parameter::Symbol=:type)

generates the manifold $\mathcal E(n,k) \subset ℝ^{n×n}$.

parameter: whether a type parameter should be used to store n and k. By default size is stored in type. Value can either be :field or :type.

ManifoldsBase.check_point — Method

check_point(M::Elliptope, q; kwargs...)

checks, whether q is a valid representation of a point $p=qq^{\mathrm{T}}$ on the Elliptope M, i.e. is a matrix of size (N,K), such that $p$ is symmetric positive semidefinite and has unit trace. Since by construction $p$ is symmetric, this is not explicitly checked. Since $p$ is by construction positive semidefinite, this is not checked. The tolerances for positive semidefiniteness and unit trace can be set using the kwargs....

ManifoldsBase.check_vector — Method

check_vector(M::Elliptope, q, Y; kwargs... )

Check whether $X = qY^{\mathrm{T}} + Yq^{\mathrm{T}}$ is a tangent vector to $p=qq^{\mathrm{T}}$ on the Elliptope M, i.e. Y has to be of same dimension as q and a $X$ has to be a symmetric matrix with zero diagonal.

The tolerance for the base point check and zero diagonal can be set using the kwargs.... Note that symmetric of $X$ holds by construction an is not explicitly checked.

ManifoldsBase.is_flat — Method

is_flat(::Elliptope)

Return false. Elliptope is not a flat manifold.

ManifoldsBase.manifold_dimension — Method

manifold_dimension(M::Elliptope)

returns the dimension of Elliptope M $=\mathcal E(n,k), n,k ∈ ℕ$, i.e.

\[\dim \mathcal E(n,k) = n(k-1) - \frac{k(k-1)}{2}.\]

ManifoldsBase.project — Method

project(M::Elliptope, q)

project q onto the manifold Elliptope M, by normalizing the rows of q.

ManifoldsBase.project — Method

project(M::Elliptope, q, Y)

Project Y onto the tangent space at q, i.e. row-wise onto the oblique manifold.

ManifoldsBase.representation_size — Method

representation_size(M::Elliptope)

Return the size of an array representing an element on the Elliptope manifold M, i.e. $n×k$, the size of such factor of $p=qq^{\mathrm{T}}$ on $\mathcal M = \mathcal E(n,k)$.

ManifoldsBase.retract — Method

retract(M::Elliptope, q, Y, ::ProjectionRetraction)

compute a projection based retraction by projecting $q+Y$ back onto the manifold.

ManifoldsBase.vector_transport_to — Method

vector_transport_to(M::Elliptope, p, X, q)

transport the tangent vector X at p to q by projecting it onto the tangent space at q.

ManifoldsBase.zero_vector — Method

zero_vector(M::Elliptope,p)

returns the zero tangent vector in the tangent space of the symmetric positive definite matrix p on the Elliptope manifold M.

Literature