Euclidean

Euclidean Space

The Euclidean space $\mathbb R^n$ is a simple model space, since it has curvature constantly zero everywhere and hence nearly all operations simplify.

Manifolds.Euclidean — Type.

Euclidean{T<:Tuple} <: Manifold

Euclidean vector space $\mathbb R^n$.

Constructor

Euclidean(n)

generates the $n$-dimensional vector space $\mathbb R^n$.

Euclidean(n₁,n₂,...,nᵢ)

generates the $n_1n_2\cdot\ldots n_i$-dimensional vector space $\mathbb R^{n_1, n_2, \ldots, n_i}$, whose elements are interpreted as $n_1 \times,n_2\times\cdots\times n_i$ arrays, e.g. for two parameters as matrices.

Manifolds.EuclideanMetric — Type.

EuclideanMetric <: RiemannianMetric

a general type for any manifold that employs the Euclidean Metric, for example the Euclidean manifold itself, or the Sphere, where every tangent space (as a plane in the embedding) uses this metric (in the embedding).

LinearAlgebra.norm — Method.

norm(M::Euclidean,x,v)

compute the norm of a tangent vector v at x on the Euclidean manifold M, i.e. since every tangent space can be identified with M itself in this case, just the (Frobenius) norm of v.

Manifolds.distance — Method.

distance(M::Euclidean,x,y)

compute the Euclidean distance between two points on the Euclidean manifold M, i.e. for vectors it's just the norm of the difference, for matrices and higher order arrays, the matrix and ternsor Frobenius norm, respectively.

Manifolds.exp! — Method.

exp!(M::Euclidean,y, x, v)

compute the exponential map on the Euclidean manifold M from x in direction v in place, i.e. in y, which in this case is just

\[ y = x + v\]

Manifolds.flat! — Method.

flat!(M::Euclidean, v, x, w)

since cotangent and tangent vectors can directly be identified in the Euclidean case, this yields just the identity for a tangent vector w in the tangent space of x on M. The result is returned also in place in v.

Manifolds.log! — Method.

log!(M::Euclidean, v, x, y)

compute the logarithmic map on the Euclidean manifold M from x tpo y, stored in the direction v in place, i.e. in v, which in this case is just

\[ v = y - x\]

Manifolds.manifold_dimension — Method.

manifold_dimension(M::Euclidean{T})

returns the manifold dimension of the Euclidean manifold M, i.e. the product of all array dimensions.

Manifolds.project_point! — Method.

project_point!(M::Euclidean, x)

project an arbitrary point x onto the Euclidean manifold M, which is of course just the identity map.

Manifolds.project_tangent! — Method.

project_tangent!(M::Euclidean, w, x, v)

project an arbitrary vector v into the tangent space of a point x on the Euclidean manifold M, which is just the identity, since any tangent space of M can be identified with all of M.

Manifolds.representation_size — Method.

representation_size(M::Euclidean{T})

returns the array dimensions required to represent an element on the Euclidean manifold M, i.e. the vector of all array dimensions.

Manifolds.sharp! — Method.

sharp!(M::Euclidean, v, x, w)

since cotangent and tangent vectors can directly be identified in the Euclidean case, this yields just the identity for a cotangent vector w in the tangent space of x on M. The result is returned also in place in v.

Manifolds.vector_transport_to! — Method.

vector_transport_to!(M::Euclidean, vto, x, v, y, ::ParallelTransport)

parallel transport the vector v from the tangent space at x to the tangent space at y on the Euclidean manifold M, i.e. the in place w is just set to v.

Manifolds.zero_tangent_vector! — Method.

zero_tangent_vector!(M::Euclidean, v, x)

compute a zero vector in the tangent space of x on the Euclidean manifold M, whcih here is just a zero filled array the same size as the in place parameter v which is assumed to have the same size as x.

Manifolds.normal_tvector_distribution — Method.

normal_tvector_distribution(S::Euclidean, x, σ)

Normal distribution in ambient space with standard deviation σ projected to tangent space at x.

Manifolds.projected_distribution — Method.

projected_distribution(M::Euclidean, d, [x])

Wraps standard distribution d into a manifold-valued distribution. Generated points will be of similar type to x. By default, the type is not changed.