Euclidean

Sphere

Manifolds.Euclidean — Type.

Euclidean{T<:Tuple} <: Manifold

Euclidean vector space $\mathbb R^n$.

Constructor

Euclidean(n)

generates the $n$-dimensional vector space $\mathbb R^n$.

Euclidean(m, n)

generates the $mn$-dimensional vector space $\mathbb R^{m \times n}$, whose elements are interpreted as $m \times n$ matrices.

Manifolds.normal_tvector_distribution — Method.

normal_tvector_distribution(S::Euclidean, x, σ)

Normal distribution in ambient space with standard deviation σ projected to tangent space at x.

Manifolds.projected_distribution — Method.

projected_distribution(M::Euclidean, d, [x])

Wraps standard distribution d into a manifold-valued distribution. Generated points will be of similar type to x. By default, the type is not changed.