Fiber bundles

Fiber bundle $E$ is a manifold that is built on top of another manifold $\mathcal M$ (base space). It is characterized by a continuous function $Π : E → \mathcal M$. For each point $p ∈ \mathcal M$ the preimage of $p$ by $Π$, $Π^{-1}(\{p\})$ is called a fiber $F$. Bundle projection can be performed using function bundle_projection.

Manifolds.jl primarily deals with the case of trivial bundles, where $E$ can be identified with a product $M \times F$.

Vector bundles is a special case of a fiber bundle. Other examples include unit tangent bundle. Note that in general fiber bundles don't have a canonical Riemannian structure but can at least be equipped with an Ehresmann connection, providing notions of parallel transport and curvature.

Documentation

Manifolds.BundleFibers — Type

BundleFibers(fiber::FiberType, M::AbstractManifold)

Type representing a family of fibers of a fiber bundle over M with vectorfiber of type fiber. In contrast with FiberBundle, operations on BundleFibers expect point-like and fiber-like parts to be passed separately instead of being bundled together. It can be thought of as a representation of fibers from a fiber bundle but without storing the point at which a fiber is attached (which is specified separately in various functions).

Manifolds.FiberAtPoint — Type

FiberAtPoint{
    𝔽,
    TFiber<:BundleFibers{<:FiberType,<:AbstractManifold{𝔽}},
    TX,
} <: AbstractManifold{𝔽}

A fiber of a FiberBundle at a point p on the manifold. This is modelled using BundleFibers with only a fiber part and fixing the point-like part to be just p.

This fiber itself is also a manifold. For vector fibers it's by default flat and hence isometric to the Euclidean manifold.

Constructor

FiberAtPoint(fiber::BundleFibers, p)

A fiber of type fiber at point p from the manifold fiber.manifold.

Manifolds.FiberType — Type

abstract type FiberType end

An abstract type for fiber types.

ManifoldsBase.allocate_result — Method

allocate_result(B::BundleFibers, f, x...)

Allocates an array for the result of function f that is an element of the fiber space of type B.fiber on manifold B.manifold and arguments x... for implementing the non-modifying operation using the modifying operation.

ManifoldsBase.allocate_result_type — Method

allocate_result_type(B::BundleFibers, f, args::NTuple{N,Any}) where N

Return type of element of the array that will represent the result of function f for representing an operation with result in the fiber fiber for manifold M on given arguments (passed at a tuple).

ManifoldsBase.zero_vector — Method

zero_vector(B::BundleFibers, p)

Compute the zero vector from the vector space of type B.fiber at point p from manifold B.manifold.

Manifolds.FiberBundle — Type

FiberBundle{𝔽,TVS<:FiberType,TM<:AbstractManifold{𝔽},TVT<:FiberBundleProductVectorTransport} <: AbstractManifold{𝔽}

Fiber bundle on a AbstractManifold M of type FiberType. Examples include vector bundles, principal bundles or unit tangent bundles.

Constructor

FiberBundle(M::AbstractManifold, type::FiberType)

Manifolds.FiberBundleInverseProductRetraction — Type

struct FiberBundleInverseProductRetraction <: AbstractInverseRetractionMethod end

Inverse retraction of the point y at point p from vector bundle B over manifold B.fiber (denoted $\mathcal M$). The inverse retraction is derived as a product manifold-style approximation to the logarithmic map in the Sasaki metric. The considered product manifold is the product between the manifold $\mathcal M$ and the topological vector space isometric to the fiber.

Notation:

The point $p = (x_p, V_p)$ where $x_p ∈ \mathcal M$ and $V_p$ belongs to the fiber $F=π^{-1}(\{x_p\})$ of the vector bundle $B$ where $π$ is the canonical projection of that vector bundle $B$. Similarly, $q = (x_q, V_q)$.

The inverse retraction is calculated as

\[\operatorname{retr}^{-1}_p q = (\operatorname{retr}^{-1}_{x_p}(x_q), V_{\operatorname{retr}^{-1}} - V_p)\]

where $V_{\operatorname{retr}^{-1}}$ is the result of vector transport of $V_q$ to the point $x_p$. The difference $V_{\operatorname{retr}^{-1}} - V_p$ corresponds to the logarithmic map in the vector space $F$.

See also FiberBundleProductRetraction.

Manifolds.FiberBundleProductRetraction — Type

struct FiberBundleProductRetraction <: AbstractRetractionMethod end

Product retraction map of tangent vector $X$ at point $p$ from vector bundle B over manifold B.fiber (denoted $\mathcal M$). The retraction is derived as a product manifold-style approximation to the exponential map in the Sasaki metric. The considered product manifold is the product between the manifold $\mathcal M$ and the topological vector space isometric to the fiber.

Notation:

The point $p = (x_p, V_p)$ where $x_p ∈ \mathcal M$ and $V_p$ belongs to the fiber $F=π^{-1}(\{x_p\})$ of the vector bundle $B$ where $π$ is the canonical projection of that vector bundle $B$.

The tangent vector $X = (V_{X,M}, V_{X,F}) ∈ T_pB$ where $V_{X,M}$ is a tangent vector from the tangent space $T_{x_p}\mathcal M$ and $V_{X,F}$ is a tangent vector from the tangent space $T_{V_p}F$ (isomorphic to $F$).

The retraction is calculated as

\[\operatorname{retr}_p(X) = (\exp_{x_p}(V_{X,M}), V_{\exp})\]

where $V_{\exp}$ is the result of vector transport of $V_p + V_{X,F}$ to the point $\exp_{x_p}(V_{X,M})$. The sum $V_p + V_{X,F}$ corresponds to the exponential map in the vector space $F$.

See also FiberBundleInverseProductRetraction.

Manifolds.FiberBundleProductVectorTransport — Type

FiberBundleProductVectorTransport{
    TMP<:AbstractVectorTransportMethod,
    TMV<:AbstractVectorTransportMethod,
} <: AbstractVectorTransportMethod

Vector transport type on FiberBundle. method_point is used for vector transport of the point part and method_fiber is used for transport of the fiber part.

The vector transport is derived as a product manifold-style vector transport. The considered product manifold is the product between the manifold $\mathcal M$ and the topological vector space isometric to the fiber.

Constructor

FiberBundleProductVectorTransport(
    method_point::AbstractVectorTransportMethod,
    method_fiber::AbstractVectorTransportMethod,
)
FiberBundleProductVectorTransport()

By default both methods are set to ParallelTransport.

Base.getindex — Method

getindex(p::ArrayPartition, M::FiberBundle, s::Symbol)
p[M::FiberBundle, s]

Access the element(s) at index s of a point p on a FiberBundle M by using the symbols :point and :vector or :fiber for the base and vector or fiber component, respectively.

Base.setindex! — Method

setindex!(p::ArrayPartition, val, M::FiberBundle, s::Symbol)
p[M::VectorBundle, s] = val

Set the element(s) at index s of a point p on a FiberBundle M to val by using the symbols :point and :fiber or :vector for the base and fiber or vector component, respectively.

Note

The content of element of p is replaced, not the element itself.

Manifolds.bundle_projection — Method

bundle_projection(B::FiberBundle, p::ArrayPartition)

Projection of point p from the bundle M to the base manifold. Returns the point on the base manifold B.manifold at which the vector part of p is attached.

ManifoldsBase.zero_vector — Method

zero_vector(B::FiberBundle, p)

Zero tangent vector at point p from the fiber bundle B over manifold B.fiber (denoted $\mathcal M$). The zero vector belongs to the space $T_{p}B$

Notation:

The point $p = (x_p, V_p)$ where $x_p ∈ \mathcal M$ and $V_p$ belongs to the fiber $F=π^{-1}(\{x_p\})$ of the vector bundle $B$ where $π$ is the canonical projection of that vector bundle $B$.

The zero vector is calculated as

\[\mathbf{0}_{p} = (\mathbf{0}_{x_p}, \mathbf{0}_F)\]

where $\mathbf{0}_{x_p}$ is the zero tangent vector from $T_{x_p}\mathcal M$ and $\mathbf{0}_F$ is the zero element of the vector space $F$.