Graph manifold

For a given graph $G(V,E)$ implemented using Graphs.jl, the GraphManifold models a PowerManifold either on the nodes or edges of the graph, depending on the GraphManifoldType. i.e., it's either a $\mathcal M^{\lvert V \rvert}$ for the case of a vertex manifold or a $\mathcal M^{\lvert E \rvert}$ for the case of a edge manifold.

Example

To make a graph manifold over $ℝ^2$ with three vertices and two edges, one can use

using Manifoldsusing GraphsM = Euclidean(2)p = [[1., 4.], [2., 5.], [3., 6.]]q = [[4., 5.], [6., 7.], [8., 9.]]x = [[6., 5.], [4., 3.], [2., 8.]]G = SimpleGraph(3)add_edge!(G, 1, 2)add_edge!(G, 2, 3)N = GraphManifold(G, M, VertexManifold())

GraphManifold
Graph:
 {3, 2} undirected simple Int64 graph
AbstractManifold on vertices:
 Euclidean(2; field=ℝ)

It supports all AbstractPowerManifold operations (it is based on NestedPowerRepresentation) and furthermore it is possible to compute a graph logarithm:

incident_log(N, p)

3-element Vector{Vector{Float64}}:
 [1.0, 1.0]
 [0.0, 0.0]
 [-1.0, -1.0]

Types and functions

Manifolds.EdgeManifold — Type

EdgeManifoldManifold <: GraphManifoldType

A type for a GraphManifold where the data is given on the edges.

Manifolds.GraphManifold — Type

GraphManifold{𝔽, G, M, T} <: AbstractPowerManifold{𝔽,M,NestedPowerRepresentation}

Build a manifold, that is a PowerManifold of the AbstractManifold M either on the edges or vertices of a graph G depending on the GraphManifoldType T.

Fields

G is an AbstractSimpleGraph

M is a AbstractManifold

Manifolds.GraphManifoldType — Type

GraphManifoldType

This type represents the type of data on the graph that the GraphManifold represents.

Manifolds.VertexManifold — Type

VectexGraphManifold <: GraphManifoldType

A type for a GraphManifold where the data is given on the vertices.

Manifolds.incident_log — Method

incident_log(M::GraphManifold, x)

Return the tangent vector on the (vertex) GraphManifold, where at each node the sum of the logs to incident nodes is computed. For a SimpleGraph, an edge is interpreted as double edge in the corresponding SimpleDiGraph

If the internal graph is a SimpleWeightedGraph the weighted sum of the tangent vectors is computed.

ManifoldsBase.check_point — Method

check_point(M::GraphManifold, p)

Check whether p is a valid point on the GraphManifold M. This requires that its length equals the number of vertices (for VertexManifolds) or the number of edges (for EdgeManifolds) and that each element of p passes the check_point test for the base manifold M.manifold.

ManifoldsBase.check_vector — Method

check_vector(M::GraphManifold, p, X; kwargs...)

Check whether X is a valid tangent vector at p on the GraphManifold M. This requires that p is a valid point, that the length of X equals the number of vertices (for VertexManifolds) or the number of edges (for EdgeManifolds), and that each element of X together with its corresponding entry of p passes the check_vector test for the base manifold M.manifold.

ManifoldsBase.manifold_dimension — Method

manifold_dimension(N::GraphManifold{G,𝔽,M,EdgeManifold})

returns the manifold dimension of the GraphManifold N on the edges of a graph $G=(V,E)$, i.e.

\[\dim(\mathcal N) = \lvert E \rvert \dim(\mathcal M),\]

where $\mathcal M$ is the manifold of the data on the edges.

ManifoldsBase.manifold_dimension — Method

manifold_dimension(N::GraphManifold{G,𝔽,M,VertexManifold})

returns the manifold dimension of the GraphManifold N on the vertices of a graph $G=(V,E)$, i.e.

\[\dim(\mathcal N) = \lvert V \rvert \dim(\mathcal M),\]

where $\mathcal M$ is the manifold of the data on the nodes.

EdgeManifoldManifold <: GraphManifoldType
A type for a GraphManifold where the data is given on the edges.

GraphManifold{𝔽, G, M, T} <: AbstractPowerManifold{𝔽,M,NestedPowerRepresentation}
Build a manifold, that is a PowerManifold of the AbstractManifold M either on the edges or vertices of a graph G depending on the GraphManifoldType T.

VectexGraphManifold <: GraphManifoldType
A type for a GraphManifold where the data is given on the vertices.

incident_log(M::GraphManifold, x)
Return the tangent vector on the (vertex) GraphManifold, where at each node the sum of the logs to incident nodes is computed.

manifold_dimension(N::GraphManifold{G,𝔽,M,EdgeManifold})
returns the manifold dimension of the GraphManifold N on the edges of a graph , i.e.

manifold_dimension(N::GraphManifold{G,𝔽,M,VertexManifold})
returns the manifold dimension of the GraphManifold N on the vertices of a graph , i.e.

manifold_dimension(M::CenteredMatrices)
Return the manifold dimension of the CenteredMatrices m-by-n matrix M over the number system 𝔽, i.e.

manifold_dimension(M::CholeskySpace)
Return the manifold dimension for the CholeskySpace M, i.e.

manifold_dimension(M::Circle)
Return the dimension of the Circle M, i.e. .

manifold_dimension(M::DeterminantOneMatrices{n,𝔽})
Return the dimension of the DeterminantOneMatrices matrix M over the number system 𝔽, which is one dimension less than its embedding, the Euclidean(n, n; field=𝔽).

manifold_dimension(M::Elliptope)
returns the dimension of Elliptope M , i.e.

manifold_dimension(M::EssentialManifold{is_signed, ℝ})
Return the manifold dimension of the EssentialManifold, which is 5[TD17].

Euclidean{T,𝔽} <: AbstractManifold{𝔽}
Euclidean vector space.

manifold_dimension(M::Euclidean)
Return the manifold dimension of the Euclidean M, i.e. the product of all array dimensions and the real_dimension of the underlying number system.

manifold_dimension(M::FixedRankMatrices)
Return the manifold dimension for the 𝔽-valued FixedRankMatrices M of dimension mxn of rank k, namely

manifold_dimension(M::Flag)
Return dimension of flag manifold .

manifold_dimension(M::GeneralizedGrassmann)
Return the dimension of the GeneralizedGrassmann(n,k,𝔽) manifold M, i.e.

manifold_dimension(M::GeneralizedStiefel)
Return the dimension of the GeneralizedStiefel manifold M=.

manifold_dimension(M::Rotations) manifold_dimension(M::OrthogonalMatrices)
Return the dimension of the manifold orthogonal matrices and of the manifold of rotations

manifold_dimension(M::SpecialUnitaryMatrices)
Return the dimension of the manifold of special unitary matrices.

manifold_dimension(M::UnitaryMatrices{n,ℂ}) where {n}
Return the dimension of the manifold unitary matrices.

manifold_dimension(M::UnitaryMatrices{<:Any,ℍ})
Return the dimension of the manifold unitary matrices.

manifold_dimension(M::Grassmann)
Return the dimension of the Grassmann(n,k,𝔽) manifold M, i.e.

manifold_dimension(M::HeisenbergMatrices)
Return the dimension of HeisenbergMatrices(n), which is equal to .

manifold_dimension(M::Hyperbolic)
Return the dimension of the hyperbolic space manifold , i.e. .

manifold_dimension(M::Hyperrectangle)
Return the manifold dimension of the Hyperrectangle M, i.e. the product of all array dimensions.

manifold_dimension(M::InvertibleMatrices{n,𝔽})
Return the dimension of the InvertibleMatrices matrix M over the number system 𝔽, which is the same dimension as its embedding, the Euclidean(n, n; field=𝔽).

manifold_dimension(M::MultinomialDoubleStochastic)
returns the dimension of the MultinomialDoubleStochastic manifold namely

manifold_dimension(M::MultinomialSymmetric)
returns the dimension of the MultinomialSymmetric manifold namely

manifold_dimension(M::ParametricSurface)
Return the dimension of the parameter space of M.

manifold_dimension(M::PositiveNumbers)
Return the dimension of the PositiveNumbers M, i.e. of the 1-dimensional hyperbolic space,

manifold_dimension(M::ProbabilitySimplex)
Returns the manifold dimension of the probability simplex in , i.e.

manifold_dimension(M::AbstractProjectiveSpace{𝔽}) where {𝔽}
Return the real dimension of the AbstractProjectiveSpace M, respectively i.e. the real dimension of the embedding minus the real dimension of the field 𝔽.

manifold_dimension(M::KendallsPreShapeSpace)
Return the dimension of the KendallsPreShapeSpace manifold M.

manifold_dimension(M::KendallsShapeSpace)
Return the dimension of the KendallsShapeSpace manifold M.

manifold_dimension(M::SkewHermitianMatrices)
Return the dimension of the SkewHermitianMatrices matrix M over the number system 𝔽, i.e.

manifold_dimension(M::Spectrahedron)
returns the dimension of Spectrahedron M, i.e.

manifold_dimension(M::AbstractSphere)
Return the dimension of the AbstractSphere M, respectively i.e. the dimension of the embedding -1.

manifold_dimension(M::SphereSymmetricMatrices{<:Any,𝔽})
Return the manifold dimension of the SphereSymmetricMatrices n-by-n symmetric matrix M of unit Frobenius norm over the number system 𝔽, i.e.

manifold_dimension(M::Stiefel)
Return the dimension of the Stiefel manifold M=.

manifold_dimension(M::SymmetricMatrices{n,𝔽})
Return the dimension of the SymmetricMatrices matrix M over the number system 𝔽, i.e.

manifold_dimension(M::SymmetricPositiveDefinite)
returns the dimension of SymmetricPositiveDefinite M , i.e.

manifold_dimension(M::SymmetricPositiveSemidefiniteFixedRank)
Return the dimension of the SymmetricPositiveSemidefiniteFixedRank matrix M over the number system 𝔽, i.e.

manifold_dimension(::SymplecticMatrices)
Returns the dimension of the symplectic manifold embedded in , i.e.

manifold_dimension(::SymplecticGrassmann)
Return the dimension of the SymplecticGrassmann(2n,2k), which is

manifold_dimension(::SymplecticStiefel)
Returns the dimension of the symplectic Stiefel manifold embedded in , i.e. [BZ21]

manifold_dimension(M::EmbeddedTorus)
Return the dimension of the EmbeddedTorus M that is 2.

manifold_dimension(::Tucker)
The dimension of the manifold of tensors of multilinear rank , i.e.