Grassmann Manifold

Manifolds.Grassmann — Type

Grassmann{n,k,T} <: Manifold

The Grassmann manifold $\operatorname{Gr}(n,k)$ consists of all subspaces spanned by $k$ linear independent vectors $\mathbb F^n$, where $\mathbb F \in \{\mathbb R, \mathbb C\}$ is either the real- (or complex-) valued vectors. This yields all $k$-dimensional subspaces of $\mathbb R^n$ for the real-valued case and all $2k$-dimensional subspaces of $\mathbb C^n$ for the second.

The manifold can be represented as

\[\operatorname{Gr}(n,k) \coloneqq \bigl\{ \operatorname{span}(x) : x \in \mathbb F^{n\times k}, \bar{x}^\mathrm{T}x = I_k\},\]

where ${\bar\cdot}^{\mathrm{T}}$ denotes the complex conjugate transpose and $I_k$ is the $k\times k$ identity matrix. This means, that the columns of $x$ form an orthonormal basis of the subspace, that is a point on $\operatorname{Gr}(n,k)$, and hence the subspace can actually be represented by a whole equivalence class of representers. Another interpretation is, that

\[\operatorname{Gr}(n,k) = \operatorname{St}(n,k) / \operatorname{O}(k),\]

i.e the Grassmann manifold is the quotient of the Stiefel manifold and the orthogonal group $\operatorname{O}(k)$ of orthogonal $k\times k$ matrices.

The tangent space at a point (subspace) $x$ is given by

\[T_x\mathrm{Gr}(n,k) = \bigl\{ v \in \mathbb{F}^{n\times k} : {\bar v}^{\mathrm{T}}x + {\bar x}^{\mathrm{T}}v = 0_{k} \bigr\},\]

where $0_{k}$ denotes the $k\times k$ zero matrix.

Note that a point $x\in \operatorname{Gr}(n,k)$ might be represented by different matrices (i.e. matrices with orthonormal column vectors that span the same subspace). Different representations of $x$ also lead to different representation matrices for the tangent space $T_x\mathrm{Gr}(n,k)$

The manifold is named after Hermann G. Graßmann (1809-1877).

Constructor

Grassmann(n,k,T=Real)

generate the Grassmann manifold $\operatorname{Gr}(n,k)$, where the real-valued case $\mathbb F = \mathbb R$ is the default.

ManifoldsBase.distance — Method

distance(M,x,y)

computes the Riemannian distance on Grassmann manifold M$= \mathrm{Gr}(n,k)$.

Let $USV = {\bar x}^\mathrm{T}y$ denote the SVD decomposition of ${\bar x}^\mathrm{T}y$, where ${\bar\cdot}^{\mathrm{T}}$ denotes the complex conjugate transposed. Then the distance is given by

\[d_{\mathrm{GR}(n,k)}(x,y) = \operatorname{norm}(\operatorname{Re}(b)).\]

where

\[b_{i}=\begin{cases} 0 & \text{if} \; S_i \geq 1\\ \operatorname{acos}(S_i) & \, \text{if} \; S_i<1 \end{cases}.\]

ManifoldsBase.exp! — Method

exp!(M, y, x, v)

compute the exponential map on the Grassmann manifold M$= \mathrm{Gr}(n,k)$ starting in x with tangent vector (direction) v and store the result in y. Let $v = USV$ denote the SVD decomposition of $v$. Then the exponential map is written using

\[z = x V\cos(S){\bar V}^\mathrm{T} + U\sin(S){\bar V}^\mathrm{T},\]

where ${\bar\cdot}^{\mathrm{T}}$ denotes the complex conjugate transposed. The cosine and sine are applied element wise to the diagonal entries of $S$. A final QR decomposition $z=QR$ is performed for numerical stability reasons, yielding the result as

\[y = Q.\]

ManifoldsBase.inner — Method

inner(M,x,v,w)

compute the inner product for two tangent vectors v, w from the tangent space of x on the Grassmann manifold M. The formula reads

\[(v,w)_x = \operatorname{trace}({\bar v}^{\mathrm{T}}w),\]

where ${\bar\cdot}^{\mathrm{T}}$ denotes the complex conjugate transposed.

ManifoldsBase.inverse_retract! — Method

inverse_retract!(M, v, x, y, ::PolarInverseRetraction)

compute the inverse retraction for the PolarRetraction

\[\operatorname{retr}_x^{-1}y = y*(\bar{x}^\mathrm{T}y)^{-1} - x,\]

where ${\bar\cdot}^{\mathrm{T}}$ denotes the complex conjugate transposed.

ManifoldsBase.inverse_retract! — Method

inverse_retract!(M, v, x, y, ::QRInverseRetraction)

compute the inverse retraction valid of the QRRetraction

\[\operatorname{retr}_x^{-1}y = y*(\bar{x}^\mathrm{T}y)^{-1} - x,\]

where ${\bar\cdot}^{\mathrm{T}}$ denotes the complex conjugate transposed.

ManifoldsBase.log! — Method

log!(M, v, x, y)

compute the logarithmic map on the Grassmann manifold $\mathcal M=\mathrm{Gr}(n,k)$, i.e. the tangent vector v whose corresponding geodesic starting from x reaches y after time 1 on M. The formula reads

\[v = V\cdot \operatorname{atan}(S) \cdot {\bar U}^\mathrm{T},\]

where ${\bar\cdot}^{\mathrm{T}}$ denotes the complex conjugate transposed, $U$ and $V$ are the unitary matrices, and $S$ is a diagonal matrix containing the singular values of the SVD-decomposition of

\[USV = ({\bar y}^\mathrm{T}x)^{-1} ( {\bar y}^\mathrm{T} - {\bar y}^\mathrm{T}x{\bar x}^\mathrm{T} ).\]

In this formula the $\operatorname{atan}$ is meant elementwise.

ManifoldsBase.manifold_dimension — Method

manifold_dimension(M)

return the dimension of the complex-valued Grassmann(n,k) manifold M, i.e.

\[2k(n-k)\]

ManifoldsBase.manifold_dimension — Method

manifold_dimension(M)

return the dimension of the real-valued Grassmann(n,k) manifold M, i.e.

\[k(n-k)\]

ManifoldsBase.retract! — Method

retract!(M, y, x, v, ::PolarRetraction)

compute the SVD-based retraction PolarRetraction on the Grassmann manifold M. With $USV = x + v$ the retraction reads

\[y = \operatorname{retr}_x v = U\bar{V}^\mathrm{T},\]

where ${\bar\cdot}^{\mathrm{T}}$ denotes the complex conjugate transposed.

ManifoldsBase.retract! — Method

retract!(M, y, x, v, ::QRRetraction )

compute the QR-based retraction QRRetraction on the Grassmann manifold M. With $QR = x + v$ the retraction reads

\[y = \operatorname{retr}_xv = QD,\]

where D is a $m\times n$ matrix with

\[D = \operatorname{diag}( \operatorname{sgn}(R_{ii}+0,5)_{i=1}^n ).\]