Group manifold

Lie groups, groups that are Manifolds with a smooth binary group operation AbstractGroupOperation, are implemented as subtypes of AbstractGroupManifold or by decorating an existing manifold with a group operation using GroupManifold.

The common addition and multiplication group operations of AdditionOperation and MultiplicationOperation are provided, though their behavior may be customized for a specific group.

Manifolds.AbstractGroupManifold — Type

AbstractGroupManifold{<:AbstractGroupOperation} <: Manifold

Abstract type for a Lie group, a group that is also a smooth manifold with a smooth binary operation. AbstractGroupManifolds must implement at least inv, group_id, compose_left, and translate_diff. Group manifolds by default assume a left-invariant canonical metric.

Manifolds.AbstractGroupOperation — Type

AbstractGroupOperation

Abstract type for smooth binary operations on elements of a Lie group.

Manifolds.ActionDirection — Type

ActionDirection

Direction of action on a manifold, either LeftAction or RightAction.

Manifolds.Identity — Type

Identity(group::AbstractGroupManifold)

The identity element of the group group.

Manifolds.LeftAction — Type

LeftAction()

Left action of a group on a manifold.

Manifolds.LeftInvariantCanonicalMetric — Type

LeftInvariantCanonicalMetric()

The left-invariant canonical metric on a Lie group.

Manifolds.RightAction — Type

RightAction()

Right action of a group on a manifold.

Manifolds.RightInvariantCanonicalMetric — Type

RightInvariantCanonicalMetric()

The right-invariant canonical metric on a Lie group.

Manifolds.compose_left — Method

compose_left(G::AbstractGroupManifold, xs...)

Compose elements of G using their left translation upon each other. This composition is left-associative.

Manifolds.group_id — Method

group_id(G::AbstractGroupManifold, x)

Identity element $e$, such that for any element $x$, $x \cdot e = e \cdot x = x$. The returned element is of a similar type to x.

Manifolds.inv! — Method

inv(G::AbstractGroupManifold, x)

Inverse $x^{-1}$ of an element $x$, such that $x \cdot x^{-1} = x^{-1} \cdot x = e$.

Manifolds.inverse_translate — Method

inverse_translate(G::AbstractGroupManifold, x, y, [conv::ActionDirection=Left()])

For group elements $x,y \in G$, inverse translate $y$ by $x$ with the specified convention, either left $L_x^{-1}$ or right $R_x^{-1}$, defined as

\[\begin{aligned} L_x^{-1} &\colon y \mapsto x^{-1} \cdot y\\ R_x^{-1} &\colon y \mapsto y \cdot x^{-1}. \end{aligned}\]

Manifolds.translate — Method

translate(G::AbstractGroupManifold, x, y[, conv::ActionDirection=LeftAction()])

For group elements $x,y \in G$, translate $y$ by $x$ with the specified convention, either left $L_x$ or right $R_x$, defined as

\[\begin{aligned} L_x &\colon y \mapsto x \cdot y\\ R_x &\colon y \mapsto y \cdot x. \end{aligned}\]

Manifolds.AdditionOperation — Type

AdditionOperation <: AbstractGroupOperation

Group operation that consists of simple addition.

Manifolds.GroupManifold — Type

GroupManifold{M<:Manifold,O<:AbstractGroupOperation} <: AbstractGroupManifold{O}

Decorator for a smooth manifold that equips the manifold with a group operation, thus making it a Lie group. See AbstractGroupManifold for more details.

Constructor

GroupManifold(manifold, op)

Manifolds.MultiplicationOperation — Type

MultiplicationOperation <: AbstractGroupOperation

Group operation that consists of multiplication.

Special orthogonal group

Manifolds.SpecialOrthogonal — Type

SpecialOrthogonal{N} <: GroupManifold{Rotations{N},MultiplicationOperation}

Special orthogonal group \mathrm{SO}(N) represented by rotation matrices.

Constructor

SpecialOrthogonal(n)

Group actions

Manifolds.AbstractActionOnManifold — Type

AbstractActionOnManifold

An abstract group action on a manifold.

Manifolds.action_on — Method

action_on(A::AbstractActionOnManifold)

The manifold the action A acts upon.

Manifolds.apply_action! — Method

apply_action(A::ActionOnManifold, x, a)

Apply action a of type A to the point x. The result is saved in y.

Manifolds.apply_action — Method

apply_action(A::AbstractActionOnManifold, x, a)

Apply action a of type A to the point x.

Manifolds.base_group — Method

base_group(A::AbstractActionOnManifold)

The group that acts in action A.

Manifolds.center_of_orbit — Function

center_of_orbit(
    A::AbstractActionOnManifold,
    pts,
    q
    [, mean_method::AbstractEstimationMethod = GradientDescentEstimation()]
)

Calculate an action element of action A that constitutes mean element of orbit of q with respect to given set of points pts. The mean is calculated using the method mean_method.

Manifolds.optimal_alignment! — Method

optimal_alignment!(A::AbstractActionOnManifold, y, x1, x2)

Calculate an action element of action A that, when acted upon x1, gets the element closest to x2. The result is written to y.

Manifolds.optimal_alignment — Method

optimal_alignment(A::AbstractActionOnManifold, x1, x2)

Calculate an action element of action A that, when acted upon x1, gets the element closest to x2.

Rotation action

Manifolds.RotationAction — Type

RotationAction(M::Manifold, SOn::SpecialOrthogonal)

Space of actions of the SpectialOrthogonal group \mathrm{SO}(N) on Euclidean-like manifold M of dimensions N.