Group manifold

Lie groups, groups that are Manifolds with a smooth binary group operation AbstractGroupOperation, are implemented as subtypes of AbstractGroupManifold or by decorating an existing manifold with a group operation using GroupManifold.

The common addition and multiplication group operations of AdditionOperation and MultiplicationOperation are provided, though their behavior may be customized for a specific group.

Manifolds.AbstractGroupManifold — Type

AbstractGroupManifold{<:AbstractGroupOperation} <: Manifold

Abstract type for a Lie group, a group that is also a smooth manifold with a smooth binary operation. AbstractGroupManifolds must implement at least inv, identity, compose, and translate_diff. Group manifolds by default forward metric-related operations to the wrapped manifold.

Manifolds.AbstractGroupOperation — Type

AbstractGroupOperation

Abstract type for smooth binary operations $\circ$ on elements of a Lie group $G$.

\[\circ \colon G \times G \to G.\]

An operation can be either defined for a specific AbstractGroupManifold or in general, by defining for an operation Op the following methods:

identity!(::AbstractGroupManifold{Op}, y, x)
identity(::AbstractGroupManifold{Op}, x)
inv!(::AbstractGroupManifold{Op}, y, x)
inv(::AbstractGroupManifold{Op}, x)
compose(::AbstractGroupManifold{Op}, x, y)
compose!(::AbstractGroupManifold{Op}, z, x, y)

Note that a manifold is connected with an operation by wrapping it with a decorator, AbstractGroupManifold. In typical cases the concrete wrapper GroupManifold can be used.

Manifolds.ActionDirection — Type

ActionDirection

Direction of action on a manifold, either LeftAction or RightAction.

Manifolds.GroupManifold — Type

GroupManifold{M<:Manifold,O<:AbstractGroupOperation} <: AbstractGroupManifold{O}

Decorator for a smooth manifold that equips the manifold with a group operation, thus making it a Lie group. See AbstractGroupManifold for more details.

Constructor

GroupManifold(manifold, op)

Manifolds.Identity — Type

Identity(group::AbstractGroupManifold)

The identity element of the group group.

Manifolds.LeftAction — Type

LeftAction()

Left action of a group on a manifold.

Manifolds.RightAction — Type

RightAction()

Right action of a group on a manifold.

Base.identity — Method

identity(G::AbstractGroupManifold, x)

Identity element $e$, such that for any element $x$, $x \circ e = e \circ x = x$. The returned element is of a similar type to x.

Base.inv — Method

inv(G::AbstractGroupManifold, x)

Inverse $x^{-1}$ of an element $x$, such that $x \circ x^{-1} = x^{-1} \circ x = e$.

Manifolds.compose! — Method

compose!(G::AbstractGroupManifold, z, x, y)

Compose elements x and y of G using their left translation upon each other. The result is saved in z.

z may alias with one or both of x and y.

Manifolds.inv! — Method

inv!(G::AbstractGroupManifold, y, x)

Inverse $x^{-1}$ of an element $x$, such that $x \circ x^{-1} = x^{-1} \circ x = e$. The result is saved to y.

Manifolds.inverse_translate! — Method

inverse_translate!(G::AbstractGroupManifold, z, x, y, [conv::ActionDirection=Left()])

For group elements $x,y \in G$, inverse translate $y$ by $x$ with the specified convention, either left $L_x^{-1}$ or right $R_x^{-1}$, defined as

\[\begin{aligned} L_x^{-1} &\colon y \mapsto x^{-1} \circ y\\ R_x^{-1} &\colon y \mapsto y \circ x^{-1}. \end{aligned}\]

Result is saved in z.

Manifolds.inverse_translate — Method

inverse_translate(G::AbstractGroupManifold, x, y, [conv::ActionDirection=Left()])

For group elements $x,y \in G$, inverse translate $y$ by $x$ with the specified convention, either left $L_x^{-1}$ or right $R_x^{-1}$, defined as

\[\begin{aligned} L_x^{-1} &\colon y \mapsto x^{-1} \circ y\\ R_x^{-1} &\colon y \mapsto y \circ x^{-1}. \end{aligned}\]

Manifolds.inverse_translate_diff — Function

inverse_translate_diff(G::AbstractGroupManifold, x, y, v[, conv::ActionDirection=Left()])

For group elements $x,y \in G$ and tangent vector $v \in T_y G$, compute the inverse of the action of the differential of the translation by $x$ on $v$, written as $((\mathrm{d}\tau_x)_y)^{-1} (v) = (\mathrm{d}\tau_{x^{-1}})_y (v)$, with the specified left or right convention. The differential transports vectors:

\[\begin{aligned} ((\mathrm{d}L_x)_y)^{-1} (v) &\colon T_y G \to T_{x^{-1} \circ y} G\\ ((\mathrm{d}R_x)_y)^{-1} (v) &\colon T_y G \to T_{y \circ x^{-1}} G\\ \end{aligned}\]

Manifolds.switch_direction — Method

switch_direction(::ActionDirection)

Returns a RightAction when given a LeftAction and vice versa.

Manifolds.translate! — Method

translate!(G::AbstractGroupManifold, z, x, y[, conv::ActionDirection=LeftAction()])

For group elements $x,y \in G$, translate $y$ by $x$ with the specified convention, either left $L_x$ or right $R_x$, defined as

\[\begin{aligned} L_x &\colon y \mapsto x \circ y\\ R_x &\colon y \mapsto y \circ x. \end{aligned}\]

Result of the operation is saved in z.

Manifolds.translate — Method

translate(G::AbstractGroupManifold, x, y[, conv::ActionDirection=LeftAction()])

For group elements $x,y \in G$, translate $y$ by $x$ with the specified convention, either left $L_x$ or right $R_x$, defined as

\[\begin{aligned} L_x &\colon y \mapsto x \circ y\\ R_x &\colon y \mapsto y \circ x. \end{aligned}\]

Manifolds.translate_diff — Function

translate_diff(G::AbstractGroupManifold, x, y, v[, conv::ActionDirection=LeftAction()])

For group elements $x,y \in G$ and tangent vector $v \in T_y G$, compute the action of the differential of the translation by $x$ on $v$, written as $(\mathrm{d}\tau_x)_y (v)$, with the specified left or right convention. The differential transports vectors:

\[\begin{aligned} (\mathrm{d}L_x)_y (v) &\colon T_y G \to T_{x \circ y} G\\ (\mathrm{d}R_x)_y (v) &\colon T_y G \to T_{y \circ x} G\\ \end{aligned}\]

Manifolds.AdditionOperation — Type

AdditionOperation <: AbstractGroupOperation

Group operation that consists of simple addition.

Manifolds.MultiplicationOperation — Type

MultiplicationOperation <: AbstractGroupOperation

Group operation that consists of multiplication.

Special orthogonal group

Manifolds.SpecialOrthogonal — Type

SpecialOrthogonal{N} <: GroupManifold{Rotations{N},MultiplicationOperation}

Special orthogonal group \mathrm{SO}(N) represented by rotation matrices.

Constructor

SpecialOrthogonal(n)

Translation group

Manifolds.TranslationGroup — Type

TranslationGroup{N} <: GroupManifold{Euclidean{N},AdditionOperation}

Translation group $\mathrm{T}(N)$ represented by translation arrays.

Constructor

TranslationGroup(n)

Group actions

Manifolds.AbstractGroupAction — Type

AbstractGroupAction

An abstract group action on a manifold.

Manifolds.apply! — Method

apply(A::AbstractGroupAction, x, a)

Apply action a to the point x with the rule specified by A. The result is saved in y.

Manifolds.apply — Method

apply(A::AbstractGroupAction, x, a)

Apply action a to the point x. The action is specified by A.

Manifolds.base_group — Method

base_group(A::AbstractGroupAction)

The group that acts in action A.

Manifolds.center_of_orbit — Function

center_of_orbit(
    A::AbstractGroupAction,
    pts,
    q,
    mean_method::AbstractEstimationMethod = GradientDescentEstimation()
)

Calculate an action element of action A that constitutes mean element of orbit of q with respect to given set of points pts. The mean is calculated using the method mean_method.

The orbit of q with respect to action of a group G is the set $O = \{ g\cdot q \colon g \in G \}$. The function is useful for computing means on quotients of manifolds by a Lie group action.

Manifolds.g_manifold — Method

g_manifold(A::AbstractGroupAction)

The manifold the action A acts upon.

Manifolds.optimal_alignment! — Method

optimal_alignment!(A::AbstractGroupAction, y, x1, x2)

Calculate an action element of action A that acts upon x1 to produce the element closest to x2. The result is written to y.

Manifolds.optimal_alignment — Method

optimal_alignment(A::AbstractGroupAction, x1, x2)

Calculate an action element of action A that acts upon x1 to produce the element closest to x2 in the metric of the G-manifold:

\[\arg\min_{g \in G} d_M(g \cdot x_1, x_2)\]

where $G$ is the group that acts on the G-manifold $M$.

Rotation action

Manifolds.RotationAction — Type

RotationAction(
    M::Manifold,
    SOn::SpecialOrthogonal,
    AD::ActionDirection = LeftAction(),
)

Space of actions of the SpecialOrthogonal group $\mathrm{SO}(N)$ on a Euclidean-like manifold M of dimension N.

Translation action

Manifolds.TranslationAction — Type

TranslationAction(
    M::Manifold,
    Rn::TranslationGroup,
    AD::ActionDirection = LeftAction(),
)

Space of left actions of the TranslationGroup $\mathrm{T}(N)$ on a Euclidean-like manifold M.