Group manifolds and actions

Lie groups, groups that are Manifolds with a smooth binary group operation AbstractGroupOperation, are implemented as subtypes of AbstractGroupManifold or by decorating an existing manifold with a group operation using GroupManifold.

The common addition and multiplication group operations of AdditionOperation and MultiplicationOperation are provided, though their behavior may be customized for a specific group.

Contents

Group manifolds and actions

Groups

Group manifold

Product group

Semidirect product group

Circle group

Special orthogonal group

Translation group

Special Euclidean group

Group actions

Group operation action

Rotation action

Translation action

Groups

Group manifold

Manifolds.AbstractGroupManifold — Type

AbstractGroupManifold{<:AbstractGroupOperation} <: Manifold

Abstract type for a Lie group, a group that is also a smooth manifold with an AbstractGroupOperation, a smooth binary operation. AbstractGroupManifolds must implement at least inv, identity, compose, and translate_diff.

Manifolds.AbstractGroupOperation — Type

AbstractGroupOperation

Abstract type for smooth binary operations $\circ$ on elements of a Lie group $G$:

\[\circ : G × G → G\]

An operation can be either defined for a specific AbstractGroupManifold or in general, by defining for an operation Op the following methods:

identity!(::AbstractGroupManifold{Op}, q, q)
identity(::AbstractGroupManifold{Op}, p)
inv!(::AbstractGroupManifold{Op}, q, p)
inv(::AbstractGroupManifold{Op}, p)
compose(::AbstractGroupManifold{Op}, p, q)
compose!(::AbstractGroupManifold{Op}, x, p, q)

Note that a manifold is connected with an operation by wrapping it with a decorator, AbstractGroupManifold. In typical cases the concrete wrapper GroupManifold can be used.

Manifolds.ActionDirection — Type

ActionDirection

Direction of action on a manifold, either LeftAction or RightAction.

Manifolds.AdditionOperation — Type

AdditionOperation <: AbstractGroupOperation

Group operation that consists of simple addition.

Manifolds.GroupManifold — Type

GroupManifold{M<:Manifold,O<:AbstractGroupOperation} <: AbstractGroupManifold{O}

Decorator for a smooth manifold that equips the manifold with a group operation, thus making it a Lie group. See AbstractGroupManifold for more details.

Group manifolds by default forward metric-related operations to the wrapped manifold.

Constructor

GroupManifold(manifold, op)

Manifolds.Identity — Type

Identity(G::AbstractGroupManifold)

The group identity element $e ∈ G$.

Manifolds.LeftAction — Type

LeftAction()

Left action of a group on a manifold.

Manifolds.MultiplicationOperation — Type

MultiplicationOperation <: AbstractGroupOperation

Group operation that consists of multiplication.

Manifolds.RightAction — Type

RightAction()

Right action of a group on a manifold.

Base.identity — Method

identity(G::AbstractGroupManifold, p)

Identity element $e ∈ G$, such that for any element $p ∈ G$, $p \circ e = e \circ p = p$. The returned element is of a similar type to p.

Base.inv — Method

inv(G::AbstractGroupManifold, p)

Inverse $p^{-1} ∈ G$ of an element $p ∈ G$, such that $p \circ p^{-1} = p^{-1} \circ p = e ∈ G$.

Manifolds.base_group — Method

base_group(M::Manifold) -> AbstractGroupManifold

Undecorate M until an AbstractGroupManifold is encountered. Return an error if the base_manifold is reached without encountering a group.

Manifolds.compose — Method

compose(G::AbstractGroupManifold, x, y)

Compose elements $x,y ∈ G$ using the group operation $x \circ y$.

Manifolds.inverse_translate — Method

inverse_translate(G::AbstractGroupManifold, p, q)
inverse_translate(G::AbstractGroupManifold, p, q, conv::ActionDirection=Left())

For group elements $p, q ∈ G$, inverse translate $q$ by $p$ with the specified convention, either left $L_p^{-1}$ or right $R_p^{-1}$, defined as

\[\begin{aligned} L_p^{-1} &: q ↦ p^{-1} \circ q\\ R_p^{-1} &: q ↦ q \circ p^{-1}. \end{aligned}\]

Manifolds.inverse_translate_diff — Method

inverse_translate_diff(G::AbstractGroupManifold, p, q, X)
inverse_translate_diff(G::AbstractGroupManifold, p, q, X, conv::ActionDirection=Left())

For group elements $p, q ∈ G$ and tangent vector $X ∈ T_q G$, compute the inverse of the action of the differential of the translation by $p$ on $X$, written as $((\mathrm{d}τ_p)_q)^{-1} (X) = (\mathrm{d}τ_{p^{-1}})_q (X)$, with the specified left or right convention. The differential transports vectors as

\[\begin{aligned} ((\mathrm{d}L_p)_q)^{-1} (X) &: T_q G → T_{p^{-1} \circ q} G\\ ((\mathrm{d}R_p)_q)^{-1} (X) &: T_q G → T_{q \circ p^{-1}} G\\ \end{aligned}\]

Manifolds.switch_direction — Method

switch_direction(::ActionDirection)

Returns a RightAction when given a LeftAction and vice versa.

Manifolds.translate — Method

translate(G::AbstractGroupManifold, p, q)
translate(G::AbstractGroupManifold, p, q, conv::ActionDirection=LeftAction()])

For group elements $p,q ∈ G$, translate $q$ by $p$ with the specified convention, either left $L_p$ or right $R_q$, defined as

\[\begin{aligned} L_p &: q ↦ p \circ q\\ R_p &: q ↦ q \circ p. \end{aligned}\]

Manifolds.translate_diff — Method

translate_diff(G::AbstractGroupManifold, p, q, X)
translate_diff(G::AbstractGroupManifold, p, q, X, conv::ActionDirection=LeftAction())

For group elements $p, q ∈ G$ and tangent vector $X ∈ T_q G$, compute the action of the differential of the translation by $p$ on $X$, written as $(\mathrm{d}τ_p)_q (X)$, with the specified left or right convention. The differential transports vectors:

\[\begin{aligned} (\mathrm{d}L_p)_q (X) &: T_q G → T_{p \circ q} G\\ (\mathrm{d}R_p)_q (X) &: T_q G → T_{q \circ p} G\\ \end{aligned}\]

Product group

Manifolds.ProductGroup — Method

ProductGroup{T} <: GroupManifold{ProductManifold{T},ProductOperation}

Decorate a product manifold with a ProductOperation.

Each submanifold must also be an AbstractGroupManifold or a decorated instance of one. This type is mostly useful for equipping the direct product of group manifolds with an Identity element.

Constructor

ProductGroup(manifold::ProductManifold)

Manifolds.ProductOperation — Type

ProductOperation <: AbstractGroupOperation

Direct product group operation.

Semidirect product group

Manifolds.SemidirectProductGroup — Method

SemidirectProductGroup(N::GroupManifold, H::GroupManifold, A::AbstractGroupAction)

A group that is the semidirect product of a normal group $N$ and a subgroup $H$, written $G = N ⋊_θ H$, where $θ: H × N → N$ is an automorphism action of $H$ on $N$. The group $G$ has the composition rule

\[g \circ g' = (n, h) \circ (n', h') = (n \circ θ_h(n'), h \circ h')\]

and the inverse

\[g^{-1} = (n, h)^{-1} = (θ_{h^{-1}}(n^{-1}), h^{-1}).\]

Manifolds.SemidirectProductOperation — Type

SemidirectProductOperation(action::AbstractGroupAction)

Group operation of a semidirect product group. The operation consists of the operation opN on a normal subgroup N, the operation opH on a subgroup H, and an automorphism action of elements of H on N. Only the action is stored.

Circle group

Manifolds.CircleGroup — Type

CircleGroup <: GroupManifold{Circle{ℂ},MultiplicationOperation}

The circle group is the complex circle (Circle(ℂ)) equipped with the group operation of complex multiplication (MultiplicationOperation).

Special orthogonal group

Manifolds.SpecialOrthogonal — Type

SpecialOrthogonal{n} <: GroupManifold{Rotations{n},MultiplicationOperation}

Special orthogonal group $\mathrm{SO}(n)$ represented by rotation matrices.

Constructor

SpecialOrthogonal(n)

Translation group

Manifolds.TranslationGroup — Type

TranslationGroup{T<:Tuple,F} <: GroupManifold{Euclidean{T,F},AdditionOperation}

Translation group $\mathrm{T}(n)$ represented by translation arrays.

Constructor

TranslationGroup(n₁,...,nᵢ; field = 𝔽)

Generate the translation group on $𝔽^{n₁,…,nᵢ}$ = Euclidean(n₁,...,nᵢ; field = 𝔽), which is isomorphic to group itself.

Special Euclidean group

Manifolds.SpecialEuclidean — Type

SpecialEuclidean(n)

Special Euclidean group $\mathrm{SE}(n)$, the group of rigid motions.

$\mathrm{SE}(n)$ is the semidirect product of the TranslationGroup on $ℝ^n$ and SpecialOrthogonal(n)

\[\mathrm{SE}(n) ≐ \mathrm{T}(n) ⋊_θ \mathrm{SO}(n),\]

where $θ$ is the canonical action of $\mathrm{SO}(n)$ on $\mathrm{T}(n)$ by vector rotation.

This constructor is equivalent to calling

Tn = TranslationGroup(n)
SOn = SpecialOrthogonal(n)
SemidirectProductGroup(Tn, SOn, RotationAction(Tn, SOn))

Points on $\mathrm{SE}(n)$ may be represented as points on the underlying product manifold $\mathrm{T}(n) × \mathrm{SO}(n)$ or as affine matrices with size (n + 1, n + 1).

Group actions

Manifolds.AbstractGroupAction — Type

AbstractGroupAction

An abstract group action on a manifold.

Manifolds.apply! — Method

apply!(A::AbstractGroupAction, q, a, p)

Apply action a to the point p with the rule specified by A. The result is saved in q.

Manifolds.apply — Method

apply(A::AbstractGroupAction, a, p)

Apply action a to the point p. The action is specified by A. Unless otherwise specified, right actions are defined in terms of the left action. For point $p ∈ \mathcal M$ and action element $a$, the right action is

\[p ⋅ a ≐ a^{-1} ⋅ p.\]

Manifolds.apply_diff — Method

apply_diff(A::AbstractGroupAction, a, p, X)

For group point $p ∈ \mathcal M$ and tangent vector $X ∈ T_p \mathcal M$, compute the action of the differential of the action of $a ∈ G$ on $X$, specified by rule A. Written as $(\mathrm{d}τ_a)_p (X)$, with the specified left or right convention, the differential transports vectors

\[\begin{aligned} (\mathrm{d}L_a)_p (X) &: T_p \mathcal M → T_{a ⋅ p} \mathcal M\\ (\mathrm{d}R_a)_p (X) &: T_p \mathcal M → T_{p ⋅ a} \mathcal M \end{aligned}\]

Manifolds.base_group — Method

base_group(A::AbstractGroupAction)

The group that acts in action A.

Manifolds.center_of_orbit — Function

center_of_orbit(
    A::AbstractGroupAction,
    pts,
    p,
    mean_method::AbstractEstimationMethod = GradientDescentEstimation()
)

Calculate an action element $g$ of action A that is the mean element of the orbit of p with respect to given set of points pts. The mean is calculated using the method mean_method.

The orbit of $p$ with respect to the action of a group $G$ is the set

\[O = \{ g ⋅ p : g ∈ G \}.\]

This function is useful for computing means on quotients of manifolds by a Lie group action.

Manifolds.direction — Method

direction(::AbstractGroupAction{AD}) -> AD

Get the direction of the action

Manifolds.g_manifold — Method

g_manifold(A::AbstractGroupAction)

The manifold the action A acts upon.

Manifolds.inverse_apply! — Method

inverse_apply!(A::AbstractGroupAction, q, a, p)

Apply inverse of action a to the point p with the rule specified by A. The result is saved in q.

Manifolds.inverse_apply — Method

inverse_apply(A::AbstractGroupAction, a, p)

Apply inverse of action a to the point p. The action is specified by A.

Manifolds.inverse_apply_diff — Method

inverse_apply_diff(A::AbstractGroupAction, a, p, X)

For group point $p ∈ \mathcal M$ and tangent vector $X ∈ T_p \mathcal M$, compute the action of the differential of the inverse action of $a ∈ G$ on $X$, specified by rule A. Written as $(\mathrm{d}τ_a)_p^{-1} (X)$, with the specified left or right convention, the differential transports vectors

\[\begin{aligned} (\mathrm{d}L_a)_p^{-1} (X) &: T_p \mathcal M → T_{a^{-1} ⋅ p} \mathcal M\\ (\mathrm{d}R_a)_p^{-1} (X) &: T_p \mathcal M → T_{p ⋅ a^{-1}} \mathcal M \end{aligned}\]

Manifolds.optimal_alignment! — Method

optimal_alignment!(A::AbstractGroupAction, y, p, q)

Calculate an action element of action A that acts upon p to produce the element closest to q. The result is written to x.

Manifolds.optimal_alignment — Method

optimal_alignment(A::AbstractGroupAction, p, q)

Calculate an action element of action A that acts upon p to produce the element closest to q in the metric of the G-manifold:

\[\arg\min_{g ∈ G} d_M(g ⋅ p, q)\]

where $G$ is the group that acts on the G-manifold $\mathcal M$.

Group operation action

Manifolds.GroupOperationAction — Type

GroupOperationAction(group::AbstractGroupManifold, AD::ActionDirection = LeftAction())

Action of a group upon itself via left or right translation.

Rotation action

Manifolds.RotationAction — Type

RotationAction(
    M::Manifold,
    SOn::SpecialOrthogonal,
    AD::ActionDirection = LeftAction(),
)

Space of actions of the SpecialOrthogonal group $\mathrm{SO}(n)$ on a Euclidean-like manifold M of dimension n.

Translation action

Manifolds.TranslationAction — Type

TranslationAction(
    M::Manifold,
    Rn::TranslationGroup,
    AD::ActionDirection = LeftAction(),
)

Space of actions of the TranslationGroup $\mathrm{T}(n)$ on a Euclidean-like manifold M.

The left and right actions are equivalent.