Hyperbolic space

Manifolds.Hyperbolic — Type

Hyperbolic{N} <: AbstractEmbeddedManifold{ℝ,DefaultIsometricEmbeddingType}

The hyperbolic space $ℍ^n$ represented by $n+1$-Tuples, i.e. embedded in the Lorentzian manifold equipped with the MinkowskiMetric $⟨\cdot,\cdot⟩_{\mathrm{M}}$. The space is defined as

\[ℍ^n = \Bigl\{p ∈ ℝ^{n+1}\ \Big|\ ⟨p,p⟩_{\mathrm{M}}= -p_{n+1}^2 + \displaystyle\sum_{k=1}^n p_k^2 = -1, p_{n+1} > 0\Bigr\},.\]

The tangent space $T_p ℍ^n$ is given by

\[T_p ℍ^n := \bigl\{ X ∈ ℝ^{n+1} : ⟨p,X⟩_{\mathrm{M}} = 0 \bigr\}.\]

Note that while the MinkowskiMetric renders the Lorentz manifold (only) pseudo-Riemannian, on the tangent bundle of the Hyperbolic space it induces a Riemannian metric. The corresponding sectional curvature is $-1$.

Constructor

Hyperbolic(n)

Generate the $ℍ^{n} ⊂ ℝ^{n+1}$

Manifolds.HyperboloidPoint — Type

HyperboloidPoint <: MPoint

In the Hyperboloid model of the Hyperbolic $ℍ^n$ points are represented as vectors in $ℝ^{n+1}$ with MinkowskiMetric equal to $-1$.

This representation is the default, i.e. vectors are assumed to have this repesentation.

Manifolds.HyperboloidTVector — Type

HyperboloidTVector <: TVector

In the Hyperboloid model of the Hyperbolic $ℍ^n$ tangent vctors are represented as vectors in $ℝ^{n+1}$ with MinkowskiMetric $⟨p,X⟩_{\mathrm{M}}=0$ to their base point $p$.

This representation is the default, i.e. vectors are assumed to have this repesentation.

Manifolds.PoincareBallPoint — Type

PoincareBallPoint <: MPoint

A point on the Hyperbolic manifold $ℍ^n$ can be represented as a vector of norm less than one in $\mathbb R^n$.

Manifolds.PoincareBallTVector — Type

PoincareBallTVector <: TVector

In the Poincaré ball model of the Hyperbolic $ℍ^n$ tangent vctors are represented as vectors in $ℝ^{n}$.

Manifolds.PoincareHalfSpacePoint — Type

PoincareHalfSpacePoint <: MPoint

A point on the Hyperbolic manifold $ℍ^n$ can be represented as a vector in the half plane, i.e. $x ∈ ℝ^n$ with $x_d > 0$.

Manifolds.PoincareHalfSpaceTVector — Type

PoincareHalfPlaneTVector <: TVector

In the Poincaré half plane model of the Hyperbolic $ℍ^n$ tangent vctors are represented as vectors in $ℝ^{n}$.

Base.convert — Method

convert(::Type{<:AbstractVector}, x::PoincareHalfSpacePoint)

convert a point PoincareHalfSpacePoint x (from $ℝ^n$) from the Poincaré half plane model of the Hyperbolic manifold $ℍ^n$ to a HyperboloidPoint $π(x) ∈ ℝ^{n+1}$.

This is done in two steps, namely transforming it to a Poincare ball point and from there further to an Poincaré half plane point.

Base.convert — Method

convert(::Type{HyperboloidPoint}, x::PoincareBallPoint)

convert a point PoincareBallPoint x (from $ℝ^n$) from the Poincaré ball model of the Hyperbolic manifold $ℍ^n$ to a HyperboloidPoint $π(x) ∈ ℝ^{n+1}$. The isometry is defined by

\[π(x) = \frac{1}{1+\lVert x \rVert^2} \begin{pmatrix}2x_1\\\vdots\\2x_n\\1+\lVert x \rVert^2\end{pmatrix}\]

Note that this is also used, when the type to convert to is a vector.

Base.convert — Method

convert(::Type{HyperboloidPoint, x::PoincareHalfSpacePoint)

convert a point PoincareHalfSpacePoint x (from $ℝ^n$) from the Poincaré half plane model of the Hyperbolic manifold $ℍ^n$ to a HyperboloidPoint $π(x) ∈ ℝ^{n+1}$.

This is done in two steps, namely transforming it to a Poincare ball point and from there further on to a Hyperboloid point.

Base.convert — Method

convert(::Type{PoincareBallPoint}, x::HyperboloidPoint)

convert a HyperboloidPoint $x∈ℝ^{n+1}$ from the hyperboloid model of the Hyperbolic manifold $ℍ^n$ to a PoincareBallPoint $π(x)∈ℝ^{n}$ in the Poincaré ball model. The isometry is defined by

\[π(x) = \frac{1}{1+x_{n+1}} \begin{pmatrix}x_1\\\vdots\\x_n\end{pmatrix}\]

Note that this is also used, when x is a vector.

Base.convert — Method

convert(::Type{PoincareBallPoint}, x::PoincareHalfSpacePoint)

convert a point PoincareHalfSpacePoint x (from $ℝ^n$) from the Poincaré half plane model of the Hyperbolic manifold $ℍ^n$ to a PoincareBallPoint $π(x) ∈ ℝ^n$. Denote by $\tilde x = (x_1,\ldots,x_{d-1})$. Then the isometry is defined by

\[π(x) = \frac{1}{\lVert \tilde x \rVert^2 + (x_n+1)^2} \begin{pmatrix}2x_1\\\vdots\\2x_{n-1}\\\lVert\tilde x\rVert^2 + x_n^2-1\end{pmatrix}.\]

Base.convert — Method

convert(::Type{PoincareHalfSpacePoint}, x::Hyperboloid)

convert a point HyperboloidPoint x (from $ℝ^{n+1}$) from the Hyperboloid model of the Hyperbolic manifold $ℍ^n$ to a PoincareHalfSpacePoint $π(x) ∈ ℝ^{n}$.

This is done in two steps, namely transforming it to a Poincare ball point and from there further on to a PoincareHalfSpacePoint point.

Base.convert — Method

convert(::Type{PoincareHalfSpacePoint}, x::PoincareBallPoint)

convert a point PoincareBallPoint x (from $ℝ^n$) from the Poincaré ball model of the Hyperbolic manifold $ℍ^n$ to a PoincareHalfSpacePoint $π(x) ∈ ℝ^n$. Denote by $\tilde x = (x_1,\ldots,x_{d-1})$. Then the isometry is defined by

\[π(x) = \frac{1}{\lVert \tilde x \rVert^2 - (x_n-1)^2} \begin{pmatrix}2x_1\\\vdots\\2x_{n-1}\\1-\lVert\tilde x\rVert^2 - x_n^2-1\end{pmatrix}.\]

Base.convert — Method

convert(::Type{PoincareHalfSpacePoint}, x)

convert a point x (from $ℝ^{n+1}$) from the Hyperboloid model of the Hyperbolic manifold $ℍ^n$ to a PoincareHalfSpacePoint $π(x) ∈ ℝ^{n}$.

This is done in two steps, namely transforming it to a Poincare ball point and from there further to a vector.

Base.exp — Method

exp(M::Hyperbolic, p, X)

Compute the exponential map on the Hyperbolic space $ℍ^n$ emanating from p towards X. The formula reads

\[\exp_p X = \cosh(\sqrt{⟨X,X⟩_{\mathrm{M}}})p + \sinh(\sqrt{⟨X,X⟩_{\mathrm{M}}})\frac{X}{\sqrt{⟨X,X⟩_{\mathrm{M}}}},\]

where $⟨\cdot,\cdot⟩_{\mathrm{M}}$ denotes the MinkowskiMetric on the embedding, the Lorentzian manifold.

Base.log — Method

log(M::Hyperbolic, p, q)

Compute the logarithmic map on the Hyperbolic space $ℍ^n$, the tangent vector representing the geodesic starting from p reaches q after time 1. The formula reads for $p ≠ q$

\[\log_p q = d_{ℍ^n}(p,q) \frac{q-⟨p,q⟩_{\mathrm{M}} p}{\lVert q-⟨p,q⟩_{\mathrm{M}} p \rVert_2},\]

where $⟨\cdot,\cdot⟩_{\mathrm{M}}$ denotes the MinkowskiMetric on the embedding, the Lorentzian manifold. For $p=q$ the logarihmic map is equal to the zero vector.

ManifoldsBase.check_manifold_point — Method

check_manifold_point(M::Hyperbolic, p; kwargs...)

Check whether p is a valid point on the Hyperbolic M.

For the HyperboloidPoint or plain vectors this means that, p is a vector of length $n+1$ with inner product in the embedding of -1, see MinkowskiMetric. The tolerance for the last test can be set using the kwargs....

For the PoincareBallPoint a valid point is a vector $p ∈ ℝ^n$ with a norm stricly less than 1.

For the PoincareHalfSpacePoint a valid point is a vector from $p ∈ ℝ^n$ with a positive last entry, i.e. $p_n>0$

ManifoldsBase.check_tangent_vector — Method

check_tangent_vector(M::Hyperbolic{n}, p, X; check_base_point = true, kwargs... )

Check whether X is a tangent vector to p on the Hyperbolic M, i.e. after check_manifold_point(M,p), X has to be of the same dimension as p. The optional parameter check_base_point indicates whether to call check_manifold_point for p. The tolerance for the last test can be set using the kwargs....

For a the hyperboloid model or vectors, X has to be orthogonal to p with respect to the inner product from the embedding, see MinkowskiMetric.

For a the Poincaré ball as well as the Poincaré half plane model, X has to be a vector from $ℝ^{n}$.

ManifoldsBase.distance — Method

distance(M::Hyperbolic, p, q)

Compute the distance on the Hyperbolic M, which reads

\[d_{ℍ^n}(p,q) = \operatorname{acosh}( - ⟨p, q⟩_{\mathrm{M}}),\]

where $⟨\cdot,\cdot⟩_{\mathrm{M}}$ denotes the MinkowskiMetric on the embedding, the Lorentzian manifold.

ManifoldsBase.injectivity_radius — Method

injectivity_radius(M::Hyperbolic)
injectivity_radius(M::Hyperbolic, p)

Return the injectivity radius on the Hyperbolic, which is $∞$.

ManifoldsBase.manifold_dimension — Method

manifold_dimension(M::Hyperbolic)

Return the dimension of the hyperbolic space manifold $ℍ^n$, i.e. $\dim(ℍ^n) = n$.

ManifoldsBase.project — Method

project(M::Hyperbolic, p, X)

Perform an orthogonal projection with respect to the Minkowski inner product of X onto the tangent space at p of the Hyperbolic space M.

The formula reads

\[Y = X + ⟨p,X⟩_{\mathrm{M}} p,\]

where $⟨\cdot, \cdot⟩_{\mathrm{M}}$ denotes the MinkowskiMetric on the embedding, the Lorentzian manifold.

Note

Projection is only available for the (default) HyperboloidTVector representation, the others don't have such an embedding

ManifoldsBase.vector_transport_to — Method

vector_transport_to(M::Hyperbolic, p, X, q, ::ParallelTransport)

Compute the paralllel transport of the X from the tangent space at p on the Hyperbolic space $ℍ^n$ to the tangent at q along the geodesic connecting p and q. The formula reads

\[\mathcal P_{q←p}X = X - \frac{⟨\log_p q,X⟩_p}{d^2_{ℍ^n}(p,q)} \bigl(\log_p q + \log_qp \bigr),\]

where $⟨\cdot,\cdot⟩_p$ denotes the inner product in the tangent space at p.

Statistics.mean — Method

mean(
    M::Hyperbolic,
    x::AbstractVector,
    [w::AbstractWeights,]
    method = CyclicProximalPointEstimation();
    kwargs...,
)

Compute the Riemannian mean of x on the Hyperbolic space using CyclicProximalPointEstimation.