Hyperbolic Space

Manifolds.Hyperbolic — Type

Hyperbolic{N} <: Manifold

The hyperbolic space $\mathbb H^n$ represented by $n+1$-Tuples, i.e. in by vectors in $\mathbb R^{n+1}$ using the Minkowsi metric, i.e.

\[\mathbb H^n = \Bigl\{x\in\mathbb R^{n+1} \ \Big|\ \langle x,x \rangle_{\mathrm{M}}= -x_{n+1}^2 + \displaystyle\sum_{k=1}^n x_k^2 = -1, x_{n+1} > 0\Bigr\},\]

where $\langle\cdot,\cdot\rangle_{\mathrm{M}}$ denotes the minkowski_dot is Minkowski inner product. The tangent space $T_x\mathbb H^n$ is given by

\[T_x\mathbb H^n \coloneqq \bigl\{ v \in \mathbb R^{n+1} \ \bigl |\ \langle x,v\rangle_{\mathrm{M}} = 0 \bigr\}.\]

The Minkowski inner product inntroduces the MinkowskiMetric, which is a Riemannian metric on the tangent bundle $T\mathbb H^n$.

Constructor

Hyperbolic(n)

Generate the $\mathbb H^{n}\subset \mathbb R^{n+1}$

Manifolds.MinkowskiMetric — Type

MinkowskiMetric <: LorentzMetric

The Minkowski metric is a LorentzMetric with, i.e.

\[\langle a,b\rangle_{\mathrm{M}} = -a_{n+1}b_{n+1} + \displaystyle\sum_{k=1}^n a_kb_k.\]

It is also the default metric e.g. for the Hyperbolic space.

Note

While the MinkowskiMetric itself is not positive definite in the whole embedded space, it is positive definite when restricted to a tangent space $T_x\mathcal M$, $x\in\mathcal M$, of the Hyperbolic space $\mathcal M$.

Base.exp — Method

exp(M::Hyperbolic, x, v)

Compute the exponential map on the Hyperbolic space $\mathbb H^n$ eminating from x towards v, which is optionally scaled by t. The formula reads

\[\exp_x v = \cosh(\sqrt{\langle v,v\rangle_{\mathrm{M}}})x + \sinh(\sqrt{\langle v,v\rangle_{\mathrm{M}}})\frac{v}{\sqrt{\langle v,v\rangle_{\mathrm{M}}}},\]

where $\langle\cdot,\cdot\rangle_{\mathrm{M}}$ denotes the minkowski_dot.

Base.log — Method

log(M::Hyperbolic, x, y)

Compute the logarithmic map on the Hyperbolic space $\mathbb H^n$, the tangent vector representing the geodesic starting from x reaches y after time 1 on the Hyperbolic space M. The formula reads for $x\neq y$

\[\log_x y = d_{\mathbb H^n}(x,y) \frac{y-\langle x,y\rangle_{\mathrm{M}} x}{\lVert y-\langle x,y\rangle_{\mathrm{M}} x \rVert_2}\]

and is zero otherwise.

ManifoldsBase.distance — Method

distance(M::Hyperbolic, x, y)

Compute the distance on the Hyperbolic M, which reads

\[d_{\mathbb H^n}(x,y) = \operatorname{acosh}( - \langle x, y \rangle_{\mathrm{M}}),\]

where $\langle\cdot,\cdot\rangle_{\mathrm{M}}$ denotes the minkowski_dot.

ManifoldsBase.injectivity_radius — Method

injectivity_radius(M::Hyperbolic[, x])

Return the injectivity radius on the Hyperbolic, which is always $\infty$.

ManifoldsBase.inner — Method

inner(M::Hyperbolic, x, v, w)

Compute the Riemannian inner product for two tangent vectors v and w from $T_x\mathbb H^n$ of the Hyperbolic space $\mathbb H^n$ given by $\langle w, v \rangle_{\mathrm{M}}$ the minkowski_dot Minkowski inner product on $\mathbb R^{n+1}$.

ManifoldsBase.manifold_dimension — Method

manifold_dimension(H::Hyperbolic)

Return the dimension of the hyperbolic space manifold $\mathbb H^n$, i.e. $n$.

ManifoldsBase.project_tangent — Method

project_tangent(M::Hyperbolic, x, v)

Perform an orthogonal projection with respect to the Minkowski inner product of v onto the tangent space at x of the Hyperbolic space M.

The formula reads

\[w = v + \langle x,v\rangle_{\mathrm{M}} x,\]

where $\langle \cdot, \cdot \rangle_{\mathrm{M}}$ denotes the Minkowski inner product in the embedding, see minkowski_dot.

ManifoldsBase.representation_size — Method

representation_size(M::Hyperbolic)

Return the representation size on the Hyperbolic, i.e. for the n-diomensional hyperbolic manifold the dimention of the embedding, i.e. n+1.

ManifoldsBase.vector_transport_to — Method

vector_transport_to(M::Hyperbolic, x, v, y, ::ParallelTransport)

Compute the paralllel transport of the v from the tangent space at x on the Hyperbolic space $\mathbb H^n$ to the tangent at y along the geodesic connecting x and y. The formula reads

\[P_{y\gets x}(v) = v - \frac{\langle \log_xy,v\rangle_x}{d^2_{\mathbb H^n}(x,y)} \bigl(\log_xy + \log_yx \bigr).\]

ManifoldsBase.zero_tangent_vector — Method

zero_tangent_vector(M::Hyperbolic, x)

Return the zero vector from the tangent space at x of the Hyperbolic M.

Statistics.mean — Method

mean(
    M::Hyperbolic,
    x::AbstractVector,
    [w::AbstractWeights,]
    method = CyclicProximalPointEstimationMethod();
    kwargs...,
)

Compute the Riemannian mean of x on the Hyperbolic space using CyclicProximalPointEstimation.

Manifolds.minkowski_dot — Method

minkowski_dot(a,b)

Compute the Minkowski inner product of two Vectors a and b of same length n+1, i.e.

\[\langle a,b\rangle_{\mathrm{M}} = -a_{n+1}b_{n+1} + \displaystyle\sum_{k=1}^n a_kb_k.\]

ManifoldsBase.check_manifold_point — Method

check_manifold_point(M::Hyperbolic, x; kwargs...)

Check whether x is a valid point on the Hyperbolic M, i.e. is a vector with minkowski_dot -1. The tolerance for the last test can be set using the kwargs....

ManifoldsBase.check_tangent_vector — Method

check_tangent_vector(M::Hyperbolic, x, v; kwargs... )

Check whether v is a tangent vector to x on the Hyperbolic M, i.e. after check_manifold_point(M,x), v has to be of same dimension as x and orthogonal to x with respect to minkowski_dot. The tolerance for the last test can be set using the kwargs....