Lorentzian Manifold

The Lorentz manifold is a pseudo-Riemannian manifold. It is named after the Dutch physicist Hendrik Lorentz (1853–1928). The default LorentzMetric is the MinkowskiMetric named after the German mathematician Hermann Minkowski (1864–1909).

Within Manifolds.jl it is used as the embedding of the Hyperbolic space.

Manifolds.Lorentz — Type

Lorentz{N} = MetricManifold{Euclidean{N,ℝ},LorentzMetric}

The Lorentz manifold (or Lorentzian) is a pseudo-Riemannian manifold.

Constructor

Lorentz(n[, metric=MinkowskiMetric()])

Generate the Lorentz manifold of dimension n with the LorentzMetric m, which is by default set to the MinkowskiMetric.

Manifolds.LorentzMetric — Type

LorentzMetric <: Metric

Abstract type for Lorentz metrics, which have a single time dimension. These metrics assume the spacelike convention with the time dimension being last, giving the signature $(++...+-)$.

Manifolds.MinkowskiMetric — Type

MinkowskiMetric <: LorentzMetric

As a special metric of signature $(++...+-)$, i.e. a LorentzMetric, see minkowski_metric for the formula.

Manifolds.minkowski_metric — Method

minkowski_metric(a,b)

Compute the minkowski metric on $\mathbb R^n$ is given by

\[⟨a,b⟩_{\mathrm{M}} = -a_{n}b_{n} + \displaystyle\sum_{k=1}^{n-1} a_kb_k.\]