Metric manifold

Metric manifold

Manifolds.HasMetric — Type.

HasMetric

A Trait to mark a Manifold M as being shorthand for a MetricManifold{M,G} with metric G. This can be used to forward functions called on the MetricManifold to the already-imlemented functions for the Manifold.

For example,

@traitfn function my_feature(M::MMT, k...) where {MT<:Manifold,
                                                  GT<:Metric,
                                                  MMT<:MetricManifold{MT,GT};
                                                  HasMetric{MT,GT}}
    return my_feature(M.manifold, k...)
end

forwards the function my_feature from M to the already-implemented my_feature on the base manifold M.manifold. A manifold with a default metric can then be written

struct MyManifold{T} <: Manifold end
struct MyMetric{S} <: Metric end
@traitimpl HasMetric{MyManifold,MyMetric}

Manifolds.LorentzMetric — Type.

LorentzMetric <: Metric

Abstract type for Lorentz metrics, which have a single time dimension. These metrics assume the spacelike convention with the time dimension being last, giving the signature $(++...+-)$.

Manifolds.Metric — Type.

Metric

Abstract type for the pseudo-Riemannian metric tensor $g$, a family of smoothly varying inner products on the tangent space. See inner.

Manifolds.MetricManifold — Type.

MetricManifold{M<:Manifold,G<:Metric} <: Manifold

Equip a manifold with a metric. Such a manifold is generally called pseudo- or semi-Riemannian. Each MetricManifold must implement local_metric.

Constructor

MetricManifold(manifold, metric)

Manifolds.RiemannianMetric — Type.

RiemannianMetric <: Metric

Abstract type for Riemannian metrics, a family of positive definite inner products. The positive definite property means that for $v \in T_x M$, the inner product $g(v, v) > 0$ whenever $v$ is not the zero vector.

Base.exp — Function.

exp(M::MetricManifold, x, v, args...)

If the HasMetric trait is defined for M, compute the exponential map of the base manifold. Otherwise, numerically integrate the exponential map assuming the Levi-Civita connection. See solve_exp_ode

Currently, the numerical integration is only accurate when using a single coordinate chart that covers the entire manifold. This excludes coordinates in an embedded space.

Manifolds.christoffel_symbols_first — Method.

christoffel_symbols_first(M::MetricManifold, x)

Compute the Christoffel symbols of the first kind in local coordinates. The Christoffel symbols are (in Einstein summation convention)

\[\Gamma_{ijk} = \frac{1}{2} \left[g_{kj,i} + g_{ik,j} - g_{ij,k}\right],\]

where $g_{ij,k}=\frac{\partial}{\partial x^k} g_{ij}$ is the coordinate derivative of the local representation of the metric tensor. The dimensions of the resulting multi-dimensional array are ordered $(i,j,k)$.

Manifolds.christoffel_symbols_second — Method.

christoffel_symbols_second(M::MetricManifold, x)

Compute the Christoffel symbols of the second kind in local coordinates. The Christoffel symbols are (in Einstein summation convention)

\[\Gamma^{l}_{ij} = g^{kl} \Gamma_{ijk},\]

where $\Gamma_{ijk}$ are the Christoffel symbols of the first kind, and $g^{kl}$ is the inverse of the local representation of the metric tensor. The dimensions of the resulting multi-dimensional array are ordered $(l,i,j)$.

Manifolds.det_local_metric — Method.

det_local_metric(M::MetricManifold, x)

Determinant of local matrix representation of the metric tensor $g$

Manifolds.einstein_tensor — Method.

einstein_tensor(M::MetricManifold, x)

Compute the Einstein tensor of the manifold M at the point x.

Manifolds.gaussian_curvature — Method.

gaussian_curvature(M::MetricManifold, x)

Compute the Gaussian curvature of the manifold M at the point x.

Manifolds.inverse_local_metric — Method.

inverse_local_metric(M::MetricManifold, x)

Local matrix representation of the inverse metric (cometric) tensor, usually written $g^{ij}$

Manifolds.local_metric — Method.

local_metric(M::MetricManifold, x)

Local matrix representation at the point x of the metric tensor $g$ on the manifold M, usually written $g_{ij}$. The matrix has the property that $g(v, w)=v^T [g_{ij}] w = g_{ij} v^i w^j$, where the latter expression uses Einstein summation convention.

Manifolds.log_local_metric_density — Method.

log_local_metric_density(M::MetricManifold, x)

Return the natural logarithm of the metric density $\rho$ of M at x, which is given by $\rho=\log \sqrt{|\det [g_{ij}]|}$.

Manifolds.metric — Method.

metric(M::MetricManifold)

Get the metric $g$ of the manifold M.

Manifolds.ricci_curvature — Method.

ricci_curvature(M::MetricManifold, x)

Compute the Ricci scalar curvature of the manifold M at the point x.

Manifolds.ricci_tensor — Method.

ricci_tensor(M::MetricManifold, x)

Compute the Ricci tensor, also known as the Ricci curvature tensor, of the manifold M at the point x.

Manifolds.riemann_tensor — Method.

riemann_tensor(M::MetricManifold, x)

Compute the Riemann tensor $R^l_{ijk}$, also known as the Riemann curvature tensor, at the point x. The dimensions of the resulting multi-dimensional array are ordered $(l,i,j,k)$.

Manifolds.local_metric_jacobian — Method.

local_metric_jacobian(M::MetricManifold, x)

Get partial derivatives of the local metric of M at x with respect to the coordinates of x, $\frac{\partial}{\partial x^k} g_{ij} = g_{ij,k}$. The dimensions of the resulting multi-dimensional array are ordered $(i,j,k)$.

Manifolds.solve_exp_ode — Method.

solve_exp_ode(M::MetricManifold,
              x,
              v,
              tspan;
              solver=AutoVern9(Rodas5()),
              kwargs...)

Approximate the exponential map on the manifold over the provided timespan assuming the Levi-Civita connection by solving the ordinary differential equation

\[\frac{d^2}{dt^2} x^k + \Gamma^k_{ij} \frac{d}{dt} x_i \frac{d}{dt} x_j = 0,\]

where $\Gamma^k_{ij}$ are the Christoffel symbols of the second kind, and the Einstein summation convention is assumed. The arguments tspan and solver follow the OrdinaryDiffEq conventions. kwargs... specify keyword arguments that will be passed to OrdinaryDiffEq.solve.

Currently, the numerical integration is only accurate when using a single coordinate chart that covers the entire manifold. This excludes coordinates in an embedded space.