Metric manifold

A Riemannian manifold always consists of a topological manifold together with a smoothly varying metric $g$.

However, often there is an implicitly assumed (default) metric, like the usual inner product on Euclidean space. This decorator takes this into account. It is not necessary to use this decorator if you implement just one (or the first) metric. If you later introduce a second, the old (first) metric can be used with the (non MetricManifold) Manifold, i.e. without an explicitly stated metric. The decorator acts transparent in that sense; see is_decorator_manifold for details.

This manifold decorator serves two purposes: to implement different metrics (e.g. in closed form) for one Manifold and to provide a way to compute geodesics on manifolds, where this Metric does not yield closed formula.

Let's first look at the provided types.

Types

Manifolds.LorentzMetric — Type

LorentzMetric <: Metric

Abstract type for Lorentz metrics, which have a single time dimension. These metrics assume the spacelike convention with the time dimension being last, giving the signature $(++...+-)$.

Manifolds.Metric — Type

Metric

Abstract type for the pseudo-Riemannian metric tensor $g$, a family of smoothly varying inner products on the tangent space. See inner.

Manifolds.MetricManifold — Type

MetricManifold{M<:Manifold,G<:Metric} <: Manifold

Equip a Manifold explicitly with a Metric G.

For a Metric Manifold, by default, assumes, that you implement the linear form from local_metric in order to evaluate the exponential map.

If the corresponding Metric G yields closed form formulae for e.g. the exponential map and this is implemented directly (without solving the ode), you can of course still implement that directly.

Constructor

MetricManifold(manifold, metric)

Manifolds.RiemannianMetric — Type

RiemannianMetric <: Metric

Abstract type for Riemannian metrics, a family of positive definite inner products. The positive definite property means that for $v \in T_x M$, the inner product $g(v, v) > 0$ whenever $v$ is not the zero vector.

Implement Different Metrics on the same Manifold

In order to distinguish different metrics on one manifold, one can introduce two Metrics and use this type to dispatch on the metric, see SymmetricPositiveDefinite. To avoid overhead, one Metric can then be marked as being the default, i.e. the one that is used, when no MetricManifold decorator is present. This avoids reimplementation of the first existing metric, access to the metric-dependent functions that were implemented using the undecorated manifold, as well as the transparent fallback of the corresponding MetricManifold with default metric to the undecorated implementations. This does not cause any runtime overhead. Introducing a default Metric serves a better readability of the code when working with different metrics.

Implementation of Metrics

For the case that a local_metric is implemented as a bilinear form that is positive definite, the following further functions are provided, unless the corresponding Metric is marked as default – then the fallbacks mentioned in the last section are used for e.g. the exp!onential map.

Manifolds.christoffel_symbols_first — Method

christoffel_symbols_first(M::MetricManifold, x; backend=:default)

Compute the Christoffel symbols of the first kind in local coordinates. The Christoffel symbols are (in Einstein summation convention)

\[\Gamma_{ijk} = \frac{1}{2} \left[g_{kj,i} + g_{ik,j} - g_{ij,k}\right],\]

where $g_{ij,k}=\frac{\partial}{\partial x^k} g_{ij}$ is the coordinate derivative of the local representation of the metric tensor. The dimensions of the resulting multi-dimensional array are ordered $(i,j,k)$.

Manifolds.christoffel_symbols_second — Method

christoffel_symbols_second(M::MetricManifold, x; backend=:default)

Compute the Christoffel symbols of the second kind in local coordinates. The Christoffel symbols are (in Einstein summation convention)

\[\Gamma^{l}_{ij} = g^{kl} \Gamma_{ijk},\]

where $\Gamma_{ijk}$ are the Christoffel symbols of the first kind, and $g^{kl}$ is the inverse of the local representation of the metric tensor. The dimensions of the resulting multi-dimensional array are ordered $(l,i,j)$.

Manifolds.christoffel_symbols_second_jacobian — Method

christoffel_symbols_second_jacobian(M::MetricManifold, x; backend = :default)

Get partial derivatives of the Christoffel symbols of the second kind for manifold M at x with respect to the coordinates of x, $\frac{\partial}{\partial x^l} \Gamma^{k}_{ij} = \Gamma^{k}_{ij,l}.$ The dimensions of the resulting multi-dimensional array are ordered $(i,j,k,l)$.

Manifolds.det_local_metric — Method

det_local_metric(M::MetricManifold, x)

Determinant of local matrix representation of the metric tensor $g$

Manifolds.einstein_tensor — Method

einstein_tensor(M::MetricManifold, x; backend = :default)

Compute the Einstein tensor of the manifold M at the point x.

Manifolds.gaussian_curvature — Method

gaussian_curvature(M::MetricManifold, x; backend = :default)

Compute the Gaussian curvature of the manifold M at the point x.

Manifolds.inverse_local_metric — Method

inverse_local_metric(M::MetricManifold, x)

Local matrix representation of the inverse metric (cometric) tensor, usually written $g^{ij}$

Manifolds.is_default_metric — Method

is_default_metric(M,G)

indicate, whether the Metric G is the default metric for the Manifold M. This means that any occurence of MetricManifold(M,G) where typeof(is_default_metric(M,G)) = Val{true} falls back to just be called with M such that the Manifold M implicitly has this metric, for example if this was the first one implemented or is the one most commonly assumed to be used.

Manifolds.is_default_metric — Method

is_default_metric(MM)

indicate, whether the Metric MM.G is the default metric for the Manifold MM.M within the MetricManifold MM. This means that any occurence of MetricManifold(MM.M,MM.G) where typeof(is_default_metric(MM.M,MM.G)) = Val{true} falls back to just be called with MM.MM such that the Manifold MM.M implicitly has the metric MM.G, for example if this was the first one implemented or is the one most commonly assumed to be used.

Manifolds.local_metric — Method

local_metric(M::MetricManifold, x)

Local matrix representation at the point x of the metric tensor $g$ on the manifold M, usually written $g_{ij}$. The matrix has the property that $g(v, w)=v^T [g_{ij}] w = g_{ij} v^i w^j$, where the latter expression uses Einstein summation convention.

Manifolds.local_metric_jacobian — Method

local_metric_jacobian(M::MetricManifold, x; backend=:default)

Get partial derivatives of the local metric of M at x with respect to the coordinates of x, $\frac{\partial}{\partial x^k} g_{ij} = g_{ij,k}$. The dimensions of the resulting multi-dimensional array are ordered $(i,j,k)$.

Manifolds.log_local_metric_density — Method

log_local_metric_density(M::MetricManifold, x)

Return the natural logarithm of the metric density $\rho$ of M at x, which is given by $\rho=\log \sqrt{|\det [g_{ij}]|}$.

Manifolds.metric — Method

metric(M::MetricManifold)

Get the metric $g$ of the manifold M.

Manifolds.ricci_curvature — Method

ricci_curvature(M::MetricManifold, x; backend = :default)

Compute the Ricci scalar curvature of the manifold M at the point x.

Manifolds.ricci_tensor — Method

ricci_tensor(M::MetricManifold, x; backend = :default)

Compute the Ricci tensor, also known as the Ricci curvature tensor, of the manifold M at the point x.

Manifolds.riemann_tensor — Method

riemann_tensor(M::MetricManifold, x)

Compute the Riemann tensor $R^l_{ijk}$, also known as the Riemann curvature tensor, at the point x. The dimensions of the resulting multi-dimensional array are ordered $(l,i,j,k)$.

ManifoldsBase.exp! — Method

exp(MM::MetricManifold, x, v, args...)

Numerically integrate the exponential map assuming the Levi-Civita connection. See solve_exp_ode if the Metric G of the MetricManifold{M,G}` is not marked as the default one.

Currently, the numerical integration is only accurate when using a single coordinate chart that covers the entire manifold. This excludes coordinates in an embedded space.

Manifolds.solve_exp_ode — Method

solve_exp_ode(M::MetricManifold, x, v, tspan; backend=:default, solver=AutoVern9(Rodas5()), kwargs...)

Approximate the exponential map on the manifold over the provided timespan assuming the Levi-Civita connection by solving the ordinary differential equation

\[\frac{d^2}{dt^2} x^k + \Gamma^k_{ij} \frac{d}{dt} x_i \frac{d}{dt} x_j = 0,\]

where $\Gamma^k_{ij}$ are the Christoffel symbols of the second kind, and the Einstein summation convention is assumed. The arguments tspan and solver follow the OrdinaryDiffEq conventions. kwargs... specify keyword arguments that will be passed to OrdinaryDiffEq.solve.

Currently, the numerical integration is only accurate when using a single coordinate chart that covers the entire manifold. This excludes coordinates in an embedded space.