Oblique manifold

The oblique manifold $\mathcal{OB}(n,m)$ is modeled as an AbstractPowerManifold of the (real-valued) Sphere and uses ArrayPowerRepresentation. Points on the torus are hence matrices, $x ∈ ℝ^{n,m}$.

Manifolds.Oblique — Type

Oblique{𝔽, T, S} <: AbstractPowerManifold{𝔽}

The oblique manifold $\mathcal{OB}(n,m)$ is the set of 𝔽-valued matrices with unit norm column endowed with the metric from the embedding. This yields exactly the same metric as considering the product metric of the unit norm vectors, i.e. PowerManifold of the $(n-1)$-dimensional Sphere.

The Sphere is stored internally within M.manifold, such that all functions of AbstractPowerManifold can be used directly.

Constructor

Oblique(n::Int, m::Int, field::AbstractNumbers=ℝ; parameter::Symbol=:type)

Generate the manifold of matrices $\mathbb R^{n×m}$ such that the $m$ columns are unit vectors, i.e. from the Sphere(n-1).

Functions

Most functions are directly implemented for an AbstractPowerManifold with ArrayPowerRepresentation except the following special cases:

ManifoldsBase.check_point — Method

check_point(M::Oblique, p)

Checks whether p is a valid point on the Oblique{m,n} M, i.e. is a matrix of m unit columns from $\mathbb R^{n}$, i.e. each column is a point from Sphere(n-1).

ManifoldsBase.check_vector — Method

check_vector(M::Oblique p, X; kwargs...)

Checks whether X is a valid tangent vector to p on the Oblique M. This means, that p is valid, that X is of correct dimension and columnswise a tangent vector to the columns of p on the Sphere.

ManifoldsBase.parallel_transport_to — Method

parallel_transport_to(M::Oblique, p, X, q)

Compute the parallel transport on the Oblique manifold by doing a column wise parallel transport on the Sphere

Oblique{𝔽, T, S} <: AbstractPowerManifold{𝔽}
The oblique manifold is the set of 𝔽-valued matrices with unit norm column endowed with the metric from the embedding.

check_vector(M::CenteredMatrices, p, X; kwargs... )
Check whether X is a tangent vector to manifold point p on the CenteredMatrices M, i.e. that X is a matrix of size (m, n) whose columns sum to zero and its values are from the correct AbstractNumbers.

check_vector(M::Circle, p, X; kwargs...)
Check whether X is a tangent vector in the tangent space of p on the Circle M.

check_vector(M::Elliptope, q, Y; kwargs... )
Check whether is a tangent vector to on the Elliptope M, i.e. Y has to be of same dimension as q and a has to be a symmetric matrix with zero diagonal.

check_vector(M::Flag, p::AbstractMatrix, X::AbstractMatrix; kwargs... )
Check whether X is a tangent vector to point p on the Flag manifold M in the Stiefel representation, i.e. that X is a matrix of the form

check_vector(M::Flag, p::OrthogonalPoint, X::OrthogonalTangentVector; kwargs... )
Check whether X is a tangent vector to point p on the Flag manifold M in the orthogonal matrix representation, i.e. that X is block-skew-symmetric with zero diagonal:

check_vector(M::UnitaryMatrices, p, X; kwargs... ) check_vector(M::OrthogonalMatrices, p, X; kwargs... ) check_vector(M::Rotations, p, X; kwargs... ) check_vector(M::GeneralUnitaryMatrices, p, X; kwargs... )
Check whether X is a tangent vector to p on the UnitaryMatrices space M, i.e. after check_point(M,p), X has to be skew symmetric (Hermitian) and orthogonal to p.

check_vector(M::GraphManifold, p, X; kwargs...)
Check whether X is a valid tangent vector at p on the GraphManifold M.

check_vector(M::ProbabilitySimplex, p, X; kwargs... )
Check whether X is a tangent vector to p on the ProbabilitySimplex M, i.e. after check_point(M,p), X has to be of same dimension as p and its elements have to sum to one.

check_vector(M::AbstractProjectiveSpace, p, X; kwargs... )
Check whether X is a tangent vector in the tangent space of p on the AbstractProjectiveSpace M.

check_vector(M::SPDFixedDeterminant, p, X; kwargs... )
Check whether X is a tangent vector to manifold point p on the SPDFixedDeterminant M, i.e. X has to be a tangent vector on SymmetricPositiveDefinite, so a symmetric matrix, and additionally fulfill .

check_vector(M::Spectrahedron, q, Y; kwargs...)
Check whether is a tangent vector to on the Spectrahedron M, i.e. after check_point of q, Y has to be of same dimension as q and a has to be a symmetric matrix with trace.

check_vector(M::AbstractSphere, p, X; kwargs... )
Check whether X is a tangent vector to p on the AbstractSphere M, i.e. after check_point(M,p), X has to be of same dimension as p and orthogonal to p.

check_vector(M::SphereSymmetricMatrices, p, X; kwargs... )
Check whether X is a tangent vector to manifold point p on the SphereSymmetricMatrices M, i.e. X has to be a symmetric matrix of size (n,n) of unit Frobenius norm.

check_vector(M::SymplecticMatrices, p, X; kwargs...)
Checks whether X is a valid tangent vector at p on the SymplecticMatrices M=, which requires that

check_vector(M::SymplecticMatrices, p, X; kwargs...)
Checks whether X is a valid tangent vector at p on the SymplecticStiefel, manifold.

check_vector(M::Torus{n}, p, X; kwargs...)
Checks whether X is a valid tangent vector to p on the Torus M.

check_vector(M::Tucker{<:Any,D}, p::TuckerPoint{T,D}, X::TuckerTangentVector) where {T,D}
Check whether a TuckerTangentVector X is is in the tangent space to the Dth order Tucker manifold M at the Dth order TuckerPoint p.