Power manifold

A power manifold is based on a AbstractManifold $\mathcal M$ to build a $\mathcal M^{n_1×n_2 ×⋯×n_m}$. In the case where $m=1$ we can represent a manifold-valued vector of data of length $n_1$, for example a time series. The case where $m=2$ is useful for representing manifold-valued matrices of data of size $n_1×n_2$, for example certain types of images.

There are three available representations for points and vectors on a power manifold:

ArrayPowerRepresentation (the default one), very efficient but only applicable when points on the underlying manifold are represented using plain AbstractArrays.

NestedPowerRepresentation, applicable to any manifold. It assumes that points on the underlying manifold are represented using mutable data types.

NestedReplacingPowerRepresentation, applicable to any manifold. It does not mutate points on the underlying manifold, replacing them instead when appropriate.

Below are some examples of usage of these representations.

Example

There are two ways to store the data: in a multidimensional array or in a nested array.

Let's look at an example for both. Let $\mathcal M$ be Sphere(2) the 2-sphere and we want to look at vectors of length 4.

ArrayPowerRepresentation

For the default, the ArrayPowerRepresentation, we store the data in a multidimensional array,

using ManifoldsM = PowerManifold(Sphere(2), 4)p = cat([1.0, 0.0, 0.0],        [1/sqrt(2.0), 1/sqrt(2.0), 0.0],        [1/sqrt(2.0), 0.0, 1/sqrt(2.0)],        [0.0, 1.0, 0.0]    ,dims=2)

3×4 Matrix{Float64}:
 1.0  0.707107  0.707107  0.0
 0.0  0.707107  0.0       1.0
 0.0  0.0       0.707107  0.0

which is a valid point i.e.

is_point(M, p)

true

This can also be used in combination with HybridArrays.jl and StaticArrays.jl, by setting

using HybridArrays, StaticArraysq = HybridArray{Tuple{3,StaticArrays.Dynamic()},Float64,2}(p)

3×4 HybridArrays.HybridMatrix{3, StaticArraysCore.Dynamic(), Float64, 2, Matrix{Float64}} with indices SOneTo(3)×Base.OneTo(4):
 1.0  0.707107  0.707107  0.0
 0.0  0.707107  0.0       1.0
 0.0  0.0       0.707107  0.0

which is still a valid point on M and PowerManifold works with these, too.

An advantage of this representation is that it is quite efficient, especially when a HybridArray (from the HybridArrays.jl package) is used to represent a point on the power manifold. A disadvantage is not being able to easily identify parts of the multidimensional array that correspond to a single point on the base manifold. Another problem is, that accessing a single point is p[:, 1] which might be unintuitive.

NestedPowerRepresentation

For the NestedPowerRepresentation we can now do

using ManifoldsM = PowerManifold(Sphere(2), NestedPowerRepresentation(), 4)p = [ [1.0, 0.0, 0.0],      [1/sqrt(2.0), 1/sqrt(2.0), 0.0],      [1/sqrt(2.0), 0.0, 1/sqrt(2.0)],      [0.0, 1.0, 0.0],    ]

4-element Vector{Vector{Float64}}:
 [1.0, 0.0, 0.0]
 [0.7071067811865475, 0.7071067811865475, 0.0]
 [0.7071067811865475, 0.0, 0.7071067811865475]
 [0.0, 1.0, 0.0]

which is again a valid point so is_point(M, p) here also yields true. A disadvantage might be that with nested arrays one loses a little bit of performance. The data however is nicely encapsulated. Accessing the first data item is just p[1].

For accessing points on power manifolds in both representations you can use get_component and set_component! functions. They work work both point representations.

using ManifoldsM = PowerManifold(Sphere(2), NestedPowerRepresentation(), 4)p = [ [1.0, 0.0, 0.0],      [1/sqrt(2.0), 1/sqrt(2.0), 0.0],      [1/sqrt(2.0), 0.0, 1/sqrt(2.0)],      [0.0, 1.0, 0.0],    ]set_component!(M, p, [0.0, 0.0, 1.0], 4)get_component(M, p, 4)

3-element Vector{Float64}:
 0.0
 0.0
 1.0

NestedReplacingPowerRepresentation

The final representation is the NestedReplacingPowerRepresentation. It is similar to the NestedPowerRepresentation but it does not perform in-place operations on the points on the underlying manifold.

Types and Functions

Manifolds.ArrayPowerRepresentation — Type

ArrayPowerRepresentation

Representation of points and tangent vectors on a power manifold using multidimensional arrays. The first dimensions are equal to representation_size of the wrapped manifold and the following ones are equal to the number of elements in each direction.

Torus uses this representation.

Manifolds.PowerAtlas — Type

PowerAtlas(atlas)

Atlas on an AbstractPowerManifold obtained by using atlas independently on every component. Chart indices have the power dimensions of the manifold and coordinates concatenate the coordinates of its components.

Manifolds.PowerMetric — Type

PowerMetric <: AbstractMetric

Represent the AbstractMetric on an AbstractPowerManifold, i.e. the inner product on the tangent space is the sum of the inner product of each elements tangent space of the power manifold.

ManifoldDiff.riemannian_Hessian — Method

Y = riemannian_Hessian(M::AbstractPowerManifold, p, G, H, X)
riemannian_Hessian!(M::AbstractPowerManifold, Y, p, G, H, X)

Compute the Riemannian Hessian $\operatorname{Hess} f(p)[X]$ given the Euclidean gradient $∇ f(\tilde p)$ in G and the Euclidean Hessian $∇^2 f(\tilde p)[\tilde X]$ in H, where $\tilde p, \tilde X$ are the representations of $p,X$ in the embedding,.

On an abstract power manifold, this decouples and can be computed elementwise.

Manifolds.flat — Method

flat(M::AbstractPowerManifold, p, X)

use the musical isomorphism to transform the tangent vector X from the tangent space at p on an AbstractPowerManifold M to a cotangent vector. This can be done elementwise for each entry of X (and p).

Manifolds.get_default_atlas — Method

get_default_atlas(M::AbstractPowerManifold)

Return the power atlas induced by the default atlas of the underlying manifold of M.

See also

PowerAtlas

Manifolds.manifold_volume — Method

manifold_volume(M::PowerManifold)

Return the manifold volume of an PowerManifold M.

Manifolds.sharp — Method

sharp(M::AbstractPowerManifold, p, ξ::RieszRepresenterCotangentVector)

Use the musical isomorphism to transform the cotangent vector ξ from the tangent space at p on an AbstractPowerManifold M to a tangent vector. This can be done elementwise for every entry of ξ (and p).

Manifolds.volume_density — Method

volume_density(M::PowerManifold, p, X)

Return volume density on the PowerManifold M, i.e. product of constituent volume densities.

RieszRepresenterCotangentVector(M::AbstractManifold, p, X)
Cotangent vector in Riesz representer form on manifold M at point p with Riesz representer X.

is_point(M::Segre{ℝ, V}, p; kwargs...)
Check whether p is a valid point on M, i.e. p[1] is a singleton containing a positive number and p[i + 1] is a point on Sphere(V[i]).

Sphere{𝔽, T} <: AbstractSphere{𝔽}
The (unit) sphere manifold is the set of all unit norm vectors in .