Power manifold

Manifolds.AbstractPowerManifold — Type

AbstractPowerManifold{M,TPR} <: Manifold

An abstract Manifold to represent manifolds that are build as powers of another Manifold M with representation type TPR, a subtype of AbstractPowerRepresentation.

Manifolds.AbstractPowerRepresentation — Type

AbstractPowerRepresentation

An abstract representation type of points and tangent vectors on a power manifold.

Manifolds.MultidimentionalArrayPowerRepresentation — Type

MultidimentionalArrayPowerRepresentation

Representation of points and tangent vectors on a power manifold using multidimensional arrays where first dimensions are equal to representation_size of the wrapped manifold and the following ones are equal to the number of elements in each direction.

Torus uses this representation.

Manifolds.NestedPowerRepresentation — Type

NestedPowerRepresentation

Representation of points and tangent vectors on a power manifold using arrays of size equal to TSize of a PowerManifold. Each element of such array stores a single point or tangent vector.

GraphManifold uses this representation.

Manifolds.PowerManifold — Type

PowerManifold{TM<:Manifold, TSize<:Tuple, TPR<:AbstractPowerRepresentation} <: AbstractPowerManifold{TM}

The power manifold $ℳ^{n_1\times n_2 \times … \times n_d}$ with power geometry TSize statically defines the number of elements along each axis.

For example, a manifold-valued time series would be represented by a power manifold with $d$ equal to 1 and $n_1$ equal to the number of samples. A manifold-valued image (for example in diffusion tensor imaging) would be represented by a two-axis power manifold ($d=2$) with $n_1$ and $n_2$ equal to width and height of the image.

While the size of the manifold is static, points on the power manifold would not be represented by statically-sized arrays. Operations on small power manifolds might be faster if they are represented as ProductManifold.

Constructor

PowerManifold(M, N_1, N_2, ..., N_n)
PowerManifold(M, NestedPowerRepresentation(), N_1, N_2, ..., N_n)

Generate the power manifold $M^{N_1 \times N_2 \times … \times N_n}$. By default, the MultidimentionalArrayPowerRepresentation of points and tangent vectors is used, although a different one, for example NestedPowerRepresentation, can be given as the second argument to the constructor.

Manifolds.PowerMetric — Type

PowerMetric <: Metric

Represent the Metric on an AbstractPowerManifold, i.e. the inner product on the tangent space is the sum of the inner product of each elements tangent space of the power manifold.

Base.exp — Method

exp(M::AbstractPowerManifold, x, v)

Compute the exponential map from x in direction v on the AbstractPowerManifold M, which can be computed using the base manifolds exponential map elementwise.

Base.log — Method

log(M::AbstractPowerManifold, x, y)

Compute the logarithmic map from x to y on the AbstractPowerManifold M, which can be computed using the base manifolds logarithmic map elementwise.

LinearAlgebra.norm — Method

norm(M::AbstractPowerManifold, x, v)

Compute the norm of v from the tangent space of x on an AbstractPowerManifold M, i.e. from the element wise norms the Frobenius norm is computed.

Manifolds.flat — Method

flat(M::AbstractPowerManifold, x, w::FVector{TangentSpaceType})

use the musical isomorphism to transform the tangent vector w from the tangent space at x on an AbstractPowerManifold M to a cotangent vector. This can be done elementwise, so r every entry of w (and x) sparately

Manifolds.sharp — Method

sharp(M::AbstractPowerManifold, x, w::FVector{CotangentSpaceType})

Use the musical isomorphism to transform the cotangent vector w from the tangent space at x on an AbstractPowerManifold M to a tangent vector. This can be done elementwise, so for every entry of w (and x) sparately

ManifoldsBase.distance — Method

distance(M::AbstractPowerManifold, x, y)

Compute the distance between x and y on an AbstractPowerManifold, i.e. from the element wise distances the Forbenius norm is computed.

ManifoldsBase.injectivity_radius — Method

injectivity_radius(M::AbstractPowerManifold[, x])

the injectivity radius on an AbstractPowerManifold is for the global case equal to the one of its base manifold. For a given point x it's equal to the minimum of all radii in the array entries.

ManifoldsBase.inner — Method

inner(M::AbstractPowerManifold, x, v, w)

Compute the inner product of v and w from the tangent space at x on an AbstractPowerManifold M, i.e. for each arrays entry the tangent vector entries from v and w are in the tangent space of the corresponding element from x. The inner product is then the sum of the elementwise inner products.

ManifoldsBase.inverse_retract — Method

inverse_retract(M::AbstractPowerManifold, x, y, m::InversePowerRetraction)

Compute the inverse retraction from x with respect to y on an AbstractPowerManifold M using an InversePowerRetraction, which by default encapsulates a inverse retraction of the base manifold. Then this method is performed elementwise, so the encapsulated inverse retraction method has to be one that is available on the base Manifold.

ManifoldsBase.manifold_dimension — Method

manifold_dimension(M::PowerManifold)

Returns the manifold-dimension of an PowerManifold M $=𝒩 = (ℳ)^{n_1,…,n_d}, i.e. with $n=(n_1,…,n_d)$ the array size of the power manifold and $d_{ℳ}$ the dimension of the base manifold $ℳ$, the manifold is of dimension

\[d_{𝒩} = d_{ℳ}\prod_{i=1}^d n_i = n_1n_2\cdot…\cdot n_d d_{ℳ}.\]

ManifoldsBase.retract — Method

retract(M::AbstractPowerManifold, x, v, m::PowerRetraction)

Compute the retraction from x with tangent vector v on an AbstractPowerManifold M using an PowerRetraction, which by default encapsulates a retraction of the base manifold. Then this method is performed elementwise, so the encapsulated retraction method has to be one that is available on the base Manifold.

Manifolds.InversePowerRetraction — Type

InversePowerRetraction(inverse_retractions::AbstractInverseRetractionMethod...)

Power inverse retraction of inverse_retractions. Works on AbstractPowerManifolds.

Manifolds.PowerFVectorDistribution — Type

PowerFVectorDistribution([type::VectorBundleFibers], [x], distr)

Generates a random vector at point x from vector space (a fiber of a tangent bundle) of type type using the power distribution of distr.

Vector space type and x can be automatically inferred from distribution distr.

Manifolds.PowerPointDistribution — Type

PowerPointDistribution(M::AbstractPowerManifold, distribution)

Power distribution on manifold M, based on distribution.

Manifolds.PowerRetraction — Type

PowerRetraction(retraction::AbstractRetractionMethod)

Power retraction based on retraction. Works on AbstractPowerManifolds.

Manifolds.PrecomputedPowerOrthonormalBasis — Type

PrecomputedPowerOrthonormalBasis(
    bases::AbstractArray{AbstractPrecomputedOrthonormalBasis},
    F::AbstractNumbers = ℝ,
)

A precomputed orthonormal basis of a tangent space of a power manifold. The array bases stores bases corresponding to particular parts of the manifold.

The type parameter F denotes the AbstractNumbers that will be used as scalars.

ManifoldsBase.check_manifold_point — Method

check_manifold_point(M::AbstractProductManifold, x; kwargs...)

Check whether x is a valid point on an AbstractPowerManifold M, i.e. each element of x has to be a valid point on the base manifold.

The tolerance for the last test can be set using the kwargs....

ManifoldsBase.check_tangent_vector — Method

check_tangent_vector(M::AbstractPowerManifold, x, v; kwargs... )

Check whether v is a tangent vector to x an the AbstractPowerManifold M, i.e. atfer check_manifold_point(M, x), and all projections to base manifolds must be respective tangent vectors.

The tolerance for the last test can be set using the kwargs....