Product Manifold

Manifolds.PowerManifold — Type

PowerManifold{TM<:Manifold, TSize<:Tuple} <: Manifold

The power manifold $\mathcal M^{n_1 \times n_2 \times \dots \times n_d}$ with power geometry represented by an array-like structure with $d$ dimensions and sizes $n_1, n_2, \dots, n_d$, along each dimension. TSize statically defines the number of elements along each axis. For example, a manifold-valued time series would be represented by a power manifold with $d$ equal to 1 and $n_1$ equal to the number of samples. A manifold-valued image (for example in diffusion tensor imaging) would be represented by a two-axis power manifold ($d=2$) with $n_1$ and $n_2$ equal to width and height of the image.

While the size of the manifold is static, points on the power manifold would not be represented by statically-sized arrays. Operations on small power manifolds might be faster if they are represented as ProductManifold.

Constructor

PowerManifold(M, N_1, N_2, ..., N_n)

Generate the power manifold $M^{N_1 \times N_2 \times \dots \times N_n}$.

Manifolds.PowerMetric — Type

PowerMetric <: Metric

Represent the Metric on the PowerManifold, i.e. the inner product on the tangent space is the sum of the inner product of each elements tangent space of the power manifold.

Base.exp — Method

exp(M::PowerManifold, x, v)

Compute the exponential map from x in direction v on the PowerManifold M, which can be computed using the base manifolds exponential map elementwise.

Base.log — Method

log(M::PowerManifold, x, y)

Compute the logarithmic map from x to y on the PowerManifold M, which can be computed using the base manifolds logarithmic map elementwise.

LinearAlgebra.norm — Method

norm(M::PowerManifold, x, v)

Compute the norm of v from the tangent space of x on the PowerManifold, i.e. from the element wise norms the Forbenius norm is computed.

Manifolds.flat — Method

flat(M::PowerManifold, x, w::FVector{TangentSpaceType})

use the musical isomorphism to transform the tangent vector w from the tangent space at x on the PowerManifold M to a cotangent vector. This can be done elementwise, so r every entry of w (and x) sparately

Manifolds.sharp — Method

sharp(M::PowerManifold, x, w::FVector{CotangentSpaceType})

Use the musical isomorphism to transform the cotangent vector w from the tangent space at x on the PowerManifold M to a tangent vector. This can be done elementwise, so for every entry of w (and x) sparately

ManifoldsBase.distance — Method

distance(M::PowerManifold, x, y)

Compute the distance between x and y on the PowerManifold, i.e. from the element wise distances the Forbenius norm is computed.

ManifoldsBase.injectivity_radius — Method

injectivity_radius(M::PowerManifold[, x])

the injectivity radius on the PowerManifold is for the global case equal to the one of its base manifold. For a given point x it's equal to the minimum of all radii in the array entries.

ManifoldsBase.inner — Method

inner(M::PowerManifold, x, v, w)

Compute the inner product of v and w from the tangent space at x on the PowerManifold M, i.e. for each arrays entry the tangent vector entries from v and w are in the tangent space of the corresponding element from x. The inner product is then the sum of the elementwise inner products.

ManifoldsBase.inverse_retract — Method

inverse_retract(M::PowerManifold, x, y, m::InversePowerRetraction)

Compute the inverse retraction from x with respect to y on the PowerManifold M using an InversePowerRetraction, which by default encapsulates a inverse retraction of the base manifold. Then this method is performed elementwise, so the encapsulated inverse retraction method has to be one that is available on the base Manifold.

ManifoldsBase.manifold_dimension — Method

manifold_dimension(M::PowerManifold)

Returns the manifold-dimension of the PowerManifold M $=\mathcal N = (\mathcal M)^{n_1,\ldots,n_d}, i.e. with $n=(n_1,\ldots,n_d)$ the array size of the power manifold and $d_{\mathcal M}$ the dimension of the base manifold $\mathcal M$, the manifold is of dimension

\[d_{\mathcal N} = d_{\mathcal M}\prod_{i=1}^d n_i = n_1n_2\cdot\ldots\cdotn_dd_{\mathcal M}.\]

ManifoldsBase.retract — Method

retract(M::PowerManifold, x, v, m::PowerRetraction)

Compute the retraction from x with tangent vector v on the PowerManifold M using an PowerRetraction, which by default encapsulates a retraction of the base manifold. Then this method is performed elementwise, so the encapsulated retraction method has to be one that is available on the base Manifold.

Manifolds.InversePowerRetraction — Type

InversePowerRetraction(inverse_retractions::AbstractInverseRetractionMethod...)

Power inverse retraction of inverse_retractions. Works on PowerManifolds.

Manifolds.PowerFVectorDistribution — Type

PowerFVectorDistribution([type::VectorBundleFibers], [x], distr)

Generates a random vector at point x from vector space (a fiber of a tangent bundle) of type type using the power distribution of distr.

Vector space type and x can be automatically inferred from distribution distr.

Manifolds.PowerPointDistribution — Type

PowerPointDistribution(M::PowerManifold, distribution)

Power distribution on manifold M, based on distribution.

Manifolds.PowerRetraction — Type

PowerRetraction(retraction::AbstractRetractionMethod)

Power retraction based on retraction. Works on PowerManifolds.

ManifoldsBase.check_manifold_point — Method

check_manifold_point(M::ProductManifold, x; kwargs...)

Check whether x is a valid point on the ProductManifold M, i.e. each element of x has to be a valid point on the base manifold.

The tolerance for the last test can be set using the kwargs....

ManifoldsBase.check_tangent_vector — Method

check_tangent_vector(M::ProductManifold, x, v; kwargs... )

Check whether v is a tangent vector to x on the ProductManifold M, i.e. atfer check_manifold_point(M, x), and all projections to base manifolds must be respective tangent vectors.

The tolerance for the last test can be set using the kwargs....