Product Manifold

Power manifold $P = M^{N_1 \times N_2 \times \dots \times N_n}$ of manifold $N$ with shape $N_1 \times N_2 \times \dots \times N_n$. Canonical projections may depend on multiple indices $i_1, i_2, \dots, i_n$. For example, manifold-valued time series require a single index ($n=1$) and manifold-valued images require two indices ($n=2$).

Manifolds.PowerManifold — Type

PowerManifold{TM<:Manifold, TSize<:Tuple} <: Manifold

Power manifold $M^{N_1 \times N_2 \times \dots \times N_n}$ with power geometry. It is represented by an array-like structure with $n$ dimensions and sizes $N_1, N_2, \dots, N_n$, along each dimension. TSize statically defines the number of elements along each axis. For example, a manifold-valued time series would be represented by a power manifold with $n$ equal to 1 and $N_1$ equal to the number of samples. A manifold-valued image (for example in diffusion tensor imaging) would be represented by a two-axis power manifold ($n=2$) with $N_1$ and $N_2$ equal to width and height of the image.

While the size of the manifold is static, points on the power manifold would not be represented by statically-sized arrays. Operations on small power manifolds might be faster if they are represented as ProductManifold.

Constructor

PowerManifold(M, N_1, N_2, ..., N_n)

generates the power manifold $M^{N_1 \times N_2 \times \dots \times N_n}$.

Manifolds.InversePowerRetraction — Type

InversePowerRetraction(inverse_retractions::AbstractInverseRetractionMethod...)

Power inverse retraction of inverse_retractions. Works on PowerManifold.

Manifolds.PowerFVectorDistribution — Type

PowerFVectorDistribution([type::VectorBundleFibers], [x], distr)

Generates a random vector at point x from vector space (a fiber of a tangent bundle) of type type using the power distribution of distr.

Vector space type and x can be automatically inferred from distribution distr.

Manifolds.PowerPointDistribution — Type

PowerPointDistribution(M::PowerManifold, distribution)

Power distribution on manifold M, based on distribution.

Manifolds.PowerRetraction — Type

PowerRetraction(retraction::AbstractRetractionMethod)

Power retraction based on retraction. Works on PowerManifold.

ManifoldsBase.check_manifold_point — Method

check_manifold_point(M::ProductManifold, x; kwargs...)

Check whether x is a valid point on the ProductManifold M.

The tolerance for the last test can be set using the ´kwargs...`.

ManifoldsBase.check_tangent_vector — Method

check_tangent_vector(M::ProductManifold, x, v; kwargs... )

Check whether v is a tangent vector to x on the ProductManifold M, i.e. atfer check_manifold_point(M, x), and all projections to base manifolds must be respective tangent vectors.

The tolerance for the last test can be set using the ´kwargs...`.