The probability simplex

Manifolds.FisherRaoMetric — Type

FisherRaoMetric <: AbstractMetric

The Fisher-Rao metric or Fisher information metric is a particular Riemannian metric which can be defined on a smooth statistical manifold. Such a manifold is a smooth manifold whose points are probability measures defined on a common probability space.

See for example the ProbabilitySimplex.

Manifolds.ProbabilitySimplex — Type

ProbabilitySimplex{T,boundary} <: AbstractDecoratorManifold{𝔽}

The (relative interior of) the probability simplex is the set

\[Δ^n := \biggl\{ p ∈ ℝ^{n+1}\ \big|\ p_i > 0 \text{ for all } i=1,…,n+1, \text{ and } ⟨\mathbb{1},p⟩ = \sum_{i=1}^{n+1} p_i = 1\biggr\},\]

where $\mathbb{1}=(1,…,1)^{\mathrm{T}}∈ ℝ^{n+1}$ denotes the vector containing only ones.

If boundary is set to :open, then the object represents an open simplex. Otherwise, that is when boundary is set to :closed, the boundary is also included:

\[\hat{Δ}^n := \biggl\{ p ∈ ℝ^{n+1}\ \big|\ p_i \geq 0 \text{ for all } i=1,…,n+1, \text{ and } ⟨\mathbb{1},p⟩ = \sum_{i=1}^{n+1} p_i = 1\biggr\},\]

This set is also called the unit simplex or standard simplex.

The tangent space is given by

\[T_pΔ^n = \biggl\{ X ∈ ℝ^{n+1}\ \big|\ ⟨\mathbb{1},X⟩ = \sum_{i=1}^{n+1} X_i = 0 \biggr\}\]

The manifold is implemented assuming the Fisher-Rao metric for the multinomial distribution, which is equivalent to the induced metric from isometrically embedding the probability simplex in the $n$-sphere of radius 2. The corresponding diffeomorphism $\varphi: \mathbb Δ^n → \mathcal N$, where $\mathcal N \subset 2𝕊^n$ is given by $\varphi(p) = 2\sqrt{p}$.

This implementation follows the notation in [APSS17].

Constructor

ProbabilitySimplex(n::Int; boundary::Symbol=:open)

Base.exp — Method

exp(M::ProbabilitySimplex, p, X)

Compute the exponential map on the probability simplex.

\[\exp_pX = \frac{1}{2}\Bigl(p+\frac{X_p^2}{\lVert X_p \rVert^2}\Bigr) + \frac{1}{2}\Bigl(p - \frac{X_p^2}{\lVert X_p \rVert^2}\Bigr)\cos(\lVert X_p\rVert) + \frac{1}{\lVert X_p \rVert}\sqrt{p}\sin(\lVert X_p\rVert),\]

where $X_p = \frac{X}{\sqrt{p}}$, with its division meant elementwise, as well as for the operations $X_p^2$ and $\sqrt{p}$.

Base.log — Method

log(M::ProbabilitySimplex, p, q)

Compute the logarithmic map of p and q on the ProbabilitySimplex M.

\[\log_pq = \frac{d_{Δ^n}(p,q)}{\sqrt{1-⟨\sqrt{p},\sqrt{q}⟩}}(\sqrt{pq} - ⟨\sqrt{p},\sqrt{q}⟩p),\]

where $pq$ and $\sqrt{p}$ is meant elementwise.

Base.rand — Method

rand(::ProbabilitySimplex; vector_at=nothing, σ::Real=1.0)

When vector_at is nothing, return a random point x on the ProbabilitySimplex manifold M. The point is uniform with respect to the Fisher-Rao metric, that is, uniform with respect to the n-sphere whose positive orthant is mapped to the simplex. It is generated according to the isometric embedding into the n-sphere by normalizing the vector length of a sample from a multivariate Gaussian. See [Mar72].

When vector_at is not nothing, return a (Gaussian) random vector from the tangent space $T_{p}\mathrm{\Delta}^n$by shifting a multivariate Gaussian with standard deviation σ to have a zero component sum.

ManifoldDiff.riemannian_gradient — Method

X = riemannian_gradient(M::ProbabilitySimplex, p, Y)
riemannian_gradient!(M::ProbabilitySimplex, X, p, Y)

Given a gradient $Y = \operatorname{grad} \tilde f(p)$ in the embedding $ℝ^{n+1}$ of the ProbabilitySimplex $Δ^n$, this function computes the Riemannian gradient $X = \operatorname{grad} f(p)$ where $f$ is the function $\tilde f$ restricted to the manifold.

The formula reads

\[ X = p ⊙ Y - ⟨p, Y⟩p,\]

where $⊙$ denotes the emelementwise product.

Manifolds.manifold_volume — Method

manifold_volume(::ProbabilitySimplex)

Return the volume of the ProbabilitySimplex, i.e. volume of the n-dimensional Sphere divided by $2^{n+1}$, corresponding to the volume of its positive orthant.

Manifolds.volume_density — Method

volume_density(M::ProbabilitySimplex, p, X)

Compute the volume density at point p on ProbabilitySimplex M for tangent vector X. It is computed using isometry with positive orthant of a sphere.

ManifoldsBase.change_metric — Method

change_metric(M::ProbabilitySimplex, ::EuclideanMetric, p, X)

Change the metric representation of the tangent vector X from the EuclideanMetric to the FisherRaoMetric on the ProbabilitySimplex M.

To change the metric, we are looking for a function $c\colon T_pΔ^n → T_pΔ^n$ such that for all $X,Y ∈ T_pΔ^n$

\[⟨X, Y⟩ = g_p(c(X), c(Y)).\]

This can be achieved by rewriting the representer change in matrix form as (Diagonal(p) - p * p') * X and taking the square root of the matrix.

ManifoldsBase.change_representer — Method

change_representer(M::ProbabilitySimplex, ::EuclideanMetric, p, X)

Given a tangent vector with respect to the metric from the embedding, the EuclideanMetric, the representer of a linear functional on the tangent space is adapted as $Z = p .* X .- p .* dot(p, X)$. The first part “compensates” for the division by $p$ in the Riemannian metric on the ProbabilitySimplex and the second part performs appropriate projection to keep the vector tangent.

For details see Proposition 2.3 in [APSS17].

ManifoldsBase.check_point — Method

check_point(M::ProbabilitySimplex, p; kwargs...)

Check whether p is a valid point on the ProbabilitySimplex M, i.e. is a point in the embedding with positive entries that sum to one The tolerance for the last test can be set using the kwargs....

ManifoldsBase.check_vector — Method

check_vector(M::ProbabilitySimplex, p, X; kwargs... )

Check whether X is a tangent vector to p on the ProbabilitySimplex M, i.e. after check_point(M,p), X has to be of same dimension as p and its elements have to sum to one. The tolerance for the last test can be set using the kwargs....

ManifoldsBase.distance — Method

distance(M, p, q)

Compute the distance between two points on the ProbabilitySimplex M. The formula reads

\[d_{Δ^n}(p,q) = 2\arccos \biggl( \sum_{i=1}^{n+1} \sqrt{p_i q_i} \biggr)\]

ManifoldsBase.injectivity_radius — Method

injectivity_radius(M::ProbabilitySimplex, p)

Compute the injectivity radius on the ProbabilitySimplex M at the point p, i.e. the distance radius to a point near/on the boundary, that could be reached by following the geodesic.

ManifoldsBase.inner — Method

inner(M::ProbabilitySimplex, p, X, Y)

Compute the inner product of two tangent vectors X, Y from the tangent space $T_pΔ^n$ at p. The formula reads

\[g_p(X,Y) = \sum_{i=1}^{n+1}\frac{X_iY_i}{p_i}\]

When M includes boundary, we can just skip coordinates where $p_i$ is equal to 0, see Proposition 2.1 in [AJLS17].

ManifoldsBase.inverse_retract — Method

inverse_retract(M::ProbabilitySimplex, p, q, ::SoftmaxInverseRetraction)

Compute a first order approximation by projection. The formula reads

\[\operatorname{retr}^{-1}_p q = \bigl( I_{n+1} - \frac{1}{n}\mathbb{1}^{n+1,n+1} \bigr)(\log(q)-\log(p))\]

where $\mathbb{1}^{m,n}$ is the size (m,n) matrix containing ones, and $\log$ is applied elementwise.

ManifoldsBase.is_flat — Method

is_flat(::ProbabilitySimplex)

Return false. ProbabilitySimplex is not a flat manifold.

ManifoldsBase.manifold_dimension — Method

manifold_dimension(M::ProbabilitySimplex)

Returns the manifold dimension of the probability simplex in $ℝ^{n+1}$, i.e.

\[ \dim_{Δ^n} = n.\]

ManifoldsBase.project — Method

project(M::ProbabilitySimplex, p, Y)

Project Y from the embedding onto the tangent space at p on the ProbabilitySimplex M. The formula reads

`math \operatorname{proj}_{Δ^n}(p,Y) = Y - \bar{Y} where $\bar{Y}$ denotes mean of $Y$.

ManifoldsBase.project — Method

project(M::ProbabilitySimplex, p)

project p from the embedding onto the ProbabilitySimplex M. The formula reads

\[\operatorname{proj}_{Δ^n}(p) = \frac{1}{⟨\mathbb 1,p⟩}p,\]

where $\mathbb 1 ∈ ℝ$ denotes the vector of ones. Not that this projection is only well-defined if $p$ has positive entries.

ManifoldsBase.representation_size — Method

representation_size(::ProbabilitySimplex)

Return the representation size of points in the $n$-dimensional probability simplex, i.e. an array size of (n+1,).

ManifoldsBase.retract — Method

retract(M::ProbabilitySimplex, p, X, ::SoftmaxRetraction)

Compute a first order approximation by applying the softmax function. The formula reads

\[\operatorname{retr}_p X = \frac{p\mathrm{e}^X}{⟨p,\mathrm{e}^X⟩},\]

where multiplication, exponentiation and division are meant elementwise.

ManifoldsBase.riemann_tensor — Method

riemann_tensor(::ProbabilitySimplex, p, X, Y, Z)

Compute the Riemann tensor $R(X,Y)Z$ at point p on ProbabilitySimplex M. It is computed using isometry with positive orthant of a sphere.

ManifoldsBase.zero_vector — Method

zero_vector(M::ProbabilitySimplex, p)

Return the zero tangent vector in the tangent space of the point p from the ProbabilitySimplex M, i.e. its representation by the zero vector in the embedding.

Statistics.mean — Method

mean(
    M::ProbabilitySimplex,
    x::AbstractVector,
    [w::AbstractWeights,]
    method = GeodesicInterpolation();
    kwargs...,
)

Compute the Riemannian mean of x using GeodesicInterpolation.

Euclidean metric

Base.rand — Method

rand(::MetricManifold{ℝ,<:ProbabilitySimplex,<:EuclideanMetric}; vector_at=nothing, σ::Real=1.0)

When vector_at is nothing, return a random (uniform) point x on the ProbabilitySimplex M with the Euclidean metric. This is done by normalizing independent exponential draws to unit sum, see [Dev86], Theorems 2.1 and 2.2 on p. 207 and 208, respectively.

When vector_at is not nothing, return a (Gaussian) random vector from the tangent space $T_{p}\mathrm{\Delta}^n$ by shifting a multivariate Gaussian with standard deviation σ to have a zero component sum.

Manifolds.manifold_volume — Method

manifold_volume(::MetricManifold{ℝ,<:ProbabilitySimplex{n},<:EuclideanMetric})) where {n}

Return the volume of the ProbabilitySimplex with the Euclidean metric. The formula reads $\frac{\sqrt{n+1}}{n!}$

Manifolds.volume_density — Method

volume_density(::MetricManifold{ℝ,<:ProbabilitySimplex,<:EuclideanMetric}, p, X)

Compute the volume density at point p on ProbabilitySimplex M for tangent vector X. It is equal to 1.

Real probability amplitudes

An isometric embedding of interior of ProbabilitySimplex in positive orthant of the Sphere is established through functions simplex_to_amplitude and amplitude_to_simplex. Some properties extend to the boundary but not all.

This embedding isometrically maps the Fisher-Rao metric on the open probability simplex to the sphere of radius 1 with Euclidean metric. More details can be found in Section 2.2 of [AJLS17].

The name derives from the notion of probability amplitudes in quantum mechanics. They are complex-valued and their squared norm corresponds to probability. This construction restricted to real valued amplitudes results in this embedding.

Manifolds.amplitude_to_simplex — Method

amplitude_to_simplex(M::ProbabilitySimplex, p)

Convert point (real) probability amplitude p on to a point on ProbabilitySimplex. The formula reads $(p_1^2, p_2^2, …, p_{N+1}^2)$. This is an isometry from the interior of the positive orthant of a sphere to interior of the probability simplex.

Manifolds.amplitude_to_simplex_diff — Method

amplitude_to_simplex_diff(M::ProbabilitySimplex, p, X)

Compute differential of amplitude_to_simplex of a point p on ProbabilitySimplex at tangent vector X from the tangent space at p from a sphere.

Manifolds.simplex_to_amplitude — Method

simplex_to_amplitude(M::ProbabilitySimplex, p)

Convert point p on ProbabilitySimplex to (real) probability amplitude. The formula reads $(\sqrt{p_1}, \sqrt{p_2}, …, \sqrt{p_{N+1}})$. This is an isometry from the interior of the probability simplex to the interior of the positive orthant of a sphere.

Manifolds.simplex_to_amplitude_diff — Method

simplex_to_amplitude_diff(M::ProbabilitySimplex, p, X)

Compute differential of simplex_to_amplitude of a point on p one ProbabilitySimplex at tangent vector X from the tangent space at p from a sphere.

Literature

[AJLS17]

N. Ay, J. Jost, H. V. Lê and L. Schwachhöfer. Information Geometry (Springer Cham, 2017).

[Dev86]

L. Devroye. Non-Uniform Random Variate Generation (Springer New York, NY, 1986).

[Mar72]

G. Marsaglia. Choosing a Point from the Surface of a Sphere. Annals of Mathematical Statistics 43, 645–646 (1972).

[APSS17]

F. Åström, S. Petra, B. Schmitzer and C. Schnörr. Image Labeling by Assignment. Journal of Mathematical Imaging and Vision 58, 211–238 (2017), arXiv:1603.05285.

ProbabilitySimplex{T,boundary} <: AbstractDecoratorManifold{𝔽}
The (relative interior of) the probability simplex is the set

manifold_volume(::MetricManifold{ℝ,<:ProbabilitySimplex{n},<:EuclideanMetric})) where {n}
Return the volume of the ProbabilitySimplex with the Euclidean metric.

manifold_volume(::ProbabilitySimplex)
Return the volume of the ProbabilitySimplex, i.e. volume of the n-dimensional Sphere divided by , corresponding to the volume of its positive orthant.

volume_density(::MetricManifold{ℝ,<:ProbabilitySimplex,<:EuclideanMetric}, p, X)
Compute the volume density at point p on ProbabilitySimplex M for tangent vector X.

volume_density(M::ProbabilitySimplex, p, X)
Compute the volume density at point p on ProbabilitySimplex M for tangent vector X.

check_point(M::ProbabilitySimplex, p; kwargs...)
Check whether p is a valid point on the ProbabilitySimplex M, i.e. is a point in the embedding with positive entries that sum to one The tolerance for the last test can be set using the kwargs....

inner(M::ProbabilitySimplex, p, X, Y)
Compute the inner product of two tangent vectors X, Y from the tangent space at p.

manifold_volume(M::AbstractManifold)
Volume of manifold M defined through integration of Riemannian volume element in a chart.

volume_density(M::AbstractManifold, p, X)
Volume density function of manifold M, i.e. determinant of the differential of exponential map exp(M, p, X).

check_point(M::CenteredMatrices, p; kwargs...)
Check whether the matrix is a valid point on the CenteredMatrices M, i.e. is an m-by-n matrix whose columns sum to zero.

check_point(M::CholeskySpace, p; kwargs...)
Check whether the matrix p lies on the CholeskySpace M, i.e. it's size fits the manifold, it is a lower triangular matrix and has positive entries on the diagonal.

inner(M::CholeskySpace, p, X, Y)
Compute the inner product on the CholeskySpace M at p for the tangent vectors X and Y.

manifold_volume(M::Circle)
Return the volume of the Circle M, i.e. .

volume_density(::Circle, p, X)
Return volume density of Circle, i.e. 1.

check_point(M::Circle, p)
Check whether p is a point on the Circle M.

inner(M::Circle, p, X, Y)
Compute the inner product of the two tangent vectors X,Y from the tangent plane at p on the Circle M using the restriction of the metric from the embedding, i.e.

check_point(M::DeterminantOneMatrices{n,𝔽}, p; kwargs...)
Check whether p is a valid manifold point on the DeterminantOneMatrices M, i.e. whether p has a determinant of .

check_point(M::Elliptope, q; kwargs...)
checks, whether q is a valid representation of a point on the Elliptope M, i.e. is a matrix of size (N,K), such that is symmetric positive semidefinite and has unit trace.

check_point(M::EssentialManifold, p; kwargs...)
Check whether the matrix is a valid point on the EssentialManifold M, i.e. a 2-element array containing SO(3) matrices.

manifold_volume(::Euclidean)
Return volume of the Euclidean manifold, i.e. infinity.

volume_density(M::Euclidean, p, X)
Return volume density function of Euclidean manifold M, i.e. 1.

inner(M::Euclidean, p, X, Y)
Compute the inner product on the Euclidean M, which is just the inner product on the real-valued or complex valued vector space of arrays (or tensors) of size , i.e.

check_point(M::FixedRankMatrices, p; kwargs...)
Check whether the matrix or SVDMPoint x ids a valid point on the FixedRankMatrices M, i.e. is an m-byn matrix of rank k.

inner(M::FixedRankMatrices, p::SVDMPoint, X::UMVTangentVector, Y::UMVTangentVector)
Compute the inner product of X and Y in the tangent space of p on the FixedRankMatrices M, which is inherited from the embedding, i.e. can be computed using dot on the elements (U, Vt, M) of X and Y.

check_point(M::GeneralizedGrassmann, p)
Check whether p is representing a point on the GeneralizedGrassmann M, i.e. its a n-by-k matrix of unitary column vectors with respect to the B inner product and of correct eltype with respect to 𝔽.

inner(M::GeneralizedGrassmann, p, X, Y)
Compute the inner product for two tangent vectors X, Y from the tangent space of p on the GeneralizedGrassmann manifold M.

check_point(M::GeneralizedStiefel, p; kwargs...)
Check whether p is a valid point on the GeneralizedStiefel M=, i.e. that it has the right AbstractNumbers type and is (approximately) the identity, where is the complex conjugate transpose.

inner(M::GeneralizedStiefel, p, X, Y)
Compute the inner product for two tangent vectors X, Y from the tangent space of p on the GeneralizedStiefel manifold M.

manifold_volume(::GeneralUnitaryMatrices{ℝ,<:Any,AbsoluteDeterminantOneMatrixType})
Volume of the manifold of real orthogonal matrices of absolute determinant one.

manifold_volume(::GeneralUnitaryMatrices{ℝ,<:Any,DeterminantOneMatrixType})
Volume of the manifold of real orthogonal matrices of determinant one.

manifold_volume(::SpecialUnitaryMatrices)
Volume of the manifold of complex general unitary matrices of determinant one.

manifold_volume(::GeneralUnitaryMatrices{ℂ,<:Any,AbsoluteDeterminantOneMatrixType})
Volume of the manifold of complex general unitary matrices of absolute determinant one.

volume_density(M::GeneralUnitaryMatrices{ℝ, TypeParameter{Tuple{2}}}, p, X)
Volume density on O(2)/SO(2) is equal to 1.

volume_density(M::GeneralUnitaryMatrices{ℝ, TypeParameter{Tuple{3}}}, p, X)
Compute the volume density on O(3)/SO(3).

volume_density(M::GeneralUnitaryMatrices{<:Any,ℝ}, p, X)
Compute volume density function of a sphere, i.e. determinant of the differential of exponential map exp(M, p, X).

check_point(M::Rotations, p; kwargs...)
Check whether p is a valid point on the UnitaryMatrices M, i.e. that has a determinant of absolute value one, i.e. that

check_point(M::UnitaryMatrices, p; kwargs...) check_point(M::OrthogonalMatrices, p; kwargs...) check_point(M::GeneralUnitaryMatrices, p; kwargs...)
Check whether p is a valid point on the UnitaryMatrices or [OrthogonalMatrices] M, i.e. that has a determinant of absolute value one.

check_point(M::GraphManifold, p)
Check whether p is a valid point on the GraphManifold M.

check_point(::Grassmann, p::ProjectorPoint; kwargs...)
Check whether an orthogonal projector is a point from the Grassmann(n,k) manifold, i.e. the ProjectorPoint , has to fulfill , , and `\operatorname{rank} p = k.

inner(M::Grassmann, p, X, Y)
Compute the inner product for two tangent vectors X, Y from the tangent space of p on the Grassmann manifold M.

check_point(M::HamiltonianMatrices, p; kwargs...)
Check whether p is a valid manifold point on the HamiltonianMatrices M, i.e. whether p is_hamiltonian.

manifold_dimension(M::Hyperbolic)
Return the volume of the hyperbolic space manifold , i.e. infinity.

volume_density(M::Hyperbolic, p, X)
Compute volume density function of the hyperbolic manifold.

check_point(M::Hyperbolic, p; kwargs...)
Check whether p is a valid point on the Hyperbolic M.

inner(M::Hyperbolic, p, X, Y) inner(M::Hyperbolic, p::HyperboloidPoint, X::HyperboloidTangentVector, Y::HyperboloidTangentVector)
Cmpute the inner product in the Hyperboloid model, i.e. the minkowski_metric in the embedding.

inner(::Hyperbolic, p::PoincareBallPoint, X::PoincareBallTangentVector, Y::PoincareBallTangentVector)
Compute the inner product in the Poincaré ball model.

inner( ::Hyperbolic, p::PoincareHalfSpacePoint, X::PoincareHalfSpaceTangentVector, Y::PoincareHalfSpaceTangentVector )
Compute the inner product in the Poincaré half space model.

manifold_volume(::Hyperrectangle)
Return volume of the Hyperrectangle manifold, i.e. infinity.

volume_density(M::Hyperrectangle, p, X)
Return volume density function of Hyperrectangle manifold M, i.e. 1.

inner(M::Hyperrectangle, p, X, Y)
Compute the inner product on the Hyperrectangle M, which is just the inner product on the real-valued vector space of arrays (or tensors) of size , i.e.

check_point(M::InvertibleMatrices{n,𝔽}, p; kwargs...)
Check whether p is a valid manifold point on the InvertibleMatrices M, i.e. whether p is an invertible matrix of size (n,n) with values from the corresponding AbstractNumbers 𝔽.

check_point(M::MultinomialMatrices, p)
Check whether p is a valid point on the MultinomialMatrices(m,n) M.

check_point(M::MultinomialDoubleStochastic, p)
Checks whether p is a valid point on the MultinomialDoubleStochastic(n) M, i.e. is a matrix with positive entries whose rows and columns sum to one.

check_point(M::MultinomialSymmetric, p)
Checks whether p is a valid point on the MultinomialSymmetric(m,n) M, i.e. is a symmetric matrix with positive entries whose rows sum to one.

check_point(M::Oblique, p)
Checks whether p is a valid point on the Oblique{m,n} M, i.e. is a matrix of m unit columns from , i.e. each column is a point from Sphere(n-1).

manifold_volume(M::PositiveNumbers)
Return volume of PositiveNumbers M, i.e. Inf.

volume_density(M::PositiveNumbers, p, X)
Compute volume density function of PositiveNumbers.

check_point(M::PositiveNumbers, p)
Check whether p is a point on the PositiveNumbers M, i.e. .

inner(M::PositiveNumbers, p, X, Y)
Compute the inner product of the two tangent vectors X,Y from the tangent plane at p on the PositiveNumbers M, i.e.

manifold_volume(M::PowerManifold)
Return the manifold volume of an PowerManifold M.

volume_density(M::PowerManifold, p, X)
Return volume density on the PowerManifold M, i.e. product of constituent volume densities.

manifold_volume(M::ProductManifold)
Return the volume of ProductManifold M, i.e. product of volumes of the manifolds M is constructed from.

volume_density(M::ProductManifold, p, X)
Return volume density on the ProductManifold M, i.e. product of constituent volume densities.

manifold_volume(M::AbstractProjectiveSpace{ℝ})
Volume of the -dimensional AbstractProjectiveSpace M.

check_point(M::AbstractProjectiveSpace, p; kwargs...)
Check whether p is a valid point on the AbstractProjectiveSpace M, i.e. that it has the same size as elements of the embedding and has unit Frobenius norm.

inner(M::Segre{ℝ, V}, p, X, Y,)
Inner product between two tangent vectors and at .

inner(M::MetricManifold{ℝ, Segre{ℝ, V}, WarpedMetric{A}}, p, X, Y)
Inner product between two tangent vectors and at :

check_point(M::KendallsPreShapeSpace, p; atol=sqrt(max_eps(X, Y)), kwargs...)
Check whether p is a valid point on KendallsPreShapeSpace, i.e. whether each row has zero mean.

check_point(M::SkewHermitianMatrices, p; kwargs...)
Check whether p is a valid manifold point on the SkewHermitianMatrices M, i.e. whether p is a skew-hermitian matrix of size (n,n) with values from the corresponding AbstractNumbers 𝔽.

check_point(M::SPDFixedDeterminant, p; kwargs...)
Check whether p is a valid manifold point on the SPDFixedDeterminant(n,d) M, i.e. whether p is a SymmetricPositiveDefinite matrix of size (n, n)

check_point(M::Spectrahedron, q; kwargs...)
Check whether q is a valid representation of a point on the Spectrahedron M.

manifold_volume(M::AbstractSphere{ℝ})
Volume of the -dimensional Sphere M.

volume_density(M::AbstractSphere{ℝ}, p, X)
Compute volume density function of a sphere, i.e. determinant of the differential of exponential map exp(M, p, X).

check_point(M::AbstractSphere, p; kwargs...)
Check whether p is a valid point on the AbstractSphere M, i.e. is a point in the embedding of unit length.

check_point(M::SphereSymmetricMatrices, p; kwargs...)
Check whether the matrix is a valid point on the SphereSymmetricMatrices M, i.e. is an n-by-n symmetric matrix of unit Frobenius norm.

check_point(M::Stiefel, p; kwargs...)
Check whether p is a valid point on the Stiefel M=, i.e. that it has the right AbstractNumbers type and is (approximately) the identity, where is the complex conjugate transpose.

inner(M::MetricManifold{ℝ,<:Stiefel{ℝ},<:StiefelSubmersionMetric}, p, X, Y)
Compute the inner product on the Stiefel manifold with respect to the StiefelSubmersionMetric.

inner(M::MetricManifold{ℝ, Stiefel{ℝ}, X, CanonicalMetric}, p, X, Y)
Compute the inner product on the Stiefel manifold with respect to the CanonicalMetric.

check_point(M::SymmetricMatrices{n,𝔽}, p; kwargs...)
Check whether p is a valid manifold point on the SymmetricMatrices M, i.e. whether p is a symmetric matrix of size (n,n) with values from the corresponding AbstractNumbers 𝔽.

manifold_volume(::SymmetricPositiveDefinite)
Return volume of the SymmetricPositiveDefinite manifold, i.e. infinity.

volume_density(::SymmetricPositiveDefinite, p, X)
Compute the volume density of the SymmetricPositiveDefinite manifold at p in direction X.

check_point(M::SymmetricPositiveDefinite, p; kwargs...)
checks, whether p is a valid point on the SymmetricPositiveDefinite M, i.e. is a matrix of size (N,N), symmetric and positive definite.

inner(::MetricManifold{ℝ,<:SymmetricPositiveDefinite,<:GeneralizedBuresWassersteinMetric}, p, X, Y)
Compute the inner product SymmetricPositiveDefinite with respect to the GeneralizedBuresWassersteinMetric given by

inner(::MetricManifold{ℝ,SymmetricPositiveDefinite,BuresWassersteinMetric}, p, X, Y)
Compute the inner product SymmetricPositiveDefinite with respect to the BuresWassersteinMetric given by

inner(M::MetricManifold{ℝ,<:SymmetricPositiveDefinite,LogCholeskyMetric}, p, X, Y)
Compute the inner product of two matrices X, Y in the tangent space of p on the SymmetricPositiveDefinite manifold M, as a MetricManifold with LogCholeskyMetric.

inner(M::SymmetricPositiveDefinite, p, X, Y) inner(M::MetricManifold{SymmetricPositiveDefinite,AffineInvariantMetric}, p, X, Y)
Compute the inner product of X, Y in the tangent space of p on the SymmetricPositiveDefinite manifold M, as a MetricManifold with AffineInvariantMetric.

check_point(M::SymmetricPositiveSemidefiniteFixedRank, q; kwargs...)
Check whether q is a valid manifold point on the SymmetricPositiveSemidefiniteFixedRank M, i.e. whether p=q*q' is a symmetric matrix of size (n,n) with values from the corresponding AbstractNumbers 𝔽.

rand(::SymplecticMatrices; vector_at=nothing, σ::Real=1.0)
Generate a random point on or a random tangent vector if vector_at is set to a point .

check_point(M::SymplecticMatrices, p; kwargs...)
Check whether p is a valid point on the SymplecticMatrices M=, i.e. that it has the right AbstractNumbers type and is (approximately) the identity, where denotes the symplectic_inverse.

inner(::SymplecticMatrices{ℝ}, p, X, Y)
Compute the canonical Riemannian inner product RealSymplecticMetric

check_point(M::SymplecticGrassmann, p; kwargs...)
Check whether p is a valid point on the SymplecticGrassmann, manifold by verifying that it is a valid representer of an equivalence class of the corresponding SymplecticStiefel manifold.

check_point(M::SymplecticGrassmann, p::ProjectorPoint; kwargs...)
Check whether p is a valid point on the SymplecticGrassmann, , that is a proper symplectic projection:

inner(::SymplecticGrassmann, p, X, Y)
Compute the Riemannian inner product on the SymplecticGrassmann manifold \mathrm{SpGr}`.

check_point(M::SymplecticStiefel, p; kwargs...)
Check whether p is a valid point on the SymplecticStiefel, manifold, that is is the identity, denotes the symplectic_inverse.

inner(M::SymplecticStiefel, p, X. Y)
Compute the Riemannian inner product at of tangent vectors .

check_point(M::EmbeddedTorus, p; kwargs...)
Check whether p is a valid point on the EmbeddedTorus M.

check_point(M::Torus{n},p)
Checks whether p is a valid point on the Torus M, i.e. each of its entries is a valid point on the Circle and the length of x is n.

check_point(M::Tucker, p; kwargs...)
Check whether the multidimensional array or TuckerPoint p is a point on the Tucker manifold, i.e. it is a Dth order N[1] × … × N[D] tensor of multilinear rank (R[1], …, R[D]).

inner(M::Tucker, p::TuckerPoint, X::TuckerTangentVector, Y::TuckerTangentVector)
The Euclidean inner product between tangent vectors X and X at the point p on the Tucker manifold.

inner(B::VectorBundle, p, X, Y)
Inner product of tangent vectors X and Y at point p from the vector bundle B over manifold B.fiber (denoted ).