Product manifold

Product Manifold

Product manifold $M = M_1 \times M_2 \times \dots M_n$ of manifolds $M_1, M_2, \dots, M_n$. Points on the product manifold can be constructed using ProductRepr with canonical projections $\Pi_i \colon M \to M_i$ for $i \in 1, 2, \dots, n$ provided by submanifold_component.

Manifolds.ProductManifold — Type.

ProductManifold{TM<:Tuple, TRanges<:Tuple, TSizes<:Tuple} <: Manifold

Product manifold $M_1 \times M_2 \times \dots \times M_n$ with product geometry. TRanges and TSizes statically define the relationship between representation of the product manifold and representations of point, tangent vectors and cotangent vectors of respective manifolds.

Constructor

ProductManifold(M_1, M_2, ..., M_n)

generates the product manifold $M_1 \times M_2 \times \dots \times M_n$. Alternatively, the same manifold can be contructed using the × operator: M_1 × M_2 × M_3.

Manifolds.is_manifold_point — Method.

is_manifold_point(M::ProductManifold, x; kwargs...)

Check whether x is a valid point on the ProductManifold M.

The tolerance for the last test can be set using the ´kwargs...`.

Manifolds.is_tangent_vector — Method.

is_tangent_vector(M::ProductManifold, x, v; kwargs... )

Check whether v is a tangent vector to x on the ProductManifold M, i.e. atfer is_manifold_point(M, x), and all projections to base manifolds must be respective tangent vectors.

The tolerance for the last test can be set using the ´kwargs...`.

Manifolds.submanifold — Method.

submanifold(M::ProductManifold, i::AbstractVector)

Extract the factor of the product manifold M indicated by indices in i. For example, for i equal to [1, 3] the product manifold constructed from the first and the third factor is returned.

This function is not type-stable, for better preformance use submanifold(M::ProductManifold, i::Val).

Manifolds.submanifold — Method.

submanifold(M::ProductManifold, i::Integer)

Extract the ith factor of the product manifold M.

Manifolds.submanifold — Method.

submanifold(M::ProductManifold, i::Val)

Extract the factor of the product manifold M indicated by indices in i. For example, for i equal to Val((1, 3)) the product manifold constructed from the first and the third factor is returned.

Manifolds.InverseProductRetraction — Method.

InverseProductRetraction(inverse_retractions::AbstractInverseRetractionMethod...)

Product inverse retraction of inverse_retractions. Works on ProductManifold.

Manifolds.ProductFVectorDistribution — Type.

ProductFVectorDistribution([type::VectorBundleFibers], [x], distrs...)

Generates a random vector at point x from vector space (a fiber of a tangent bundle) of type type using the product distribution of given distributions.

Vector space type and x can be automatically inferred from distributions distrs.

Manifolds.ProductPointDistribution — Type.

ProductPointDistribution(M::ProductManifold, distributions)

Product distribution on manifold M, combined from distributions.

Manifolds.ProductRetraction — Type.

ProductRetraction(retractions::AbstractRetractionMethod...)

Product retraction of retractions. Works on ProductManifold.