Rotations

Rotations

The manifold $\mathrm{SO}(n)$ of orthogonal matrices with determinant $+1$ in $\mathbb R^{n\times n}$, i.e.

\[\mathrm{SO}(n) = \bigl\{R \in \mathbb{R}^{n\times n} \big| RR^{\mathrm{T}} = R^{\mathrm{T}}R = \mathrm{I}_n, \det(R) = 1 \bigr\}\]

$\mathrm{SO}(n)$ is a subgroup of the orthogonal group $\mathrm{O}(n)$ and also known as the special orthogonal group or the set of rotations group.

Tangent vectors are represented by elements of the corresponding Lie algebra, which is the tangent space at the identity element. This convention allows for more efficient operations on tangent vectors. Tangent spaces at different points are different vector spaces.

Let $L_R\colon \mathrm{SO}(n) \to \mathrm{SO}(n)$ where $R \in \mathrm{SO}(n)$ be the left-multiplication by $R$, that is $L_R(S) = RS$. The tangent space at rotation $R$, $T_R \mathrm{SO}(n)$, is related to the tangent space at the identity rotation $\mathrm{I}_n$ by the differential of $L_R$ at identity, $(\mathrm{d}L_R)_{\mathrm{I}_n} \colon T_{\mathrm{I}_n} \mathrm{SO}(n) \to T_R \mathrm{SO}(n)$. For a tangent vector at the identity rotation $v \in T_{\mathrm{I}_n} \mathrm{SO}(n)$ the matrix representation of the corresponding tangent vector $w$ at a rotation $R$ can be obtained by matrix multiplication: $w=Rv \in T_R \mathrm{SO}(n)$. You can compare the functions log!(::Manifolds.Rotations, v, x, y) and exp!(::Manifolds.Rotations, y, x, v) to see how it works in practice.

Manifolds.exp! — Method.

exp!(M::Rotations{4}, y, x, v)

Exponential map of tangent vector v at point x from $\mathrm{SO}(4)$ manifold M. Result is saved to y.

The algorithm used is a more numerically stable form of those proposed in [Gallier, 2003] and [Andrica, 2013].

Manifolds.hat! — Method.

hat!(M::Rotations, Ω, x, ω)

Convert the unique tangent vector components $\omega$ at point $x$ on rotations group $\mathrm{SO}(n)$ to the matrix representation $\Omega$ of the tangent vector. See vee! for the conventions used.

Manifolds.inner — Method.

inner(M, x, w, v)

compute the inner product of the two tangent vectors w, v from the tangent plane at x on the special orthogonal space M=$\mathrm{SO}(n)$ using the restriction of the metric from the embedding, i.e.

\[(v, w)_x = \operatorname{tr}(v^T w)\]

.

Tangent vectors are represented by matrices.

Manifolds.inverse_retract! — Method.

inverse_retract!(M, v, x, y, ::PolarInverseRetraction)

Compute a vector from the tagent space $T_x\mathrm{SO}(n)$ of the point x on the Rotations manifold M with which the point y can be reached by the PolarRetraction from the point x after time 1.

The formula reads $v = -\frac{1}{2}(x^{\mathrm{T}}ys - (x^{\mathrm{T}}ys)^{\mathrm{T}})$

where $s$ is the solution to the Sylvester equation

\[x^{\mathrm{T}}ys + s(x^{\mathrm{T}}y)^{\mathrm{T}} + 2\mathrm{I}_n = 0.\]

Manifolds.inverse_retract! — Method.

inverse_retract!(M::Rotations, x, y, ::QRInverseRetraction)

Compute a vector from the tagent space $T_x\mathrm{SO}(n)$ of the point x on the Rotations manifold M with which the point y can be reached by the QRRetraction from the point x after time 1.

Manifolds.is_manifold_point — Method.

is_manifold_point(M,x; kwargs...)

checks, whether x is a valid point on the Rotations M, i.e. is an array of size manifold_dimension(M) and represents a valid rotation. The tolerance for the last test can be set using the ´kwargs...`.

Manifolds.is_tangent_vector — Method.

is_tangent_vector(M,x,v; kwargs... )

checks whether v is a tangent vector to x on the Rotations space M, i.e. after is_manifold_point(M,x), v has to be of same dimension as x and orthogonal to x. The tolerance for the last test can be set using the ´kwargs...`.

Manifolds.log! — Method.

log!(M::Rotations, v, x, y)

compute the logarithmic map on the Rotations manifold M$=\mathrm{SO}(n)$, which is given by

\[\log_{x} y = \frac{1}{2} \bigl(\operatorname{Log}(x^{\mathrm{T}}y) - (\operatorname{Log} x^{\mathrm{T}}y)^{\mathrm{T}}),\]

where $\operatorname{Log}$ denotes the matrix logarithm, and save the result to v.

For antipodal rotations the function returns one of the tangent vectors that point at y.

Manifolds.manifold_dimension — Method.

manifold_dimension(M::Rotations)

returns the dimension of the manifold $\mathrm{SO}(n)$, i.e. $\frac{n(n-1)}{2}$.

Manifolds.retract! — Method.

retract_polar!(M::Rotations, y, x, v, method::PolarRetraction)

Compute the SVD-based retraction PolarRetraction, a second-order approximation of the exponential map.

Manifolds.retract! — Method.

retract!(M, y, x, v, method::QRRetraction)

This QR-based retraction is a first-order approximation of the exponential map.

Manifolds.vee! — Method.

vee!(M::Rotations, ω, x, Ω)

Extract the unique tangent vector components $\omega$ at point $x$ on rotations group $\mathrm{SO}(n)$ from the matrix representation $\Omega$ of the tangent vector.

The basis on the Lie algebra $\mathfrak{so}(n)$ is chosen such that for $\mathrm{SO}(2)$, $\omega=\theta=\Omega_{21}$ is the angle of rotation, and for $\mathrm{SO}(3)$, $\omega = (\Omega_{32}, \Omega_{13}, \Omega_{21}) = \theta u$ is the angular velocity and axis-angle representation, where $u$ is the unit vector along the axis of rotation.

For $\mathrm{SO}(n)$ where $n \ge 4$, the additional elements of $\omega$ are $\omega_{i (i - 3)/2 + j + 1} = \Omega_{ij}$, for $i \in [4, n], j \in [1,i)$.

Manifolds.NormalRotationDistribution — Type.

NormalRotationDistribution(M::Rotations, d::Distribution, x::TResult)

Distribution that returns a random point on the manifold Rotations M. Random point is generated using base distribution d and the type of the result is adjusted to TResult.

See normal_rotation_distribution for details.

Manifolds.PolarInverseRetraction — Type.

PolarInverseRetraction

Inverse retraction on the rotations manifold using the polar method.

Manifolds.PolarRetraction — Type.

PolarRetraction

Retraction on the rotations manifold using the polar method. This SVD-based retraction is a second-order approximation of the exponential map. Let

\[USV = x + xv\]

be the singular value decomposition, then the formula reads

\[\operatorname{retr}_x v = UV^\mathrm{T}\]

Retraction is performed by the function retract!(::Rotations, y, x, v, ::PolarRetraction)

Manifolds.QRInverseRetraction — Type.

QRInverseRetraction

Inverse retraction on the rotations manifold using the QR method.

Manifolds.QRRetraction — Type.

QRRetraction

Retraction on the rotations manifold using the QR method, a first order approximation of the exponential map.

Manifolds.Rotations — Type.

Rotations{N} <: Manifold

Special orthogonal manifold $\mathrm{SO}(n)$ represented by $n\times n$ real-valued orthogonal matrices with determinant $+1$.

Constructor

Rotations(n)

generates the $\mathrm{SO}(n) \subset \mathbb R^{n\times n}$

Manifolds.angles_4d_skew_sym_matrix — Method.

angles_4d_skew_sym_matrix(A)

The Lie algebra of $\mathrm{SO}(4)$ consists of 4x4 skew-symmetric matrices. The unique imaginary components of their eigenvalues are the angles of the two plane rotations. This function computes these more efficiently than eigvals.

By convention, the returned values are sorted in decreasing order (corresponding to the same ordering of angles as cos_angles_4d_rotation_matrix).

Manifolds.cos_angles_4d_rotation_matrix — Method.

cos_angles_4d_rotation_matrix(R)

4D rotations can be described by two orthogonal planes that are unchanged by the action of the rotation (vectors within a plane rotate only within the plane). The cosines of the two angles of rotation about these planes may be obtained from the distinct real parts of the eigenvalues of the rotation matrix. This function computes these more efficiently by solving the system

\[\begin{aligned} \cos\alpha + \cos\beta &= \frac{1}{2} \operatorname{tr}(R)\\ \cos\alpha + \cos\beta &= \frac{1}{8} \operatorname{tr}(R)^2 - \frac{1}{16} \operatorname{tr}((R - R^T)^2) - 1. \end{aligned}\]

By convention, the returned values are sorted in increasing order. See angles_4d_skew_sym_matrix. For derivation of the above, see [Gallier, 2013].

Manifolds.normal_rotation_distribution — Method.

normal_rotation_distribution(M::Rotations, x, σ::Real)

return a random point on the manifold Rotations M by generating a (Gaussian) random orthogonal matrix with determinant $+1$. Let

\[QR = A\]

be the QR decomposition of a random matrix $A$, then the formula reads

\[x = QD\]

where $D$ is a diagonal matrix with the signs of the diagonal entries of $R$, i.e.

\[D_{ij}=\begin{cases} \operatorname{sgn}(R_{ij}) & \text{if} \; i=j \\ 0 & \, \text{otherwise} \end{cases}.\]

It can happen that the matrix gets -1 as a determinant. In this case, the first and second columns are swapped.

The argument x is used to determine the type of returned points.

Manifolds.normal_tvector_distribution — Method.

normal_tvector_distribution(M::Rotations, x, σ)

Normal distribution in ambient space with standard deviation σ projected to tangent space at x.

Literature

[Gallier, 2003] J. Gallier and D. Xu, “Computing exponentials of skew-symmetric matrices and logarithms of orthogonal matrices,” International Journal of Robotics and Automation, vol. 18, no. 1, pp. 10–20, 2003.

[Andrica, 2013] D. Andrica and R.-A. Rohan, “Computing the Rodrigues coefficients of the exponential map of the Lie groups of matrices,” Balkan Journal of Geometry and Its Applications, vol. 18, no. 2, pp. 1–10, 2013.