Rotations

The manifold $\mathrm{SO}(n)$ of orthogonal matrices with determinant $+1$ in $\mathbb R^{n\times n}$, i.e.

\[\mathrm{SO}(n) = \bigl\{R \in \mathbb{R}^{n\times n} \big| RR^{\mathrm{T}} = R^{\mathrm{T}}R = \mathrm{I}_n, \det(R) = 1 \bigr\}\]

$\mathrm{SO}(n)$ is a subgroup of the orthogonal group $\mathrm{O}(n)$ and also known as the special orthogonal group or the set of rotations group.

Tangent vectors are represented by elements of the corresponding Lie algebra, which is the tangent space at the identity element. This convention allows for more efficient operations on tangent vectors. Tangent spaces at different points are different vector spaces.

Let $L_R\colon \mathrm{SO}(n) \to \mathrm{SO}(n)$ where $R \in \mathrm{SO}(n)$ be the left-multiplication by $R$, that is $L_R(S) = RS$. The tangent space at rotation $R$, $T_R \mathrm{SO}(n)$, is related to the tangent space at the identity rotation $\mathrm{I}_n$ by the differential of $L_R$ at identity, $(\mathrm{d}L_R)_{\mathrm{I}_n} \colon T_{\mathrm{I}_n} \mathrm{SO}(n) \to T_R \mathrm{SO}(n)$. For a tangent vector at the identity rotation $v \in T_{\mathrm{I}_n} \mathrm{SO}(n)$ the matrix representation of the corresponding tangent vector $w$ at a rotation $R$ can be obtained by matrix multiplication: $w=Rv \in T_R \mathrm{SO}(n)$. You can compare the functions log!(::Manifolds.Rotations, v, x, y) and exp!(::Manifolds.Rotations, y, x, v) to see how it works in practice.

Manifolds.Rotations — Type

Rotations{N} <: Manifold

Special orthogonal manifold $\mathrm{SO}(n)$ represented by $n\times n$ real-valued orthogonal matrices with determinant $+1$.

Constructor

Rotations(n)

Generate the $\mathrm{SO}(n) \subset \mathbb R^{n\times n}$

Base.exp — Method

exp(M::Rotations, x, v)

Compute the exponential map on the Rotations from x into direction v, i.e.

\[\exp_xv = x \operatorname{Exp}(v),\]

where $\operatorname{Exp}(v)$ denotes the matrix exponential of $v$.

Base.exp — Method

exp(M::Rotations{4}, x, v)

Compute the exponential map of tangent vector v at point x from $\mathrm{SO}(4)$ manifold M.

The algorithm used is a more numerically stable form of those proposed in [Gallier2002] and [Andrica2013].

Base.log — Method

log(M::Rotations, x, y)

Compute the logarithmic map on the Rotations manifold M$=\mathrm{SO}(n)$, which is given by

\[\log_{x} y = \frac{1}{2} \bigl(\operatorname{Log}(x^{\mathrm{T}}y) - (\operatorname{Log} x^{\mathrm{T}}y)^{\mathrm{T}}),\]

where $\operatorname{Log}$ denotes the matrix logarithm.

For antipodal rotations the function returns deterministically one of the tangent vectors that point at y.

LinearAlgebra.norm — Method

norm(M::Rotations, x, v)

Compute the norm of a tangent vector v from the tangent space at x on the Rotations M. The formula reads

\[\lVert v \rVert_x = \lVert v \rVert,\]

i.e. the Frobenius norm of v, where tangent vectors are represented by elements from the Lie group.

Manifolds.hat! — Method

hat!(M::Rotations, Ω, x, ω)

Convert the unique tangent vector components $\omega$ at point $x$ on rotations group $\mathrm{SO}(n)$ to the matrix representation $\Omega$ of the tangent vector. See vee! for the conventions used.

Manifolds.normal_tvector_distribution — Method

normal_tvector_distribution(M::Rotations, x, σ)

Normal distribution in ambient space with standard deviation σ projected to tangent space at x.

Manifolds.vee! — Method

vee!(M::Rotations, ω, x, Ω)

Extract the unique tangent vector components $\omega$ at point $x$ on rotations group $\mathrm{SO}(n)$ from the matrix representation $\Omega$ of the tangent vector.

The basis on the Lie algebra $\mathfrak{so}(n)$ is chosen such that for $\mathrm{SO}(2)$, $\omega=\theta=\Omega_{21}$ is the angle of rotation, and for $\mathrm{SO}(3)$, $\omega = (\Omega_{32}, \Omega_{13}, \Omega_{21}) = \theta u$ is the angular velocity and axis-angle representation, where $u$ is the unit vector along the axis of rotation.

For $\mathrm{SO}(n)$ where $n \ge 4$, the additional elements of $\omega$ are $\omega_{i (i - 3)/2 + j + 1} = \Omega_{ij}$, for $i \in [4, n], j \in [1,i)$.

ManifoldsBase.injectivity_radius — Method

injectivity_radius(M::Rotations, x, ::PolarRetraction)

Return the radius of injectivity for the PolarRetraction on the Rotations M which is $\frac{\pi}{\sqrt{2}}$.

ManifoldsBase.injectivity_radius — Method

injectivity_radius(M::Rotations)
injectivity_radius(M::Rotations, x)

Return the injectivity radius on the Rotations M, which is globally

\[ \operatorname{inj}_{\mathrm{SO}(n)}(x) = \pi\sqrt{2}.\]

ManifoldsBase.inner — Method

inner(M::Rotations, x, w, v)

Compute the inner product of the two tangent vectors w, v from the tangent plane at x on the special orthogonal space M=$\mathrm{SO}(n)$ using the restriction of the metric from the embedding, i.e.

\[(v, w)_x = \operatorname{tr}(v^T w)\]

.

Tangent vectors are represented by matrices.

ManifoldsBase.inverse_retract — Method

inverse_retract(M, x, y, ::PolarInverseRetraction)

Compute a vector from the tagent space $T_x\mathrm{SO}(n)$ of the point x on the Rotations manifold M with which the point y can be reached by the PolarRetraction from the point x after time 1.

The formula reads

\[\operatorname{retr}^{-1}_x(y) = -\frac{1}{2}(x^{\mathrm{T}}ys - (x^{\mathrm{T}}ys)^{\mathrm{T}})\]

where $s$ is the solution to the Sylvester equation

\[x^{\mathrm{T}}ys + s(x^{\mathrm{T}}y)^{\mathrm{T}} + 2\mathrm{I}_n = 0.\]

ManifoldsBase.inverse_retract — Method

inverse_retract(M::Rotations, x, y, ::QRInverseRetraction)

Compute a vector from the tagent space $T_x\mathrm{SO}(n)$ of the point x on the Rotations manifold M with which the point y can be reached by the QRRetraction from the point x after time 1.

ManifoldsBase.manifold_dimension — Method

manifold_dimension(M::Rotations)

Return the dimension of the manifold $\mathrm{SO}(n)$, i.e. $\frac{n(n-1)}{2}$.

ManifoldsBase.project_point — Method

project_point(M::Rotations, x; check_det = true)

Project x to the nearest point on manifold M.

Given the singular value decomposition $x = U \Sigma V^\mathrm{T}$, with the singular values sorted in descending order, the projection is

\[\operatorname{proj}_{\mathrm{SO}(n)}(x) = U\operatorname{diag}\left[1,1,\dots,\det(U V^\mathrm{T})\right] V^\mathrm{T}\]

The diagonal matrix ensures that the determinant of the result is $+1$. If x is expected to be almost special orthogonal, then you may avoid this check with check_det = false.

ManifoldsBase.project_tangent — Method

project_tangent(M::Rotations, x, v)

Project the matrix v onto the tangent space by making v skew symmetric,

\[\operatorname{proj}_x(v) = \frac{v-v^{\mathrm{T}}}{2},\]

where tangent vectors are represented by elements from the Lie group

ManifoldsBase.representation_size — Method

representation_size(M::Rotations)

Return the size() of a point on the Rotations M, i.e. for the $\mathrm{SO}(n)$ it's (n,n).

ManifoldsBase.retract — Method

retract(M::Rotations, x, v, ::PolarRetraction)

Compute the SVD-based retraction on the Rotations M from x in direction v (as an element of the Lie group) and is a second-order approximation of the exponential map. Let

\[USV = x + xv\]

be the singular value decomposition, then the formula reads

\[\operatorname{retr}_x v = UV^\mathrm{T}.\]

ManifoldsBase.retract — Method

retract(M, x, v)
retract(M, x, v, ::QRRetraction)

Compute the SVD-based retraction on the Rotations M from x in direction v (as an element of the Lie group) and is a first-order approximation of the exponential map.

This is also the default retraction on the Rotations

ManifoldsBase.zero_tangent_vector — Method

zero_tangent_vector(M::Rotations, x)

Return the zero tangent vector from the tangent space art x on the Rotations as an element of the Lie group, i.e. the zero matrix.

Statistics.mean — Method

mean(
    M::Rotations,
    x::AbstractVector,
    [w::AbstractWeights,]
    method = GeodesicInterpolationWithinRadius(π/2/√2);
    kwargs...,
)

Compute the Riemannian mean of x using GeodesicInterpolationWithinRadius.

Manifolds.NormalRotationDistribution — Type

NormalRotationDistribution(M::Rotations, d::Distribution, x::TResult)

Distribution that returns a random point on the manifold Rotations M. Random point is generated using base distribution d and the type of the result is adjusted to TResult.

See normal_rotation_distribution for details.

Manifolds.angles_4d_skew_sym_matrix — Method

angles_4d_skew_sym_matrix(A)

The Lie algebra of $\mathrm{SO}(4)$ consists of 4x4 skew-symmetric matrices. The unique imaginary components of their eigenvalues are the angles of the two plane rotations. This function computes these more efficiently than eigvals.

By convention, the returned values are sorted in decreasing order (corresponding to the same ordering of angles as cos_angles_4d_rotation_matrix).

Manifolds.cos_angles_4d_rotation_matrix — Method

cos_angles_4d_rotation_matrix(R)

4D rotations can be described by two orthogonal planes that are unchanged by the action of the rotation (vectors within a plane rotate only within the plane). The cosines of the two angles of rotation about these planes may be obtained from the distinct real parts of the eigenvalues of the rotation matrix. This function computes these more efficiently by solving the system

\[\begin{aligned} \cos\alpha + \cos\beta &= \frac{1}{2} \operatorname{tr}(R)\\ \cos\alpha + \cos\beta &= \frac{1}{8} \operatorname{tr}(R)^2 - \frac{1}{16} \operatorname{tr}((R - R^T)^2) - 1. \end{aligned}\]

By convention, the returned values are sorted in increasing order. See angles_4d_skew_sym_matrix. For derivation of the above, see [Gallier, 2013].

Manifolds.normal_rotation_distribution — Method

normal_rotation_distribution(M::Rotations, x, σ::Real)

Return a random point on the manifold Rotations M by generating a (Gaussian) random orthogonal matrix with determinant $+1$. Let

\[QR = A\]

be the QR decomposition of a random matrix $A$, then the formula reads

\[x = QD\]

where $D$ is a diagonal matrix with the signs of the diagonal entries of $R$, i.e.

\[D_{ij}=\begin{cases} \operatorname{sgn}(R_{ij}) & \text{if} \; i=j \\ 0 & \, \text{otherwise} \end{cases}.\]

It can happen that the matrix gets -1 as a determinant. In this case, the first and second columns are swapped.

The argument x is used to determine the type of returned points.

ManifoldsBase.check_manifold_point — Method

check_manifold_point(M,x; kwargs...)

Check whether x is a valid point on the Rotations M, i.e. is an array of size manifold_dimension(M) and represents a valid rotation. The tolerance for the last test can be set using the kwargs....

ManifoldsBase.check_tangent_vector — Method

check_tangent_vector(M,x,v; kwargs... )

Check whether v is a tangent vector to x on the Rotations space M, i.e. after check_manifold_point(M,x), v has to be of same dimension as x and orthogonal to x. The tolerance for the last test can be set using the kwargs....

Literature

Gallier2002

Gallier J.; Xu D.; Computing exponentials of skew-symmetric matrices and logarithms of orthogonal matrices. International Journal of Robotics and Automation (2002), 17(4), pp. 1-11. pdf.

Andrica2013

Andrica D.; Rohan R.-A.; Computing the Rodrigues coefficients of the exponential map of the Lie groups of matrices. Balkan Journal of Geometry and Its Applications (2013), 18(2), pp. 1-2. pdf.