The Segre manifold

Manifolds.Segre ÔÇö Type

Segre{­ŁöŻ,V} <: AbstractManifold{­ŁöŻ}

The Segre manifold

\[\mathcal{S} = \operatorname{Seg}(­ŁöŻ^{n_1} \times \dots \times ­ŁöŻ^{n_d})\]

is the set of rank-one tensors in $­ŁöŻ^{n_1} \otimes \dots \otimes ­ŁöŻ^{n_d}$.

When $­ŁöŻ = ÔäŁ$, the Segre manifold is a normal Riemannian covering of

\[\mathcal{P} = ÔäŁ^{+} \times \mathbb{S}^{n_1 - 1} \times \dots \times \mathbb{S}^{n_d - 1}\]

equipped with a warped product metric. The tuple $(n_1, \dots, n_d)$ is called the valence of the manifold.

The geometry of the Segre manifold is summarized in [JSVV24]. It is named after Corrado Segre(1863ÔÇô1924).

Constructor

Segre(n::Int...; field::AbstractNumbers=ÔäŁ)

Generate a valence (n, ...) Segre manifold. Segre(n) is the same as $\mathbb{R}^{n} \setminus \{ 0 \}$.

Base.exp ÔÇö Method

exp(M::Segre{ÔäŁ, V}, p, X)

Exponential map on the Segre manifold.

Let $p ÔëÉ (╬╗, x_1, ÔÇŽ, x_d) Ôłł \mathcal{S}$ and $X = (╬Ż, u_1, ÔÇŽ, u_d) Ôłł T_p \mathcal{S}$. The exponential map is given by

\[ \operatorname{exp}_p(X) ÔëÉ \left( \sqrt{(╬╗ + ╬Ż)^2 + (╬╗ m)^2},\\ x_1 \cos\mathopen{\Big(} \frac{f \lVert u_1 \rVert_{x_1}}{m} \mathclose{\Big)} + \frac{u_1}{\lVert u_1 \rVert_{x_1}} \sin\mathopen{\Big(} \frac{f \lVert u_1 \rVert_{x_1}}{m} \mathclose{\Big)},\\ ÔÇŽ,\\ x_d \cos\mathopen{\Big(} \frac{f \lVert u_d \rVert_{x_d}}{m} \mathclose{\Big)} + \frac{u_d}{\lVert u_d \rVert_{x_d}} \sin\mathopen{\Big(} \frac{f \lVert u_d \rVert_{x_d}}{m} \mathclose{\Big)} \right),\]

where

\[ \begin{aligned} f &= \frac{\pi}{2} - \tan^{-1}\mathopen{\Big(} \frac{╬╗ + ╬Ż}{╬╗ m} \mathclose{\Big)},\\ m &= \sqrt{\lVert u_1 \rVert_{x_1}^2 + ÔÇŽ + \lVert u_d \rVert_{x_d}^2}. \end{aligned}\]

If $m = 0$ and $-╬╗ < ╬Ż$, then $\operatorname{exp}_p(v) = p + X$.

The formula is derived in proposition 3.1 in [JSVV24].

Base.log ÔÇö Method

log(M::Segre{ÔäŁ, V}, p, q)

Logarithmic map on the Segre manifold.

Let $p ÔëÉ (╬╗, x_1, ÔÇŽ, x_d)$, $q ÔëÉ (╬╝, y_1, ÔÇŽ, y_d) Ôłł \mathcal{S}$. Assume $p$ and $q$ are connected by a geodesic.

Let

\[ m = \sqrt{\sphericalangle(x_1, y_1)^2 + ÔÇŽ + \sphericalangle(x_d, y_d)^2}\]

and assume $(╬╝, y_1, ÔÇŽ, y_d)$ is the representative of $q$ that minimizes $m$. Then

\[ \operatorname{log}_p(q) = \left( \mu \cos{m} - \lambda, (y_1 - ÔčĘx_1, y_1Ôčę x_1) \frac{\mu \sphericalangle(x_1, y_1) \sin{m}}{\lambda m \sin{\sphericalangle(x_1, y_1)}}, \dots, (y_d - ÔčĘx_d, y_dÔčę x_d) \frac{\mu \sphericalangle(x_d, y_d) \sin{m}}{\lambda m \sin{\sphericalangle(x_d, y_d)}} \right).\]

The formula is derived in theorem 4.4 in [JSVV24].

Base.rand ÔÇö Method

rand(M::Segre{ÔäŁ, V}; vector_at=nothing)

If vector_at is nothing, return a random point on

\[ ÔäŁ^{+} ├Ś \mathbb{S}^{n_1 - 1} ├ŚÔő»├Ś \mathbb{S}^{n_d - 1}\]

from a log-normal distribution on $ÔäŁ^{+}$ and a uniform distribution on $\mathbb{S}^{n_1 - 1} ├ŚÔő»├Ś \mathbb{S}^{n_d - 1}$.

If vector_at is not nothing, return a random tangent vector from a normal distribution on the tangent space.

Manifolds.closest_representative! ÔÇö Method

closest_representative!(M::Segre{ÔäŁ, V}, p, q)

$\mathcal{S}$ is a $2^d$-sheeted Riemannian covering of

\[\mathcal{P} = ÔäŁ^{+} \times \mathbb{S}^{n_1 - 1} \times \dots \times \mathbb{S}^{n_d - 1}\]

with a warped product metric. Every equivalence class $q \in \mathcal{S}$ has $2^d$ representatives in $\mathcal{P}$. closest_representative!(M, q, p) changes representative of q to the one that is closest to p in $\mathcal{P}$.

Manifolds.connected_by_geodesic ÔÇö Method

connected_by_geodesic(M::Segre{ÔäŁ, V}, p, q)

$\mathcal{S}$ is not a complete manifold, i.e. not every pair p and q of points are connected by a geodesic in $\mathcal{S}$. connected_by_geodesic(M, p, q) returns true if two points, p and q, are connected by a geodesic, and otherwise returns false.

Manifolds.spherical_angle_sum ÔÇö Method

spherical_angle_sum(M::Segre{ÔäŁ, V}, p, q)

Let $p ÔëÉ (╬╗, x_1, ÔÇŽ, x_d)$, $q ÔëÉ (╬╝, y_1, ÔÇŽ, y_d) Ôłł \mathcal{S}$. Then this is

\[ \sqrt{\sphericalangle(x_1, y_1)^2 + ÔÇŽ + \sphericalangle(x_d, y_d)^2},\]

where $\sphericalangle(x_i, y_i)$ is the distance between $x_i$ and $y_i$ on the sphere $\mathbb{S}^{n_i - 1}$.

ManifoldsBase.distance ÔÇö Method

distance(M::Segre{ÔäŁ, V}, p, q)

Riemannian distance between two points p and q on the Segre manifold.

Assume $p ÔëÉ (╬╗, x_1, ÔÇŽ, x_d)$, $q ÔëÉ (╬╝, y_1, ÔÇŽ, y_d) Ôłł \mathcal{S}$ are connected by a geodesic. Let

\[ m = \sqrt{\sphericalangle(x_1, y_1)^2 + ÔÇŽ + \sphericalangle(x_d, y_d)^2}\]

and assume $(╬╝, y_1, ÔÇŽ, y_d)$ is the representation of $q$ that minimizes $m$. Then

\[ \operatorname{dist}_{\mathcal{S}}(p, q) = \sqrt{╬╗^2 - 2 ╬╗╬╝\cos(m) + ╬╝^2}.\]

ManifoldsBase.embed! ÔÇö Method

embed!(M::Segre{­ŁöŻ, V}, p, X)

Embed tangent vector $X = (╬Ż, u_1, ÔÇŽ, u_d)$ at $p ÔëÉ (╬╗, x_1, ÔÇŽ, x_d)$ in $­ŁöŻ^{n_1 ├ŚÔő»├Ś n_d}$ using the Kronecker product

\[ (╬Ż, u_1, ÔÇŽ, u_d) ÔćŽ ╬Ż x_1 ÔŐŚÔő»ÔŐŚ x_d + ╬╗ u_1 ÔŐŚ x_2 ÔŐŚÔő»ÔŐŚ x_d + ÔÇŽ + ╬╗ x_1 ÔŐŚÔő»ÔŐŚ x_{d - 1} ÔŐŚ u_d.\]

ManifoldsBase.embed ÔÇö Method

embed(M::Segre{­ŁöŻ, V}, p)
embed!(M::Segre{­ŁöŻ, V}, q, p)

Embed $p ÔëÉ (╬╗, x_1, ÔÇŽ, x_d)$ in $­ŁöŻ^{n_1 ├ŚÔő»├Ś n_d}$ using the Kronecker product

\[(╬╗, x_1, ÔÇŽ, x_d) ÔćŽ ╬╗ x_1 ÔŐŚÔő»ÔŐŚ x_d.\]

ManifoldsBase.get_coordinates ÔÇö Method

get_coordinates(M::Segre{­ŁöŻ, V}, p, X, ::DefaultOrthonormalBasis; kwargs...)

Get coordinates of X in the tangent space $T_{(╬╗, x_1, ÔÇŽ, x_d)} \mathcal{S} = \mathbb{R} ├Ś T_{x_1} \mathbb{S}^{n_1 - 1} ├ŚÔÇŽ├Ś T_{x_d} \mathbb{S}^{n_d - 1}$ using a DefaultOrthonormalBasis on each factor.

ManifoldsBase.get_embedding ÔÇö Method

get_embedding(M::Segre{­ŁöŻ,V})

Return the embedding of the Segre manifold $\mathcal{S}$, which is $­ŁöŻ^{n_1 ├ŚÔő»├Ś n_d}$.

ManifoldsBase.get_vector ÔÇö Method

get_vector( M::Segre{­ŁöŻ, V}, p, c, DefaultOrthonormalBasis; kwargs...)

Get tangent vector X from coordinates in the tangent space $T_{(╬╗, x_1, ÔÇŽ, x_d)} \mathcal{S} = \mathbb{R} ├Ś T_{x_1} \mathbb{S}^{n_1 - 1} ├ŚÔő»├Ś T_{x_d} \mathbb{S}^{n_d - 1}$ using a DefaultOrthonormalBasis on each factor.

ManifoldsBase.inner ÔÇö Method

inner(M::Segre{ÔäŁ, V}, p, X, Y,)

Inner product between two tangent vectors $X = (╬Ż, u_1, ÔÇŽ, u_d)$ and $Y = (╬ż, v_1, ÔÇŽ, v_d)$ at $p ÔëÉ (╬╗, x_1, \dots, x_d)$. This inner product is obtained by embedding the Segre manifold in the space of tensors equipped with the Euclidean metric:

\[ \langle X, Y \rangle_{p} = \nu \xi + \lambda^2 (\langle u_1, v_1 \rangle_{x_1} + \dots + \langle u_d, v_d \rangle_{x_d}),\]

where $╬Ż, ╬ż Ôłł T_{╬╗} ÔäŁ^{+} = ÔäŁ$ and $u_i$, $v_i Ôłł T_{x_i} \mathbb{S}^{n_i - 1} ÔŐé ÔäŁ^{n_i}$.

ManifoldsBase.is_point ÔÇö Method

is_point(M::Segre{ÔäŁ, V}, p; kwargs...)

Check whether p is a valid point on M, i.e. p[1] is a singleton containing a positive number and p[i + 1] is a point on Sphere(V[i]). The tolerance can be set using the kwargs....

ManifoldsBase.is_vector ÔÇö Method

is_vector(M::Segre{ÔäŁ, V}, p, X, kwargs...)

Check whether X is a tangent vector to p on M, i.e. X has to be of same dimension as p and orthogonal to p. The tolerance can be set using the kwargs....

ManifoldsBase.riemann_tensor ÔÇö Method

riemann_tensor(M::Segre{ÔäŁ, V}, p, X, Y, Z)

Riemann tensor of the Segre manifold at $p$.

$\mathcal{S}$ is locally a warped product of $ÔäŁ^{+}$ and $\mathbb{S}^{n_1 - 1} ├ŚÔő»├Ś \mathbb{S}^{n_d - 1}$. If $p ÔëÉ (╬╗, x_1, ÔÇŽ, x_d) Ôłł \mathcal{S}$ and $X$, $Y$, $Z Ôłł T_p (\mathbb{S}^{n_1 - 1} ├ŚÔő»├Ś \mathbb{S}^{n_d - 1}) ÔŐé T_p \mathcal{S}$, then

\[ R_{\mathcal{S}}(X, Y) Z = R_{\mathbb{S}^{n_1 - 1} ├ŚÔő»├Ś \mathbb{S}^{n_d - 1}}(X, Y) Z + ╬╗^{-2}(ÔčĘX, ZÔčę_p Y - ÔčĘY, ZÔčę_p X).\]

$R_{\mathcal{S}}$ is zero in the remaining (orthogonal) directions.

ManifoldsBase.sectional_curvature ÔÇö Method

sectional_curvature(M::Segre{ÔäŁ, V}, p, u, v)

Sectional curvature of the Segre manifold at $p$.

$\mathcal{S}$ is locally a warped product of $ÔäŁ^{+}$ and $\mathbb{S}^{n_1 - 1} ├ŚÔő»├Ś \mathbb{S}^{n_d - 1}$ If $p ÔëÉ (╬╗, x_1, ÔÇŽ, x_d) Ôłł \mathcal{S}$, $u_i Ôłł T_{x_i} \mathbb{S}^{n_i - 1}$, and $v_j Ôłł T_{x_j} \mathbb{S}^{n_j - 1}$, then

\[ K_{\mathcal{S}}(u_i, v_j) = \frac{\delta_{i j} - 1}{\lambda^2}.\]

$K_{\mathcal{S}}$ is zero in the remaining (orthogonal) directions.

ManifoldsBase.vector_transport_to ÔÇö Method

vector_transport_to(M::Segre{ÔäŁ, V}, Y, p, X, q, ::ProjectionTransport)

Compute projection vector transport on the Segre manifold by projecting the embedded tangent vector $D\Phi_p[X]$ onto the tangent space of M at q.

Let

\[p = (╬╝, y_1,ÔÇŽ,y_d), \qquad X = (╬Ż,u_1,ÔÇŽ,u_d) Ôłł T_pM,\]

and let $q = (╬╗,x_1,ÔÇŽ,x_d)$.

Under the Segre parametrization

\[╬Ž(╬╗,x_1,ÔÇŽ,x_d) = ╬╗\, x_1 ÔŐŚÔő»ÔŐŚ x_d,\]

the embedded tangent vector at $p$ is

\[D╬Ž_p[X] = ╬Ż\, y_1 ÔŐŚÔő»ÔŐŚ y_d + ╬╝ \sum_{k=1}^d y_1 ÔŐŚÔő»ÔŐŚ u_k ÔŐŚÔő»ÔŐŚ y_d .\]

Projection vector transport from $T_pM$ to $T_qM$ is given by

\[Y = ╬á_{T_qM}\bigl(D╬Ž_p[X]\bigr),\]

where

\[Y = (\dot{\lambda},v_1,ÔÇŽ,v_d) Ôłł T_qM.\]

For each rank-one term $c\,z_1 ÔŐŚÔő»ÔŐŚ z_d$, its contribution to $Y$ is

\[\dot{\lambda} = \sum_{r=0}^{d} c_r \prod_{j=1}^{d} ÔčĘ z_{r,j},\, x_j Ôčę\]

and, for each $k=1,ÔÇŽ,d$,

\[v_k = \sum_{r=0}^{d} \frac{c_r}{\lambda} \left( ÔłĆ_{jÔëák} ÔčĘ z_{r,j}, x_j Ôčę \right) \left( z_{r,k} - ÔčĘ z_{r,k}, x_kÔčę x_k \right).\]

The implementation uses this rank-one structure and therefore avoids explicitly forming the ambient tensor.

A warped metric

Manifolds.WarpedMetric ÔÇö Type

WarpedMetric{A} <: AbstractMetric

The $A$-warped metric on the Segre manifold $\mathcal{S}$ is a generalization of the Euclidean metric on $\mathcal{S}$. We denote this manifold by $\mathcal{S}_A$.

Similarly to $\mathcal{S}$, when $­ŁöŻ = ÔäŁ$, $\mathcal{S}_A$ is a normal Riemannian covering of the product manifold

\[ ÔäŁ^{+} ├Ś \mathbb{S}^{n_1 - 1} ├ŚÔő»├Ś \mathbb{S}^{n_d - 1}\]

with a warped product metric, but the warping function now depends on the warping factor $A$. $A = 1$ corresponds to the usual Segre manifold. The Segre manifold is a cone in the sense that if $p \in \mathcal{S}$, then $r p \in \mathcal{S}$ for all $r \neq 0$. The tangent subspace at $p$ defined $\mathrm{d} (r p) / \mathrm{d} r$ is called the radial direction. $A < 1$ puts less weight on the directions orthogonal to the radial direction compared to $\mathcal{S}$, while $A > 1$ puts more weight on those directions.

The geometry is summarized in [JSVV24].

Constructor

WarpedMetric(A::Real)

Generate a warped product metric with warping factor A.

Base.exp ÔÇö Method

exp(M::MetricManifold{ÔäŁ, Segre{ÔäŁ, V}, WarpedMetric{A}}, p, X)

Exponential map on the warped Segre manifold.

Let $p ÔëÉ (╬╗, x_1,ÔÇŽ, x_d) Ôłł \mathcal{S}_A$ and $X = (╬Ż, u_1,ÔÇŽ, u_d) Ôłł T_p \mathcal{S}_A$. Then the exponential map is given by

\[ \operatorname{exp}_p(X) ÔëÉ \left( \sqrt{(╬╗ + ╬Ż)^2 + (╬╗ A m)^2},\\ x_1 \cos\mathopen{\Big(} \frac{f \lVert u_1 \rVert_{x_1}}{A m} \mathclose{\Big)} + \frac{u_1}{\lVert u_1 \rVert_{x_1}} \sin\mathopen{\Big(} \frac{f \lVert u_1 \rVert_{x_1}}{A m} \mathclose{\Big)},\\ ÔÇŽ,\\ x_d \cos\mathopen{\Big(} \frac{f \lVert u_d \rVert_{x_d}}{A m} \mathclose{\Big)} + \frac{u_d}{\lVert u_d \rVert_{x_d}} \sin\mathopen{\Big(} \frac{f \lVert u_d \rVert_{x_d}}{A m} \mathclose{\Big)} \right),\]

where

\[ \begin{aligned} f &= \frac{\pi}{2} - \tan^{-1}\mathopen{\Big(} \frac{╬╗ + ╬Ż}{╬╗ A m} \mathclose{\Big)},\\ m &= \sqrt{\lVert u_1 \rVert_{x_1}^2 + ÔÇŽ + \lVert u_d \rVert_{x_d}^2}. \end{aligned}\]

If $m = 0$ and $-╬╗ < ╬Ż$, then $\operatorname{exp}_p(v) = p + X$.

The formula is derived in proposition 3.1 in [JSVV24].

Base.log ÔÇö Method

log(M::MetricManifold{ÔäŁ, Segre{ÔäŁ,V}, WarpedMetric{A}}, p, q)

Logarithmic map on the warped Segre manifold.

Let $p ÔëÉ (╬╗, x_1,ÔÇŽ, x_d)$, $q ÔëÉ (╬╝, y_1,ÔÇŽ, y_d) Ôłł \mathcal{S}_A$. Assume $p$ and $q$ are connected by a geodesic. Let

\[ m = \sqrt{\sphericalangle(x_1, y_1)^2 +ÔÇŽ + \sphericalangle(x_d, y_d)^2}\]

and assume $(╬╝, y_1,ÔÇŽ, y_d)$ is the representative of $q$ that minimizes $m$. Then

\[ \operatorname{log}_p(q) = \left( \mu \cos{m} - \lambda, (y_1 - ÔčĘx_1, y_1Ôčę x_1) \frac{\mu \sphericalangle(x_1, y_1) \sin(A m)}{\lambda A m \sin{\sphericalangle(x_1, y_1)}}, \dots, (y_d - ÔčĘx_d, y_dÔčę x_d) \frac{\mu \sphericalangle(x_d, y_d) \sin(A m)}{\lambda A m \sin{\sphericalangle(x_d, y_d)}} \right).\]

The formula is derived in theorem 4.4 in [JSVV24].

ManifoldsBase.distance ÔÇö Method

distance(M::MetricManifold{ÔäŁ, Segre{ÔäŁ,V}, WarpedMetric{A}}, p, q)

Riemannian distance between two points p and q on the warped Segre manifold.

Assume $p ÔëÉ (╬╗, x_1,ÔÇŽ, x_d)$, $q ÔëÉ (╬╝, y_1,ÔÇŽ, y_d) Ôłł \mathcal{S}_A$ are connected by a geodesic. Let

\[ m = \sqrt{\sphericalangle(x_1, y_1)^2 +ÔÇŽ + \sphericalangle(x_d, y_d)^2}\]

and assume $(╬╝, y_1,ÔÇŽ, y_d)$ is the representation of $q$ that minimizes $m$. Then

\[ \operatorname{dist}_{\mathcal{S}_A}(p, q) = \sqrt{╬╗^2 - 2 ╬╗ ╬╝ \cos(A m) + ╬╝^2}.\]

ManifoldsBase.get_coordinates ÔÇö Method

get_coordinates(M::Segre{­ŁöŻ, V}, p, v, ::DefaultOrthonormalBasis; kwargs...)

Get coordinates of X in the tangent space $T_{(╬╗, x_1,ÔÇŽ, x_d)} \mathcal{S}_A = \mathbb{R} ├Ś T_{x_1} \mathbb{S}^{n_1 - 1} ├ŚÔő»├Ś T_{x_d} \mathbb{S}^{n_d - 1}$ using a DefaultOrthonormalBasis on each factor.

ManifoldsBase.get_vector ÔÇö Method

get_vector(M::Segre{­ŁöŻ, V}, p, c, ::DefaultOrthonormalBasis; kwargs...)

Get tangent vector X from coordinates in the tangent space $T_{(╬╗, x_1,ÔÇŽ, x_d)} \mathcal{S}_A = \mathbb{R} ├Ś T_{x_1} \mathbb{S}^{n_1 - 1} ├ŚÔő»├Ś T_{x_d} \mathbb{S}^{n_d - 1}$ using a DefaultOrthonormalBasis on each factor.

ManifoldsBase.inner ÔÇö Method

inner(M::MetricManifold{ÔäŁ, Segre{ÔäŁ, V}, WarpedMetric{A}}, p, X, Y)

Inner product between two tangent vectors $X = (╬Ż, u_1,ÔÇŽ, u_d)$ and $Y = (╬ż, v_1,ÔÇŽ, v_d)$ at $p \doteq (╬╗, x_1,ÔÇŽ, x_d)$:

\[ ÔčĘX, YÔčę_{p} = ╬Ż ╬ż + (A ╬╗)^2 (ÔčĘ u_1, v_1 Ôčę_{x_1} +ÔÇŽ + ÔčĘu_d, v_dÔčę_{x_d}),\]

where $╬Ż$, $╬ż Ôłł T_{╬╗} ÔäŁ^{+} = ÔäŁ$ and $u_i$, $v_i Ôłł T_{x_i} \mathbb{S}^{n_i - 1} \subset ÔäŁ^{n_i}$.

ManifoldsBase.riemann_tensor ÔÇö Method

riemann_tensor(M::MetricManifold{ÔäŁ, Segre{ÔäŁ,V}, WarpedMetric{A}}, p, X, Y)

Riemann tensor of the warped Segre manifold at $p$.

$\mathcal{S}_A$ is locally a warped product of $ÔäŁ^{+}$ and $\mathbb{S}^{n_1 - 1} ├ŚÔő»├Ś \mathbb{S}^{n_d - 1}$. If $p ÔëÉ (╬╗, x_1,ÔÇŽ, x_d) Ôłł \mathcal{S}_A$ and $X$, $Y$, $Z Ôłł T_{(x_1,ÔÇŽ, x_d)} (\mathbb{S}^{n_1 - 1} ├ŚÔő»├Ś \mathbb{S}^{n_d - 1}) \subset T_p \mathcal{S}_A$ then

\[ R_{\mathcal{S}_A}(X, Y) Z = R_{\mathbb{S}^{n_1 - 1} ├ŚÔő»├Ś \mathbb{S}^{n_d - 1}}(X, Y) Z + ╬╗^{-2}(ÔčĘ X, Z Ôčę_{p} Y - ÔčĘ Y, Z Ôčę_{p} X).\]

$R_{\mathcal{S}_A}$ is zero in the remaining (orthogonal) directions.

ManifoldsBase.sectional_curvature ÔÇö Method

sectional_curvature(M::MetricManifold{ÔäŁ, Segre{ÔäŁ,V}, WarpedMetric{A}}, p, X, Y)

Sectional curvature of the warped Segre manifold at $p$.

$\mathcal{S}_A$ is locally a warped product of $ÔäŁ^{+}$ and $\mathbb{S}^{n_1 - 1} ├ŚÔő»├Ś \mathbb{S}^{n_d - 1}$ If $p = (╬╗, x_1,ÔÇŽ, x_d) Ôłł \mathcal{S}$, $u_i Ôłł T_{x_i} \mathbb{S}^{n_i - 1}$, and $v_j Ôłł T_{x_j} \mathbb{S}^{n_j - 1}$, then

\[ K_{\mathcal{S}_A}(u_i, v_j) = \frac{A^{-2} \delta_{i j} - 1}{╬╗^{2}}.\]

$K_{\mathcal{S}_A}$ is zero in the remaining (orthogonal) directions.

Literature

[JSVV24]

S.┬áJacobsson, L.┬áSwijsen, J.┬áV.┬áVeken and N.┬áVannieuwenhoven. Warped geometries of Segre-Veronese manifolds (2024), arXiv:2410.00664 [math.NA].