Sphere

Sphere

Manifolds.Sphere — Type.

Sphere{N} <: Manifold

The unit sphere manifold $\mathbb S^n$ represented by $n+1$-Tuples, i.e. in by vectors in $\mathbb R^{n+1}$ of unit length

Constructor

Sphere(n)

generates the $\mathbb S^{n}\subset \mathbb R^{n+1}$

Manifolds.inner — Method.

inner(S::Sphere, x, w, v)

compute the inner product of the two tangent vectors w,v from the tangent plane at x on the sphere S=$\mathbb S^n$ using the restriction of the metric from the embedding, i.e. $ (v,w)_x = v^\mathrm{T}w $.

Manifolds.is_manifold_point — Method.

is_manifold_point(S,x; kwargs...)

checks, whether x is a valid point on the Sphere S, i.e. is a vector of length manifold_dimension(S)+1 (approximately) of unit length. The tolerance for the last test can be set using the ´kwargs...`.

Manifolds.is_tangent_vector — Method.

is_tangent_vector(S,x,v; kwargs... )

checks whether v is a tangent vector to x on the Sphere S, i.e. atfer is_manifold_point(S,x), v has to be of same dimension as x and orthogonal to x. The tolerance for the last test can be set using the ´kwargs...`.

Manifolds.manifold_dimension — Method.

manifold_dimension(S::Sphere)

Return the dimension of the manifold $\mathbb S^n$, i.e. $n$.

Manifolds.normal_tvector_distribution — Method.

normal_tvector_distribution(S::Sphere, x, σ)

Normal distribution in ambient space with standard deviation σ projected to tangent space at x.

Manifolds.uniform_distribution — Method.

uniform_distribution(S::Sphere, x)

Uniform distribution on given sphere. Generated points will be of similar type to x.