Stiefel

Manifolds.Stiefel — Type

Stiefel{n,k,T} <: Manifold

The Stiefel manifold consists of all $n\times k$, $n\geq k$ orthonormal matrices, i.e.

\[\mathcal M = \{ x \in \mathbb F^{n\times k} : x^{\mathrm{H}}x = I_k \},\]

where $\mathbb F \in \{\mathbb R, \mathbb C\}$, $\cdot^{\mathrm{H}}$ denotes the complex conjugate transpose or Hermitian, and $I_n \in \mathbb R^{n\times n}$ denotes the $k \times k$ identity matrix.

The tangent space at a point $x\in\mathcal M$ is given by

\[T_x\mathcal M = \{ v \in \mathbb{F}^{n\times k} : x^{\mathrm{H}}v + v^{\mathrm{H}}x=0_n\},\]

where $0_n$ is the $k\times k$ zero matrix.

The metric is either inherited from $\mathbb R^{n,k}$ for the real-valued case or the one inherited from interpreting the complex valued entries in the Gaussian plane $\mathbb R^2$ and then over all entries as before, i.e. the latter may be called an Hermitian metric in the complex-valued matrices.

The manifold is named after Eduard L. Stiefel (1909–1978).

Constructor

Stiefel(n,k,F=ℝ)

Generate the (real-valued) Stiefel manifold of $n\times k$ dimensional orthonormal matrices.

Base.exp — Method

exp(M, x, v)

Compute the exponential map on the Stiefel{n,k,T}() manifold M eminating from x into tangent direction v.

\[\operatorname{exp}_{x} v = \begin{pmatrix} x\\v \end{pmatrix} \exp \left( \begin{pmatrix} x^{\mathrm{H}}v & - v^{\mathrm{H}}v\\ I_n & x^{\mathrm{H}}v\end{pmatrix} \right) \begin{pmatrix} \exp( -x^{\mathrm{H}}v) \\ 0_n\end{pmatrix}\]

where $\exp$ denotes matrix exponential, $\cdot^{\mathrm{H}}$ denotes the complex conjugate transpose or Hermitian, and $I_k$ and $0_k$ are the identity matrix and the zero matrix of dimension $k \times k$, respectively.

ManifoldsBase.inner — Method

inner(M::Stiefel, x, v, w)

Compute the inner product for two tangent vectors v, w from the tangent space of x on the Stiefel manifold M. The formula reads

\[(v,w)_x = \operatorname{trace}(v^{\mathrm{H}}w),\]

i.e. the EuclideanMetric from the embedding restricted to the tangent space. For the complex-valued case this is the Hermitian metric, to be precise.

ManifoldsBase.inverse_retract — Method

inverse_retract(M::Stiefel, x, y, ::PolarInverseRetraction)

Compute the inverse retraction based on a singular value decomposition for two points x, y on the Stiefel manifold M and return the resulting tangent vector in v. This follows the folloing approach: From the Polar retraction we know that

\[\operatorname{retr}_x^{-1}(y) = ys - q\]

if such a symmetric positive definite $k\times k$ matrix exists. Since $ys-q$ is also a tangent vector at $x$ we obtain

\[x^{\mathrm{H}}ys + s(x^{\mathrm{H}}y)^{\mathrm{H}} + 2I_k = 0,\]

which can either be solved by a Lyapunov approach or a continuous-time algebraic Riccati equation as described in [KanekoFioriTanaka2013]

This implementation follows the Lyapunov approach.

ManifoldsBase.inverse_retract — Method

inverse_retract(M, x, y, ::QRInverseRetraction)

Compute the inverse retraction based on a qr decomposition for two points x, y on the Stiefel manifold M and return the resulting tangent vector in v. The computation follows Algorithm 1 in [KanekoFioriTanaka2013].

ManifoldsBase.manifold_dimension — Method

manifold_dimension(M::Stiefel)

Return the dimension of the Stiefel manifold M=$\operatorname{St}(n,k)$. The dimension for $\mathbb{F}=\mathbb{R}$ is given by

\[nk - \frac{1}{2}k(k+1)\]

and for $\mathbb{F}=\mathbb{C}$

\[2nk - k^2.\]

ManifoldsBase.project_tangent — Method

project_tangent(M, x, v)

Project v onto the tangent space of x to the Stiefel manifold M. The formula reads

\[\operatorname{proj}_{\mathcal M}(x,v) = v - x \operatorname{Sym}(x^{\mathrm{H}}v),\]

where $\operatorname{Sym}(y)$ is the symmetrization of $y$, e.g. by $\operatorname{Sym}(y) = \frac{y^{\mathrm{H}}+y}{2}$.

ManifoldsBase.representation_size — Method

representation_size(M::Stiefel)

Returns the representation size of the Stiefel M=$\operatorname{St}(n,k)$, i.e. (n,k), which is the matrix dimensions.

ManifoldsBase.retract — Method

retract(M, x, v, ::PolarRetraction)

Compute the SVD-based retraction PolarRetraction on the Stiefel manifold M. With $USV = x + v$ the retraction reads

\[\operatorname{retr}_x(v) = U\bar{V}^\mathrm{H}.\]

ManifoldsBase.retract — Method

retract(M, x, v, ::QRRetraction )

Compute the QR-based retraction QRRetraction on the Stiefel manifold M. With $QR = x + v$ the retraction reads

\[\operatorname{retr}_x(v) = QD,\]

where D is a $n\times k$ matrix with

\[D = \operatorname{diag}\bigl(\operatorname{sgn}(R_{ii}+0,5)_{i=1}^k \bigr),\]

where $\operatorname{sgn}(x) = \begin{cases} 1 & \text{ for } x > 0,\\ 0 & \text{ for } x = 0,\\ -1& \text{ for } x < 0. \end{cases}$

ManifoldsBase.zero_tangent_vector — Method

representation_size(M::Stiefel, x)

Returns the zero tangent vector from the tangent space at x on the Stiefel M=$\operatorname{St}(n,k)$, i.e. an (n,k) zero matrix.

ManifoldsBase.check_manifold_point — Method

check_manifold_point(M::Stiefel, x; kwargs...)

Check whether x is a valid point on the Stiefel M=$\operatorname{St}(n,k)$, i.e. that it has the right AbstractField type and $x^{\mathrm{H}}x$ is (approximatly) the identity, where $\cdot^{\mathrm{H}}$ is the complex conjugate transpose. The settings for approximately can be set with kwargs....

ManifoldsBase.check_tangent_vector — Method

check_tangent_vector(M::Stiefel, x, v; kwargs...)

Check whether v is a valid tangent vector at x on the Stiefel M=$\operatorname{St}(n,k)$, i.e. the AbstractField fits and it (approximtly) holds that $x^{\mathrm{H}}v + v^{\mathrm{H}}x = 0$, where kwargs... is passed to the isapprox.

Literature

H. Sato and T. Iawi, “A complex singular value decomposition algorithm based on the Riemannian Newton method”, in: IEEE 52nd Annual Conference on Decision and Control (CDC), 2013, p. 2972--2978, doi: 10.1109/CDC.2013.6760335

H. Sato: “Riemannian conjugate gradient method for complex singular value decomposition problem”, in: IEEE 53rd Annual COnference on Decision and Control (CDC), 2014, p. 5849–5854, doi: 10.1109/CDC.2014.7040305

KanekoFioriTanaka2013

T. Kaneko, S. Fiori, T. Tanaka: "Empirical Arithmetic Averaging over the Compact Stiefel Manifold", IEEE Transactions on Signal Processing, 2013, doi: 10.1109/TSP.2012.2226167.

KanekoFioriTanaka2013

T. Kaneko, S. Fiori, T. Tanaka: "Empirical Arithmetic Averaging over the Compact Stiefel Manifold", IEEE Transactions on Signal Processing, 2013, doi: 10.1109/TSP.2012.2226167.