Symmetric matrices

Manifolds.SymmetricMatrices — Type

SymmetricMatrices{n,T} <: Manifold

The Manifold $ \operatorname{Sym} (n)$ consisting of the T-valued, symmetric matrices of size $ n\times n$, i.e. the set

\[\operatorname{Sym}(n) = \bigl\{A \in \mathbb F^{n\times n} \big| A^{\mathrm{T}} = A \bigr\}\]

where $\mathbb F \in \{ \mathbb R, \mathbb C\}$. For T, there is a distinction between

T = Real: real-valued matrix entries

T = Complex: complex-valued matrix entries

Being slightly inefficient, the matrices are safed as $n\times n$ arrays despite the redundance of the lower triangle.

Constructor

SymmetricMatrices(n::Int,T::Type = Real)

Manifolds.flat! — Method

flat!(M::SymmetricMatrices, v::FVector{CotangentSpaceType}, x, w::FVector{TangentSpaceType})

Computes the flat! isomorphism of the SymmetricMatrices manifold M on the manifold point x and tangent vector w. The result is stored in v.

As M is already a vector space over $\mathbb R$, this returns just the vector w.

Manifolds.sharp! — Method

sharp!(M::SymmetricMatrices, v::FVector{TangentSpaceType}, x, w::FVector{CotangentSpaceType})

Computes the sharp! isomorphism of the SymmetricMatrices manifold M on the manifold point x and cotangent vector w. The result is stored in v.

As M is already a vector space over $\mathbb R$, this returns just the vector w.

ManifoldsBase.exp! — Method

exp!(M::SymmetricMatrices, y, x, v)

Computes the exponential map exp! of the tangent vector v at manifold point x lying on the SymmetricMatrices manifold M. The result is saved in y. For the SymmetricMatrices, the exponential map is simply the sum, i.e.

\[ y = x + v\]

ManifoldsBase.inner — Method

inner(M::SymmetricMatrices, x, w, v)

compute the inner product of the two tangent vectors w,v from the tangent plane at x on the SymmetricMatrices M using the restriction of the metric from the embedding, i.e.

\[(v,w)_x = \operatorname{tr}(v^{\mathrm{T}}w)\]

ManifoldsBase.log! — Method

log!(M::SymmetricMatrices, v, x, y)

Computes the logarithmic map from the manifold point x to the manifold point y, both lying on the SymmetricMatrices manifold M. The result is saved in v. For the SymmetricMatrices, the logarithmic map is simply the difference, i.e.

\[ v = y-x\]

ManifoldsBase.manifold_dimension — Method

manifold_dimension(M::SymmetricMatrices{n,ℂ})

Return the dimension of the SymmetricMatrices matrix M with complex-valued entries, i.e.

\[ n(n+1)\]

ManifoldsBase.manifold_dimension — Method

manifold_dimension(M::SymmetricMatrices{n,ℝ})

Return the dimension of the SymmetricMatrices matrix M with real-valued entries, i.e.

\[ \frac{n(n+1)}{2}\]

ManifoldsBase.project_point! — Method

project_point!(M::SymmetricMatrices,x)

Implements the project_point of an arbitrary matrix x onto the manifold of SymmetricMatrices, i.e.

\[x_{new} = \frac{1}{2} \left( x_{old} + x_{old}^{\mathrm{T}} \right).$\]

The value of the variable x is overwritten.

ManifoldsBase.project_tangent! — Method

project_tangent!(M::SymmetricMatrices, w, x, v)

Implements the project_tangent! of the matrix v onto the tangent space of the manifold of SymmetricMatrices, i.e. the SymmetricMatrices itself. The result is stored in w:

\[w = \frac{1}{2} \left( v + v^{\mathrm{T}} \right)\]

ManifoldsBase.vector_transport_to! — Method

vector_transport_to!(M::SymmetricMatrices, vto, x, v, y)

Computes the vector_transport_to! of the tangent vector v from the manifold point x on the SymmetricMatrices M to the manifold point $y\in M$. The result is stored in vto. Since the metric is inherited from the embedding space, this is just the identity.

ManifoldsBase.check_manifold_point — Method

check_manifold_point(M::SymmetricMatrices,x; kwargs...)

Checks whether x is a valid manifold point on the SymmetricMatrices M, i.e. is a symmetric matrix of size (N,N) with values of type T.

The tolerance for the symmetry of x can be set using kwargs.

ManifoldsBase.check_tangent_vector — Method

check_tangent_vector(M::SymmetricMatrices,x,v; kwargs... )

Checks whether v is a valid tangent vector to manifold point x on the SymmetricMatrices matrices M, i.e. after is_manifold_point(M,x), v has to be a symmetric matrix of dimension (N,N).

The tolerance for the symmetry of x can be set using kwargs.