Torus

The torus $𝕋^d ≅ [-π,π)^d$ is modeled as an AbstractPowerManifold of the (real-valued) Circle and uses ArrayPowerRepresentation. Points on the torus are hence row vectors, $x ∈ ℝ^{d}$.

Example

The following code can be used to make a three-dimensional torus $𝕋^3$ and compute a tangent vector:

using ManifoldsM = Torus(3)p = [0.5, 0.0, 0.0]q = [0.0, 0.5, 1.0]X = log(M, p, q)

3-element Vector{Float64}:
 -0.5
  0.5
  1.0

Types and functions

Most functions are directly implemented for an AbstractPowerManifold with ArrayPowerRepresentation except the following special cases:

Manifolds.Torus — Type

Torus{N} <: AbstractPowerManifold

The n-dimensional torus is the $n$-dimensional product of the Circle.

The Circle is stored internally within M.manifold, such that all functions of AbstractPowerManifold can be used directly.

ManifoldsBase.check_point — Method

check_point(M::Torus{n}, p)

Checks whether p is a valid point on the Torus M, i.e. each of its entries is a valid point on the Circle and the length of x is n.

ManifoldsBase.check_vector — Method

check_vector(M::Torus{n}, p, X; kwargs...)

Checks whether X is a valid tangent vector to p on the Torus M. This means, that p is valid, that X is of correct dimension and elementwise a tangent vector to the elements of p on the Circle.

Embedded Torus

Two-dimensional torus embedded in $ℝ^3$.

Manifolds.DefaultTorusAtlas — Type

DefaultTorusAtlas()

Atlas for torus with charts indexed by two angles numbers $θ₀, φ₀ ∈ [-π, π)$. Inverse of a chart $(θ₀, φ₀)$ is given by

\[x(θ, φ) = (R + r\cos(θ + θ₀))\cos(φ + φ₀) \\ y(θ, φ) = (R + r\cos(θ + θ₀))\sin(φ + φ₀) \\ z(θ, φ) = r\sin(θ + θ₀)\]

Manifolds.EmbeddedTorus — Type

EmbeddedTorus{TR<:Real} <: AbstractDecoratorManifold{ℝ}

Surface in ℝ³ described by parametric equations:

\[x(θ, φ) = (R + r\cos θ)\cos φ \\ y(θ, φ) = (R + r\cos θ)\sin φ \\ z(θ, φ) = r\sin θ\]

for θ, φ in $[-π, π)$. It is assumed that $R > r > 0$.

Alternative names include anchor ring, donut and doughnut.

Constructor

EmbeddedTorus(R, r)

Manifolds.affine_connection — Method

affine_connection(M::EmbeddedTorus, A::DefaultTorusAtlas, i, a, Xc, Yc)

Affine connection on EmbeddedTorus M.

Manifolds.check_chart_switch — Method

check_chart_switch(::EmbeddedTorus, A::DefaultTorusAtlas, i, a; ϵ = pi/3)

Return true if parameters a lie closer than ϵ to chart boundary.

Manifolds.gaussian_curvature — Method

gaussian_curvature(M::EmbeddedTorus, p)

Gaussian curvature at point p from EmbeddedTorus M.

Manifolds.inverse_chart_injectivity_radius — Method

inverse_chart_injectivity_radius(M::AbstractManifold, A::AbstractAtlas, i)

Injectivity radius of get_point for chart i from the DefaultTorusAtlas A of the EmbeddedTorus.

Manifolds.normal_vector — Method

normal_vector(M::EmbeddedTorus, p)

Outward-pointing normal vector on the EmbeddedTorus at the point p.

ManifoldsBase.check_point — Method

check_point(M::EmbeddedTorus, p; kwargs...)

Check whether p is a valid point on the EmbeddedTorus M. The tolerance for the last test can be set using the kwargs....

The method checks if $(p_1^2 + p_2^2 + p_3^2 + R^2 - r^2)^2$ is approximately equal to $4R^2(p_1^2 + p_2^2)$.

ManifoldsBase.check_vector — Method

check_vector(M::EmbeddedTorus, p, X; atol=eps(eltype(p)), kwargs...)

Check whether X is a valid vector tangent to p on the EmbeddedTorus M. The method checks if the vector X is orthogonal to the vector normal to the torus, see normal_vector. Absolute tolerance can be set using atol.

ManifoldsBase.inner — Method

inner(M::EmbeddedTorus, ::DefaultTorusAtlas, i, a, Xc, Yc)

Inner product on EmbeddedTorus in chart i in the DefaultTorusAtlas. between vectors with coordinates Xc and Yc tangent at point with parameters a. Vector coordinates must be given in the induced basis.

ManifoldsBase.is_flat — Method

is_flat(::EmbeddedTorus)

Return false. EmbeddedTorus is not a flat manifold.

ManifoldsBase.manifold_dimension — Method

manifold_dimension(M::EmbeddedTorus)

Return the dimension of the EmbeddedTorus M that is 2.

DefaultTorusAtlas()
Atlas for torus with charts indexed by two angles numbers .

EmbeddedTorus{TR<:Real} <: AbstractDecoratorManifold{ℝ}
Surface in ℝ³ described by parametric equations:

Torus{N} <: AbstractPowerManifold
The n-dimensional torus is the -dimensional product of the Circle.

AbstractAtlas{𝔽}
An abstract class for atlases with charts that have values in the vector space 𝔽ⁿ for some value of n.

log(M::CholeskySpace, X, p, q)
Compute the logarithmic map on the CholeskySpace M for the geodesic emanating from the lower triangular matrix with positive diagonal p towards q.

log(M::Circle, p, q)
Compute the logarithmic map on the Circle M.

log(M::EssentialManifold, p, q)
Compute the logarithmic map on the EssentialManifold M, i.e. the tangent vector, whose geodesic starting from p reaches q after time 1.

log(M::Euclidean, p, q)
Compute the logarithmic map on the Euclidean M from p to q, which in this case is just

log(M::GeneralizedGrassmann, p, q)
Compute the logarithmic map on the GeneralizedGrassmann M , i.e. the tangent vector X whose corresponding geodesic starting from p reaches q after time 1 on M.

log(M::Rotations, p, X) log(M::OrthogonalMatrices, p, X) log(M::UnitaryMatrices, p, X)
Compute the logarithmic map, that is, since the resulting is represented in the Lie algebra,

log(M::Rotations, p, q)
Compute the logarithmic map on the Rotations manifold M which is given by

log(M::Grassmann, p, q)
Compute the logarithmic map on the Grassmann M , i.e. the tangent vector X whose corresponding geodesic starting from p reaches q after time 1 on M.

log(M::Hyperbolic, p, q)
Compute the logarithmic map on the Hyperbolic space , the tangent vector representing the geodesic starting from p reaches q after time 1.

log(M::Hyperrectangle, p, q)
Compute the logarithmic map on the Hyperrectangle M from p to q, which in this case is just

log(M::PositiveNumbers, p, q)
Compute the logarithmic map on the PositiveNumbers M.

log(M::ProbabilitySimplex, p, q)
Compute the logarithmic map of p and q on the ProbabilitySimplex M.

log(M::AbstractProjectiveSpace, p, q)
Compute the logarithmic map on AbstractProjectiveSpace M, i.e. the tangent vector whose corresponding geodesic starting from p reaches q after time 1 on M.

log(M::MetricManifold{ℝ, Segre{ℝ,V}, WarpedMetric{A}}, p, q)
Logarithmic map on the warped Segre manifold.

log(M::Segre{ℝ, V}, p, q)
Logarithmic map on the Segre manifold.

log(M::KendallsShapeSpace, p, q)
Compute the logarithmic map on KendallsShapeSpace M.

log(M::AbstractSphere, p, q)
Compute the logarithmic map on the AbstractSphere M, i.e. the tangent vector, whose geodesic starting from p reaches q after time 1.

log(M::MetricManifold{ℝ,<:Stiefel{ℝ},<:StiefelSubmersionMetric}, p, q; kwargs...)
Compute the logarithmic map on the Stiefel(n,k) manifold with respect to the StiefelSubmersionMetric.

log(::MetricManifold{ℝ,<:SymmetricPositiveDefinite,<:GeneralizedBuresWassersteinMetric}, p, q)
Compute the logarithmic map on SymmetricPositiveDefinite with respect to the BuresWassersteinMetric given by

log(::MetricManifold{SymmetricPositiveDefinite,BuresWassersteinMetric}, p, q)
Compute the logarithmic map on SymmetricPositiveDefinite with respect to the BuresWassersteinMetric given by

log(M::MetricManifold{ℝ,<:SymmetricPositiveDefinite,LogCholeskyMetric}, p, q)
Compute the logarithmic map on SymmetricPositiveDefinite M with respect to the LogCholeskyMetric emanating from p to q.

log(M::SymmetricPositiveDefinite, p, q) log(M::MetricManifold{SymmetricPositiveDefinite,AffineInvariantMetric}, p, q)
Compute the logarithmic map from p to q on the SymmetricPositiveDefinite as a MetricManifold with AffineInvariantMetric.

log(M::SymmetricPositiveSemidefiniteFixedRank, q, p)
Compute the logarithmic map on the SymmetricPositiveSemidefiniteFixedRank manifold by minimizing with respect to .