Tucker manifold

Manifolds.Tucker — Type

Tucker{T, D, 𝔽} <: AbstractManifold{𝔽}

The manifold of $N_1×\dots×N_D$ real-valued or complex-valued tensors of fixed multilinear rank $(R_1, \dots, R_D)$ . If $R_1 = \dots = R_D = 1$, this is the Segre manifold, i.e., the set of rank-1 tensors.

Representation in HOSVD format

Let $𝔽$ be the real or complex numbers. Any tensor $p$ on the Tucker manifold can be represented as a multilinear product in HOSVD [LMV00] form

\[p = (U_1,\dots,U_D) ⋅ \mathcal{C}\]

where $\mathcal C \in 𝔽^{R_1×\dots×R_D}$ and, for $d=1,\dots,D$, the matrix $U_d \in 𝔽^{N_d×R_d}$ contains the singular vectors of the $d$th unfolding of $\mathcal{A}$

Tangent space

The tangent space to the Tucker manifold at $p = (U_1,\dots,U_D) ⋅ \mathcal{C}$ is [KL10]

\[T_p \mathcal{M} = \bigl\{ (U_1,\dots,U_D) ⋅ \mathcal{C}^\prime + \sum_{d=1}^D \bigl( (U_1, \dots, U_{d-1}, U_d^\prime, U_{d+1}, \dots, U_D) ⋅ \mathcal{C} \bigr) \bigr\}\]

where $\mathcal{C}^\prime$ is arbitrary, $U_d^{\mathrm{H}}$ is the Hermitian adjoint of $U_d$, and $U_d^{\mathrm{H}} U_d^\prime = 0$ for all $d$.

Constructor

Tucker(N::NTuple{D, Int}, R::NTuple{D, Int}[, field=ℝ]; parameter::Symbol=:type)

Generate the manifold of field-valued tensors of dimensions N[1] × … × N[D] and multilinear rank R = (R[1], …, R[D]).

Manifolds.TuckerPoint — Type

TuckerPoint{T,D}

An order D tensor of fixed multilinear rank and entries of type T, which makes it a point on the Tucker manifold. The tensor is represented in HOSVD form.

Constructors:

TuckerPoint(core::AbstractArray{T,D}, factors::Vararg{<:AbstractMatrix{T},D}) where {T,D}

Construct an order D tensor of element type T that can be represented as the multilinear product (factors[1], …, factors[D]) ⋅ core. It is assumed that the dimensions of the core are the multilinear rank of the tensor and that the matrices factors each have full rank. No further assumptions are made.

TuckerPoint(p::AbstractArray{T,D}, mlrank::NTuple{D,Int}) where {T,D}

The low-multilinear rank tensor arising from the sequentially truncated the higher-order singular value decomposition of the D-dimensional array p of type T. The singular values are truncated to get a multilinear rank mlrank [VVM12].

Manifolds.TuckerTangentVector — Type

TuckerTangentVector{T, D} <: AbstractTangentVector

Tangent vector to the D-th order Tucker manifold at $p = (U_1,\dots,U_D) ⋅ \mathcal{C}$. The numbers are of type T and the vector is represented as

\[X = (U_1,\dots,U_D) ⋅ \mathcal{C}^\prime + \sum_{d=1}^D (U_1,\dots,U_{d-1},U_d^\prime,U_{d+1},\dots,U_D) ⋅ \mathcal{C}\]

where $U_d^\mathrm{H} U_d^\prime = 0$.

Constructor

TuckerTangentVector(C′::Array{T,D}, U′::NTuple{D,Matrix{T}}) where {T,D}

Constructs a Dth order TuckerTangentVector of number type T with $C^\prime$ and $U^\prime$, so that, together with a TuckerPoint $p$ as above, the tangent vector can be represented as $X$ in the above expression.

Functions

Base.convert — Method

Base.convert(::Type{Matrix{T}}, basis::CachedBasis{𝔽,DefaultOrthonormalBasis{𝔽, TangentSpaceType},HOSVDBasis{T, D}}) where {𝔽, T, D}
Base.convert(::Type{Matrix}, basis::CachedBasis{𝔽,DefaultOrthonormalBasis{𝔽, TangentSpaceType},HOSVDBasis{T, D}}) where {𝔽, T, D}

Convert a HOSVD-derived cached basis from [DBV21] of the Dth order Tucker manifold with number type T to a matrix. The columns of this matrix are the vectorisations of the embeddings of the basis vectors.

ManifoldsBase.check_point — Method

check_point(M::Tucker, p; kwargs...)

Check whether the multidimensional array or TuckerPoint p is a point on the Tucker manifold, i.e. it is a Dth order N[1] × … × N[D] tensor of multilinear rank (R[1], …, R[D]). The keyword arguments are passed to the matrix rank function applied to the unfoldings. For a TuckerPoint it is checked that the point is in correct HOSVD form.

ManifoldsBase.check_vector — Method

check_vector(M::Tucker{<:Any,D}, p::TuckerPoint{T,D}, X::TuckerTangentVector) where {T,D}

Check whether a TuckerTangentVector X is is in the tangent space to the Dth order Tucker manifold M at the Dth order TuckerPoint p. This is the case when the dimensions of the factors in X agree with those of p and the factor matrices of X are in the orthogonal complement of the HOSVD factors of p.

ManifoldsBase.default_inverse_retraction_method — Method

default_inverse_retraction_method(M::Tucker)

Return PolarInverseRetraction as the default inverse retraction for the Tucker manifold.

ManifoldsBase.default_retraction_method — Method

default_retraction_method(M::Tucker)

Return PolarRetraction as the default retraction for the Tucker manifold.

ManifoldsBase.default_vector_transport_method — Method

default_vector_transport_method(M::Tucker)

Return the ProjectionTransport as the default vector transport method for the Tucker manifold.

ManifoldsBase.embed — Method

embed(::Tucker, p::TuckerPoint, X::TuckerTangentVector)

Convert a tangent vector X with base point p on the rank R Tucker manifold to a full tensor, represented as an N[1] × … × N[D]-array.

ManifoldsBase.embed — Method

embed(::Tucker, p::TuckerPoint)

Convert a TuckerPoint p on the rank R Tucker manifold to a full N[1] × … × N[D]-array by evaluating the Tucker decomposition.

ManifoldsBase.get_basis — Method

get_basis(:: Tucker, p::TuckerPoint, basisType::DefaultOrthonormalBasis{𝔽, TangentSpaceType}) where 𝔽

An implicitly stored basis of the tangent space to the Tucker manifold. Assume $p = (U_1,\dots,U_D) ⋅ \mathcal{C}$ is in HOSVD format and that, for $d=1,\dots,D$, the singular values of the $d$'th unfolding are $\sigma_{dj}$, with $j = 1,\dots,R_d$. The basis of the tangent space is as follows: [DBV21]

\[\bigl\{ (U_1,\dots,U_D) e_i \bigr\} \cup \bigl\{ (U_1,\dots, \sigma_{dj}^{-1} U_d^{\perp} e_i e_j^T,\dots,U_D) ⋅ \mathcal{C} \bigr\}\]

for all $d = 1,\dots,D$ and all canonical basis vectors $e_i$ and $e_j$. Every $U_d^\perp$ is such that $[U_d \quad U_d^{\perp}]$ forms an orthonormal basis of $ℝ^{N_d}$.

ManifoldsBase.inner — Method

inner(M::Tucker, p::TuckerPoint, X::TuckerTangentVector, Y::TuckerTangentVector)

The Euclidean inner product between tangent vectors X and X at the point p on the Tucker manifold. This is equal to embed(M, p, X) ⋅ embed(M, p, Y).

inner(::Tucker, A::TuckerPoint, X::TuckerTangentVector, Y)inner(::Tucker, A::TuckerPoint, X, Y::TuckerTangentVector)

The Euclidean inner product between X and Y where X is a vector tangent to the Tucker manifold at p and Y is a vector in the ambient space or vice versa. The vector in the ambient space is represented as a full tensor, i.e., a multidimensional array.

ManifoldsBase.inverse_retract — Method

inverse_retract(M::Tucker, p::TuckerPoint, q::TuckerPoint, ::ProjectionInverseRetraction)

The projection inverse retraction on the Tucker manifold interprets q as a point in the ambient Euclidean space (see embed) and projects it onto the tangent space at to M at p.

ManifoldsBase.is_flat — Method

is_flat(::Tucker)

Return false. Tucker is not a flat manifold.

ManifoldsBase.manifold_dimension — Method

manifold_dimension(::Tucker)

The dimension of the manifold of $N_1×\dots×N_D$ tensors of multilinear rank $(R_1, \dots, R_D)$, i.e.

\[\mathrm{dim}(\mathcal{M}) = \prod_{d=1}^D R_d + \sum_{d=1}^D R_d (N_d - R_d).\]

ManifoldsBase.project — Method

project(M::Tucker, p::TuckerPoint, X)

The least-squares projection of a dense tensor X onto the tangent space to M at p.

ManifoldsBase.retract — Method

retract(::Tucker, p::TuckerPoint, X::TuckerTangentVector, ::PolarRetraction)

The truncated HOSVD-based retraction [KSV13] to the Tucker manifold, i.e. the result is the sequentially truncated HOSVD approximation of $p + X$.

In the exceptional case that the multilinear rank of $p + X$ is lower than that of $p$, this retraction produces a boundary point, which is outside the manifold.

ManifoldsBase.vector_transport_to — Method

vector_transport_to(M::Tucker, Y, p, X, q, ::ProjectionTransport)

Compute the projection vector transport on the Tucker manifold by projecting the tangent vector X onto the tangent space of M at q.

Let

\[p = (C_p, U_p^1, …, U_p^D), \qquad X = (C_X, U_X^1, …, U_X^D) ∈ T_p M, \qquad q = (C_q, U_q^1, …, U_q^1).\]

This means that

\[ X = C_X ×_{j = 1, …, D} U_p^j + \sum_{k = 1}^D C_p ×_{j ≠ k} U_p^j ×_k U_X^k\]

as an embedded vector. Let

\[C_Y = X ×_{j} (U_q^j)^T = C_X ×_{j = 1, …, D} U_p^j (U_q^j)^T + \sum_{k = 1}^D C_p ×_{j ≠ k} U_p^j (U_q^j)^T ×_k U_X^k(U_q^k)^T\]

and

\[U_Y^i = P_{U_q^i}^⊥ \left[X ×_{j ≠ i} (U_q^j)^T\right]_{(i)} (C_q)_{(i)}^+ \\ = \left[C_X ×_{j ≠ i} U_p^j (U_q^j)^T ×_i U_p^i (I - U_q^i (U_q^i)^T)\right]_{(i)} (C_q)_{(i)}^+ + \sum_{k ≠ i} \left[ C_p ×_{j ≠ k ≠ i} U_p^j (U_q^j)^T ×_k U_X^k(U_q^k)^T ×_i U_p^i (I - U_q^i (U_q^i)^T)\right]_{(i)} (C_q)_{(i)}^+ + \left[ C_p ×_{j ≠ i} U_p^j (U_q^j)^T ×_i U_X^i (I - U_q^i (U_q^i)^T)\right]_{(i)} (C_q)_{(i)}^+ \]

Note that $(C_q)_{(i)}^+ = ((C_q)_{(i)}^∗ (C_q)_{(i)})^{-1} (C_q)_{(i)}^∗ = Σ_{(i)}^{-2} (C_q)_{(i)}^∗$, where $Σ_{(i)}$ is the diagonal Matrix containing the singular values of the mode-$i$ unfolding of $C_q$.

Then the transported tangent vector is represented by

\[Y = (C_Y, U_Y^1, …, U_Y^D).\]

For details, see [KSV13].

ManifoldsBase.zero_vector — Method

zero_vector(::Tucker, p::TuckerPoint)

The zero element in the tangent space to p on the Tucker manifold, represented as a TuckerTangentVector.

Internal Functions

Base.foreach — Function

Base.foreach(f, M::Tucker, p::TuckerPoint, basis::AbstractBasis, indices=1:manifold_dimension(M))

Let basis be and AbstractBasis at a point p on M. Suppose f is a function that takes an index and a vector as an argument. This function applies f to i and the ith basis vector sequentially for each i in indices. Using a CachedBasis may speed up the computation.

NOTE: The i'th basis vector is overwritten in each iteration. If any information about the vector is to be stored, f must make a copy.

Base.ndims — Method

Base.ndims(p::TuckerPoint{T,D}) where {T,D}

The order of the tensor corresponding to the TuckerPoint p, i.e., D.

Base.size — Method

Base.size(p::TuckerPoint)

The dimensions of a TuckerPoint p, when regarded as a full tensor (see embed).

Manifolds.compute_projection_summand! — Method

compute_projection_summand!(result::Matrix{T}, buffer::NTuple{5, Array{T}}, core::Array{T, D}, Uᵢ::Matrix{T}, q::TuckerPoint{T, D}, Σ⁻¹::Vector{T}, factors::Vector{Matrix{T}}, i::Int64, add_to_result::Bool) where {T, D}

Helper function for the computation of summands in vector_transport_to_project! Let $q = (C_q, U_q^1, …, U_q^D)$ and $\mathrm{factors} = (F_1, F_2, \ldots, F_D)\, .$ This function computes

\[[\mathrm{core} ×_{j ≠ i} F_j ×_i U_i (I - U_q^i (U_q^i)^T)]_{(i)} \cdot (C_q)_{(i)}^T \cdot Σ^{-1}\, .\]

If add_to_result == true, the result is added to the matrix result, otherwise, result is overwritten. Note that buffer is a 5-tuple of arrays that are used only to store intermediate results of the computation. Let dims be the dimensions of q and ranks its ranks. Then the contents of buffer must be arrays with sizes:

(dims[i], ranks[i]), (dims[i], ranks[i]), (ranks[i], ranks[i]), (ranks...), (ranks...).

Manifolds.contract_with_factors! — Method

contract_with_factors!(result::Array{T, D}, B1::Array{T, D}, B2::Array{T, D}, core::Array{T, D}, factors::Vector{Matrix{T}}, add_to_result::Bool) where {T, D}

This function compute the contraction of the order-D tensor core with the quadratic factor matrices factors. That is, if $\mathrm{factors} = (F_1, F_2, \ldots, F_D)\, ,$ this function computes

\[\mathrm{result}= \mathrm{core} \times_{j = 1, \ldots, D} F_j\]

For D = 3, e.g.

result[n1, n2, n3] = core[r1, r2, r3] * factors[1][r1, n1] * factors[2][r2, n2] * factors[3][r3, n3].

If add_to_result == true, adds to the result array, otherwise overwrites the result array. The arrays B1 and B2 are buffer arrays used to store intermediate results and must have the same size as core.

Manifolds.contract_with_partial_factors! — Method

contract_with_partial_factors!(result::Array{T, D}, B::Array{T, D}, core::Array{T, D}, factors::Vector{Matrix{T}}, k::Int64) where {T, D}

Computes the contraction of the order-$D$ tensor core with the quadratic factor matrices factors in all except the k-th mode. That is, if $\mathrm{factors} = (F_1, F_2, \ldots, F_D)\, ,$ this computes

\[\mathrm{result} = \mathrm{core} \times_{j ≠ k} F_j \, .\]

For D = 3, k = 1, e.g.

result[r1, n1, n2] = core[r1, r2, r3] * factors[2][r2, n2] * factors[3][r3, n3]. The array B is a buffer array used to store intermediate results and must have the same size as core.

Literature

[DBV21]

N. Dewaele, P. Breiding and N. Vannieuwenhoven. The condition number of many tensor decompositions is invariant under Tucker compression, arXiv Preprint (2021), arXiv:2106.13034.

[KL10]

O. Koch and C. Lubich. Dynamical Tensor Approximation. SIAM Journal on Matrix Analysis and Applications 31, 2360–2375 (2010).

[KSV13]

D. Kressner, M. Steinlechner and B. Vandereycken. Low-rank tensor completion by Riemannian optimization. BIT Numerical Mathematics 54, 447–468 (2013).

[LMV00]

L. D. Lathauwer, B. D. Moor and J. Vandewalle. A Multilinear Singular Value Decomposition. SIAM Journal on Matrix Analysis and Applications 21, 1253–1278 (2000).

[VVM12]

N. Vannieuwenhoven, R. Vandebril and K. Meerbergen. A New Truncation Strategy for the Higher-Order Singular Value Decomposition. SIAM Journal on Scientific Computing 34, A1027–A1052 (2012).

Tucker{T, D, 𝔽} <: AbstractManifold{𝔽}
The manifold of real-valued or complex-valued tensors of fixed multilinear rank .

TuckerPoint{T,D}
An order D tensor of fixed multilinear rank and entries of type T, which makes it a point on the Tucker manifold.

TuckerTangentVector{T, D} <: AbstractTangentVector
Tangent vector to the D-th order Tucker manifold at .

check_vector(M::Tucker{<:Any,D}, p::TuckerPoint{T,D}, X::TuckerTangentVector) where {T,D}
Check whether a TuckerTangentVector X is is in the tangent space to the Dth order Tucker manifold M at the Dth order TuckerPoint p.

manifold_dimension(::Tucker)
The dimension of the manifold of tensors of multilinear rank , i.e.

check_vector(M::CenteredMatrices, p, X; kwargs... )
Check whether X is a tangent vector to manifold point p on the CenteredMatrices M, i.e. that X is a matrix of size (m, n) whose columns sum to zero and its values are from the correct AbstractNumbers.

manifold_dimension(M::CenteredMatrices)
Return the manifold dimension of the CenteredMatrices m-by-n matrix M over the number system 𝔽, i.e.

project(M::CenteredMatrices, p, X)
Project the matrix X onto the tangent space at p on the CenteredMatrices M, i.e.

check_vector(M::CholeskySpace, p, X; kwargs... )
Check whether v is a tangent vector to p on the CholeskySpace M, i.e. after check_point(M,p), X has to have the same dimension as p and a symmetric matrix.

manifold_dimension(M::CholeskySpace)
Return the manifold dimension for the CholeskySpace M, i.e.

check_vector(M::Circle, p, X; kwargs...)
Check whether X is a tangent vector in the tangent space of p on the Circle M.

manifold_dimension(M::Circle)
Return the dimension of the Circle M, i.e. .

project(M::Circle, p, X)
Project a value X onto the tangent space of the point p on the Circle M.

check_vector(M::DeterminantOneMatrices{n,𝔽}, p, X; kwargs... )
Check whether X is a tangent vector to manifold point p on the DeterminantOneMatrices M, which are all matrices of size with trace 0.

manifold_dimension(M::DeterminantOneMatrices{n,𝔽})
Return the dimension of the DeterminantOneMatrices matrix M over the number system 𝔽, which is one dimension less than its embedding, the Euclidean(n, n; field=𝔽).

project(G::DeterminantOneMatrices, p, X) project!(G::DeterminantOneMatrices, Y, p, X)
Orthogonally project onto the tangent space of to the DeterminantOneMatrices.

check_vector(M::Elliptope, q, Y; kwargs... )
Check whether is a tangent vector to on the Elliptope M, i.e. Y has to be of same dimension as q and a has to be a symmetric matrix with zero diagonal.

manifold_dimension(M::Elliptope)
returns the dimension of Elliptope M , i.e.

project(M::Elliptope, q, Y)
Project Y onto the tangent space at q, i.e. row-wise onto the oblique manifold.

check_vector(M::EssentialManifold, p, X; kwargs... )
Check whether X is a tangent vector to manifold point p on the EssentialManifold M, i.e. X has to be a 2-element array of 3-by-3 skew-symmetric matrices.

manifold_dimension(M::EssentialManifold{is_signed, ℝ})
Return the manifold dimension of the EssentialManifold, which is 5[TD17].

project(M::EssentialManifold, p, X)
Project the matrix X onto the tangent space

manifold_dimension(M::Euclidean)
Return the manifold dimension of the Euclidean M, i.e. the product of all array dimensions and the real_dimension of the underlying number system.

project(M::Euclidean, p, X)
Project an arbitrary vector X into the tangent space of a point p on the Euclidean M, which is just the identity, since any tangent space of M can be identified with all of M.

check_vector(M::FixedRankMatrices, p, X; kwargs...)
Check whether the tangent UMVTangentVector X is from the tangent space of the SVDMPoint p on the FixedRankMatrices M.

inverse_retract(M, p, q, ::OrthographicInverseRetraction)
Compute the orthographic inverse retraction FixedRankMatrices M by computing

manifold_dimension(M::FixedRankMatrices)
Return the manifold dimension for the 𝔽-valued FixedRankMatrices M of dimension mxn of rank k, namely

project(M, p, A)
Project the matrix or from the embedding the tangent space at on the FixedRankMatrices M, further decomposing the result into , i.e. a UMVTangentVector.

check_vector(M::Flag, p::AbstractMatrix, X::AbstractMatrix; kwargs... )
Check whether X is a tangent vector to point p on the Flag manifold M in the Stiefel representation, i.e. that X is a matrix of the form

check_vector(M::Flag, p::OrthogonalPoint, X::OrthogonalTangentVector; kwargs... )
Check whether X is a tangent vector to point p on the Flag manifold M in the orthogonal matrix representation, i.e. that X is block-skew-symmetric with zero diagonal:

inverse_retract(M::Flag, p, q, ::PolarInverseRetraction)
Compute the inverse retraction for the PolarRetraction, on the Flag manifold M.

manifold_dimension(M::Flag)
Return dimension of flag manifold .

project(::Flag, p, X)
Project vector X in the Euclidean embedding to the tangent space at point p on Flag manifold.

project(M::Flag, p::OrthogonalPoint, X::OrthogonalTangentVector)
Project vector X to tangent space at point p from Flag manifold M , in the orthogonal matrix representation.

check_vector(M::GeneralizedGrassmann, p, X; kwargs...)
Check whether X is a tangent vector in the tangent space of p on the GeneralizedGrassmann M, i.e. that X is of size and type as well as that

manifold_dimension(M::GeneralizedGrassmann)
Return the dimension of the GeneralizedGrassmann(n,k,𝔽) manifold M, i.e.

project(M::GeneralizedGrassmann, p, X)
Project the n-by-k X onto the tangent space of p on the GeneralizedGrassmann M, which is computed by

check_vector(M::GeneralizedStiefel, p, X; kwargs...)
Check whether X is a valid tangent vector at p on the GeneralizedStiefel M=.

manifold_dimension(M::GeneralizedStiefel)
Return the dimension of the GeneralizedStiefel manifold M=.

project(M:GeneralizedStiefel, p, X)
Project X onto the tangent space of p to the GeneralizedStiefel manifold M.

check_vector(M::UnitaryMatrices, p, X; kwargs... ) check_vector(M::OrthogonalMatrices, p, X; kwargs... ) check_vector(M::Rotations, p, X; kwargs... ) check_vector(M::GeneralUnitaryMatrices, p, X; kwargs... )
Check whether X is a tangent vector to p on the UnitaryMatrices space M, i.e. after check_point(M,p), X has to be skew symmetric (Hermitian) and orthogonal to p.

manifold_dimension(M::Rotations) manifold_dimension(M::OrthogonalMatrices)
Return the dimension of the manifold orthogonal matrices and of the manifold of rotations

manifold_dimension(M::SpecialUnitaryMatrices)
Return the dimension of the manifold of special unitary matrices.

manifold_dimension(M::UnitaryMatrices{n,ℂ}) where {n}
Return the dimension of the manifold unitary matrices.

manifold_dimension(M::UnitaryMatrices{<:Any,ℍ})
Return the dimension of the manifold unitary matrices.

project(M::OrthogonalMatrices, p, X) project(M::Rotations, p, X) project(M::UnitaryMatrices, p, X)
Orthogonally project the tangent vector , to the tangent space of M at p, and change the representer to use the corresponding Lie algebra, i.e. we compute

check_vector(M::GraphManifold, p, X; kwargs...)
Check whether X is a valid tangent vector at p on the GraphManifold M.

manifold_dimension(N::GraphManifold{G,𝔽,M,EdgeManifold})
returns the manifold dimension of the GraphManifold N on the edges of a graph , i.e.

manifold_dimension(N::GraphManifold{G,𝔽,M,VertexManifold})
returns the manifold dimension of the GraphManifold N on the vertices of a graph , i.e.

check_vector(::Grassmann, p::ProjectorPoint, X::ProjectorTangentVector; kwargs...)
Check whether the ProjectorTangentVector X is from the tangent space at the ProjectorPoint p on the Grassmann manifold .

inverse_retract(M::Grassmann, p, q, ::PolarInverseRetraction)
Compute the inverse retraction for the PolarRetraction, on the Grassmann manifold M, i.e.,

inverse_retract(M, p, q, ::QRInverseRetraction)
Compute the inverse retraction for the QRRetraction, on the Grassmann manifold M, i.e.,

manifold_dimension(M::Grassmann)
Return the dimension of the Grassmann(n,k,𝔽) manifold M, i.e.

project(M::Grassmann, p, X)
Project the n-by-k X onto the tangent space of p on the Grassmann M, which is computed by

check_vector(M::HamiltonianMatrices, p, X; kwargs... )
Check whether X is a tangent vector to manifold point p on the HamiltonianMatrices M, i.e. X has to be a Hamiltonian matrix The tolerance for is_hamiltonian X can be set using kwargs....

manifold_dimension(M::HeisenbergMatrices)
Return the dimension of HeisenbergMatrices(n), which is equal to .

check_vector(M::Hyperbolic, p, X; kwargs... )
Check whether X is a tangent vector to p on the Hyperbolic M, i.e. after check_point(M,p), X has to be of the same dimension as p.

manifold_dimension(M::Hyperbolic)
Return the dimension of the hyperbolic space manifold , i.e. .

project(M::Hyperbolic, p, X)
Perform an orthogonal projection with respect to the Minkowski inner product of X onto the tangent space at p of the Hyperbolic space M.

project(::Hyperbolic, ::PoincareBallPoint, ::PoincareBallTangentVector)
projection of tangent vectors in the Poincaré ball model is just the identity, since the tangent space consists of all .

manifold_dimension(M::Hyperrectangle)
Return the manifold dimension of the Hyperrectangle M, i.e. the product of all array dimensions.

project(M::Hyperrectangle, p, X)
Project an arbitrary vector X into the tangent space of a point p on the Hyperrectangle M.

check_vector(M::InvertibleMatrices{n,𝔽}, p, X; kwargs... )
Check whether X is a tangent vector to manifold point p on the InvertibleMatrices M, which are all matrices of size its values have to be from the correct AbstractNumbers.

manifold_dimension(M::InvertibleMatrices{n,𝔽})
Return the dimension of the InvertibleMatrices matrix M over the number system 𝔽, which is the same dimension as its embedding, the Euclidean(n, n; field=𝔽).

check_vector(M::MultinomialMatrices p, X; kwargs...)
Checks whether X is a valid tangent vector to p on the MultinomialMatrices M.

check_vector(M::MultinomialDoubleStochastic p, X; kwargs...)
Checks whether X is a valid tangent vector to p on the MultinomialDoubleStochastic M.

manifold_dimension(M::MultinomialDoubleStochastic)
returns the dimension of the MultinomialDoubleStochastic manifold namely

project(M::MultinomialDoubleStochastic, p, Y)
Project Y onto the tangent space at p on the MultinomialDoubleStochastic M, return the result in X.

check_vector(M::MultinomialSymmetric p, X; kwargs...)
Checks whether X is a valid tangent vector to p on the MultinomialSymmetric M.

manifold_dimension(M::MultinomialSymmetric)
returns the dimension of the MultinomialSymmetric manifold namely

project(M::MultinomialSymmetric, p, Y)
Project Y onto the tangent space at p on the MultinomialSymmetric M, return the result in X.

check_vector(M::Oblique p, X; kwargs...)
Checks whether X is a valid tangent vector to p on the Oblique M.

check_vector(M::ParametricSurface, p, X; atol::Real=sqrt(eps(float(number_eltype(p)))), kwargs...)
Check whether X is tangent to the parametric surface M at embedded point p.

manifold_dimension(M::ParametricSurface)
Return the dimension of the parameter space of M.

check_vector(M::PositiveNumbers, p, X; kwargs...)
Check whether X is a tangent vector in the tangent space of p on the PositiveNumbers M.

manifold_dimension(M::PositiveNumbers)
Return the dimension of the PositiveNumbers M, i.e. of the 1-dimensional hyperbolic space,

project(M::PositiveNumbers, p, X)
Project a value X onto the tangent space of the point p on the PositiveNumbers M, which is just the identity.

check_vector(M::ProbabilitySimplex, p, X; kwargs... )
Check whether X is a tangent vector to p on the ProbabilitySimplex M, i.e. after check_point(M,p), X has to be of same dimension as p and its elements have to sum to one.

inverse_retract(M::ProbabilitySimplex, p, q, ::SoftmaxInverseRetraction)
Compute a first order approximation by projection.

manifold_dimension(M::ProbabilitySimplex)
Returns the manifold dimension of the probability simplex in , i.e.

project(M::ProbabilitySimplex, p, Y)
Project Y from the embedding onto the tangent space at p on the ProbabilitySimplex M.

check_vector(M::AbstractProjectiveSpace, p, X; kwargs... )
Check whether X is a tangent vector in the tangent space of p on the AbstractProjectiveSpace M.

inverse_retract(M::AbstractProjectiveSpace, p, q, method::ProjectionInverseRetraction) inverse_retract(M::AbstractProjectiveSpace, p, q, method::PolarInverseRetraction) inverse_retract(M::AbstractProjectiveSpace, p, q, method::QRInverseRetraction)
Compute the equivalent inverse retraction ProjectionInverseRetraction, PolarInverseRetraction on the AbstractProjectiveSpace manifold M, i.e.

manifold_dimension(M::AbstractProjectiveSpace{𝔽}) where {𝔽}
Return the real dimension of the AbstractProjectiveSpace M, respectively i.e. the real dimension of the embedding minus the real dimension of the field 𝔽.

project(M::AbstractProjectiveSpace, p, X)
Orthogonally project the point X onto the tangent space at p on the AbstractProjectiveSpace M:

inverse_retract(M, p, q, ::PolarInverseRetraction)
Compute a vector from the tangent space of the point p on the Rotations manifold M with which the point q can be reached by the PolarRetraction from the point p after time 1.

inverse_retract(M::Rotations, p, q, ::QRInverseRetraction)
Compute a vector from the tangent space of the point p on the Rotations manifold M with which the point q can be reached by the QRRetraction from the point q after time 1.

check_vector(M::KendallsPreShapeSpace, p, X; kwargs... )
Check whether X is a valid tangent vector on KendallsPreShapeSpace, i.e. whether each row has zero mean.

manifold_dimension(M::KendallsPreShapeSpace)
Return the dimension of the KendallsPreShapeSpace manifold M.

manifold_dimension(M::KendallsShapeSpace)
Return the dimension of the KendallsShapeSpace manifold M.

project(M::KendallsPreShapeSpace, p, X)
Project tangent vector X at point p from the embedding to KendallsPreShapeSpace by selecting the right element from the tangent space to orthogonal section representing the quotient manifold M.

check_vector(M::SkewHermitianMatrices, p, X; kwargs... )
Check whether X is a tangent vector to manifold point p on the SkewHermitianMatrices M, i.e. X must be a skew-hermitian matrix of size (n,n) and its values have to be from the correct AbstractNumbers.

manifold_dimension(M::SkewHermitianMatrices)
Return the dimension of the SkewHermitianMatrices matrix M over the number system 𝔽, i.e.

project(M::SkewHermitianMatrices, p, X)
Project the matrix X onto the tangent space at p on the SkewHermitianMatrices M,

check_vector(M::SPDFixedDeterminant, p, X; kwargs... )
Check whether X is a tangent vector to manifold point p on the SPDFixedDeterminant M, i.e. X has to be a tangent vector on SymmetricPositiveDefinite, so a symmetric matrix, and additionally fulfill .

Y = project(M::SPDFixedDeterminant, p, X) project!(M::SPDFixedDeterminant, Y, p, X)
Project the symmetric matrix X onto the tangent space at p of the (sub-)manifold of s.p.d. matrices of determinant M.d (in place of Y), by setting its diagonal (and hence its trace) to zero.

check_vector(M::Spectrahedron, q, Y; kwargs...)
Check whether is a tangent vector to on the Spectrahedron M, i.e. after check_point of q, Y has to be of same dimension as q and a has to be a symmetric matrix with trace.

manifold_dimension(M::Spectrahedron)
returns the dimension of Spectrahedron M, i.e.

project(M::Spectrahedron, q, Y)
Project Y onto the tangent space at q, i.e. row-wise onto the Spectrahedron manifold.

check_vector(M::AbstractSphere, p, X; kwargs... )
Check whether X is a tangent vector to p on the AbstractSphere M, i.e. after check_point(M,p), X has to be of same dimension as p and orthogonal to p.

inverse_retract(M::AbstractSphere, p, q, ::ProjectionInverseRetraction)
Compute the inverse of the projection based retraction on the AbstractSphere M, i.e. rearranging yields since and when that

manifold_dimension(M::AbstractSphere)
Return the dimension of the AbstractSphere M, respectively i.e. the dimension of the embedding -1.

project(M::AbstractSphere, p, X)
Project the point X onto the tangent space at p on the Sphere M.

check_vector(M::SphereSymmetricMatrices, p, X; kwargs... )
Check whether X is a tangent vector to manifold point p on the SphereSymmetricMatrices M, i.e. X has to be a symmetric matrix of size (n,n) of unit Frobenius norm.

manifold_dimension(M::SphereSymmetricMatrices{<:Any,𝔽})
Return the manifold dimension of the SphereSymmetricMatrices n-by-n symmetric matrix M of unit Frobenius norm over the number system 𝔽, i.e.

project(M::SphereSymmetricMatrices, p, X)
Project the matrix X onto the tangent space at p on the SphereSymmetricMatrices M, i.e.

check_vector(M::Stiefel, p, X; kwargs...)
Checks whether X is a valid tangent vector at p on the Stiefel M=, i.e. the AbstractNumbers fits and it (approximately) holds that .

inverse_retract( M::MetricManifold{ℝ,<:Stiefel{ℝ},<:StiefelSubmersionMetric}, p, q, method::ShootingInverseRetraction, )
Compute the inverse retraction using ShootingInverseRetraction.

X = inverse_retract(M::MetricManifold{ℝ, Stiefel{ℝ}, CanonicalMetric}, p, q, a::ApproximateLogarithmicMap) inverse_retract!(M::MetricManifold{ℝ, Stiefel{ℝ}, X, CanonicalMetric}, p, q, a::ApproximateLogarithmicMap)
Compute an approximation to the logarithmic map on the Stiefel(n, k) manifold with respect to the CanonicalMetric.

inverse_retract(M::Stiefel, p, q, ::PolarInverseRetraction)
Compute the inverse retraction based on a singular value decomposition for two points p, q on the Stiefel manifold M.

inverse_retract(M::Stiefel, p, X, ::PolarLightInverseRetraction)
Compute an SVD-based inverse retraction PolarLightInverseRetraction on the Stiefel(n,k) manifold M following [JZ26], equation (3.5)

inverse_retract(M::Stiefel, p, q, method::ProjectionInverseRetraction)
Compute a projection-based inverse retraction.

inverse_retract(M::Stiefel, p, q, ::QRInverseRetraction)
Compute the inverse retraction based on a qr decomposition for two points p, q on the Stiefel manifold M and return the resulting tangent vector in X.

manifold_dimension(M::Stiefel)
Return the dimension of the Stiefel manifold M=.

project(M::Stiefel, p)
Projects p from the embedding onto the Stiefel M, i.e. compute q as the polar decomposition of such that is the identity, where denotes the hermitian, i.e. complex conjugate transposed.

project(M::Stiefel, p, X)
Project X onto the tangent space of p to the Stiefel manifold M.

check_vector(M::SymmetricMatrices{n,𝔽}, p, X; kwargs... )
Check whether X is a tangent vector to manifold point p on the SymmetricMatrices M, i.e. X has to be a symmetric matrix of size (n,n) and its values have to be from the correct AbstractNumbers.

manifold_dimension(M::SymmetricMatrices{n,𝔽})
Return the dimension of the SymmetricMatrices matrix M over the number system 𝔽, i.e.

project(M::SymmetricMatrices, p, X)
Project the matrix X onto the tangent space at p on the SymmetricMatrices M,

check_vector(M::SymmetricPositiveDefinite, p, X; kwargs... )
Check whether X is a tangent vector to p on the SymmetricPositiveDefinite M.

manifold_dimension(M::SymmetricPositiveDefinite)
returns the dimension of SymmetricPositiveDefinite M , i.e.

project(M::SymmetricPositiveDefinite, p, X)
project a matrix from the embedding onto the tangent space of the SymmetricPositiveDefinite matrices, i.e. the set of symmetric matrices.

check_vector(M::SymmetricPositiveSemidefiniteFixedRank, p, X; kwargs... )
Check whether X is a tangent vector to manifold point p on the SymmetricPositiveSemidefiniteFixedRank M.

manifold_dimension(M::SymmetricPositiveSemidefiniteFixedRank)
Return the dimension of the SymmetricPositiveSemidefiniteFixedRank matrix M over the number system 𝔽, i.e.

check_vector(M::SymplecticMatrices, p, X; kwargs...)
Checks whether X is a valid tangent vector at p on the SymplecticMatrices M=, which requires that

inverse_retract(M::SymplecticMatrices, p, q, ::CayleyInverseRetraction)
Compute the Cayley Inverse Retraction such that the Cayley Retraction from along lands at , i.e. [BZ21].

manifold_dimension(::SymplecticMatrices)
Returns the dimension of the symplectic manifold embedded in , i.e.

project(::SymplecticMatrices, p, A) project!(::SymplecticMatrices, Y, p, A)
Given a point , project an element onto the tangent space relative to the euclidean metric of the embedding .

check_vector(M::SymplecticGrassmann, p, X; kwargs...)
Check whether X is a valid tangent vector at p on the SymplecticGrassmann manifold M.

check_vector(M::SymplecticGrassmann, p::ProjectorPoint, X::ProjectorTangentVector; kwargs...)
Check whether X is a valid tangent vector at p on the SymplecticGrassmann, manifold by verifying that it

inverse_retract(::SymplecticGrassmann, p, q, ::CayleyInverseRetraction) inverse_retract!(::SymplecticGrassmann, X, p, q, ::CayleyInverseRetraction)
Compute the Cayley Inverse Retraction on the Symplectic Grassmann manifold, when the points are represented as symplectic bases, i.e. on the SymplecticStiefel.

manifold_dimension(::SymplecticGrassmann)
Return the dimension of the SymplecticGrassmann(2n,2k), which is

check_vector(M::SymplecticMatrices, p, X; kwargs...)
Checks whether X is a valid tangent vector at p on the SymplecticStiefel, manifold.

inverse_retract(::SymplecticStiefel, p, q, ::CayleyInverseRetraction) inverse_retract!(::SymplecticStiefel, X, p, q, ::CayleyInverseRetraction)
Compute the Cayley Inverse Retraction such that the Cayley Retraction from along lands at , i.e. [BZ21].

manifold_dimension(::SymplecticStiefel)
Returns the dimension of the symplectic Stiefel manifold embedded in , i.e. [BZ21]

project(::SymplecticStiefel, p, A) project!(::SymplecticStiefel, Y, p, A)
Given a point , project an element onto the tangent space relative to the euclidean metric of the embedding .

check_vector(M::EmbeddedTorus, p, X; atol=eps(eltype(p)), kwargs...)
Check whether X is a valid vector tangent to p on the EmbeddedTorus M.

check_vector(M::Torus{n}, p, X; kwargs...)
Checks whether X is a valid tangent vector to p on the Torus M.

manifold_dimension(M::EmbeddedTorus)
Return the dimension of the EmbeddedTorus M that is 2.

inverse_retract(M::VectorBundle, p, q, ::FiberBundleInverseProductRetraction)
Compute the allocating variant of the FiberBundleInverseProductRetraction, which by default allocates and calls inverse_retract_product!.

project(B::VectorBundle, p, X)
Project the element X of the ambient space of the tangent space to the tangent space .