Vector bundles

Vector bundle $E$ is a special case of a fiber bundle where each fiber is a vector space.

Tangent bundle is a simple example of a vector bundle, where each fiber is the tangent space at the specified point $p$. An object representing a tangent bundle can be obtained using the constructor called TangentBundle.

There is also another type, VectorSpaceFiber, that represents a specific fiber at a given point. This is also considered a manifold.

FVector

For cases where confusion between different types of vectors is possible, the type FVector can be used to express which type of vector space the vector belongs to. It is used for example in musical isomorphisms (the flat and sharp functions) that are used to go from a tangent space to cotangent space and vice versa.

Documentation

Manifolds.TensorProductType — Type

TensorProductType(spaces::VectorSpaceType...)

Vector space type corresponding to the tensor product of given vector space types.

Manifolds.TangentBundle — Type

TangentBundle{𝔽,M} = VectorBundle{𝔽,TangentSpaceType,M} where {𝔽,M<:AbstractManifold{𝔽}}

Tangent bundle for manifold of type M, as a manifold with the Sasaki metric [Sas58].

Exact retraction and inverse retraction can be approximated using FiberBundleProductRetraction, FiberBundleInverseProductRetraction and SasakiRetraction. FiberBundleProductVectorTransport can be used as a vector transport.

Constructors

TangentBundle(M::AbstractManifold)TangentBundle(M::AbstractManifold, vtm::FiberBundleProductVectorTransport)

Manifolds.VectorBundle — Type

VectorBundle{𝔽,TVS,TM,VTV} = FiberBundle{𝔽,TVS,TM,TVT} where {TVS<:VectorSpaceType}

Alias for FiberBundle when fiber type is a TVS of type VectorSpaceType.

VectorSpaceFiberType is used to encode vector spaces as fiber types.

Manifolds.fiber_bundle_transport — Method

fiber_bundle_transport(M::AbstractManifold, fiber::FiberType)

Determine the vector transport used for exp and log maps on a vector bundle with fiber type fiber and manifold M.

ManifoldsBase.injectivity_radius — Method

injectivity_radius(M::TangentBundle)

Injectivity radius of TangentBundle manifold is infinite if the base manifold is flat and 0 otherwise. See https://mathoverflow.net/questions/94322/injectivity-radius-of-the-sasaki-metric.

ManifoldsBase.inner — Method

inner(B::VectorBundle, p, X, Y)

Inner product of tangent vectors X and Y at point p from the vector bundle B over manifold B.fiber (denoted $\mathcal M$).

Notation:

The point $p = (x_p, V_p)$ where $x_p ∈ \mathcal M$ and $V_p$ belongs to the fiber $F=π^{-1}(\{x_p\})$ of the vector bundle $B$ where $π$ is the canonical projection of that vector bundle $B$.

The tangent vector $v = (V_{X,M}, V_{X,F}) ∈ T_{x}B$ where $V_{X,M}$ is a tangent vector from the tangent space $T_{x_p}\mathcal M$ and $V_{X,F}$ is a tangent vector from the tangent space $T_{V_p}F$ (isomorphic to $F$). Similarly for the other tangent vector $w = (V_{Y,M}, V_{Y,F}) ∈ T_{x}B$.

The inner product is calculated as

$⟨X, Y⟩_p = ⟨V_{X,M}, V_{Y,M}⟩_{x_p} + ⟨V_{X,F}, V_{Y,F}⟩_{V_p}.$

ManifoldsBase.inverse_retract — Method

inverse_retract(M::VectorBundle, p, q, ::FiberBundleInverseProductRetraction)

Compute the allocating variant of the FiberBundleInverseProductRetraction, which by default allocates and calls inverse_retract_product!.

ManifoldsBase.is_flat — Method

is_flat(::VectorBundle)

Return true if the underlying manifold of VectorBundle M is flat.

ManifoldsBase.project — Method

project(B::VectorBundle, p, X)

Project the element X of the ambient space of the tangent space $T_p B$ to the tangent space $T_p B$.

Notation:

The point $p = (x_p, V_p)$ where $x_p ∈ \mathcal M$ and $V_p$ belongs to the fiber $F=π^{-1}(\{x_p\})$ of the vector bundle $B$ where $π$ is the canonical projection of that vector bundle $B$.

The vector $x = (V_{X,M}, V_{X,F})$ where $x_p$ belongs to the ambient space of $T_{x_p}\mathcal M$ and $V_{X,F}$ belongs to the ambient space of the fiber $F=π^{-1}(\{x_p\})$ of the vector bundle $B$ where $π$ is the canonical projection of that vector bundle $B$.

The projection is calculated by projecting $V_{X,M}$ to tangent space $T_{x_p}\mathcal M$ and then projecting the vector $V_{X,F}$ to the fiber $F$.

ManifoldsBase.project — Method

project(B::VectorBundle, p)

Project the point p from the ambient space of the vector bundle B over manifold B.fiber (denoted $\mathcal M$) to the vector bundle.

Notation:

The point $p = (x_p, V_p)$ where $x_p$ belongs to the ambient space of $\mathcal M$ and $V_p$ belongs to the ambient space of the fiber $F=π^{-1}(\{x_p\})$ of the vector bundle $B$ where $π$ is the canonical projection of that vector bundle $B$.

The projection is calculated by projecting the point $x_p$ to the manifold $\mathcal M$ and then projecting the vector $V_p$ to the tangent space $T_{x_p}\mathcal M$.

ManifoldsBase.retract — Method

retract(M::VectorBundle, p, q, ::FiberBundleProductRetraction)

Compute the allocating variant of the FiberBundleProductRetraction, which by default allocates and calls retract_product!.

ManifoldsBase.vector_transport_to — Method

vector_transport_to(M::VectorBundle, p, X, q, m::FiberBundleProductVectorTransport)

Compute the vector transport the tangent vector Xat p to q on the VectorBundle M using the FiberBundleProductVectorTransport m.

Example

The following code defines a point on the tangent bundle of the sphere $S^2$ and a tangent vector to that point.

using Manifolds, RecursiveArrayToolsM = Sphere(2)TB = TangentBundle(M)p = ArrayPartition([1.0, 0.0, 0.0], [0.0, 1.0, 3.0])X = ArrayPartition([0.0, 1.0, 0.0], [0.0, 0.0, -2.0])

([0.0, 1.0, 0.0], [0.0, 0.0, -2.0])

An approximation of the exponential in the Sasaki metric using 1000 steps can be calculated as follows.

q = retract(TB, p, X, SasakiRetraction(1000))println("Approximation of the exponential map: ", q)

Approximation of the exponential map: RecursiveArrayTools.ArrayPartition{Float64, Tuple{Vector{Float64}, Vector{Float64}}}(([0.6759570857309888, 0.35241486404386485, 0.6472138609849252], [-1.0318269583261073, 0.6273324630574116, 0.7360618920075961]))

TangentBundle{𝔽,M} = VectorBundle{𝔽,TangentSpaceType,M} where {𝔽,M<:AbstractManifold{𝔽}}
Tangent bundle for manifold of type M, as a manifold with the Sasaki metric [Sas58].

VectorBundle{𝔽,TVS,TM,VTV} = FiberBundle{𝔽,TVS,TM,TVT} where {TVS<:VectorSpaceType}
Alias for FiberBundle when fiber type is a TVS of type VectorSpaceType.

injectivity_radius(M::TangentBundle)
Injectivity radius of TangentBundle manifold is infinite if the base manifold is flat and 0 otherwise.

retract(M::VectorBundle, p, q, ::FiberBundleProductRetraction)
Compute the allocating variant of the FiberBundleProductRetraction, which by default allocates and calls retract_product!.

injectivity_radius(M::Circle[, p])
Return the injectivity radius on the Circle M, i.e. .

retract(M::Elliptope, q, Y, ::ProjectionRetraction)
compute a projection based retraction by projecting back onto the manifold.

injectivity_radius(M::Euclidean)
Return the injectivity radius on the Euclidean M, which is .

FiberBundle{𝔽,TVS<:FiberType,TM<:AbstractManifold{𝔽},TVT<:FiberBundleProductVectorTransport} <: AbstractManifold{𝔽}
Fiber bundle on a AbstractManifold M of type FiberType.

struct FiberBundleInverseProductRetraction <: AbstractInverseRetractionMethod end
Inverse retraction of the point y at point p from vector bundle B over manifold B.fiber (denoted ).

struct FiberBundleProductRetraction <: AbstractRetractionMethod end
Product retraction map of tangent vector at point from vector bundle B over manifold B.fiber (denoted ).

FiberBundleProductVectorTransport{ TMP<:AbstractVectorTransportMethod, TMV<:AbstractVectorTransportMethod, } <: AbstractVectorTransportMethod
Vector transport type on FiberBundle.

injectivity_radius(::FixedRankMatrices)
Return the incjectivity radius of the manifold of FixedRankMatrices, i.e. 0.

retract(M::FixedRankMatrices, p, X, ::OrthographicRetraction)
Compute the OrthographicRetraction on the FixedRankMatrices M by finding the nearest point to in

retract(M, p, X, ::PolarRetraction)
Compute an SVD-based retraction on the FixedRankMatrices M by computing

injectivity_radius(M::Flag) injectivity_radius(M::Flag, p)
Return the injectivity radius on the Flag M, which is .

retract(M::Flag, p, X, ::PolarRetraction)
Compute the SVD-based retraction PolarRetraction on the Flag M.

retract(M::Flag, p::OrthogonalPoint, X::OrthogonalTangentVector, ::QRRetraction)
Compute the QR retraction on the Flag in the orthogonal matrix representation as the first order approximation to the exponential map.

injectivity_radius(M::GeneralizedGrassmann) injectivity_radius(M::GeneralizedGrassmann, p)
Return the injectivity radius on the GeneralizedGrassmann M, which is .

retract(M::GeneralizedGrassmann, p, X, ::PolarRetraction)
Compute the SVD-based retraction PolarRetraction on the GeneralizedGrassmann M, by projecting onto M.

retract(M::GeneralizedStiefel, p, X) retract(M::GeneralizedStiefel, p, X, ::PolarRetraction) retract(M::GeneralizedStiefel, p, X, ::ProjectionRetraction)
Compute the SVD-based retraction PolarRetraction on the GeneralizedStiefel manifold M.

injectivity_radius(G::GeneraliUnitaryMatrices)
Return the injectivity radius for general unitary matrix manifolds, which is

injectivity_radius(G::SpecialUnitaryMatrices)
Return the injectivity radius for general complex unitary matrix manifolds, where the determinant is , which is

injectivity_radius(G::SpecialOrthogonal) injectivity_radius(G::Orthogonal) injectivity_radius(M::Rotations) injectivity_radius(M::Rotations, ::ExponentialRetraction)
Return the radius of injectivity on the Rotations manifold M, which is .

retract(M::Rotations, p, X, ::PolarRetraction) retract(M::OrthogonalMatrices, p, X, ::PolarRetraction)
Compute the SVD-based retraction on the Rotations and OrthogonalMatrices M from p in direction X (as an element of the Lie group) and is a second-order approximation of the exponential map.

retract(M::Rotations, p, X, ::QRRetraction) retract(M::OrthogonalMatrices, p. X, ::QRRetraction)
Compute the QR-based retraction on the Rotations and OrthogonalMatrices M from p in direction X (as an element of the Lie group), which is a first-order approximation of the exponential map.

injectivity_radius(M::Grassmann) injectivity_radius(M::Grassmann, p)
Return the injectivity radius on the Grassmann M, which is .

retract(M::Grassmann, p, X, ::PolarRetraction)
Compute the SVD-based retraction PolarRetraction on the Grassmann M.

retract(M::Grassmann, p, X, ::QRRetraction )
Compute the QR-based retraction QRRetraction on the Grassmann M.

injectivity_radius(M::Hyperbolic) injectivity_radius(M::Hyperbolic, p)
Return the injectivity radius on the Hyperbolic, which is .

retract(M::MultinomialDoubleStochastic, p, X, ::ProjectionRetraction)
compute a projection based retraction by projecting back onto the manifold, where are elementwise multiplication and division, respectively.

retract(M::MultinomialSymmetric, p, X, ::ProjectionRetraction)
compute a projection based retraction by projecting back onto the manifold, where are elementwise multiplication and division, respectively.

injectivity_radius(M::PositiveNumbers[, p])
Return the injectivity radius on the PositiveNumbers M, i.e. .

retract(M::ProbabilitySimplex, p, X, ::SoftmaxRetraction)
Compute a first order approximation by applying the softmax function.

retract(M::AbstractProjectiveSpace, p, X, method::ProjectionRetraction) retract(M::AbstractProjectiveSpace, p, X, method::PolarRetraction) retract(M::AbstractProjectiveSpace, p, X, method::QRRetraction)
Compute the equivalent retraction ProjectionRetraction, and QRRetraction on the AbstractProjectiveSpace manifold M, i.e.

retract(M::Spectrahedron, q, Y, ::ProjectionRetraction)
compute a projection based retraction by projecting back onto the manifold.

Sphere{𝔽, T} <: AbstractSphere{𝔽}
The (unit) sphere manifold is the set of all unit norm vectors in .

injectivity_radius(M::AbstractSphere[, p, ::ExponentialRetraction])
Return the injectivity radius for the AbstractSphere M, which is globally .

retract(M::AbstractSphere, p, X, ::ProjectionRetraction)
Compute the retraction that is based on projection, i.e.

injectivity_radius(M::Stiefel)
Return the injectivity radius for the Stiefel manifold M, which is globally [ZS24].

retract(::Stiefel, p, X, ::CayleyRetraction)
Compute the retraction on the Stiefel that is based on the Cayley transform[Zhu16].

retract(M::Stiefel, p, X, ::PadeRetraction{m})
Compute the retraction on the Stiefel manifold M based on the Padé approximation of order [ZD18].

retract(M::Stiefel, p, X, ::PolarLightRetraction)
Compute an SVD-based retraction PolarLightRetraction on the Stiefel(n,k) manifold M following [JZ26], equation (3.4)

retract(M::Stiefel, p, X, ::PolarRetraction)
Compute the SVD-based retraction PolarRetraction on the Stiefel manifold M.

retract(M::Stiefel, p, X, method::ProjectionRetraction)
Compute a projection-based retraction.

retract(M::Stiefel, p, X, ::QRRetraction)
Compute the QR-based retraction QRRetraction on the Stiefel manifold M.

injectivity_radius(M::SymmetricPositiveDefinite[, p]) injectivity_radius(M::MetricManifold{SymmetricPositiveDefinite,AffineInvariantMetric}[, p]) injectivity_radius(M::MetricManifold{SymmetricPositiveDefinite,LogCholeskyMetric}[, p])
Return the injectivity radius of the SymmetricPositiveDefinite.

retract(::SymplecticGrassmann, p, X, ::CayleyRetraction) retract!(::SymplecticGrassmann, q, p, X, ::CayleyRetraction)
Compute the Cayley retraction on the Symplectic Grassmann manifold, when the points are represented as symplectic bases, i.e. on the SymplecticStiefel.

retract(::SymplecticStiefel, p, X, ::CayleyRetraction) retract!(::SymplecticStiefel, q, p, X, ::CayleyRetraction)
Compute the Cayley retraction on the Symplectic Stiefel manifold, from p along X (computed inplace of q).

retract(::Tucker, p::TuckerPoint, X::TuckerTangentVector, ::PolarRetraction)
The truncated HOSVD-based retraction [KSV13] to the Tucker manifold, i.e. the result is the sequentially truncated HOSVD approximation of .