Vector bundle

Vector bundles

Vector bundle $E$ is a manifold that is built on top of another manifold $M$ (base space). It is characterized by a continuous function $\Pi \colon E \to M$, such that for each point $x \in M$ the preimage of $x$ by $\Pi$, $\Pi^{-1}(\{x\})$, has a structure of a vector space. These vector spaces are called fibers. Bundle projection can be performed using function bundle_projection.

Tangent bundle is a simple example of a vector bundle, swhere each fiber is the tangent space at the specified point $x$. An object representing a tangent bundle can be obtained using the constructor called TangentBundle.

Fibers of a vector bundle are represented by the type VectorBundleFibers. The important difference between functions operating on VectorBundle and VectorBundleFibers is that in the first case both a point on the underlying manifold and the vector are represented together (by a single argument) while in the second case only the vector part is present, while the point is supplied in a different argument where needed.

VectorBundleFibers refers to the whole set of fibers of a vector bundle. There is also another type, VectorSpaceAtPoint, that represents a specific fiber at a given point. This distinction is made to reduce the need to repeatedly construct objects of type VectorSpaceAtPoint in certain usage scenarios.

FVector

For cases where confusion between different types of vectors is possible, the type FVector can be used to express which type of vector space the vector belongs to. It is used for example in musical isomorphisms (the flat! and sharp! functions) that are used to go from a tangent space to cotangent space and vice versa.

Manifolds.FVector — Type.

FVector(type::VectorSpaceType, data)

Decorator indicating that the vector data from a fiber of a vector bundle is of type type.

Manifolds.VectorBundle — Type.

VectorBundle(M::Manifold, type::VectorSpaceType)

Vector bundle on manifold M of type type.

Manifolds.VectorBundleFibers — Type.

VectorBundleFibers(VS::VectorSpaceType, M::Manifold)

Type representing a family of vector spaces (fibers) of a vector bundle over M with vector spaces of type VS. In contrast with VectorBundle, operations on VectorBundleFibers expect point-like and vector-like parts to be passed separately instead of being bundled together. It can be thought of as a representation of vector spaces from a vector bundle but without storing the point at which a vector space is attached (which is specified separately in various functions).

Manifolds.VectorSpaceAtPoint — Type.

VectorSpaceAtPoint(fiber::VectorBundleFibers, x)

A vector space (fiber type fiber of a vector bundle) at point x from the manifold fiber.M.

Manifolds.VectorSpaceType — Type.

VectorSpaceType

Abstract type for tangent spaces, cotangent spaces, their tensor products, exterior products, etc.

Every vector space VS is supposed to provide:

a method of constructing vectors,

basic operations: addition, subtraction, multiplication by a scalar and negation (unary minus),

zero_vector!(VS, v, x) to construct zero vectors at point x,

similar(v, T) for vector v,

copyto!(v, w) for vectors v and w,

eltype(v) for vector v,

vector_space_dimension(::VectorBundleFibers{<:typeof(VS)}) where VS.

Optionally:

inner product via inner (used to provide Riemannian metric on vector bundles),

flat! and sharp!,

norm (by default uses inner),

project_vector! (for embedded vector spaces),

representation_size (if support for ProductArray is desired),

broadcasting for basic operations.

LinearAlgebra.norm — Method.

norm(B::VectorBundleFibers, x, v)

Norm of the vector v from the vector space of type B.VS at point x from manifold B.M.

Manifolds.CotangentSpaceAtPoint — Method.

CotangentSpaceAtPoint(M::Manifold, x)

Return an object of type VectorSpaceAtPoint representing cotangent space at x.

Manifolds.TangentSpaceAtPoint — Method.

TangentSpaceAtPoint(M::Manifold, x)

Return an object of type VectorSpaceAtPoint representing tangent space at x.

Manifolds.bundle_projection — Method.

bundle_projection(B::VectorBundle, x::ProductRepr)

Projection of point x from the bundle M to the base manifold. Returns the point on the base manifold B.M at which the vector part of x is attached.

Manifolds.distance — Method.

distance(B::VectorBundle, x, y)

Distance between points $x$ and $y$ from the vector bundle B over manifold B.VS (denoted $M$).

Notation:

The point $x = (p_x, \xi_x)$ where $p_x \in M$ and $\xi_x$ belongs to the fiber $F=\pi^{-1}(\{p_x\})$ of the vector bundle $B$ where $\pi$ is the canonical projection of that vector bundle $B$. Similarly, $y = (p_y, \xi_y)$.

The distance is calculated as

\[d_B(x, y) = \sqrt{d_M(p_x, p_y)^2 + d_F(\xi_x, \xi_{y\to x})^2}\]

where $d_M$ is the distance on manifold $M$, $d_F$ is the distance between two vectors from the fiber $F$ and $\xi_{y\to x}$ is the result of parallel transport of vector $\xi_y$ to point $p_x$. The default behavior of vector_transport_to is used to compute the vector transport.

Manifolds.distance — Method.

distance(B::VectorBundleFibers, x, v, w)

Distance between vectors v and w from the vector space at point x from the manifold M.M, that is the base manifold of M.

Manifolds.exp! — Method.

exp!(B::VectorBundle, y, x, v)

Exponential map of tangent vector $v$ at point $x$ from vector bundle B over manifold B.VS (denoted $M$). The result overwrites the point y.

Notation:

The point $x = (p_x, \xi_x)$ where $p_x \in M$ and $\xi_x$ belongs to the fiber $F=\pi^{-1}(\{p_x\})$ of the vector bundle $B$ where $\pi$ is the canonical projection of that vector bundle $B$.

The tangent vector $v = (\xi_{v,M}, \xi_{v,F}) \in T_{x}B$ where $\xi_{v,M}$ is a tangent vector from the tangent space $T_{p_x}M$ and $\xi_{v,F}$ is a tangent vector from the tangent space $T_{\xi_x}F$ (isomorphic to $F$).

The exponential map is calculated as

\[\exp_{x}(v) = (\exp_{p_x}(\xi_{v,M}), \xi_{\exp})\]

where $\xi_{\exp}$ is the result of vector transport of $\xi_x + \xi_{v,F}$ to the point $\exp_{p_x}(\xi_{v,M})$. The sum $\xi_x + \xi_{v,F}$ corresponds to the exponential map in the vector space $F$.

Manifolds.flat! — Method.

flat!(M::Manifold, v::FVector, x, w::FVector)

Compute the flat isomorphism (one of the musical isomorphisms) of vector w from the vector space of type M at point x from manifold M.M.

The function can be used for example to transform vectors from the tangent bundle to vectors from the cotangent bundle $\flat \colon TM \to T^{*}M$

Manifolds.inner — Method.

inner(B::VectorBundle, x, v, w)

Inner product of tangent vectors v and w at point x from the vector bundle B over manifold B.VS (denoted $M$).

Notation:

The point $x = (p_x, \xi_x)$ where $p_x \in M$ and $\xi_x$ belongs to the fiber $F=\pi^{-1}(\{p_x\})$ of the vector bundle $B$ where $\pi$ is the canonical projection of that vector bundle $B$.

The tangent vector $v = (\xi_{v,M}, \xi_{v,F}) \in T_{x}B$ where $\xi_{v,M}$ is a tangent vector from the tangent space $T_{p_x}M$ and $\xi_{v,F}$ is a tangent vector from the tangent space $T_{\xi_x}F$ (isomorphic to $F$). Similarly for the other tangent vector $w = (\xi_{w,M}, \xi_{w,F}) \in T_{x}B$.

The inner product is calculated as

\[\langle v, w \rangle_{B} = \langle \xi_{v,M}, \xi_{w,M} \rangle_{M} + \langle \xi_{v,F}, \xi_{w,F} \rangle_{F}.\]

Manifolds.inner — Method.

inner(B::VectorBundleFibers, x, v, w)

Inner product of vectors v and w from the vector space of type B.VS at point x from manifold B.M.

Manifolds.log! — Method.

log!(B::VectorBundle, v, x, y)

Logarithmic map of the point $y$ at point $x$ from vector bundle B over manifold B.VS (denoted $M$). The result overwrites the tangent vector v.

Notation:

The point $x = (p_x, \xi_x)$ where $p_x \in M$ and $\xi_x$ belongs to the fiber $F=\pi^{-1}(\{p_x\})$ of the vector bundle $B$ where $\pi$ is the canonical projection of that vector bundle $B$. Similarly, $y = (p_y, \xi_y)$.

The logarithmic map is calculated as

\[\log_{x}(y) = (\log_{p_x}(p_y), \xi_{\log} - \xi_x)\]

where $\xi_{\log}$ is the result of vector transport of $\xi_y$ to the point $p_x$. The difference $\xi_{\log} - \xi_x$ corresponds to the logarithmic map in the vector space $F$.

Manifolds.project_point! — Method.

project_point!(B::VectorBundle, x)

Project the point $x$ from the ambient space of the vector bundle B over manifold B.VS (denoted $M$) to the vector bundle.

Notation:

The point $x = (p_x, \xi_x)$ where $p_x$ belongs to the ambient space of $M$ and $\xi_x$ belongs to the ambient space of the fiber $F=\pi^{-1}(\{p_x\})$ of the vector bundle $B$ where $\pi$ is the canonical projection of that vector bundle $B$.

The projection is calculated by projecting the point $p_x$ to the manifold $M$ and then projecting the vector $\xi_x$ to the tangent space $T_{p_x}M$.

Manifolds.project_tangent! — Method.

project_tangent!(B::VectorBundle, w, x, v)

Project the element $v$ of the ambient space of the tangent space $T_x B$ to the tangent space $T_x B$. The results overwrites $w$.

Notation:

The point $x = (p_x, \xi_x)$ where $p_x \in M$ and $\xi_x$ belongs to the fiber $F=\pi^{-1}(\{p_x\})$ of the vector bundle $B$ where $\pi$ is the canonical projection of that vector bundle $B$.

The vector $x = (\xi_{v,M}, \xi_{v,F})$ where $p_x$ belongs to the ambient space of $T_{p_x}M$ and $\xi_{v,F}$ belongs to the ambient space of the fiber $F=\pi^{-1}(\{p_x\})$ of the vector bundle $B$ where $\pi$ is the canonical projection of that vector bundle $B$.

The projection is calculated by projecting $\xi_{v,M}$ to tangent space $T_{p_x}M$ and then projecting the vector $\xi_{v,F}$ to the fiber $F$.

Manifolds.sharp! — Method.

sharp!(M::Manifold, v::FVector, x, w::FVector)

Compute the sharp isomorphism (one of the musical isomorphisms) of vector w from the vector space of type M at point x from manifold M.M.

The function can be used for example to transform vectors from the cotangent bundle to vectors from the tangent bundle $\sharp \colon T^{*}M \to TM$

Manifolds.vector_space_dimension — Method.

vector_space_dimension(B::VectorBundleFibers)

Dimension of the vector space of type B.

Manifolds.zero_tangent_vector! — Method.

zero_tangent_vector!(B::VectorBundle, v, x)

Zero tangent vector at point $x$ from the vector bundle B over manifold B.VS (denoted $M$). The zero vector belongs to the space $T_{x}B$ The result overwrites the tangent vector v.

Notation:

The point $x = (p_x, \xi_x)$ where $p_x \in M$ and $\xi_x$ belongs to the fiber $F=\pi^{-1}(\{p_x\})$ of the vector bundle $B$ where $\pi$ is the canonical projection of that vector bundle $B$.

The zero vector is calculated as

\[\mathbf{0}_{x} = (\mathbf{0}_{p_x}, \mathbf{0}_F)\]

where $\mathbf{0}_{p_x}$ is the zero tangent vector from $T_{p_x}M$ and $\mathbf{0}_F$ is the zero element of the vector space $F$.

Manifolds.zero_vector! — Method.

zero_vector!(B::VectorBundleFibers, v, x)

Save the zero vector from the vector space of type B.VS at point x from manifold B.M to v.

Manifolds.zero_vector — Method.

zero_vector(B::VectorBundleFibers, x)

Compute the zero vector from the vector space of type B.VS at point x from manifold B.M.

Manifolds.TensorProductType — Type.

TensorProductType(spaces::VectorSpaceType...)

Vector space type corresponding to the tensor product of given vector space types.

Manifolds.project_vector! — Method.

project_vector!(B::VectorBundleFibers, v, x, w)

Project vector w from the vector space of type B.VS at point x and save the result to v.

Manifolds.similar_result — Method.

similar_result(B::VectorBundleFibers, f, x...)

Allocates an array for the result of function f that is an element of the vector space of type B.VS on manifold B.M and arguments x... for implementing the non-modifying operation using the modifying operation.

Manifolds.similar_result_type — Method.

similar_result_type(B::VectorBundleFibers, f, args::NTuple{N,Any}) where N

Returns type of element of the array that will represent the result of function f for representing an operation with result in the vector space VS for manifold M on given arguments (passed at a tuple).