Vector bundle

Vector bundles

Vector bundle $E$ is a manifold that is built on top of another manifold $M$ (base space). It is characterized by a continuous function $\Pi \colon E \to M$, such that for each point $x \in M$ the preimage of $x$ by $\Pi$, $\Pi^{-1}(\{x\})$, has a structure of a vector space. These vector spaces are called fibers. Bundle projection can be performed using function bundle_projection.

Tangent bundle is a simple example of a vector bundle, swhere each fiber is the tangent space at the specified point $x$. An object representing a tangent bundle can be obtained using the constructor called TangentBundle.

Fibers of a vector bundle are represented by the type VectorBundleFibers. The important difference between functions operating on VectorBundle and VectorBundleFibers is that in the first case both a point on the underlying manifold and the vector are represented together (by a single argument) while in the second case only the vector part is present, while the point is supplied in a different argument where needed.

VectorBundleFibers refers to the whole set of fibers of a vector bundle. There is also another type, VectorSpaceAtPoint, that represents a specific fiber at a given point. This distinction is made to reduce the need to repeatedly construct objects of type VectorSpaceAtPoint in certain usage scenarios.

FVector

For cases where confusion between different types of vectors is possible, the type FVector can be used to express which type of vector space the vector belongs to. It is used for example in musical isomorphisms (the flat! and sharp! functions) that are used to go from a tangent space to cotangent space and vice versa.

Manifolds.FVector — Type.

FVector(type::VectorSpaceType, data)

Decorator indicating that the vector data from a fiber of a vector bundle is of type type.

Manifolds.VectorBundle — Type.

VectorBundle(M::Manifold, type::VectorSpaceType)

Vector bundle on manifold M of type type.

Manifolds.VectorBundleFibers — Type.

VectorBundleFibers(VS::VectorSpaceType, M::Manifold)

Type representing a family of vector spaces (fibers) of a vector bundle over M with vector spaces of type VS. In contrast with VectorBundle, operations on VectorBundleFibers expect point-like and vector-like parts to be passed separately instead of being bundled together. It can be thought of as a representation of vector spaces from a vector bundle but without storing the point at which a vector space is attached (which is specified separately in various functions).

Manifolds.VectorSpaceAtPoint — Type.

VectorSpaceAtPoint(fiber::VectorBundleFibers, x)

A vector space (fiber type fiber of a vector bundle) at point x from the manifold fiber.M.

Manifolds.VectorSpaceType — Type.

VectorSpaceType

Abstract type for tangent spaces, cotangent spaces, their tensor products, exterior products, etc.

Every vector space VS is supposed to provide:

a method of constructing vectors,

basic operations: addition, subtraction, multiplication by a scalar and negation (unary minus),

zero_vector!(VS, v, x) to construct zero vectors at point x,

similar(v, T) for vector v,

copyto!(v, w) for vectors v and w,

eltype(v) for vector v,

vector_space_dimension(::VectorBundleFibers{<:typeof(VS)}) where VS.

Optionally:

inner product via inner (used to provide Riemannian metric on vector bundles),

flat! and sharp!,

norm (by default uses inner),

project_vector! (for embedded vector spaces),

representation_size (if support for ProductArray is desired),

broadcasting for basic operations.

Manifolds.CotangentSpaceAtPoint — Method.

CotangentSpaceAtPoint(M::Manifold, x)

Return an object of type VectorSpaceAtPoint representing cotangent space at x.

Manifolds.TangentSpaceAtPoint — Method.

TangentSpaceAtPoint(M::Manifold, x)

Return an object of type VectorSpaceAtPoint representing tangent space at x.

Manifolds.bundle_projection — Method.

bundle_projection(B::VectorBundle, x::ProductRepr)

Projection of point x from the bundle M to the base manifold. Returns the point on the base manifold B.M at which the vector part of x is attached.

Manifolds.distance — Method.

distance(B::VectorBundleFibers, x, v, w)

Distance between vectors v and w from the vector space at point x from the manifold M.M, that is the base manifold of M.

Manifolds.flat! — Method.

flat!(M::Manifold, v::FVector, x, w::FVector)

Compute the flat isomorphism (one of the musical isomorphisms) of vector w from the vector space of type M at point x from manifold M.M.

The function can be used for example to transform vectors from the tangent bundle to vectors from the cotangent bundle $\flat \colon TM \to T^{*}M$

Manifolds.inner — Method.

inner(B::VectorBundleFibers, x, v, w)

Inner product of vectors v and w from the vector space of type B.VS at point x from manifold B.M.

Manifolds.sharp! — Method.

sharp!(M::Manifold, v::FVector, x, w::FVector)

Compute the sharp isomorphism (one of the musical isomorphisms) of vector w from the vector space of type M at point x from manifold M.M.

The function can be used for example to transform vectors from the cotangent bundle to vectors from the tangent bundle $\sharp \colon T^{*}M \to TM$

Manifolds.vector_space_dimension — Method.

vector_space_dimension(B::VectorBundleFibers)

Dimension of the vector space of type B.

Manifolds.zero_vector! — Method.

zero_vector!(B::VectorBundleFibers, v, x)

Save the zero vector from the vector space of type B.VS at point x from manifold B.M to v.

Manifolds.zero_vector — Method.

zero_vector(B::VectorBundleFibers, x)

Compute the zero vector from the vector space of type B.VS at point x from manifold B.M.

Manifolds.TensorProductType — Type.

TensorProductType(spaces::VectorSpaceType...)

Vector space type corresponding to the tensor product of given vector space types.

Manifolds.project_vector! — Method.

project_vector!(B::VectorBundleFibers, v, x, w)

Project vector w from the vector space of type B.VS at point x and save the result to v.

Manifolds.similar_result — Method.

similar_result(B::VectorBundleFibers, f, x...)

Allocates an array for the result of function f that is an element of the vector space of type B.VS on manifold B.M and arguments x... for implementing the non-modifying operation using the modifying operation.

Manifolds.similar_result_type — Method.

similar_result_type(B::VectorBundleFibers, f, args::NTuple{N,Any}) where N

Returns type of element of the array that will represent the result of function f for representing an operation with result in the vector space VS for manifold M on given arguments (passed at a tuple).