Statistics

Statistics.mean — Function

mean(M, x, weights=Weights(ones(n)); x0=x[1], stop_iter=100, kwargs... )

computes the Riemannian center of mass also known as Karcher mean of the vector x of n points on the Manifold M. This function uses the gradient descent scheme. Optionally one can provide weights $w_i$ for the weighted Riemannian center of mass. The general formula to compute the minimizer reads

\[\argmin_{y\in\mathcal M} \frac{1}{2}\sum_{i=1}^n w_i\mathrm{d}_{\mathcal M}^2(y,x_i),\]

where $\mathrm{d}_{\mathcal M}$ denotes the Riemannian distance.

Note that any provided weights are normalized to $\displaystyle\sum_{i=1}^n w_i=1$.

Optionally you can provide x0, the starting point (by default set to the first data point). stop_iter denotes the maximal number of iterations to perform and the kwargs... are passed to isapprox to stop, when the minimal change between two iterates is small. For more stopping criteria check the Manopt.jl package and use a solver therefrom.

The algorithm is further described in

Afsari, B; Tron, R.; Vidal, R.: On the Convergence of Gradient Descent for Finding the Riemannian Center of Mass, SIAM Journal on Control and Optimization (2013), 51(3), pp. 2230–2260, doi: 10.1137/12086282X, arxiv: 1201.0925

Statistics.mean! — Method

mean!(M,y,x,w)

computes the mean in-place in y where the initial value of y is the starting point of the algorithm.

Statistics.median — Function

median(M, x, weights=Weights(ones(n)); stop_iter=10000)

computes the Riemannian median of the vector x of n points on the Manifold M. This function is nonsmooth (i.e nondifferentiable) and uses a cyclic proximal point scheme. Optionally one can provide weights $w_i$ for the weighted Riemannian median. The general formula to compute the minimizer reads

\[\argmin_{y\in\mathcal M}\sum_{i=1}^n w_i\mathrm{d}_{\mathcal M}(y,x_i),\]

where $\mathrm{d}_{\mathcal M}$ denotes the Riemannian distance.

Optionally you can provide x0, the starting point (by default set to the first data point). stop_iter denotes the maximal number of iterations to perform and the kwargs... are passed to isapprox to stop, when the minimal change between two iterates is small. For more stopping criteria check the Manopt.jl package and use a solver therefrom.

The algorithm is further described in Algorithm 4.3 and 4.4 in

Bačák, M: Computing Medians and Means in Hadamard Spaces. SIAM Journal on Optimization (2014), 24(3), pp. 1542–1566, doi: 10.1137/140953393, arxiv: 1210.2145

Statistics.median! — Method

median!(M,y,x,w)

computes the median in-place in y where the initial value of y is the starting point of the algorithm.

Statistics.std — Function

std(M, x, m=nothing; corrected=true, kwargs...)

compute the variance of a Vector x of n data points on the Manifold M, i.e.

\[\sqrt{\frac{1}{c} \sum_{i=1}^n d_{\mathcal M}^2 (x_i,m)},\]

where c is a correction term, see Statistics.var. and m is the provideed mean of x.

Statistics.std — Function

std(M, x, w::AbstractWeights, m=nothing; corrected=false, kwargs...)

compute the (weighted) standard deviation of a Vector x of n data points on the Manifold M, i.e.

\[\sqrt{\frac{1}{c} \sum_{i=1}^n w_i d_{\mathcal M}^2 (x_i,m)},\]

where c is a correction term, see Statistics.var. The (weighted) mean of x can be specified as m, and the corrected variance can be activated by setting corrected=true. All further kwargs... are passed to the computation of the mean (if that is not provided).

Statistics.var — Function

var(M, x, m=nothing; corrected=true, kwargs...)

compute the variance of a Vector x of n data points on the Manifold M, i.e.

\[\frac{1}{c} \sum_{i=1}^n d_{\mathcal M}^2 (x_i,m),\]

where c is a correction term, see Statistics.var. and m is the provideed mean of x. The uncorrected variance can be activated by setting corrected=false.

Statistics.var — Function

var(M, x, w::AbstractWeights, m=nothing; corrected=false, kwargs...)

compute the (weighted) variance of a Vector x of n data points on the Manifold M, i.e.

\[\frac{1}{c} \sum_{i=1}^n w_i d_{\mathcal M}^2 (x_i,m),\]

where c is a correction term, see Statistics.var. The (weighted) mean of x can be specified as m, and the corrected variance can be activated by setting corrected=true. All further kwargs... are passed to the computation of the mean (if that is not provided).

StatsBase.mean_and_std — Method

mean_and_std(M, x, w; corrected = false, kwargs...) -> (mean, std)
mean_and_std(M, x; corrected = true, kwargs...) -> (mean, std)

compute the mean m and the standard deviation std s simultaneously.

StatsBase.mean_and_var — Method

mean_and_var(M, x, w; corrected = false, kwargs...) -> (mean, var)
mean_and_var(M, x; corrected = true, kwargs...) -> (mean, var)

compute the mean m and the variance v simultaneously.